% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_datasets.R
\docType{data}
\name{Anime}
\alias{Anime}
\title{Data: Anime}
\format{An object of class \code{data.frame} with 183 rows and 13 columns.}
\source{
Original source: \href{https://github.com/ISS-Analytics/pls-predict}{github.com/ISS-Analytics/pls-predict}
}
\usage{
Anime
}
\description{
A data frame with 183 observations and 13 variables.
}
\details{
The data set for the example on \href{https://github.com/ISS-Analytics/pls-predict}{github.com/ISS-Analytics/pls-predict}
with irrelevant variables removed.
}
\keyword{datasets}
