% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_dotdotdot.R
\name{args_default}
\alias{args_default}
\title{Show possible "..." arguments and their defaults}
\usage{
args_default()
}
\value{
A named list of argument names and their defaults.
}
\description{
The function \code{args_default} lists all possible arguments to be passed to lower
level functions via the \code{...} argument of the \link{csem} or \link{cca} function.
}
\details{
This is the complete list of arguments, their current defaults, and their
possible alternatives. Each default may be changed by passing a \code{arg_name = value}
pair to \link{csem} or \link{cca}. Arguments are processed by the internal \link{handleArgs}
function and, if valid, passed down to their corresponding function
via the \link{workhorse} function.

\strong{Arguments accepted by \link{csem}}
\describe{
\item{\code{.approach_cf}}{\emph{Default: \strong{"dist_euclid"}}. A character string
of the name of the approach used to obtain the correction factor for PLSc.

Currently seven methods are available (see: \insertCite{Dijkstra2013}{cSEM}
for details on the different approaches):

\itemize{
\item "dist_euclid"`
\item "dist_euclid_weighted"
\item "fisher_transformed"
\item "mean_geometric"
\item "mean_harmonic"
\item "mean_arithmetic"
\item "geo_of_harmonic" (not yet implemented)
}

The argument is passed to: \link{calculateCorrectionFactors}}
\item{\code{.ignore_structural_model}}{\emph{Default: FALSE}. Should the structural (path)
model be ignored when calculating the inner weights of the PLS algorithm?

The argument is passed to: \link{calculateInnerWeightsPLS}}
\item{\code{.iter_max}}{\emph{Default: \strong{100}}. An integer indicating the maximum number of
iterations of the PLS algorithm. If \code{iter_max = 1} one-step weights
are returned. If the algorithm exceeds the specified number, weights of the
previous iteration step will be returned with a warning.

The argument is passed to: \link{calculateWeightsPLS}}
\item{\code{.tolerance}}{\emph{Default: \strong{1e-06}}. A numeric value specifying the tolerance
criterion for convergence of the PLS algorithm.

The argument is passed to: \link{calculateWeightsPLS}}
}
}
\examples{
## List all possible arguments to be passed to lower level functions via
   the ... argument of the csem or cca function.

args_default()

}
\references{
\insertRef{Dijkstra2013}{cSEM}
}
\seealso{
\link{handleArgs}, \link{cca}, \link{csem}, \link{workhorse}
}
