% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_assess.R
\name{assess}
\alias{assess}
\alias{assess.cSEMResults_default}
\alias{assess.cSEMResults_multi}
\alias{assess.cSEMResults_2ndorder}
\title{Assess model}
\usage{
assess(.object = NULL, .only_common_factors = TRUE, ...)

\method{assess}{cSEMResults_default}(.object = NULL,
  .only_common_factors = TRUE, ...)

\method{assess}{cSEMResults_multi}(.object = NULL,
  .only_common_factors = TRUE, ...)

\method{assess}{cSEMResults_2ndorder}(.object, ...)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}

\item{.only_common_factors}{Logical. Should only constructs modeled as common
factors be included? Defaults to \code{TRUE}.}

\item{...}{Further arguments passed to functions called by \code{assess()}.
See \link{args_assess_dotdotdot} for a complete list of available arguments.}
}
\value{
A named list of quality criteria. Note that if only a single quality
criteria is computed the return value is still a list!
}
\description{
Assess a model using common evaluation criteria and fit measures.
See the \href{https://m-e-rademaker.github.io/cSEM/articles/Using-assess.html}{Postestimation: Assessing a model}
article on the
\href{https://m-e-rademaker.github.io/cSEM/index.html}{cSEM website} for details.
}
\details{
The function is essentially a wraper around a number of internal functions
that perform an "assessment task" (called a \strong{quality criterion} in \pkg{cSEM}
parlance) like computing the (congeneric) reliability,
the effect size, the heterotrait-monotrait ratio of correlations (HTMT) etc.

By default every possible quality criterion is calculated (\code{.what = "all"}).
If only a subset of quality criteria needs to be computed a single character string
or a vector of character strings naming the quantity to compute may be
supplied to \code{assess()} via the \code{.what} argument.

Some of the quality criteria are inherently tied to the classical common
factor model and therefore only meaningfully interpreted within a common
factor model (see the
\href{https://m-e-rademaker.github.io/cSEM/articles/Using-assess.html}{Postestimation: Assessing a model}
article for details).
It is possible to force computation of all quality criteria for constructs
modeled as composites by setting \code{.only_common_factors = FALSE}, however,
we explicitly warn to interpret quality criteria in this case with caution,
as they may not even have a conceptual meaning.
}
\section{Methods (by class)}{
\itemize{
\item \code{cSEMResults_default}: (TODO)

\item \code{cSEMResults_multi}: (TODO)

\item \code{cSEMResults_2ndorder}: (TODO)
}}

\seealso{
\code{\link[=csem]{csem()}}, \code{\link[=foreman]{foreman()}}, \link{cSEMResults}
}
