% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_new.R
\name{AVE}
\alias{AVE}
\title{AVE}
\usage{
AVE(
 .object              = args_default()$.object,
 .only_common_factors = args_default()$.only_common_factors
)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}

\item{.only_common_factors}{Logical. Should only common factors be included?
Defaults to \code{FALSE}.}
}
\description{
Computes the average variance extracted \insertCite{Raykov1997}{cSEM} based on the standardized loadings.
}
\examples{
\dontrun{
# still to implement
}

}
\references{
\insertAllCited{}
}
\seealso{
\link{csem}, \link{cSEMResults}
}
