% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_estimators_weights.R
\name{bdiagGSCA}
\alias{bdiagGSCA}
\title{Block Diagonalize GSCA Parameter Estimates and Scores}
\usage{
bdiagGSCA(.object)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}
}
\value{
Single group cSEMResults of Block diagonalized path estimates, loadingestimates, weight estaimes, construct scores and unique scores for GSCA type models.
}
\description{
Block diagonalizes the estimated paramater matrices as shown on Equations
3.28-3.29 on page 111 of \insertCite{Hwang2014;textual}{cSEM}. This is used to compute a single FIT-type statistic given a multigroup GSCA-type model.
}
\details{
This function should be called through calculateFIT, calculateFIT_s, etc... and is not intended to be user-facing.
}
