% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_igsca.R
\name{bdiagonalizeMultiGroupIgscaEstimates}
\alias{bdiagonalizeMultiGroupIgscaEstimates}
\title{Block Diagonalize Estimated Parameter Matrices to Facilitate Computation of FIT Statistics}
\usage{
bdiagonalizeMultiGroupIgscaEstimates(x)
}
\arguments{
\item{x}{Fitted \code{csem} model with class \code{cSEMResults} as returned from \code{\link[=csem]{csem()}}}
}
\value{
cSEMResults in single-group data structure with block diagonalized parameter estimates
}
\description{
Block diagonalizes the estimated paramater matrices as shown on Equations
3.28-3.29 on page 111 of \insertCite{Hwang2014;textual}{cSEM}. Should only be used on multi-group models
}
