% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_datasets.R
\docType{data}
\name{Benitezetal2020}
\alias{Benitezetal2020}
\title{Data: Benitezetal2020}
\format{
An object of class \code{data.frame} with 300 rows and 22 columns.
}
\source{
The dataset is provided as supplementary material by \insertCite{Benitez2020;textual}{cSEM}.
}
\usage{
Benitezetal2020
}
\description{
A data frame containing 22 variables with 300 observations.
}
\details{
The simulated data contains variables about the social executive and employee behavior.
Moreover, it contains variables about the social media capability and business performance.
The dataset was used as an illustrative example in \insertCite{Benitez2020;textual}{cSEM}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
