% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.step3.regression.R
\name{calc.step3.regression}
\alias{calc.step3.regression}
\title{Polynomial Regression Model for CSEM}
\usage{
calc.step3.regression(squared.diff, half1, half2)
}
\arguments{
\item{squared.diff}{Vector of adjusted squared differences}

\item{half1}{First half of the items}

\item{half2}{Second half of the items}
}
\value{
Fitted polynomial regression model
}
\description{
Polynomial Regression Model for CSEM
}
