% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_csem.R
\name{calculate2ndOrder}
\alias{calculate2ndOrder}
\title{Second order constructs}
\usage{
calculate2ndOrder(
  .csem_model, 
  .first_stage_results)
}
\arguments{
\item{.csem_model}{A (possibly incomplete) \link{cSEMModel}-list.}
}
\value{
A cSEMResults object.
}
\description{
Performs the second and third stage for a model containing second order
constructs.
}
\keyword{internal}
