% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_assess.R
\name{calculateAIC}
\alias{calculateAIC}
\title{Akaike information criterion}
\usage{
calculateAIC(.object = NULL)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}
}
\value{
A single numeric value. If \code{.object} is a list
of \code{cSEMResults} objects, a list of AICs is returned.
}
\description{
Calculate the Akaike information criterion (AIC) as proposed by
\insertCite{Akaike1974;textual}{cSEM}.
}
\details{
The AIC is typically used as a model selection criteria.
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=assess]{assess()}}, \link{cSEMResults}
}
