% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_foreman.R
\name{calculateCompositeConstructCV}
\alias{calculateCompositeConstructCV}
\title{Internal: Calculate composite-construct correlation}
\usage{
calculateCompositeConstructCV(
 .W                  = args_default()$.W,
 .csem_model         = args_default()$.csem_model,
 .disattenuate       = args_default()$.disattenuate,
 .modes              = args_default()$.modes,
 .correction_factors = args_default()$.correction_factors,
 .reliabilities      = args_default()$.reliabilities
 )
}
\arguments{
\item{.W}{A (J x K) matrix of weights.}

\item{.csem_model}{A (possibly incomplete) \link{cSEMModel}-list.}

\item{.disattenuate}{Logical. If possible, should composite correlations be disattenuated
if the construct is modeled as a common factor? Defaults to \code{TRUE}.}

\item{.modes}{A vector giving the mode for each construct in the form \code{"name" = "mode"}.
Only used internally.}

\item{.reliabilities}{A character vector of \code{"name" = value} pairs,
where \code{value} is a number between 0 and 1 and \code{"name"} a character string
of the corresponding construct name, or \code{NULL}. Reliabilities
may be given for a subset of the constructs. Defaults to \code{NULL} in which case
reliabilities are estimated by \code{csem()}.}
}
\value{
A vector of composite-construct correlations.
}
\description{
Calculate the correlation between composite and construct (usually called Q).
Note: Q_i^2 := R^2(eta_i; eta_bar_i) is also called the reliability coefficient
rho_A \insertCite{Dijkstra2015a}{cSEM}.
}
\references{
\insertAllCited{}
}
\keyword{internal}
