% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_foreman.R
\name{calculateComposites}
\alias{calculateComposites}
\title{Internal: Calculate composites}
\usage{
calculateComposites(
 .X  = args_default()$.X,
 .W  = args_default()$.W
 )
}
\arguments{
\item{.X}{A matrix of processed data (scaled, cleaned and ordered).}

\item{.W}{A (J x K) matrix of weights.}
}
\value{
The (N x J) matrix of proxy values with column names equal to
the names of the J constructs.
}
\description{
Calculates composite values for the J constructs of the structural model.
}
\keyword{internal}
