% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_foreman.R
\name{calculateCompositeVCV}
\alias{calculateCompositeVCV}
\title{Internal: Calculate composite variance-covariance matrix}
\usage{
calculateCompositeVCV(
 .S  = args_default()$.S,
 .W  = args_default()$.W
 )
}
\arguments{
\item{.S}{The (K x K) empirical indicator correlation matrix.}

\item{.W}{A (J x K) matrix of weights.}
}
\value{
A (J x J) composite VCV matrix.
}
\description{
Calculate the variance-covariance (VCV) matrix of the composites.
}
\keyword{internal}
