% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_foreman.R
\name{calculateConstructVCV}
\alias{calculateConstructVCV}
\title{Internal: Calculate construct VCV matrix}
\usage{
calculateConstructVCV(
 .C          = args_default()$.C, 
 .Q          = args_default()$.Q, 
 .csem_model = args_default()$.csem_model
 )
}
\arguments{
\item{.C}{A (J x J) composite variance-covariance matrix.}

\item{.Q}{A vector of composite-construct correlations with element names equal to
the names of the J construct names used in the measurement model. Note
Q^2 is also called the reliability coefficient.}

\item{.csem_model}{A (possibly incomplete) \link{cSEMModel}-list.}
}
\value{
The (J x J) construct VCV matrix. Disattenuated if requested.
}
\description{
Calculate the covariance matrix of the constructs.
}
\keyword{internal}
