% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_foreman.R
\name{calculateCorTheilSen}
\alias{calculateCorTheilSen}
\title{Internal: Calculate indicator correlation matrix using Theil-Sen}
\usage{
calculateCorTheilSen(.X= args_default()$.X)
}
\arguments{
\item{.X}{A matrix of processed data (scaled, cleaned and ordered).}
}
\value{
The (K x K) indicator correlation matrix.
}
\description{
Calculate the indicator correlation matrix using Theil-Sen correlation.
}
\keyword{internal}
