% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_assess.R
\name{calculateDf}
\alias{calculateDf}
\title{Internal: Degrees of freedom}
\usage{
calculateDf(
  .object     = NULL,
  .null_model = args_default()$.null_model,
  ...
  )
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}

\item{.null_model}{Logical. Should the degrees of freedom for the null model
be computed? Defaults to \code{FALSE}.}

\item{...}{Ignored.}
}
\value{
A single numeric value.
}
\description{
Calculate the degrees of freedom for a given model from a \link{cSEMResults} object.
}
\seealso{
\code{\link[=assess]{assess()}}, \link{cSEMResults}
}
\keyword{internal}
