% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_distance_measures.R
\name{calculateDistance}
\alias{calculateDistance}
\title{Internal: Calcualte matrix difference using distance measure}
\usage{
calculateDistance(.matrices = NULL, .distance = "geodesic")
}
\value{
A numeric vector of length one containing the (arithmetic mean) difference
between all matrices.
}
\description{
Calculates the differences between symmetric matrices using a given
distance measure. This is typically used to calculate the difference between
the model-implied and the empirical indicator covariance matrix.
}
\details{
\code{.matrices} must be a list of at least two matrices. If more than two matrices
are supplied the arithmetic mean distance over all possible matrix
distances is computed. Hence, supplying a large number of matrices will
quickly become computationally challenging. Use with care.

Currently two distance measures are supported:

\describe{
\item{\code{geodesic}}{The gedesic distance. Details here.}
\item{\code{squared_euclidian}}{The squared euclidian distance. Details here.}
}
}
\examples{
\dontrun{
# still to implement
}


}
