% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_assess.R
\name{calculateEffectSize}
\alias{calculateEffectSize}
\title{Internal: Calculate effect size}
\usage{
calculateEffectSize(.object = NULL)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}
}
\description{
Calculate the effect size
}
\seealso{
\link{assess}, \link{csem}, \link{cSEMResults}
}
\keyword{internal}
