% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_assess.R
\name{calculatef2}
\alias{calculatef2}
\title{Calculate Cohen's f^2}
\usage{
calculatef2(.object = NULL)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}
}
\value{
A matrix with as many rows as there are structural equations. The
number of columns is equal to the total number of right-hand side variables
of these equations.
}
\description{
Calculate the effect size for regression analysis \insertCite{Cohen1992}{cSEM}
known as Cohen's f^2.
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=assess]{assess()}}, \link{csem}, \link{cSEMResults}
}
