% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_doTrees.R
\name{calculateFITForSplit}
\alias{calculateFITForSplit}
\title{Calculate the difference in FIT}
\usage{
calculateFITForSplit(
  data,
  idx,
  .model,
  .id,
  .approach_weights,
  .conv_criterion,
  .disattenuate,
  .dominant_indicators,
  .iter_max,
  .tolerance
)
}
\arguments{
\item{data}{Data passed by boot}

\item{idx}{Row-indices passed by boot}

\item{.model}{A model in \link[lavaan:model.syntax]{lavaan model syntax}
or a \link{cSEMModel} list.}

\item{.id}{Character string or integer. A character string giving the name or
an integer of the position of the column of \code{.data} whose levels are used
to split \code{.data} into groups. Defaults to \code{NULL}.}

\item{.approach_weights}{Character string. Approach used to
obtain composite weights. One of: "\emph{PLS-PM}", "\emph{SUMCORR}", "\emph{MAXVAR}",
"\emph{SSQCORR}", "\emph{MINVAR}", "\emph{GENVAR}", "\emph{GSCA}", "\emph{IGSCA}", "\emph{PCA}", "\emph{unit}", "\emph{bartlett}",
or "\emph{regression}". Defaults to "\emph{PLS-PM}".}

\item{.conv_criterion}{Character string. The criterion to use for the convergence check.
One of: "\emph{diff_absolute}", "\emph{diff_squared}", "\emph{diff_relative}", or "\emph{sum_diff_absolute}". Defaults
to "\emph{diff_absolute}".}

\item{.disattenuate}{Logical. Should composite/proxy correlations
be disattenuated to yield consistent loadings and path estimates if at least
one of the construct is modeled as a common factor? Defaults to \code{TRUE}.}

\item{.dominant_indicators}{A character vector of \code{"construct_name" = "indicator_name"} pairs,
where \code{"indicator_name"} is a character string giving the name of the dominant indicator
and \code{"construct_name"} a character string of the corresponding construct name.
Dominant indicators may be specified for a subset of the constructs.
Default to \code{NULL}.}

\item{.iter_max}{Integer. The maximum number of iterations allowed.
If \code{iter_max = 1} and \code{.approach_weights = "PLS-PM"} one-step weights are returned.
If the algorithm exceeds the specified number, weights of iteration step
\code{.iter_max - 1}  will be returned with a warning. Defaults to \code{100}.}

\item{.tolerance}{Double. The tolerance criterion for convergence.
Defaults to \code{1e-05}.}
}
\value{
Named vector of the delta between the FIT of the multi-group model
versus the FIT of the single-group model--positive values favor
multi-group, negative, single-group; the FIT of the single-group model; and
the FIT of the multi-group model.
}
\description{
Calculate the difference in FIT
}
\keyword{internal}
