% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_doModelSearch.R
\name{calculateFitness}
\alias{calculateFitness}
\title{Internal: Function that calculates the fitness value}
\usage{
calculateFitness(
  .matrix_vector,
  .data,
  .model_org,
  .only_structural,
  .ms_criterion,
  .fbar
)
}
\value{
numeric
}
\description{
Function that calculates the fitness value of a model.
This function adheres to the necessary requirements  for a function to be supplied to \code{\link[GA]{ga}}.
}
\keyword{internal}
