% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_test_MGD.R
\name{calculateFR}
\alias{calculateFR}
\title{Internal: ANOVA F-test statistic}
\usage{
calculateFR(
 .parameter,
 .id
 )
}
\value{
A named scaler, the test statistic of the ANOVA F-test
}
\description{
Calculate the ANOVA F-test statistic suggested by
\insertCite{Sarstedt2011;textual}{cSEM}.
}
\references{
\insertAllCited{}
}
\keyword{internal}
