% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsca_fit_measures.R
\name{calculateGFI}
\alias{calculateGFI}
\title{Calculate the GFI}
\usage{
calculateWeightsGSCA(.data, .model)
}
\arguments{
\item{.object}{An R object of class \code{cSEM<class>} with corresponding method.}
}
\value{
A list with the elements
\describe{
\item{\code{$W}}{A (J x K) matrix of estimated weights.}
\item{\code{$E}}{A (J x J) matrix of inner weights.}
\item{\code{$Modes}}{A named vector of Modes used for the outer estimation.}
\item{\code{$Conv_status}}{The convergence status. One of \code{TRUE}or \code{FASLE}. If
one-step weights are used via \code{.iter_max = 1} or no iterative algorithm was used,
the convergence status is set to \code{NULL}.}
\item{\code{$Iterations}}{The number of iterations used.}
}
}
\description{
Calculates GFI
}
\details{
Some more description...
}
