% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_estimators_weights.R
\name{calculateGSCADegreesOfFreedom}
\alias{calculateGSCADegreesOfFreedom}
\title{Calculate the degrees of freedom for GSCA-type models}
\usage{
calculateGSCADegreesOfFreedom(.object)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}
}
\value{
A named numeric vector of length 3 with elements \code{npar}, \code{d_0} and
\code{d_1}.
}
\description{
Computes the number of model parameters (\code{npar}), the total number of
degrees of freedom (\code{d_0}) and the residual degrees of freedom
(\code{d_1 = d_0 - npar}) of a GSCA-type model. Following
\insertCite{Hwang2014;textual}{cSEM} and the \code{gesca.mg} function of the
gesca package, the number of path, loading, weight and unique loading
estimates is multiplied by the number of groups for multigroup models.
}
\details{
This function is not intended to be user-facing.
}
\keyword{internal}
