% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_assess.R
\name{calculateHTMT2}
\alias{calculateHTMT2}
\title{HTMT2}
\usage{
calculateHTMT2(
 .object               = NULL,
 .absolute             = TRUE,
 .alpha                = 0.05,
 .handle_inadmissibles = c("drop", "ignore", "replace"),
 .inference            = FALSE,
 .only_common_factors  = TRUE,
 .R                    = 499,
 .seed                 = NULL
)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}

\item{.only_common_factors}{Logical. Should only concepts modeled as common
factors be included when calculating one of the following quality critera:
AVE, the Fornell-Larcker criterion, HTMT, and all reliability estimates.
Defaults to \code{TRUE}.}
}
\value{
A lower tringular matrix of HTMT values. If \code{.inference = TRUE}
the upper tringular part is the .alpha\%-quantile of the HTMT's bootstrap
distribution.
}
\description{
HTMT2
}
\seealso{
\code{\link[=assess]{assess()}}, \link{csem}, \link{cSEMResults}
}
