% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_doTrees.R
\name{calculateIgscaObjectiveFunction}
\alias{calculateIgscaObjectiveFunction}
\title{Calculate I-GSCA's Objective Function}
\usage{
calculateIgscaObjectiveFunction(.object = NULL)
}
\arguments{
\item{.object}{}
}
\value{
Sum of Squares of Unexplained Variance
}
\description{
Numerator of the fraction of unexplained variance in the data of the FIT statistic for GSCA models.
}
\keyword{internal}
