% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators_weights.R
\name{calculateInnerWeightsPLS}
\alias{calculateInnerWeightsPLS}
\title{Calculate the inner weights for PLS}
\usage{
calculateInnerWeightsPLS(
  .S                           = NULL,
  .W                           = NULL,
  .csem_model                  = NULL,
  .PLS_ignore_structural_model = NULL,
  .PLS_weight_scheme_inner     = NULL
)
}
\arguments{
\item{.S}{The (K x K) empirical indicator correlation matrix.}

\item{.W}{A list with elements \code{$W}, \code{$E}, \code{Modes}, \code{Conv_status}, and
\code{Iterations}.}

\item{.csem_model}{A (possibly incomplete) \link{cSEMModel}-list.}

\item{.PLS_ignore_structural_model}{Logical. Should the structural model be ignored
when calculating the inner weights of the PLS algorithm? Defaults to \code{FALSE}.
Ignored if \code{.approach_weights} is not PLS.}

\item{.PLS_weight_scheme_inner}{Character string. The inner weighting scheme
used in PLS. One of: "\emph{centroid}", "\emph{factorial}", or "\emph{path}".
Defaults to "\emph{centroid}". Ignored if \code{.approach_weight} is not PLS.}
}
\value{
The (J x J) matrix \code{E} of inner weights.
}
\description{
Calculates inner weights
}
\details{
More description here.
}
