% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_loadings.R
\name{calculateLoadings}
\alias{calculateLoadings}
\title{Calculate factor and composite loadings and cross-loadings}
\usage{
calculateLoadings(.S = NULL, .W = NULL, .Q = NULL, .csem_model = NULL,
  .disattenuate = NULL, .modes = NULL)
}
\arguments{
\item{.S}{The (scaled) empirical (K x K) indicator covariance/correlation matrix,
where K is the number of indicators.}

\item{.W}{A list with elements \code{$W}, \code{$E}, \code{Modes}, \code{Conv_status}, and
\code{Iterations}.}

\item{.Q}{A vector of proxy-construct correlations with element names equal to
the names of the J construct names used in the measurement model. Note
Q^2 is also called the reliability coefficient.}

\item{.csem_model}{A (possibly incomplete) \link{cSEMModel} list.}

\item{.disattenuate}{Logical. Should proxy correlations be disattenuated
if the construct is modeled as a common factor? Defaults to \code{TRUE}.}

\item{.modes}{A named vector giving the mode used to obtain new outer weights.}
}
\value{
The (J x K) matrix of loadings and "cross loadings" (attenuated if required).
}
\description{
Calculates factor loadings (for constructs modeled as common factors) and
composite loadings (for constructs modeled as composites).
}
\details{
Here more details
}
