% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_predict.R
\name{calculateMAE}
\alias{calculateMAE}
\title{Internal: Calculate prediction metrics}
\usage{
calculateMAE(resid)
}
\value{
A vector of the prediction measures for the observed variables
belonging to endogenous constructs
}
\description{
Currently, the following prediction measures are available:
}
\details{
\itemize{
\item Mean absolute error
\item Mean absolute percentage error
\item Mean squared error (MSE)
\item Root mean squared error
\item Theil's forecast accuracy
\item Theil's forecast quality
\item Bias proportion of MSE
\item Regression proportion of MSE
\item Disturbance proportion of MSE
}
}
\keyword{internal}
