% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators_weights.R
\name{calculateOuterWeightsPLS}
\alias{calculateOuterWeightsPLS}
\title{Calculate the outer weights for PLS}
\usage{
calculateOuterWeightsPLS(
   .S      = NULL,
   .W      = NULL,
   .E      = NULL,
   .modes  = NULL
   )
}
\arguments{
\item{.S}{The (K x K) empirical indicator correlation matrix.}

\item{.W}{A list with elements \code{$W}, \code{$E}, \code{Modes}, \code{Conv_status}, and
\code{Iterations}.}

\item{.E}{A (J x J) matrix of inner weights.}

\item{.modes}{A named vector giving the mode used to obtain new outer weights.}
}
\value{
A (J x K) matrix of outer weights.
}
\description{
Calculates outer weights
}
\details{
More description here..
}
