% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_test_MGD.R
\name{calculateParameterDifference}
\alias{calculateParameterDifference}
\title{Parameter differences across groups}
\usage{
calculateParameterDifference(
  .object     = NULL,
  .comparison = args_default()$.comparison
)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}

\item{.comparison}{A model in \link[lavaan:model.syntax]{lavaan model syntax} indicating which
parameters are compared.}
}
\value{
A list of length equal to the number of possible combinations of
groups in .object (basically n choose k = 2), e.g., 3 if there are three
groups and 6 if there are 4 groups. Each list elements contains the
values of the difference between the parameter estimates based on the
data of group i and and group j.
}
\description{
Calculates the difference between one or more paramater estimates based
one different groups (data sets).
}
\keyword{internal}
