% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_test_MGD.R
\name{calculatePr}
\alias{calculatePr}
\title{Internal: Calculation of the CDF used in Henseler et al. (2009)}
\usage{
calculatePr(.resample_centered = NULL, .parameters_to_compare = NULL)
}
\arguments{
\item{.parameters_to_compare}{A model in \link[lavaan:model.syntax]{lavaan model syntax} indicating which
parameters (i.e, path (\code{~}), loadings (\verb{=~}), weights (\verb{<~}), or correlations (\verb{~~})) should be
compared across groups. Defaults to \code{NULL} in which case all parameters of the
originally specified model are compared.}
}
\value{
A named vector
}
\description{
Calculates the probability that theta^1 is smaller than or equal to theta^2.
See Equation (6) in \insertCite{Sarstedt2011;textual}{cSEM}.
}
\references{
\insertAllCited{}
}
\keyword{internal}
