% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_auxillary.R
\name{calculateProxies}
\alias{calculateProxies}
\title{Internal: Calculate proxies}
\usage{
calculateProxies(.X = NULL, .W = NULL)
}
\arguments{
\item{.W}{A list with elements \code{$W}, \code{$E}, \code{PLS_mode} and xxx.}
}
\value{
The (N x J) matrix of proxy values with column names equal to
the names of the J constructs.
}
\description{
Calculates proxy values for the J constructs of the structural model.
}
\details{
More on that
}
