% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_auxillary.R
\name{calculateProxyVCV}
\alias{calculateProxyVCV}
\title{Internal: Calculate proxy variance-covariance matrix}
\usage{
calculateProxyVCV(.S = NULL, .W = NULL)
}
\arguments{
\item{.S}{The (K x K) empirical indicator correlation matrix.}

\item{.W}{A list with elements \code{$W}, \code{$E}, \code{Modes}, \code{Conv_status}, and
\code{Iterations}.}
}
\value{
A (J x J) proxy VCV matrix.
}
\description{
Calculates the variance-covariance (VCV) matrix of the proxies.
}
\details{
More on that
}
