% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_assess.R
\name{calculateRA}
\alias{calculateRA}
\title{Internal: Calculate a redundancy analysis}
\usage{
calculateRA(.object = NULL)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}
}
\value{
A named numeric vector of correlations obtained by the RA. If
the weightning approach is not \code{"PLS-PM"} or non of the PLS outer modes
was mode B, the function silently returns \code{NA}.
}
\description{
Calculate/do a redundancy analysis (RA) as proposed by Hair (2014) with
reference to Chin (1998).
}
\details{
According to Hair (2014), redundancy analysis (RA)
is the process of regressing the scores of a reflectivly measured construct
on the scores of a formatively measured construct in order to gain empirical
evidence for convergent validity of a formatively measured construct.
RA is therefore confined to PLS, specifically PLS with at least one construct
whose mode is Mode B. This is the case if the construct is modeled as a
composite or if the construct was explicitly given Mode B.
Hence RA is only conducated if \code{.approach_weights = "PLS-PM"} and if at least
one construct's mode is Mode B.

The principal idea of RA is to take two different measures of the
same construct and regress the scores obtained for each measure on each
other. If they are similar they are likely to measure the same "thing"
which is then taken as evidence that both measurement models actually
measure what they are supposed to measure (validity).

There are several issues with the termionlogy and the reasoning behind this logic.
RA is therefore only implemented since reviewers are likely to demand
its compuation, however, its actual application for validitiy assessment
is strongly discouraged.

The function is only applicable to objects of class \code{cSEMResults_default}.
}
\references{
\insertAllCited{}
}
\seealso{
\link{cSEMResults}
}
\keyword{internal}
