% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_assess.R
\name{calculateRelativeGoF}
\alias{calculateRelativeGoF}
\title{Relative Goodness of Fit (relative GoF)}
\usage{
calculateRelativeGoF(
 .object              = NULL
)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}
}
\value{
A single numeric value.
}
\description{
Calculate the Relative Goodness of Fit (GoF) proposed by \insertCite{Vinzi2010a;textual}{cSEM}.
Note that, contrary to what the name suggests, the Relative GoF is \strong{not} a
measure of model fit in the sense of SEM. See e.g. \insertCite{Henseler2012a;textual}{cSEM}
for a discussion.
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=assess]{assess()}}, \link{cSEMResults}
}
