% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_foreman.R
\name{calculateReliabilities}
\alias{calculateReliabilities}
\title{Internal: Calculate Croon Reliabilities}
\usage{
calculateReliabilities(.csem_model = args_default()$.csem_model,
  .W = args_default()$.W, .X = args_default()$.X)
}
\arguments{
\item{.csem_model}{A (possibly incomplete) \link{cSEMModel}-list.}

\item{.W}{A (J x K) matrix of weights.}

\item{.X}{A matrix of processed data (scaled, cleaned and ordered).}
}
\description{
(TODO)
}
\keyword{internal}
