% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_assess.R
\name{calculateRMS}
\alias{calculateRMS}
\title{Internal: RMS}
\usage{
RMS(.object)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}
}
\description{
Compute the RMS.
}
\seealso{
\link{assess}, \link{csem}, \link{cSEMResults}
}
\keyword{internal}
