% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_new.R
\name{calculateVIFPLS}
\alias{calculateVIFPLS}
\alias{calculateVIFPLS.cSEMResults_default}
\alias{calculateVIFPLS.cSEMResults_multi}
\alias{calculateVIFPLS.cSEMResults_2ndorder}
\title{Variance Inflation Factor for weight estimates obtained by Mode B}
\usage{
calculateVIFPLS(
 .object              = args_default()$.object,
)


  \method{calculateVIFPLS}{cSEMResults_default}(.object = args_default()$.object)


  \method{calculateVIFPLS}{cSEMResults_multi}(.object = args_default()$.object)


  \method{calculateVIFPLS}{cSEMResults_2ndorder}(.object = args_default()$.object)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}
}
\description{
(TODO)
}
\section{Methods (by class)}{
\itemize{
\item \code{cSEMResults_default}: (TODO)

\item \code{cSEMResults_multi}: (TODO)

\item \code{cSEMResults_2ndorder}: (TODO)
}}

\examples{
\dontrun{
# still to implement
}

}
\seealso{
\link{csem}, \link{cSEMResults}
}
