% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators_weights.R
\name{calculateWeightsKettenring}
\alias{calculateWeightsKettenring}
\title{Calculate weights using one of Kettenrings's approaches}
\usage{
calculateWeightsKettenring(.data, .model, .criteria)
}
\arguments{
\item{.data}{A data frame or a matrix containing the raw data.}

\item{.model}{A model in \href{http://lavaan.ugent.be/tutorial/syntax1.html}{lavaan model syntax}
or a \link{cSEMModel} object.}

\item{.criteria}{Character string. The criteria to use. One of "\emph{SUMCOR}", "\emph{MAXVAR}",
"\emph{SSQCOR}", "\emph{MINVAR}" or "\emph{GENVAR}".}
}
\value{
The (J x K) matrix of estimated weights.
}
\description{
Calculates weights according to on of the the five criteria
"\emph{SUMCOR}", "\emph{MAXVAR}", "\emph{SSQCOR}", "\emph{MINVAR}", and "\emph{GENVAR}"
suggested by \insertCite{Kettenring1971;textual}{cSEM}.
}
\details{
Some more description...
}
