% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_cca.R
\name{cca}
\alias{cca}
\title{Confirmatory composite analysis}
\usage{
cca(.data, .model, ...)
}
\arguments{
\item{.data}{A \code{data.frame} or a \code{matrix} containing the raw data.}

\item{.model}{A model in \href{http://lavaan.ugent.be/tutorial/syntax1.html}{lavaan model syntax}
or a \link{cSEMModel} object.}

\item{...}{Further arguments to be passed down to other functions.
See \code{\link[=args_default]{args_default()}} with \code{.only_dots = TRUE} for a complete list
of possible arguments and their defaults.}
}
\value{
An object of class \code{cSEMResults} for which the following methods exist:
\describe{
\item{\code{print.cSEMResults}}{Prints a message to inform the user
whether estimation has been successful and what functions may be used
to examine the object.}
\item{\code{summary.cSEMResults}}{Print and return a comprehensive summary of the results.}
}
Technically \code{cSEMResults} is a named list containing the following list elements:
\describe{
\item{\code{$Estimates}}{A list containing the estimated quantities.}
\item{\code{$Information}}{A list of additional information. (incomplete)}
}
}
\description{
Conduct a confirmatory composite analysis (CCA).
}
\details{
More details here.
}
\examples{
\dontrun{
# still to implement
}

}
