% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.alpha.R
\name{check.alpha}
\alias{check.alpha}
\title{Calculate Cronbach's Alpha for Two Halves}
\usage{
check.alpha(half1, half2, B = 1000, conf = 0.95)
}
\arguments{
\item{half1}{First half of the items}

\item{half2}{Second half of the items}

\item{B}{Number of bootstrap resamples (default = 1000)}

\item{conf}{Confidence level for the confidence interval (default = 0.95)}
}
\value{
Data frame with Cronbach's alpha estimates and confidence intervals
}
\description{
Calculate Cronbach's Alpha for Two Halves
}
\details{
`check.alpha` calcula el coeficient alfa (Cronbach, 1951) para cada mitad, más
intervalos de confianza usando bootstrap resamples.
}
\examples{
### Example 1 ----------
check.alpha(half1 = dbase[, 1:5], half2 = dbase[, 6:10], B = 1000, conf = .95)



}
\references{
Cronbach, L. J. (1951). Coefficient alpha and the internal structure of tests.
 Psychometrika, 16, 297-334.
}
\seealso{
\code{\link[lavaan]{cfa}}, \code{\link[scripty]{mimic}}, \code{\link[scripty]{prods}}
}
\author{
Cesar Merino-Soto \email{sikayax@yahoo.com.ar}
}
