% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.angoff.R
\name{check.angoff}
\alias{check.angoff}
\title{Calculate Angoff Coefficient}
\usage{
check.angoff(half1, half2, B = 1000, conf = 0.95)
}
\arguments{
\item{half1}{First half of the items}

\item{half2}{Second half of the items}

\item{B}{Number of bootstrap resamples (default = 1000)}

\item{conf}{Confidence level for the confidence interval (default = 0.95)}
}
\value{
Data frame with the Angoff coefficient and confidence interval
}
\description{
Calculate Angoff Coefficient
}
