% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.distribution.R
\name{check.distribution}
\alias{check.distribution}
\title{Distribution Tests: Anderson-Darling, Overlapping Index, Kendall's W}
\usage{
check.distribution(half1, half2, B = 2000, conf = 0.95)
}
\arguments{
\item{half1}{First half of the items}

\item{half2}{Second half of the items}

\item{B}{Number of bootstrap resamples (default = 2000)}

\item{conf}{Confidence level for the confidence interval (default = 0.95)}
}
\value{
Data frame with the results of the distribution tests
}
\description{
Distribution Tests: Anderson-Darling, Overlapping Index, Kendall's W
}
