% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_check.R
\name{check}
\alias{check}
\title{Check fit}
\usage{
check(.object)
}
\arguments{
\item{.object}{An R object of class \code{cSEM<class>} with corresponding method.}
}
\value{
(TODO)
}
\description{
Check model fit using common fit measures.
}
\details{
The following fit measures are used (TODO).
}
\seealso{
\code{\link[=csem]{csem()}}, \code{\link[=cca]{cca()}}, \code{\link[=foreman]{foreman()}}, \link{cSEMResults}
}
