% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.split.R
\name{check.split}
\alias{check.split}
\title{Dividir ítems en mitades}
\usage{
check.split(
  data,
  method = c("ritc", "optimized", "diffict"),
  n_iter = 1000,
  seed = 123,
  half.eq = c("equal", "unequal")
)
}
\arguments{
\item{data}{Dataframe con los ítems}

\item{method}{Método de división ("ritc", "diffict", "optimized")}

\item{seed}{Semilla para la reproducibilidad}

\item{half.eq}{Indica si las mitades deben tener igual número de ítems}
}
\value{
Una lista con las mitades y estadísticas relacionadas
}
\description{
Esta función divide los ítems en dos mitades usando distintos métodos.
}
