% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_plot.cSEMResults.R
\name{check_connection}
\alias{check_connection}
\title{Internal: Check whether two indicators belong to the same construct.}
\usage{
check_connection(
  .indicator1,
  .indicator2,
  .model_measurement,
  .model_error_cor
)
}
\arguments{
\item{.indicator1, }{.indicator2 Character string. The names of the indicators.}

\item{.model_measurement}{a matrix. The measurement matrix mimicking the relationship
between constructs and indicators.}

\item{.model_error_cor}{a matrix. The matrix mimicking the error correlation structure.}
}
\value{
TRUE if both indicators belong to the same construct, FALSE otherwise.
}
\description{
Checks whether two indicators belong to the same construct.
}
\keyword{internal}
