% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_doModelSearch.R
\name{check_matrix_criteria}
\alias{check_matrix_criteria}
\title{Internal:  --- Matrix utils ------------------------------}
\usage{
check_matrix_criteria(.matrix, .model_org)
}
\description{
Internal:  --- Matrix utils ------------------------------
}
\keyword{internal}
