% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_doModelSearch.R
\name{checkCycles}
\alias{checkCycles}
\title{Internal: Function that search for cycles}
\usage{
checkCycles(.matrix)
}
\value{
Logical vector \code{'TRUE'}: Cycle detected; \code{'FALSE'} No cycle detected
}
\description{
Function that searches for cycles in a binary indicator matrix using the depth first search (DFS).
}
\keyword{internal}
