% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_doModelSearch.R
\name{checkIsolatedConstruct}
\alias{checkIsolatedConstruct}
\title{Internal: Check whether constructs in the structural model are isolated}
\usage{
checkIsolatedConstruct(.matrix)
}
\value{
logical: \code{'TRUE'}: Isolated; \code{'FALSE'}:connected to at least one variable via a path.
}
\description{
Internal: Check whether constructs in the structural model are isolated
}
\keyword{internal}
