% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_auxillary.R
\name{classifyConstructs}
\alias{classifyConstructs}
\title{Internal: Classify interaction terms by type}
\usage{
classifyConstructs(.terms)
}
\arguments{
\item{.terms}{A vector of construct names to be classified.}
}
\value{
A named list of length equal to the number of terms provided containing
a data frame with columns "\emph{Term_class}", "\emph{Component}",
"\emph{Component_type}", and "\emph{Component_freq}".
}
\description{
Classifies interaction between constructs according to their type.
}
\details{
Classification is required to estimate nonlinear structural relationships.
Note that exponential terms are modeled as "interactions with itself"
(i.e. x^3 = x.x.x).  The following terms and classifications are currently
supported:

\itemize{
\item Single term: e.g., \code{eta1}
\item Quadratic term: e.g., \code{eta1.eta1}
\item Cubic term: e.g., \code{eta1.eta1.eta1}
\item Two-way interactions term: e.g., \code{eta1.eta2}
\item Three-way interactions term: e.g., \code{eta1.eta2.eta3}
\item Quadratic and two-way interaction term: e.g., \code{eta1.eta1.eta3}"
}
}
\examples{
x <- c("eta1.eta2", "eta1", "eta3.eta1.eta1")

classifyConstructs(x)

}
