% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_plot.cSEMResults.R
\name{cleanNode}
\alias{cleanNode}
\title{Internal: Clean a node name.}
\usage{
cleanNode(node)
}
\arguments{
\item{node}{A node name.}
}
\value{
A cleaned node name.
}
\description{
Removes a trailing "_temp" from a node name.
}
\keyword{internal}
