% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_foreman.R
\name{compute_cor}
\alias{compute_cor}
\title{Internal: Calculate correlation matrix between the indicators (Bravais-Pearson
or Theil-Sen)}
\usage{
compute_cor(X_cleaned, .cor_kind)
 )
}
\value{
The (N x N) indicator correlation matrix
}
\description{
Internal: Calculate correlation matrix between the indicators (Bravais-Pearson
or Theil-Sen)
}
\keyword{internal}
