% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_estimators_weights.R
\name{constructGSCAObjectiveParts}
\alias{constructGSCAObjectiveParts}
\title{Constructs the different matrices of the GSCA-type Least Squares Objective function}
\usage{
constructGSCAObjectiveParts(.object)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}
}
\value{
List of Psi, Z, Eta, A, Lambda, B, S and UD matrices
}
\description{
This function should be called through \code{calculateFIT}, \code{calculateFIT_s}, etc... and is not intended to be user-facing.
}
