% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igsca.R, R/igsca_translation.R
\name{convert_matlab2R}
\alias{convert_matlab2R}
\title{Convert Matlab to R}
\usage{
convert_matlab2R(mat_env, vectorness)

convert_matlab2R(mat_env, vectorness)
}
\arguments{
\item{mat_env}{The (piped-in) matlab lab environment from readMat()}

\item{vectorness}{Vector object that lists the different objects in the pre-swap environment that were vectors}
}
\value{
Convert particular Matlab objects to be compatible with idiomatic R programming.

Convert particular Matlab objects to be compatible with idiomatic R programming.
}
\description{
Matlab vectors are matrices when read into R, so here we perform the conversions to make them more R-friendly. This is also a placeholder for any future needed adjustments to make the Matlab environment swappable with R

Matlab vectors are matrices when read into R, so here we perform the conversions to make them more R-friendly. This is also a placeholder for any future needed adjustments to make the Matlab environment swappable with R
}
