% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csem_model.R
\name{convertModel}
\alias{convertModel}
\title{Convert second order cSEMModel}
\usage{
convertModel(
 .csem_model        = NULL, 
 .approach_2ndorder = "2stage",
 .stage             = "first"
 )
}
\arguments{
\item{.csem_model}{A (possibly incomplete) \link{cSEMModel}-list.}

\item{.approach_2ndorder}{Character string. Approach used for models containing
second order constructs. One of: "\emph{2stage}", "\emph{RI_original}", "\emph{RI_extended}"
or "\emph{mixed}". Defaults to "\emph{2stage}".}

\item{.stage}{Character string. The stage the model is need for.
One of "\emph{first}" or "\emph{second}". Defaults to "\emph{first}".}
}
\value{
A \link{cSEMModel} list that may be passed to any function requiring
\code{.csem_model} as a mandatory argument.
}
\description{
Uses a \link{cSEMModel} containg second order constructs and turns it into an
estimable model using either the "repeated indicators" approach or a
two-step procedure (TODO; link to literature)
}
\keyword{internal}
