% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_foreman.R
\name{Cov_theilsen}
\alias{Cov_theilsen}
\title{Internal: Calculate correlation matrix of a data set using the Theil-Sen
estimator}
\usage{
Cov_theilsen(X_cleaned)
 )
}
\value{
The (J x J) construct VCV matrix. Disattenuated if requested.
}
\description{
Internal: Calculate correlation matrix of a data set using the Theil-Sen
estimator
}
\keyword{internal}
