% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_new.R
\name{CR}
\alias{CR}
\title{CR}
\usage{
CR(
 .object              = args_default()$.object,
 .only_common_factors = args_default()$.only_common_factors
)
}
\arguments{
\item{.object}{An R object of class \code{cSEM<class>} with corresponding method.}

\item{.only_common_factors}{Logical. Should only common factors be included?
Defaults to \code{FALSE}.}
}
\description{
Computes the composite reliability \insertCite{Raykov1997} based on standardized loading.
}
\examples{
\dontrun{
# still to implement
}

}
\references{
\insertAllCited{}
}
\seealso{
\link{csem}, \link{cSEMResults}
}
