% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_package.R
\docType{package}
\name{cSEM-package}
\alias{cSEM}
\alias{cSEM-package}
\title{cSEM: A package for composite-based structural equation modeling}
\description{
Estimate, analyze, test, and study linear, nonlinear, hierarchical and
multigroup structural equation models using composite-based approaches and procedures including estimation
techniques such as partial least squares path modeling (PLS) and its derivatives
(PLSc, ordPLSc, robustPLSc), generalized structured component analysis (GSCA),
generalized structured component analysis with uniqueness terms (GSCAm),
generalized canonical correlation analysis (GCCA) unit weights (sum score) and
fixed weights, as well as several tests and typical postestimation
procedures (e.g., assess the model fit, compute direct, indirect and total effects).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FloSchuberth/cSEM/}
  \item \url{https://floschuberth.github.io/cSEM/}
  \item Report bugs at \url{https://github.com/FloSchuberth/cSEM/issues/}
}

}
\author{
\strong{Maintainer}: Florian Schuberth \email{f.schuberth@utwente.nl} (\href{https://orcid.org/0000-0002-2110-9086}{ORCID})

Authors:
\itemize{
  \item Florian Schuberth \email{f.schuberth@utwente.nl} (\href{https://orcid.org/0000-0002-2110-9086}{ORCID})
  \item Manuel E. Rademaker (\href{https://orcid.org/0000-0002-8902-3561}{ORCID})
}

Other contributors:
\itemize{
  \item Tamara Schamberger (\href{https://orcid.org/0000-0002-7845-784X}{ORCID}) [contributor]
  \item Michael Klesel (\href{https://orcid.org/0000-0002-2884-1819}{ORCID}) [contributor]
  \item Huu Phuc Nguyen (\href{https://orcid.org/0009-0000-9344-7132}{ORCID}) [contributor]
  \item Theo K. Dijkstra [contributor]
  \item Jörg Henseler (\href{https://orcid.org/0000-0002-9736-3048}{ORCID}) [contributor]
  \item Gloria Pietropolli (\href{https://orcid.org/0000-0001-7623-8419}{ORCID}) [contributor]
  \item Kjell S. Slupphaug (\href{https://orcid.org/0009-0005-8324-2834}{ORCID}) [contributor]
  \item Jason J. Berger (\href{https://orcid.org/0009-0002-5207-9032}{ORCID}) [contributor]
  \item Michael S. Truong \email{emstruonger@gmail.com} (\href{https://orcid.org/0009-0006-4521-6528}{ORCID}) [contributor]
}

}
\keyword{internal}
