% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_templates.R, R/zz_utils.R
\name{csem_resultssummary}
\alias{csem_resultssummary}
\alias{cSEMResultssummary}
\alias{csem_resultssummary}
\alias{cSEMResultssummary}
\title{cSEMResultssummary}
\value{
An object of class \code{cSEMResultssummary} for which the following methods exist:
\describe{
\item{\code{print.cSEMResultssummary}}{Prints a summary.}
}
Technically \code{cSEMResultssummary} is a named list containing the following list elements:
\describe{
\item{`...}{Not finished yet.}
}

An object of class \code{cSEMResultssummary} for which the following methods exist:
\describe{
\item{\code{print.cSEMResultssummary}}{Prints a summary.}
}
Technically \code{cSEMResultssummary} is a named list containing the following list elements:
\describe{
\item{`...}{Not finished yet.}
}
}
\description{
cSEMResultssummary

cSEMResultssummary
}
