% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_templates.R
\name{csem_testresults}
\alias{csem_testresults}
\alias{cSEMTestResults}
\title{cSEMTestResults}
\value{
A standarized list of class \code{cSEMTestResults} for which the following methods exists:
\describe{
\item{\code{print.cSEMTestResults}}{A formated summary of the test.}
\item{\code{summary.cSEMTestResults}}{A formated summary of the test (identical to \code{print.cSEMTestResults}).}
}
Technically \code{cSEMTestResults} is a named list containing the following list elements:
\describe{
\item{\code{$Test_statistic}}{The value of test statistic.}
\item{\code{$Critical_value}}{The critical value.}
\item{\code{$Decision}}{The test decision. One of: \strong{Reject} or \emph{Do not reject}}
\item{\code{$Number_addmissibles}}{The number of admissible runs. See xxx for what
constitutes an \strong{inadmissible run}.}
}
}
\description{
cSEMTestResults
}
