% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_polycor.R
\name{dbinorm}
\alias{dbinorm}
\title{Internal: Bivariate standard normal density}
\usage{
dbinorm(u, v, rho, force.zero = FALSE, rho.lim = 0.9999)
}
\description{
Evaluate the density of the bivariate standard normal distribution with
correlation \code{rho} at \verb{(u, v)}. Used internally by \code{\link[=polychor]{polychor()}} to
compute the gradient of the polychoric log-likelihood. Implemented by Kjell S. Slupphaug.
}
\references{
\insertAllCited{}
}
\keyword{internal}
