% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_test_MGD.R
\name{decide_by_critical}
\alias{decide_by_critical}
\title{Decision fot test of multigroup difference is based on the critical values instead
of p-values}
\usage{
decide_by_critical=function(
.test_MGD = NULL,
.approach_alpha_adjust = args_default()$.approach_alpha_adjust,
.alpha = args_default()$.alpha)
}
\arguments{
\item{.alpha}{An integer or a numeric vector of significance levels.
Defaults to \code{0.05}.}
}
\description{
Decision fot test of multigroup difference is based on the critical values instead
of p-values
}
\references{
\insertAllCited{}
}
\seealso{
\link{cSEMResults}
}
