% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_doIPA.R
\name{doIPA}
\alias{doIPA}
\title{Do a importance performance analysis}
\usage{
doIPA(
 .object             = NULL
 )
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}
}
\value{
A list of class \code{cSEMIPA} with a corresponding method for \code{plot()}.
See: \code{\link[=plot.cSEMIPA]{plot.cSEMIPA()}}.
}
\description{
Description
}
\seealso{
\code{\link[=csem]{csem()}}, \link{cSEMResults}, \code{\link[=plot.cSEMIPA]{plot.cSEMIPA()}}
}
