% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_doModelSearch.R
\name{doModelSearch}
\alias{doModelSearch}
\title{Automated model specification search}
\usage{
doModelSearch(.object = NULL, 
                     .pop_size = 20, 
                     .n_generations= 20,
                     .prob_mutation = 0.5,
                     .prob_crossover = 0.8,
                     .fbar = -100000,
                     .ms_criterion = c('bic','aic','hq'),
                     .seed = NULL,
                     .only_structural = FALSE)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}

\item{.pop_size}{Integer. Value giving the population size used in the genetic algorithm.
Defaults to \code{20}.}

\item{.n_generations}{Integer. A value giving the number of generations used in the genetic
algorithm. Defaults to \code{20}.}

\item{.prob_mutation}{Scalar. Mutation probability used in the genetic algorithm.
Defaults to \code{0.5}.}

\item{.prob_crossover}{Scalar. Crossover probability used in the genetic algorithm.
Defaults to \code{0.8}.}

\item{.fbar}{Integer. Low fitness value that is used to penalize inadmissible models.
Defaults to -100000.}

\item{.ms_criterion}{Character string. Either a single character string or a vector
of character strings naming the model selection criterion to compute.
Defaults to \code{"all"}.}

\item{.seed}{Integer or \code{NULL}. The random seed to use. Defaults to \code{NULL} in which
case an arbitrary seed is chosen. Note that the scope of the seed is limited
to the body of the function it is used in. Hence, the global seed will
not be altered!}

\item{.only_structural}{Should the the log-likelihood be based on the
structural model? Ignored if \code{.by_equation == TRUE}. Defaults to \code{TRUE}.}
}
\value{
A list containing a list of model(s) showing the 'best' value of the model
selection criterion, the value of the model selection criteria, and a list of cSEM_models
}
\description{
\lifecycle{experimental}
}
\details{
Performs an automated model specification search using a genetic algorithm \insertCite{Holland1992}{cSEM} in combination
with partial least squares path modeling (PLS-PM).
In particular, the function implements the approach presented in \insertCite{Trinchera2026}{cSEM}.
}
\references{
\insertAllCited{}

\insertAllCited{}
}
\seealso{
\link{cSEMResults}
}
