% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_assess.R
\name{doRedundancyAnalysis}
\alias{doRedundancyAnalysis}
\title{Internal: Do a redundancy analysis}
\usage{
doRedundancyAnalysis(.object = NULL)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}
}
\value{
A named numeric vector of correlations. If
the weighting approach is not \code{"PLS-PM"} or non of the PLS outer modes
was mode B, the function silently returns \code{NA}.
}
\description{
Calculate/do a redundancy analysis (RA) as proposed by \insertCite{Hair2016;textual}{cSEM}
with reference to \insertCite{Chin1998;textual}{cSEM}.
}
\details{
RA is therefore confined to PLS-PM, specifically PLS-PM with at least one construct
whose weight are obtained by Mode B. This is the case if the construct is modeled as a
composite or if the construct was explicitly given Mode B.
Hence RA is only conducated if \code{.approach_weights = "PLS-PM"} and if at least
one construct's mode is Mode B.

The principal idea of RA is to take two different measures of the
same construct and regress the scores obtained for each measure on each
other. If they are similar they are likely to measure the same "thing"
which is then taken as evidence that both measurement models actually
measure what they are supposed to measure (validity).

There are several issues with the termionlogy and the reasoning behind this logic.
RA is therefore only implemented since reviewers are likely to demand
its compuation, however, its actual application for validitiy assessment
is discouraged.

The function is only applicable to objects of class \code{cSEMResults_default}.
}
\references{
\insertAllCited{}
}
\seealso{
\link{cSEMResults}
}
\keyword{internal}
