% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_doSpecificationSearch.R
\name{doSpecificationSearch}
\alias{doSpecificationSearch}
\title{Specification search}
\usage{
doSpecificationSearch(
.object = NULL)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}
}
\value{
List
}
\description{
\lifecycle{experimental}
}
\details{
The \code{doSpecificationSearch()}
}
\examples{
model_IT_Flex="
# Composite models
ITComp  <~ ITCOMP1 + ITCOMP2 + ITCOMP3 + ITCOMP4
Modul   <~ MOD1 + MOD2 + MOD3 + MOD4
ITConn  <~ ITCONN1 + ITCONN2 + ITCONN3 + ITCONN4
ITPers  <~ ITPSF1 + ITPSF2 + ITPSF3 + ITPSF4

# Saturated structural model
ITPers ~ ITComp + Modul + ITConn
Modul  ~ ITComp + ITConn 
ITConn ~ ITComp 
"

# Estimate the model using PLS:
out <- csem(.data = ITFlex, 
            .model = model_IT_Flex,
            .resample_method = "none")


doSpecificationSearch(out)


}
\seealso{
\code{\link[=csem]{csem()}}, \link{cSEMResults}, \link[DiagrammeR]{grViz}
}
