% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_doModelSearch.R
\name{.agas_fitness}
\alias{.agas_fitness}
\title{Internal:  --- Fitness ---------------------------------------------------}
\usage{
.agas_fitness(
  .matrix_vector,
  .dataset_generated,
  .n_exogenous,
  .measurement_model,
  .variables,
  .only_structural
)
}
\description{
Internal:  --- Fitness ---------------------------------------------------
}
\keyword{internal}
