% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_doModelSearch.R
\name{.matrix_to_string}
\alias{.matrix_to_string}
\title{Internal:  --- Matrix utils ------------------------------}
\usage{
.matrix_to_string(.matrix)
}
\description{
Internal:  --- Matrix utils ------------------------------
}
\keyword{internal}
