% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_doTrees.R
\name{doTrees}
\alias{doTrees}
\title{Recursive Partitioning Algorithm for \emph{csem} GSCA Models}
\usage{
doTrees(
  data,
  model,
  covariates,
  influence = influence_ssr,
  splitter = NULL,
  control = igsca_tree_control()
)
}
\arguments{
\item{data}{A \code{data.frame} or a \code{matrix} of standardized or unstandardized
data (indicators/items/manifest variables). Possible column types or classes
of the data provided are: "\code{logical}", "\code{numeric}" ("\code{double}" or "\code{integer}"),
"\code{factor}" ("\code{ordered}" and/or "\code{unordered}"), "\code{character}" (converted to factor),
or a mix of several types.}

\item{model}{A model in \link[lavaan:model.syntax]{lavaan model syntax}
or a \link{cSEMModel} list.}

\item{covariates}{Character vector. List of variables for \code{\link[=doTrees]{doTrees()}}
to consider splitting on. See \link[partykit:mob]{partykit::mob}}
}
\value{
cSEM Tree model of class \code{modelparty} and \code{party}.
}
\description{
Recursive Partitioning Algorithm for \emph{csem} GSCA Models
}
\references{
\insertAllCited{}
}
