% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_doTrees.R
\name{doTreesBeta}
\alias{doTreesBeta}
\title{Prototype of doTrees}
\usage{
doTreesBeta(.object, .splitvars, .model, .R = 100)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}

\item{.splitvars}{List of the column names to try splitting on.}

\item{.model}{A model in \link[lavaan:model.syntax]{lavaan model syntax}}

\item{.R}{Integer. The number of bootstrap replications. Defaults to \code{499}.}
}
\value{
List of whether split occured and with what variable, the associated
paired t-test and the resulting model
}
\description{
Pending the computation of the gradient, here I use a manual multi-group
versus single group hypothesis test of the difference in FIT statistic,
similar to semtrees.
}
\details{
The current implementation assumes that once you use a split-variable, you
cannot re-use it. The final implementation will re-use variables.

\code{.splitvars} currently only supports binary variables
\code{.maxdepth} doesn't do anything
\code{.minsize} does not currently work
}
