% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_methods.R
\name{effects.cSEMResults}
\alias{effects.cSEMResults}
\title{Effects}
\usage{
effects(object)
}
\arguments{
\item{object}{An R object of class \link{cSEMResults}.
#' @param .only_dots Logical. Should only those arguments to be passed to lower level functions via the
\code{...} argument of the \link{csem} or \link{cca} function be returned. Defaults to \code{FALSE}.}
}
\description{
Compute direct, indirect and total effects.
}
\examples{
\dontrun{
# still to implement
}

}
\seealso{
\link{csem}, \link{cSEMResults}
}
