% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_foreman.R
\name{estimaterobustcor}
\alias{estimaterobustcor}
\title{Internal: Calculate the robust correlation#'}
\usage{
estimateTheilSen(
  .x = args_default()$.x, 
  .y = args_default()$.y)
}
\value{
A vector of length one.
}
\description{
Internal: Calculate the robust correlation#'
}
\keyword{internal}
