% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_foreman.R
\name{estimateTheilSen}
\alias{estimateTheilSen}
\title{Internal: Calculate the Theil-Sen estimator}
\usage{
estimateTheilSen(
  .x = args_default()$.x, 
  .y = args_default()$.y)
}
\value{
A vector of length one.
}
\description{
Calculate the Theil-Sen estimator (TODO, Reference) between two vectors.
}
\details{
(TODO) Description of what Theil-Sen does exactly.
}
\keyword{internal}
