% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_igsca.R
\name{extract_parseModel}
\alias{extract_parseModel}
\title{A parseModel extractor function for the purposes of running I-GSCA code example}
\usage{
extract_parseModel(model, data)
}
\arguments{
\item{model}{Specified Model in lavaan style}

\item{data}{Input dataframe}
}
\value{
Returns a list of matrices/vectors required for igsca_sim() to run:
\itemize{
\item Z0
\item W0
\item B0
\item C0
\item con_type
\item indicator_type
}
}
\description{
In the context of igsca, this function prepares: (1) the initial indicators
(Z0), weights (W0), structural (B0), loadings(C0) matrices; (2) whether a
construct is a latent or composite variable (con_type); (3) whether an
indicator corresponds to a latent or composite variable (indicator_type); and
(4) the dominant indicator of each construct (.dominant_indicators).
}
\examples{


tutorial_igsca_model <- "
# Composite Model
NetworkingBehavior <~ Behavior1 + Behavior2 + Behavior3 + Behavior5 + Behavior7 + Behavior8 + Behavior9
Numberofjobinterviews <~ Interview1 + Interview2
Numberofjoboffers <~ Offer1 + Offer2 

# Reflective Measurement Model
HonestyHumility =~ Honesty1 + Honesty2 + Honesty3 + Honesty4 + Honesty5 + Honesty6 + Honesty7 + Honesty8 + Honesty9 + Honesty10
Emotionality =~ Emotion1 + Emotion2 + Emotion3 + Emotion4 + Emotion5 + Emotion6 + Emotion8 + Emotion10
Extraversion =~ Extraver2 + Extraver3 + Extraver4 + Extraver5 + Extraver6 + Extraver7 + Extraver8 + Extraver9 + Extraver10
Agreeableness =~ Agreeable1 + Agreeable3 + Agreeable4 + Agreeable5 + Agreeable7 + Agreeable8 + Agreeable9 + Agreeable10
Conscientiousness =~ Conscientious1 + Conscientious3 + Conscientious4 + Conscientious6 + Conscientious7 + Conscientious8 + Conscientious9 + Conscientious10
OpennesstoExperience =~ Openness1 + Openness2 + Openness3 + Openness5 + Openness7 + Openness8 + Openness9 + Openness10

# Structural Model
NetworkingBehavior ~ HonestyHumility + Emotionality + Extraversion + Agreeableness + Conscientiousness + OpennesstoExperience
Numberofjobinterviews ~ NetworkingBehavior
Numberofjoboffers ~ NetworkingBehavior
"

data("LeDang2022")

extract_parseModel(model = tutorial_igsca_model, data = LeDang2022)
}
