% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_methods.R
\name{fitted.cSEMResults}
\alias{fitted.cSEMResults}
\title{Model implied indicator variance-covariance matrix}
\usage{
fitted(object, test = NULL)
}
\arguments{
\item{object}{An R object of class \link{cSEMResults}.
#' @param .only_dots Logical. Should only those arguments to be passed to lower level functions via the
\code{...} argument of the \link{csem} or \link{cca} function be returned. Defaults to \code{FALSE}.}

\item{test}{not used}
}
\description{
Compute the model implied indicator variance-covariance (VCV) matrix. Usually
this is called \eqn{\hat\Sigma}. Currently only the model implied VCV for
linear model for constructs modeled as common factors or composites are
implemented. An error is given if the model is not linear.
}
\seealso{
\link{csem}, \link{cSEMResults}
}
