% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_new.R
\name{Fornell_Larcker}
\alias{Fornell_Larcker}
\title{Fornell_Larcker}
\usage{
Fornell_Larcker(
 .object              = args_default()$.object,
 .only_common_factors = args_default()$.only_common_factors
)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}

\item{.only_common_factors}{Logical. Should only common factors be included?
Defaults to \code{FALSE}.}
}
\description{
Computes the Fornell-Larcker criterion \insertCite{Fornell1981}{cSEM}.
}
\examples{
\dontrun{
# still to implement
}
}
\references{
\insertAllCited{}
}
\seealso{
\link{csem}, \link{cSEMResults}
}
