% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_doSpecificationSearch.R
\name{ga}
\alias{ga}
\title{Internal: perform GA}
\usage{
get_significance_stars(
X
)
}
\arguments{
\item{X}{A matrix of processed data (scaled, cleaned and ordered).}
}
\value{
Character string. A p-value transformed into a star.
}
\description{
Transforms a p-value into stars.
}
\keyword{internal}
