% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igsca_translation.R
\name{get_lavaan_table_igsca_gscapro}
\alias{get_lavaan_table_igsca_gscapro}
\title{Takes GSCAPro input and Creates a Lavaan-style Table}
\usage{
get_lavaan_table_igsca_gscapro(gscapro_in, model)
}
\arguments{
\item{gscapro_in}{}

\item{model}{}
}
\description{
Assumes that every indicator loads onto only one latent variable (composite/factor)
}
