% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_plot.cSEMResults.R
\name{get_significance_stars}
\alias{get_significance_stars}
\title{Internal: get significance stars}
\usage{
get_significance_stars(
.pvalue
)
}
\value{
Character string. A p-value transformed into a star.
}
\description{
Transforms a p-value into stars.
}
\details{
\code{.pvalue} Numeric. A p-value that is transformed into a star.
}
\keyword{internal}
