% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_getConstructScores.R
\name{getConstructScores}
\alias{getConstructScores}
\title{Get construct scores}
\usage{
getConstructScores(
 .object        = NULL,
 .standardized  = TRUE
 )
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}

\item{.standardized}{Logical. Should standardized scores be returned? Defaults
to \code{TRUE}.}
}
\value{
A list of three with elements \code{Construct_scores}, \code{W_used},
\code{Indicators_used}.
}
\description{
\lifecycle{stable}
}
\details{
Get the standardized or unstandardized construct scores.
}
\seealso{
\code{\link[=csem]{csem()}}, \link{cSEMResults}
}
