% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_igsca.R
\name{getIgscaInputs}
\alias{getIgscaInputs}
\title{Gets the Relevant Inputs for IGSCA}
\usage{
getIgscaInputs(.data, .model)
}
\arguments{
\item{.data}{A \code{data.frame} or a \code{matrix} of standardized or unstandardized
data (indicators/items/manifest variables).
Additionally, a \code{list} of data sets (data frames or matrices) is accepted in which
case estimation is repeated for each data set. Possible column types or classes
of the data provided are: "\code{logical}", "\code{numeric}" ("\code{double}" or "\code{integer}"),
"\code{factor}" ("\code{ordered}" and/or "\code{unordered}"), "\code{character}" (will be converted to factor),
or a mix of several types.}

\item{.model}{A model in \link[lavaan:model.syntax]{lavaan model syntax}
or a \link{cSEMModel} list.}
}
\value{
Returns a list of matrices/vectors required for igsca_sim() to run:
\itemize{
\item Z0
\item W0
\item B0
\item C0
\item con_type
\item indicator_type
}
}
\description{
In the context of igsca, this function prepares: (1) the initial indicators
(Z0), weights (W0), structural (B0), loadings(C0) matrices; (2) whether a
construct is a latent or composite variable (con_type); (3) whether an
indicator corresponds to a latent or composite variable (indicator_type); and
(4) the dominant indicator of each construct (.dominant_indicators).
}
