% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_test_MGD.R
\name{getParameterNames}
\alias{getParameterNames}
\title{Get parameter names}
\usage{
getParameterNames(.object = NULL, .model = NULL)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}

\item{.model}{A model in \link[lavaan:model.syntax]{lavaan model syntax} indicating which
parameters (i.e, path (\code{~}), loadings (\code{=~}), or weights (\code{<~})) should be
compared across groups. Defaults to \code{NULL} in which case all parameters of the model
are compared.}
}
\value{
A list containing the names of the structural parameters,
the loadings and weight to be compared.
}
\description{
Based on a cSEmodelDetermines the names of the parameters to be used for testing
}
\keyword{internal}
