% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_getTestMGDDecisions.R
\name{getTestMGDDecisions}
\alias{getTestMGDDecisions}
\title{Get decision from cSEMTestMGD object.}
\usage{
getTestMGDDecisions(.object)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}
}
\value{
A data.frame with the decisions.
}
\description{
This function summarizes the decisions of all tests passed in the .object.
}
\details{
This function is particularily useful to get an overview of all decisions
made.
}
\seealso{
\code{\link[=csem]{csem()}}, \code{\link[=testMGD]{testMGD()}}
}
