% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_new.R
\name{GoF}
\alias{GoF}
\alias{GoF.cSEMResults_default}
\alias{GoF.cSEMResults_multi}
\alias{GoF.cSEMResults_2ndorder}
\title{Goodness of fit}
\usage{
GoF(
 .object              = args_default()$.object,
)

\method{GoF}{cSEMResults_default}(.object)

\method{GoF}{cSEMResults_multi}(.object = args_default()$.object)

\method{GoF}{cSEMResults_2ndorder}(.object = args_default()$.object)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}
}
\description{
(TODO)
}
\section{Methods (by class)}{
\itemize{
\item \code{cSEMResults_default}: (TODO)

\item \code{cSEMResults_multi}: (TODO)

\item \code{cSEMResults_2ndorder}: (TODO)
}}

\examples{
\dontrun{
# still to implement
}

}
\seealso{
\link{csem}, \link{cSEMResults}
}
