% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igsca.R
\name{gsca_inione}
\alias{gsca_inione}
\title{R Implementation of gsca_inione.m from Heungsun Hwang}
\usage{
gsca_inione(Z0, W0, B0)
}
\arguments{
\item{Z0}{Data Matrix}

\item{W0}{Indicator matrix of weights}

\item{B0}{Indicator matrix of Path Coefficients.}
}
\value{
When used in the context of igsca_sim(), it returns a list of the starting values for Weights, Loadings and Path Coefficients. In principle, otherwise, it is a slightly modified implementation of ordinary Generalised Structured Component Analysis (GSCA).
}
\description{
Internal I-GSCA Function
}
\details{
Initializes the values for I-GSCA
}
\author{
Michael S. Truong
}
