% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_dotdotdot.R
\name{handleArgs}
\alias{handleArgs}
\title{Internal: Handle "..." arguments}
\usage{
handleArgs(.dotdotdot)
}
\arguments{
\item{.dotdotdot}{A named list of argument names and defaults usually captured by \code{list(...)}.}
}
\value{
The function returns the \link{args_default} list, with defaults
changed if necessary.
}
\description{
Internal helper function to handle arguments passed via the \code{...} argument
to \link{csem} or \link{cca}.
}
