% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_assess.R
\name{HTMT}
\alias{HTMT}
\title{HTMT}
\usage{
HTMT(
 .object              = args_default()$.object,
 .only_common_factors = args_default()$.only_common_factors
)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}

\item{.only_common_factors}{Logical. Should only common factors be included?
Defaults to \code{FALSE}.}
}
\description{
Compute the heterotrait-monotrait ratio of correlations \insertCite{Henseler2015}{cSEM}.
}
\examples{
\dontrun{
# still to implement
}
}
\references{
\insertAllCited{}
}
\seealso{
\link{csem}, \link{cSEMResults}
}
