% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igsca.R
\name{igsca}
\alias{igsca}
\title{R Implementation of igsca_sim.m}
\usage{
igsca(Z0, W0, C0, B0, con_type, ov_type, ind_domi, itmax = 100, ceps = 0.001)
}
\arguments{
\item{Z0}{Input matrix. Should contain named columns in this form.}

\item{W0}{Indicator matrix of weights: indicators (rows) and their corresponding construct variable (columns).}

\item{C0}{Indicator matrix of loadings: indicators (rows) and their corresponding construct variable (columns).}

\item{B0}{Square indicator matrix of path coefficients: from-construct-variable (rows) and to-construct-variable (columns). The order of construct variables should match the order in C0 and W0.}

\item{con_type}{A vector that denotes whether each construct variable (columns in W0 and C0) is a common factor or composite. Its length should be equal to the number of columns of W0 and C0.}

\item{ov_type}{An indicator vector that indices whether an indicator (rows of W0 and C0) corresponds to a common factor variable (1) or a composite variable (0). This vector is important for computing the uniqueness terms (D) because it zeros the entries for composite indicators.}

\item{ind_domi}{A numeric vector that indices the dominant indicator for each construct variable. \emph{It is to be clarified whether this should only apply to factor latent variables or also composite latent variables.} This is important for ensuring that the signs of the path-coefficients and loadings are consistent. It is sometimes the case in composite-based structural equation modelling methods that loadings/path-coefficients may have the opposite sign. The length of this vector should be equal to the number of construct variables and each value should represent the row number of the dominant indicator for that construct variable.}

\item{itmax}{Maximum number of iterations of the Alternating Least Squares (ALS) algorithm.}

\item{ceps}{Minimum amount of absolute change in the estimates of the path-coefficients (if B0 is non-zero) or the loadings (if B0 is all zero, meaning ther are no path-coefficients) between ALS iterations before ending the optimization.}
}
\value{
List of 4 matrices that make up a fitted I-GSCA Model: (1) Weights, (2) Loadings, (3) Uniqueness Terms D^2, and (4) Path Coefficients.
TODO: Cite the GSCA Pro SEM reference instead of GSCA Pro's Example
}
\description{
This is a streamlined R implementation of I-GSCA, based on the matlab implementation in igsca_sim.m by Heungsun Hwang.
}
\details{
The code here generally assumes that every indicator corresponds to a single construct. (A construct is a superset consisting of latent and emergent variables. Latent variables are sometimes referred to as common-factor/factor variables; and emergent variables are sometimes referred to as component/composite variables.)
}
\examples{

# Specify the model according to GSCA Pro's example
tutorial_igsca_model <- "
# Composite Model
NetworkingBehavior <~ Behavior1 + Behavior2 + Behavior3 + Behavior5 + Behavior7 + Behavior8 + Behavior9
Numberofjobinterviews <~ Interview1 + Interview2
Numberofjoboffers <~ Offer1 + Offer2 

# Reflective Measurement Model
HonestyHumility =~ Honesty1 + Honesty2 + Honesty3 + Honesty4 + Honesty5 + Honesty6 + Honesty7 + Honesty8 + Honesty9 + Honesty10
Emotionality =~ Emotion1 + Emotion2 + Emotion3 + Emotion4 + Emotion5 + Emotion6 + Emotion8 + Emotion10
Extraversion =~ Extraver2 + Extraver3 + Extraver4 + Extraver5 + Extraver6 + Extraver7 + Extraver8 + Extraver9 + Extraver10
Agreeableness =~ Agreeable1 + Agreeable3 + Agreeable4 + Agreeable5 + Agreeable7 + Agreeable8 + Agreeable9 + Agreeable10
Conscientiousness =~ Conscientious1 + Conscientious3 + Conscientious4 + Conscientious6 + Conscientious7 + Conscientious8 + Conscientious9 + Conscientious10
OpennesstoExperience =~ Openness1 + Openness2 + Openness3 + Openness5 + Openness7 + Openness8 + Openness9 + Openness10

# Structural Model
NetworkingBehavior ~ HonestyHumility + Emotionality + Extraversion + Agreeableness + Conscientiousness + OpennesstoExperience
Numberofjobinterviews ~ NetworkingBehavior
Numberofjoboffers ~ NetworkingBehavior
"

# Load the data
data(LeDang2022)

# Pre-generate the required matrices for the algorithm and assume that the first indicator for each latent variable is the dominant indicator.
igsca_in <- extract_parseModel(model = tutorial_igsca_model,
                                   data = LeDang2022,
                                   ind_domi_as_first = TRUE)

# Fit the I-GSCA model
(igsca_out <- with(igsca_in, igsca(Z0 = Z0,
                            W0 = W0,
                            C0 = C0,
                            B0 = B0,
                            con_type = con_type,
                            ov_type = ov_type,
                            ind_domi = ind_domi)
                            ))
}
\author{
Michael S. Truong
}
