% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_doTrees.R
\name{influence_mat}
\alias{influence_mat}
\title{Casewise squared-residual matrix (n x q, multivariate) -- COIN_mat.}
\usage{
influence_mat(E)
}
\arguments{
\item{E}{}
}
\description{
Casewise squared-residual matrix (n x q, multivariate) -- COIN_mat.
}
