% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_doTrees.R
\name{influence_vec}
\alias{influence_vec}
\title{Casewise sum of squared GSCA residuals (n x 1) -- COIN_ssr influence.}
\usage{
influence_vec(E)
}
\arguments{
\item{E}{}
}
\description{
Casewise sum of squared GSCA residuals (n x 1) -- COIN_ssr influence.
}
