% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_igsca.R
\name{initializeIgscaEstimates}
\alias{initializeIgscaEstimates}
\title{R Implementation of gsca_inione.m from Heungsun Hwang}
\usage{
initializeIgscaEstimates(Z0, W0, B0, .S = args_default()$.S)
}
\arguments{
\item{Z0}{Data matrix of N cases (measurements) x J indicators with named
columns, unstandardized.}

\item{W0}{Indicator matrix of weights: J indicators (rows) and their
corresponding Gamma construct variables (columns).}

\item{B0}{Square indicator matrix of path coefficients:
from-construct-variable (rows) and to-construct-variable (columns). The
order of Gamma construct variables should match the order in C0 and W0.}

\item{.S}{The (K x K) empirical indicator correlation matrix.}
}
\value{
Returns a list of starting values for:
\itemize{
\item Weights (W)
\item Loadings (C)
\item Path Coefficients (B)
}
}
\description{
Initializes the values for I-GSCA by using modified Generalized Structured
Component Analysis (GSCA) to estimate the model.
}
\author{
Michael S. Truong
}
