% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_matrix.R
\name{isPositiveSemiDefinite}
\alias{isPositiveSemiDefinite}
\title{Internal: Check if matrix is positive semi definite}
\usage{
isPositiveSemiDefinite(.matrix, .tolerance = 1e-08)
}
\arguments{
\item{.matrix}{Numeric matrix to be checked.}

\item{.tolerance}{Tolerance for close to zero (negative) values.}
}
\value{
A logical scalar: \code{TRUE} if \code{X} is PSD, otherwise \code{FALSE}.
}
\description{
Check if a matrix is positive semi definite (PSD).
A matrix is PSD if all of its eigenvalues are greater than or
equal to zero.
}
\keyword{internal}
