% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_datasets.R
\docType{data}
\name{LeDang2022}
\alias{LeDang2022}
\title{Data: LeDang2022}
\format{
An object of class \code{data.frame} with 773 rows and 107 columns.
}
\source{
The data were made available in \insertCite{Le2022;textual}{cSEM}.
}
\usage{
LeDang2022
}
\description{
A data frame of 107 variables and 773 participants. Labelled
factors were converted from numerical to character representation through
\code{haven::as_factor()}. The provided data frame is a fusion of the converted
labelled factors for the demographic data and the unconverted Likert scale
data for the psychometric data.
}
\details{
This data is based on self-report surveys collected from graduates of universities
in Northern Vietnam \insertCite{Le2022}{cSEM}. The data has been used to analyze
the relationship between measures of HEXACO personality traits and two job-search outcomes:
(1) number of job interviews; and (2) job offers through job search
networking behavior \insertCite{Le2022}{cSEM}.

The data were also analyzed using Integrated-Generalized Structured Component
Analysis (IGSCA) in GSCAPro \insertCite{Hwang2021a}{cSEM}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
