% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_doTrees.R
\name{node_group_data}
\alias{node_group_data}
\title{Title}
\usage{
node_group_data(mf, subset, indicators, goes_left)
}
\arguments{
\item{goes_left}{}
}
\description{
Node data for one candidate partition: children as \code{group} levels 1/2.
}
