% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_estimators_paths.R
\name{nonlinear_estimation_utilities}
\alias{nonlinear_estimation_utilities}
\alias{f1}
\alias{f2}
\alias{f3}
\alias{f4}
\alias{f5}
\title{Internal: Utility functions for the estimation of nonlinear models}
\usage{
f1(.i, .j)

f2(.i, .j, .select_from, .Q, .H)

f3(.i, .j, .Q, .H)

f4(.i, .j, .Q, .H, .var_struc_error, .temp = NULL)

f5(.i, .j, .H, .Q, .var_struc_error)
}
\arguments{
\item{.i}{Row index}

\item{.j}{Column index}

\item{.select_from}{matrix to select from}

\item{.Q}{A vector of composite-construct correlations with element names equal to
the names of the J construct names used in the measurement model. Note
Q^2 is also called the reliability coefficient.}

\item{.H}{The (N x J) matrix of construct scores.}
}
\description{
Internal: Utility functions for the estimation of nonlinear models
}
\keyword{internal}
