% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_test_MGD.R
\name{parameter_difference}
\alias{parameter_difference}
\title{Function that calculates the parameter differences across groups}
\usage{
parameter_difference(
.object=args_default()$.object,
.comparison=args_default()$.comparison)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}

\item{.comparison}{A model in \link[lavaan:model.syntax]{lavaan model syntax} indicating which
parameters are compared.}
}
\description{
Function that calculates the parameter differences across groups
}
\seealso{
\link{cSEMResults}
}
