% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igsca_translation.R
\name{parse_GSCAPro_FullResults}
\alias{parse_GSCAPro_FullResults}
\title{Helper Function for parsing GSCA Pro V1.2.1 Full Results Files}
\usage{
parse_GSCAPro_FullResults(
  file_path = list("tests", "comparisons", "igsca_translation", "GSCAPro_1_2_1Output",
    "Tutorial_IGSCA_model1_full_result_ver0.csv"),
  headers = c("Weights", "Loadings", "Path coefficients", "Unique D^2")
)
}
\value{
List of extracted matrices from GSCA Pro Full results
}
\description{
The commented headers require more work to get working
}
