% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csem_model.R
\name{parseModel}
\alias{parseModel}
\title{Parse lavaan model}
\usage{
parseModel(.model)
}
\arguments{
\item{.model}{A model in \link[lavaan:model.syntax]{lavaan model syntax}
or a \link{cSEMModel} list.}
}
\value{
An object of class \code{cSEMModel} is a standardized list containing the following
components. J stands for the number of constructs and K for the number
of indicators.
\describe{
\item{\code{$structural}}{A matrix mimicking the structural relationship between
constructs. If constructs are only linearly related, \code{structural} is
of dimension (J x J) with row- and column names equal to the construct
names. If the structural model contains nonlinear relationships
\code{structural} is (J x (J + J*)) where J* is the number of
nonlinear terms. Rows are ordered such that exogenous constructs are always
first, followed by constructs that only depend on exogenous constructs and/or
previously ordered constructs.}
\item{\code{$measurement}}{A (J x K) matrix mimicking the measurement relationship
between constructs and their related indicators. Rows are in the same
order as the matrix \code{$structural} with row names equal to
the construct names. The order of the columns is such that \code{$measurement}
forms a block diagonal matrix.}
\item{\code{$error_cor}}{A (K x K) matrix mimicking the measurement error
correlation relationship. The row and column order is identical to \code{$measurement}.}
\item{\code{$construct_type}}{A named vector containing the names of each construct
and their respective type ("Common factor" or "Composite").}
\item{\code{$construct_order}}{A named vector containing the names of each construct
and their respective order ("First order" or "Second order").}
\item{\code{$model_type}}{The type of model ("Linear" or "Nonlinear").}
}
Note: it is possible to supply an incomplete \code{cSEMModel} list
to all functions that require \code{.csem_model} as a mandatory argument. Currently,
only the structural and the measurement matrix are required.
However, specifying an incomplete \code{cSEMModel} list may lead to unexpected behavior
and errors. Use with care.
}
\description{
Turns a model written in \link[lavaan:model.syntax]{lavaan model syntax} into a
\link{cSEMModel} list.
}
\examples{
model <- '
# Structural model
y1 ~ y2 + y3

# Measurement model
y1 =~ x1 + x2 + x3
y2 =~ x4 + x5
y3 =~ x6 + x7

# Error correlation
x1 ~~ x2
'

(m <- parseModel(model))

# If the model is already a cSEMModel object, the model is returned as is.

identical(m, parseModel(m)) # TRUE
}
