% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cSEMIPA.R
\name{plot.cSEMIPA}
\alias{plot.cSEMIPA}
\title{\code{cSEMIPA} method for \code{plot()}}
\usage{
\method{plot}{cSEMIPA}(x, .dependent)
}
\arguments{
\item{x}{An R object of class \code{cSEMSurface}.}

\item{.plot_type}{A character vector indicating the plot package used. Options are
"\emph{plotly}", "\emph{rsm}", and "\emph{persp}". Defaults to "\emph{plotly}".}

\item{...}{Additional parameters that can be passed to
\code{\link[graphics:persp]{graphics::persp}}, e.g., to rotate the plot.}
}
\description{
Plot the importance-performance matrix
}
\seealso{
\code{\link[=doIPA]{doIPA()}}
}
