% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cSEMIPMA.R
\name{plot.cSEMIPMA}
\alias{plot.cSEMIPMA}
\title{\code{cSEMIPMA} method for \code{plot()}}
\usage{
\method{plot}{cSEMIPMA}(x = NULL, .dependent = NULL, .attributes = NULL, .level = "construct")
}
\arguments{
\item{x}{An R object of class \code{cSEMIPMA}.}

\item{.dependent}{Character. Name of the target construct for which the
importance performance matrix should be created.}

\item{.attributes}{Character string. A vector containing indicator/construct names that
should be plotted in the importance performance matrix. It must be at least of length 2.}

\item{.level}{Character. It indicates whether the importance performance matrix
should be created on indicator or construct level. Options are \code{'construct'} and \code{'indicator'}.
Defaults to \code{'construct'}.}
}
\description{
Plot the importance-performance matrix
}
\seealso{
\code{\link[=doIPMA]{doIPMA()}}
}
