% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_plot.R
\name{plot.cSEMPredict}
\alias{plot.cSEMPredict}
\title{\code{cSEMPredict} method for \code{plot()}}
\usage{
\method{plot}{cSEMPredict}(x, ...)
}
\description{
The \code{cSEMPredict} method for the generic function \code{\link[=plot]{plot()}}.
}
\seealso{
\code{\link[=csem]{csem()}}, \code{\link[=predict]{predict()}}
}
\keyword{internal}
