% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_new.R
\name{plot.Two_Way_Effect}
\alias{plot.Two_Way_Effect}
\title{Plot the output of floodlight analysis}
\usage{
plot(.TWobject)
}
\description{
It plots the direct effect of an independent variable depending on the levels of the
moderator variable. Moreover, the confidence interval are plotted including the
Johnson-Neyman points.
}
