% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_plotModel.R
\name{plotModelSave}
\alias{plotModelSave}
\title{Save plotMOdel object to a file}
\usage{
plotModelSave(plot_object, .file)
}
\arguments{
\item{plot_object}{The DiagrammeR plot object to be saved. If not provided, the function will try to retrieve \code{latest_plot} from the global environment.}

\item{.file}{Character string. The name of the file to save the plot to (supports 'pdf', 'png', 'svg', and 'dot' formats).}
}
\value{
NULL (the function saves the plot to a file and provides messages on completion)
}
\description{
This function saves a given plot of a cSEM model to a specified file format.
}
\details{
@usage plotModelSave(
plot_object,
.file)
}
\examples{
\dontrun{

# Example: Save a plot object to a pdf file
plot_object <- plotModel(.object)
plotModelSave(plot_object, .file = "sem_plot.pdf")
}
}
