% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igsca_translation.R
\name{prepare_for_ALS}
\alias{prepare_for_ALS}
\title{Prepare for ALS Algorithm}
\usage{
prepare_for_ALS(z0, W0, B0, nvar, ncase, nlv, ov_type)
}
\arguments{
\item{z0}{}

\item{W0}{}

\item{B0}{}

\item{nvar}{}

\item{ncase}{}

\item{nlv}{}

\item{ov_type}{}
}
\description{
Internal I-GSCA function
}
