% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_print.R
\name{print.cSEMAssess}
\alias{print.cSEMAssess}
\title{\code{cSEMAssess} method for \code{print()}}
\usage{
\method{print}{cSEMAssess}(x, ...)
}
\description{
The \code{cSEMAssess} method for the generic function \code{\link[=print]{print()}}.
}
\seealso{
\code{\link[=csem]{csem()}}, \code{\link[=assess]{assess()}}
}
\keyword{internal}
