% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cSEMModelSearch.R
\name{print.cSEMModelSearch}
\alias{print.cSEMModelSearch}
\title{\code{cSEMModelSearch} method for \code{print()}}
\usage{
\method{print}{cSEMModelSearch}(x, ...)
}
\description{
The \code{cSEMModelSearch} method for the generic function \code{\link[=print]{print()}}.
}
\seealso{
\code{\link[=csem]{csem()}}, \link{cSEMResults}, \code{\link[=doModelSearch]{doModelSearch()}}
}
\keyword{internal}
