% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cSEMNonlinearEffects.R
\name{print.cSEMNonlinearEffects}
\alias{print.cSEMNonlinearEffects}
\title{\code{cSEMNonlinearEffectsAnalysis} method for \code{print()}}
\usage{
\method{print}{cSEMNonlinearEffects}(x, ...)
}
\description{
The \code{cSEMNonlinearEffectsAnalysis} method for the generic function \code{\link[=print]{print()}}.
}
\seealso{
\code{\link[=csem]{csem()}}, \link{cSEMResults}, \code{\link[=doNonlinearEffectsAnalysis]{doNonlinearEffectsAnalysis()}}, \code{\link[=plot.NonlinearEffectsAnalysis]{plot.NonlinearEffectsAnalysis()}}
}
\keyword{internal}
