% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cSEMPredict.R
\name{print.cSEMPlotPredict}
\alias{print.cSEMPlotPredict}
\title{\code{cSEMPlotPredict} method for \code{print()}}
\usage{
\method{print}{cSEMPlotPredict}(x, ...)
}
\arguments{
\item{x}{An R object of class \code{cSEMPlotPredict}.}

\item{...}{Currently ignored.}
}
\description{
The \code{cSEMPlotPredict} method for the generic function \code{\link[=print]{print()}}.
}
\seealso{
\code{\link[=predict]{predict()}}, \code{\link[=plot.cSEMPredict]{plot.cSEMPredict()}}
}
\keyword{internal}
