% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_print.R
\name{print.cSEMSummarize_default}
\alias{print.cSEMSummarize_default}
\title{\code{cSEMSummarize_default} method for \code{print()}}
\usage{
\method{print}{cSEMSummarize_default}(x, ...)
}
\description{
The \link{cSEMSummary} method for the generic function \code{\link[=print]{print()}}.
}
\seealso{
\code{\link[=csem]{csem()}}, \code{\link[=foreman]{foreman()}}, \link{cSEMResults}, \code{\link[=summarize]{summarize()}}
}
\keyword{internal}
