% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cSEMTestCVPAT.R
\name{print.cSEMTestCVPAT}
\alias{print.cSEMTestCVPAT}
\title{\code{cSEMTestCVPAT} method for \code{print()}}
\usage{
\method{print}{cSEMTestCVPAT}(x, ...)
}
\description{
The \code{cSEMTestCVAT} method for the generic function \code{\link[=print]{print()}}.
}
\seealso{
\code{\link[=csem]{csem()}}, \link{cSEMResults}, \code{\link[=testCVPAT]{testCVPAT()}}
}
\keyword{internal}
