% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cSEMTestHausman.R
\name{print.cSEMTestHausman}
\alias{print.cSEMTestHausman}
\title{\code{cSEMTestHausman} method for \code{print()}}
\usage{
\method{print}{cSEMTestHausman}(x, ...)
}
\description{
The \code{cSEMTestHausman} method for the generic function \code{\link[=print]{print()}}.
}
\seealso{
\code{\link[=csem]{csem()}}, \link{cSEMResults}, \code{\link[=testHausman]{testHausman()}}
}
\keyword{internal}
