% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cSEMTestMICOM.R
\name{print.cSEMTestMICOM}
\alias{print.cSEMTestMICOM}
\title{\code{cSEMTestMICOM} method for \code{print()}}
\usage{
\method{print}{cSEMTestMICOM}(x, ...)
}
\description{
The \code{cSEMTestMICOM} method for the generic function \code{\link[=print]{print()}}.
}
\seealso{
\code{\link[=csem]{csem()}}, \link{cSEMResults}, \code{\link[=testMICOM]{testMICOM()}}
}
\keyword{internal}
