% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cSEMTestOMF.R
\name{print.cSEMTestOMF}
\alias{print.cSEMTestOMF}
\title{\code{cSEMTestOMF} method for \code{print()}}
\usage{
\method{print}{cSEMTestOMF}(x, ...)
}
\description{
The \code{cSEMTestOMF} method for the generic function \code{\link[=print]{print()}}.
}
\seealso{
\code{\link[=csem]{csem()}}, \link{cSEMResults}, \code{\link[=testOMF]{testOMF()}}
}
\keyword{internal}
