% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_print.R
\name{print.cSEMVerify_2ndorder}
\alias{print.cSEMVerify_2ndorder}
\title{\code{cSEMVerify_2ndorder} method for \code{print()}}
\usage{
\method{print}{cSEMVerify_2ndorder}(x, ...)
}
\description{
The \code{cSEMVerify_default} method for the generic function \code{\link[=print]{print()}}.
}
\seealso{
\code{\link[=csem]{csem()}}, \code{\link[=foreman]{foreman()}}, \link{cSEMResults}
}
\keyword{internal}
