% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_doTrees.R
\name{prune.cSEMResults}
\alias{prune.cSEMResults}
\title{Prune a grown tree from doTrees}
\usage{
prune.cSEMResults(.tree)
}
\arguments{
\item{.tree}{Fitted tree}
}
\value{
A pruned tree
}
\description{
Prune a grown tree from doTrees
}
