% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_new.R
\name{REBUS_cluster}
\alias{REBUS_cluster}
\title{Implementation of REBUS-PLS
Code obtained by Laura Trinchera.}
\usage{
REBUS_cluster(.object)
}
\description{
Implementation of REBUS-PLS
Code obtained by Laura Trinchera.
}
