% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csem_resample.R
\name{resample}
\alias{resample}
\title{Internal: Resample}
\usage{
resample(
 .object        = args_default()$.object,
 .data          = NULL,
 .method        = c("Bootstrap", "Jackknife", "Permutation", "Cross-validation"),
 .draws         = 499,
 .id            = NULL,
 .cv_folds      = 10      
)
}
\arguments{
\item{.object}{An R object of class \code{cSEM<class>} with corresponding method.}

\item{.data}{A \code{data.frame} or a \code{matrix} containing the raw data. Possible
data column types or classes are: logical, numeric (double or integer), factor
(ordered and unordered) or a mix of several types.}

\item{.id}{Character string. The name of the column of \code{.data} used to split
the data into groups. Defaults to \code{NULL}.}
}
\description{
Resample from a given cSEMResults object or data.frame using common
resampling methods.
}
\examples{
\dontrun{
# still to implement
}

}
\seealso{
\link{csem}, \link{cSEMResults}
}
