% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_datasets.R
\docType{data}
\name{satisfaction}
\alias{satisfaction}
\title{Satisfaction data}
\format{A data frame with 250 rows and 27 variables.}
\usage{
satisfaction
}
\description{
European costumer satisfaction dataset. See
the \href{https://m-e-rademaker.github.io/cSEM/index.html}{cSEM website} for
details.
}
\keyword{datasets}
