% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_saveplotModel.R
\name{save_single_plot}
\alias{save_single_plot}
\title{Internal: save_single_plot
Helper function to save a single DiagrammeR plot based on the file extension}
\usage{
save_single_plot(
diagrammer_obj, 
out_file
)
}
\value{
NULL.
}
\description{
\code{diagrammer_obj} DiagrammeR plot object to be saved.
\code{out_file} The name of the file to save the plot to (supports 'pdf', 'png', 'svg', and 'dot' formats).
}
\keyword{internal}
