% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_savePlot.R
\name{savePlot}
\alias{savePlot}
\title{savePlot}
\usage{
savePlot(
 .plot_object,
 .file)
}
\arguments{
\item{.plot_object}{Object returned by one of the following functions \code{\link[=plot.cSEMResults_default]{plot.cSEMResults_default()}}, \code{\link[=plot.cSEMResults_multi]{plot.cSEMResults_multi()}}, or
\code{\link[=plot.cSEMResults_2ndorder]{plot.cSEMResults_2ndorder()}}.}

\item{.file}{Character string. The name of the file to save the plot to (supports 'pdf', 'png', 'svg', and 'dot' formats).}
}
\description{
This function saves a given plot of a cSEMResults object to a specified file format.
}
\examples{
\dontrun{
model_Bergami_int="
  # Common factor and composite models
  OrgPres <~ cei1 + cei2 + cei3 + cei4 + cei5 + cei6 + cei7 + cei8 
  OrgIden =~ ma1 + ma2 + ma3 + ma4 + ma5 + ma6
  AffJoy =~ orgcmt1 + orgcmt2 + orgcmt3 + orgcmt7
  AffLove  =~ orgcmt5 + orgcmt6 + orgcmt8

  # Structural model 
  OrgIden ~ OrgPres 
  AffLove ~ OrgPres+OrgIden+OrgPres.OrgIden
  AffJoy  ~ OrgPres+OrgIden
  "
  
  outBergamiInt <- csem(.data = BergamiBagozzi2000,.model = model_Bergami_int,
                        .disattenuate = T,
                        .PLS_weight_scheme_inner = 'factorial',
                        .tolerance = 1e-6,
                        .resample_method = 'none')
  
  outPlot <- plot(outBergamiInt)
  outPlot
  savePlot(outPlot,.file='plot.pdf')
  savePlot(outPlot,.file='plot.png')
  savePlot(outPlot,.file='plot.svg')
  savePlot(outPlot,.file='plot.dot')
}
}
\seealso{
\code{\link[=plot.cSEMResults_default]{plot.cSEMResults_default()}} \code{\link[=plot.cSEMResults_multi]{plot.cSEMResults_multi()}} \code{\link[=plot.cSEMResults_2ndorder]{plot.cSEMResults_2ndorder()}}
}
