% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_saveplotModel.R
\name{saveplotModel}
\alias{saveplotModel}
\title{Save plotMOdel object to a file}
\usage{
saveplotModel(.plot_object, .file)
}
\arguments{
\item{.plot_object}{Object returned by the \code{\link[=plotModel]{plotModel()}} function.}

\item{.file}{Character string. The name of the file to save the plot to (supports 'pdf', 'png', 'svg', and 'dot' formats).}
}
\value{
NULL (the function saves the plot to a file and provides messages on completion)
}
\description{
This function saves a given plot of a cSEM model to a specified file format.
}
\details{
@usage saveplotModel(
.plot_object,
.file)
}
\examples{
\dontrun{

# Example: Save a plot object to a pdf file
plot_object <- plotModel(.object)
saveplotModel(plot_object, .file = "sem_plot.pdf")
}
}
