% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_auxillary.R
\name{scaleWeights}
\alias{scaleWeights}
\title{Internal: Scale weights}
\usage{
scaleWeights(.S = NULL, .W = NULL)
}
\arguments{
\item{.S}{The (K x K) empirical indicator correlation matrix.}

\item{.W}{A list with elements \code{$W}, \code{$E}, \code{Modes}, \code{Conv_status}, and
\code{Iterations}.}
}
\value{
The (J x K) matrix of scaled weights.
}
\description{
Scales weights to ensure that the proxies have variance of one.
}
\details{
More description here
}
