% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_auxillary.R
\name{setdiff2}
\alias{setdiff2}
\title{Internal: A symmetric version of setdiff()}
\usage{
setdiff2(x, y)
}
\arguments{
\item{x, y}{elements to compare.}
}
\value{
A vector containing the non-overlapping elements.
}
\description{
A symmetric version of setdiff()
}
\details{
More description here
}
