% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_polycor.R
\name{snlminb}
\alias{snlminb}
\title{Internal: Safer alternative to \code{\link[stats:nlminb]{stats::nlminb()}}}
\usage{
snlminb(
  start,
  objective,
  gradient = NULL,
  hessian = NULL,
  ...,
  scale = 1,
  control = list(),
  lower = -Inf,
  upper = Inf,
  warn.on.failure = TRUE
)
}
\description{
A safer alternative to \code{\link[stats:nlminb]{stats::nlminb()}} which doesn't hard error.
In the case of an error in \code{\link[stats:nlminb]{stats::nlminb()}} (e.g.,. a NA/NaN gradient)
the error is caught, and a list structured like the return value
of \code{\link[stats:nlminb]{stats::nlminb()}} is returned. Used in \code{\link[=polychor]{polychor()}} and
\code{\link[=polyserial]{polyserial()}}. Implemented by Kjell S. Slupphaug.
}
\references{
\insertAllCited{}
}
\keyword{internal}
