% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_datasets.R
\docType{data}
\name{SQ}
\alias{SQ}
\title{Data: SQ}
\format{
An object of class \code{data.frame} with 411 rows and 23 columns.
}
\source{
The dataset is provided by Joerg Henseler.
}
\usage{
SQ
}
\description{
A data frame containing 23 variables with 411 observations. The original
indicators were measured on a 6-point scale. In this version of the dataset,
the indidcators are scaled to be between 0 and 100.
}
\details{
The data comes from a European manufacturer of durable consumer goods and was
studied by \insertCite{Bliemel2004;textual}{cSEM} who focused on service quality.
It is also used in \insertCite{Henseler2020;textual}{cSEM}
for demonstration purposes (Tutorial 13).
}
\references{
\insertAllCited{}
}
\keyword{datasets}
