% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_methods.R
\name{status}
\alias{status}
\title{\code{cSEMResults} method to check the validity of the results}
\usage{
status(object)
}
\arguments{
\item{object}{An R object of class \link{cSEMResults}.
#' @param .only_dots Logical. Should only those arguments to be passed to lower level functions via the
\code{...} argument of the \link{csem} or \link{cca} function be returned. Defaults to \code{FALSE}.}
}
\value{
A NULL vector or a vector of status codes pointing to specific problems.
Problems are:
\itemize{
\item 1 Algorithm has not not converged
\item 2: at least one absolute standardized loading estimate is larger than 1,
which implies either a negative veriance of the measurement error or
a correlation larger than 1
\item 3: construct VCV is not positive semi-definit
\item 4 model-implied indicators VCV is not positive semi-definit
}
}
\description{
Check the validity of the estimated quantities.
}
\seealso{
\link{csem}, \link{cSEMResults}
}
