% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_test_MGD.R
\name{structureTestMGDDecisions}
\alias{structureTestMGDDecisions}
\title{Internal: get structured cSEMTestMGD results}
\usage{
structureTestMGDDecisions(.object)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}
}
\value{
A data.frame.
}
\description{
Convenience function to summarize the results of all tests resulting from a
call to \code{\link[=testMGD]{testMGD()}} in a user-friendly way.
}
\keyword{internal}
