% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_summarize.R
\name{summarize}
\alias{summarize}
\title{Summarize model}
\usage{
summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}

\item{.alpha}{An integer or a numeric vector of significance levels.
Defaults to \code{0.05}.}

\item{.ci}{A vector of character strings naming the confidence interval to compute.
For possible choices see \code{\link[=infer]{infer()}}.}

\item{...}{Further arguments to \code{summarize()}. Currently ignored.}
}
\value{
The return object has the exact same structure as \code{.object}.
As opposed to \code{.object}, the results for the Path coefficients, the loadings,
the weights, the indirect effect, the total effects, and (if present) the
residual correlations are returned as standarized data.frames! This is usually
much more convenient if users intend to present the results in e.g., a paper.
}
\description{
Summarize an estimated model.
}
\seealso{
\link{csem}, \link{cSEMResults}
}
