% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_igsca.R
\name{summarize_FIT_idx}
\alias{summarize_FIT_idx}
\title{Compute FIT statistic for I-GSCA models}
\usage{
summarize_FIT_idx(igsca_results)
}
\arguments{
\item{igsca_results}{Results from igsca()}
}
\value{
FIT index statistic
}
\description{
TBD
}
