% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_datasets.R
\docType{data}
\name{Switching}
\alias{Switching}
\title{Data: Switching}
\format{An object of class \code{data.frame} with 767 rows and 26 columns.}
\source{
The dataset was provided by J\"org Henseler.
}
\usage{
Switching
}
\description{
A data frame containing 26 variables with 767 observations.
}
\details{
The data was studied where they examined the consumers’ intention to switch a
service provider. It is also used in \insertCite{Henseler2020;textual}{cSEM}
for demonstration purposes (Tutorial 12).
}
\examples{
#============================================================================
# Example is taken from Henseler (2020)
#============================================================================
model_Int="
# Measurement models
INV =~ INV1 + INV2 + INV3 +INV4
SAT=~ SAT1 + SAT2 + SAT3
INT =~ INT1 + INT2
# Structrual model containing an interaction term.
INT ~ INV + SAT + INV.SAT
"

out <- csem(.data = Switching, .model = model_Int,
            .PLS_weight_scheme_inner = 'factorial',
            .tolerance = 1e-06,
            .resample_method = 'bootstrap')

}
\references{
\insertAllCited{}
}
\keyword{datasets}
