% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_new.R
\name{testHausman}
\alias{testHausman}
\title{Bootstrap-based Hausman test}
\usage{
testHausman=function(.object2SLS,
                            .seed=1234,
                            .alpha = args_default()$.alpha,
                            .R = args_default()$.R)
}
\arguments{
\item{.seed}{Integer or FALSE. The random seed to use. Defaults to \code{FALSE} in which
case an arbitrary seed is choosen.}

\item{.alpha}{An integer or a numeric vector of significance levels.
Defaults to \code{0.05}.}

\item{.R}{Integer. The number of bootstrap replications. Defaults to \code{499}.}
}
\description{
Calculates a bootstrap-based Hausman test that can be used to compare OLS to 2SLS estimates (Wong, 1996)
or 2SLS to 3SLS estimates.
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=csem]{csem()}}, \code{\link[=foreman]{foreman()}}, \link{cSEMResults}
}
