% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_new.R
\name{testHausman2}
\alias{testHausman2}
\title{Regression-based Hausman-test}
\usage{
testHausman2(.object = NULL, .alpha = args_default()$.alpha,
  .eval_plan = args_default()$.eval_plan,
  .handle_inadmissibles = args_default()$.handle_inadmissibles,
  .R = args_default()$.R,
  .resample_method = args_default()$.resample_method,
  .seed = args_default()$.seed)
}
\description{
The regression based Hausman test
}
