% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_test_OMF.R
\name{testOMF}
\alias{testOMF}
\title{Test for overall model fit}
\usage{
testOMF(
 .object                = args_default()$.object, 
 .alpha                 = args_default()$.alpha, 
 .handle_inadmissibles  = args_default()$.handle_inadmissibles, 
 .iter_max              = 10000
 .R                     = args_default()$.R, 
 .saturated             = args_default()$.saturated,
 .seed                  = args_default()$.seed,
 .verbose               = args_default()$.verbose
)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}

\item{.alpha}{An integer or a numeric vector of significance levels.
Defaults to \code{0.05}.}

\item{.handle_inadmissibles}{Character string. How should inadmissible results
be treated? One of "\emph{drop}", "\emph{ignore}", or "\emph{replace}". If "\emph{drop}", all
replications/resamples yielding an inadmissible result will be dropped
(i.e. the number of results returned will be less than .R). For "\emph{ignore}"
all results are returned even if they are inadmissible (i.e.
number of results returned = .R). For "\emph{replace}" resampling continues until
there are exactly .R admissible solutions.
Defaults to "\emph{drop}".}

\item{.iter_max}{Integer. The maximum number of iterations allowed.
If \code{iter_max = 1} and \code{.approach_weights = "PLS-PM"} one-step weights are returned.
If the algorithm exceeds the specified number, weights of iteration step
\code{.iter_max - 1}  will be returned with a warning. Defaults to \code{100}.}

\item{.R}{Integer. The number of bootstrap replications. Defaults to \code{499}.}

\item{.saturated}{Logical. Should a saturated structural model be used?
Defaults to \code{FALSE}.}

\item{.seed}{Integer or NULL. The random seed to use. Defaults to \code{NULL} in which
case an arbitrary seed is choosen.}

\item{.verbose}{Logical. Should information be printed to the console? Defaults
to \code{TRUE}.}
}
\value{
A standardized list of class \code{cSEMTest}. Technically \code{cSEMTest} is a named
list containing the following list elements:
\describe{
\item{\code{$Test_statistic}}{The value of test statistic(s).}
\item{\code{$Critical_value}}{The critical value(s).}
\item{\code{$Decision}}{The test decision. One of: \strong{Reject} or \strong{Do not reject}}
\item{\code{$Number_admissibles}}{The number of admissible runs. See \code{\link[=verify]{verify()}} for
what constitues and inadmissible run.}
}
}
\description{
Bootstrap-based test for overall model fit. It tests the null hypothesis that the
population indicator correlation matrix equals the population model-implied
correlation matirx. The geodesic distance, the suqared Euclidean distance between the
sample indicator correlation matrix and the estimated model-implied correlation matrix
and the SRMR serve as test statistic. The reference distribution is obtained by
bootstrap \insertCite{Beran1985;textual}{cSEM}. See also
\insertCite{Dijkstra2015;textual}{cSEM} who suggested it in the context of PLS-PM.
After 10000 iterations it stops automatically.
}
\examples{
\dontrun{
# TODO
}

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=csem]{csem()}}, \code{\link[=foreman]{foreman()}}, \link{cSEMResults}
}
