% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_test_MGA.R
\name{testOverallMGA}
\alias{testOverallMGA}
\title{Test for group differences}
\usage{
testOverallMGA(.object = args_default()$.model,
  .drop_inadmissibles = args_default()$.drop_inadmissibles,
  .alpha = args_default()$.alpha, .runs = args_default()$.runs, ...)
}
\arguments{
\item{.object}{An R object of class \code{cSEM<class>} with corresponding method.}

\item{.drop_inadmissibles}{Logical. Should inadmissible solutions be dropped?
Defaults to \code{TRUE}.}

\item{.alpha}{An integer or a numeric vector of significance levels.
Defaults to \code{0.05}.}

\item{.runs}{Integer. How many runs should be performed? Defaults to \code{499}.}
}
\value{
A standardized list of class \code{cSEMTest}. Technically \code{cSEMTest} is a named
list containing the following list elements:
\describe{
\item{\code{Hypothesis}}{The null hypothesis.}
\item{\code{$Test_statistic}}{The value of test statistic(s).}
\item{\code{$Critical_value}}{The critical value(s).}
\item{\code{$Decision}}{The test decision. One of: \strong{Reject} or \strong{Do not reject}}
\item{\code{$Number_admissibles}}{The number of admissible runs. See \code{\link[=verify]{verify()}} for
what constitues and inadmissble run.}
}
}
\description{
What it does.
}
\details{
More details here.
}
\examples{
\dontrun{
# still to implement
}

}
\seealso{
\link{cSEMResults}
}
