% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_foreman.R
\name{theil_sen}
\alias{theil_sen}
\title{Internal: Calculate Theil-Sen estimator between two vectors}
\usage{
theil_sen(x,y)
 )
}
\value{
The regression estimator from x on y
}
\description{
Internal: Calculate Theil-Sen estimator between two vectors
}
\keyword{internal}
