% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_datasets.R
\docType{data}
\name{threecommonfactors}
\alias{threecommonfactors}
\title{Threecommonfactors}
\format{A matrix with 500 rows and 9 variables:
\describe{
\item{y11-y13}{indicators attachted to the first common factor}
\item{y21-y23}{indicators attachted to the second common factor}
\item{y21-y23}{indicators attachted to the third common factor}
}}
\usage{
threecommonfactors
}
\description{
A dataset containing 500 standardized observations on 9 indicator generated from a
population model with three concepts modeled as common factors. See
the \href{https://m-e-rademaker.github.io/cSEM/index.html}{cSEM website} for
details.
}
\keyword{datasets}
