% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_csem-tidiers.R
\name{tidy.cSEMResults}
\alias{tidy.cSEMResults}
\title{tidy Method for cSEMResults Objects}
\usage{
\method{tidy}{cSEMResults}(
  x,
  parameters = c("all", "Path_estimates", "Loadings_estimates", "Weight_estimates",
    "Unique_loading_estimates", "Effect_estimates", "Exo_construct_correlation"),
  conf.int = FALSE,
  conf.level = 0.95,
  conf.type = NULL,
  ...
)
}
\arguments{
\item{x}{Fitted \code{csem} model with class \code{cSEMResults} as returned from \code{\link[=csem]{csem()}}}

\item{parameters}{Character string. Selects the desired parameter estimates as outputted from \code{\link[=summarize]{summarize()}}. Defaults to \code{all}.}

\item{conf.int}{Boolean. Controls whether confidence intervals are returned or not. Defaults to \code{FALSE}.}

\item{conf.level}{Controls the size of the confidence interval returned. Passed to the \code{.alpha} arguments of \code{\link[=summarize]{summarize()}} and \code{\link[=infer]{infer()}} as \code{1 - conf.level}.}

\item{conf.type}{Type of confidence interval returned. See  the \code{.ci} argument of \code{\link[=summarize]{summarize()}}. Unlike \code{\link[=summarize]{summarize()}}, only one type of confidence interval at-a-time is supported.}

\item{...}{
  Arguments passed on to \code{\link{summarize}}
  \describe{
    \item{\code{.alpha}}{An integer or a numeric vector of significance levels.
Defaults to \code{0.05}.}
    \item{\code{.ci}}{A vector of character strings naming the confidence interval to compute.
For possible choices see \code{\link[=infer]{infer()}}.}
    \item{\code{.object}}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}
  }}
}
\value{
A \code{data.frame} with one row for each estimated parameter. The
description of the column names are as follows:

\item{term}{The result of \code{paste(lhs, op, rhs)}}
\item{op}{The operator in the model syntax (e.g., \verb{~~} for correlation, \verb{<~} for weights, \verb{=~} for loadings, and \code{~} for path estimates). Indicators of common factors that load onto themselves denote the unique loading of the common factor (e.g., cf_ind_1 =~ cf_ind_1).}
\item{group}{The group as specified by \code{.id} in the \code{\link[=csem]{csem()}}.}
\item{estimate}{The parameter estimate}
\item{construct.type}{Whether the modelled construct is a common factor or composite variable}
\item{std.error}{}
\item{statistic}{The t-statistic as returned from \code{\link[=summarize]{summarize()}} and \code{cSEM:::addInfer()}}
\item{p_value}{}
}
\description{
This method and documentation is based on the \code{tidy} method for \code{lavaan}, as found in \code{broom:::tidy.lavaan()}.
}
\details{
\strong{Cautionary Note}: As with all \code{tidy()} methods, mis-specified arguments may be silently ignored.

The attribute \code{verification} shows whether the model is admissible.
}
\examples{
\dontrun{
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

# Single Group Example
res_boot <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

tidy(res_boot, conf.int = TRUE, conf.level = .95, conf.type = "CI_percentile")
tidy(res_boot, conf.int = TRUE, conf.level = .95, conf.type = NULL)
# Multi-Group Example
threecommonfactors_id <- cbind(
 "id" = sample(1:3, nrow(threecommonfactors), replace = TRUE),
 threecommonfactors
)

res_mg_boot <- csem(
 threecommonfactors_id,
 model,
 .resample_method = "bootstrap",
 .R = 40,
 .id = "id"
)

tidy(res_mg_boot, conf.int = TRUE)
}
}
\seealso{
\code{\link[=summarize]{summarize()}}
}
\author{
Michael S. Truong
}
