% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_matrix.R
\name{trace}
\alias{trace}
\title{Internal: Trace of a matrix}
\usage{
trace(X)
}
\arguments{
\item{X}{Square numeric matrix.}
}
\value{
A numeric scalar: value with the trace of \code{X}.
}
\description{
Compute the trace of \code{X}, i.e., the sum of the diagonal elements of \code{X}.
}
\keyword{internal}
