% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_igsca.R
\name{update_C_B_D_U}
\alias{update_C_B_D_U}
\title{Update Loadings, Path-Coefficients and Uniqueness Terms After Updating Latent Variables}
\usage{
update_C_B_D_U(
  Z,
  X,
  Gamma,
  C,
  B,
  D,
  indicator_type,
  n_indicators,
  c_index,
  n_constructs,
  b_index,
  n_case
)
}
\arguments{
\item{Z}{Standardized data matrix}

\item{X}{Pseudo-weights for composites}

\item{Gamma}{Construct Scores}

\item{C}{Matrix of estimated loadings}

\item{B}{Matrix of path coefficients}

\item{D}{Matrix of estimated unique error}

\item{indicator_type}{Vector of whether each indicator corresponds to a common factor or composite}

\item{n_indicators}{Number of indicators}

\item{c_index}{Index of loadings}

\item{n_constructs}{Number of constructs}

\item{b_index}{Index of Path Coefficients}

\item{n_case}{Number of Cases}
}
\value{
List of matrices:
\itemize{
\item (1) Uniqueness terms (D) and (U)
\item (2) Estimated Path Coefficients matrix (B)
\item (3) Estimated Loadings matrix (C)
}
}
\description{
Update Loadings, Path-Coefficients and Uniqueness Terms After Updating Latent Variables
}
