% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igsca.R
\name{update_common_factor}
\alias{update_common_factor}
\title{Update Common Factor Variable}
\usage{
update_common_factor(WW, windex_gamma_idx, gamma_idx)
}
\arguments{
\item{WW}{}

\item{windex_gamma_idx}{}

\item{gamma_idx}{}
}
\value{
theta: Used to update factor latent variables -- after accounting for loadings and path-coefficients.
}
\description{
Update Common Factor Variable
}
