% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igsca.R
\name{update_composite}
\alias{update_composite}
\title{Update Composite Variables}
\usage{
update_composite(
  n_total_var,
  tot,
  n_constructs,
  gamma_idx,
  W,
  A,
  V,
  X,
  windex_gamma_idx
)
}
\arguments{
\item{n_total_var}{}

\item{tot}{}

\item{n_constructs}{}

\item{gamma_idx}{}

\item{W}{}

\item{A}{}

\item{V}{}

\item{X}{}

\item{windex_gamma_idx}{}
}
\value{
theta: A matrix that will later be used to update the weights for the composite variable.
TODO: Double check that this is for weights only?
}
\description{
Update Composite Variables
}
