% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igsca_translation.R
\name{update_composite_LV}
\alias{update_composite_LV}
\title{Update Composite Latent Variables}
\usage{
update_composite_LV(ntv, tot, nlv, j, W, A, V, X, windex_j)
}
\arguments{
\item{ntv}{}

\item{tot}{}

\item{nlv}{}

\item{j}{}

\item{W}{}

\item{A}{}

\item{V}{}

\item{X}{}

\item{windex_j}{}
}
\description{
Update Composite Latent Variables
}
