% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igsca_translation.R
\name{update_factor_LV}
\alias{update_factor_LV}
\title{Update Factor Latent Variable}
\usage{
update_factor_LV(WW, windex_j, j)
}
\arguments{
\item{WW}{}

\item{windex_j}{}

\item{j}{}
}
\description{
Update Factor Latent Variable
}
