% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igsca.R
\name{update_X_weights}
\alias{update_X_weights}
\title{Update Weights and X (?)}
\usage{
update_X_weights(Z, U, D, C, nlv, B)
}
\arguments{
\item{Z}{}

\item{U}{}

\item{D}{}

\item{C}{}

\item{nlv}{}

\item{B}{}
}
\value{
Two matrices:
\itemize{
\item \emph{X}: Remaining part of data (Z) after accounting for uniqueness terms (U) and (D)(?)
\item \emph{WW}: \emph{presumably} Weights after accounting for current Loading and Path-Coefficients values.
TODO: Double check that I've got this right.
}
}
\description{
Update Weights and X (?)
}
