% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igsca.R
\name{update_X_WW_pseudo_weights}
\alias{update_X_WW_pseudo_weights}
\title{Update Pseudo Weights for Composites and Common-Factors}
\usage{
update_X_WW_pseudo_weights(Z, U, D, C, n_constructs, B)
}
\arguments{
\item{Z}{Standardized Data}

\item{U}{Related to Unique error}

\item{D}{Unique Error}

\item{C}{Loadings}

\item{n_constructs}{Number of constructs}

\item{B}{Path Coefficients}
}
\value{
Two matrices:
\itemize{
\item \emph{X}: Remaining part of data (Z) after accounting for uniqueness terms (U) and (D), used for estimating composite loadings. Also used for standardizing theta when updating Gamma, W and V
\item \emph{WW}: Weights after accounting for current Loading and Path-Coefficients values, used for estimating common-factor loadings
}
}
\description{
Update Pseudo Weights for Composites and Common-Factors
}
