% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_estimators_weights.R
\name{updateCB}
\alias{updateCB}
\title{Update Loadings and Path-Coefficients for IGSCA}
\usage{
updateCB(
  X,
  Eta,
  Lambda,
  B,
  .indicator_type,
  n_indicators,
  lambda_index,
  n_constructs,
  b_index,
  n_case,
  .con_type,
  modes
)
}
\arguments{
\item{X}{Indicators with measurement error removed}

\item{Eta}{Construct Scores}

\item{Lambda}{Loadings matrix}

\item{B}{Path coefficients matrix}

\item{.indicator_type}{Vector of whether each indicator corresponds to a common factor or composite}

\item{n_indicators}{Number of indicators}

\item{lambda_index}{Index of loadings}

\item{n_constructs}{Number of oncstructs}

\item{b_index}{Index of Path Coefficients}

\item{n_case}{Number of Cases}

\item{.con_type}{A vector that denotes whether each construct variable
(columns in W0 and C0) is a common factor or composite. Its length should
be equal to the number of columns of W0 and C0.}

\item{modes}{Named vector of whether the construct is a Common factor, nomological composite or canonical composite.}
}
\value{
List of matrices:
\itemize{
\item (1) Estimated Loadings matrix (C)
\item (2) Estimated Path Coefficients matrix (B)
}
}
\description{
Update Loadings and Path-Coefficients for IGSCA
}
