% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_igsca.R
\name{updateCBDU}
\alias{updateCBDU}
\title{Update Loadings, Path-Coefficients and Uniqueness Terms After Updating Latent Variables}
\usage{
updateCBDU(
  Z,
  X,
  Gamma,
  C,
  B,
  D,
  indicator_type,
  n_indicators,
  c_index,
  n_constructs,
  b_index,
  n_case,
  con_type
)
}
\arguments{
\item{Z}{Standardized data matrix}

\item{X}{Pseudo-weights for composites}

\item{Gamma}{Construct Scores}

\item{C}{Matrix of estimated loadings}

\item{B}{Matrix of path coefficients}

\item{D}{Matrix of estimated unique error}

\item{indicator_type}{Vector of whether each indicator corresponds to a common factor or composite}

\item{n_indicators}{Number of indicators}

\item{c_index}{Index of loadings}

\item{n_constructs}{Number of constructs}

\item{b_index}{Index of Path Coefficients}

\item{n_case}{Number of Cases}

\item{con_type}{A vector that denotes whether each construct variable
(columns in W0 and C0) is a common factor or composite. Its length should
be equal to the number of columns of W0 and C0.}
}
\value{
List of matrices:
\itemize{
\item (1) Uniqueness terms (D) and (U)
\item (2) Estimated Path Coefficients matrix (B)
\item (3) Estimated Loadings matrix (C)
}
}
\description{
Update Loadings, Path-Coefficients and Uniqueness Terms After Updating Latent Variables
}
