% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_igsca.R
\name{updateCommonFactorTheta}
\alias{updateCommonFactorTheta}
\title{Update Theta for Common Factor Variable}
\usage{
updateCommonFactorTheta(WW, windex_gamma_idx, gamma_idx)
}
\arguments{
\item{WW}{Pseudo-weights for Common Factors}

\item{windex_gamma_idx}{Index of weights related to the indicators for the construct of interest}

\item{gamma_idx}{Index of which construct we are examining}
}
\value{
Theta: Used to update factor latent variables -- after accounting for loadings and path-coefficients.
}
\description{
Update Theta for Common Factor Variable
}
