% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_igsca.R
\name{updateXAndWW}
\alias{updateXAndWW}
\title{Update X and WW (Pseudo Weights) for Composites and Common-Factors}
\usage{
updateXAndWW(Z, U, D, C, B, n_constructs)
}
\arguments{
\item{Z}{Standardized data matrix}

\item{U}{Matrix of estimates related to unique error}

\item{D}{Matrix of estimated unique error}

\item{C}{Matrix of estimated loadings}

\item{B}{Matrix of path coefficients}

\item{n_constructs}{Number of constructs}
}
\value{
Two matrices:
\itemize{
\item X: Remaining part of data (Z) after accounting for uniqueness terms (U) and (D), used for estimating composite loadings. Also used for standardizing Theta when updating Gamma, W and V
\item WW: Weights after accounting for current Loading and Path-Coefficients values, used for estimating common-factor loadings
}
}
\description{
Computation of WW matrix was converted between Matlab to R based on page 44 of \insertCite{Hiebeler2015;textual}{cSEM}.
}
\references{
\insertAllCited{}
}
