% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_igsca.R
\name{warnIgscaFeatures}
\alias{warnIgscaFeatures}
\title{Warning Function for I-GSCA}
\usage{
warnIgscaFeatures()
}
\description{
Warning message to URL for what is currently supported and unsupported when using IGSCA
}
\author{
Michael S. Truong
}
