% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_test_MGD.R
\name{adjustAlpha}
\alias{adjustAlpha}
\title{Multiple testing correction}
\usage{
adjustAlpha <- function(
 .alpha                 = args_default()$.alpha,
 .approach_alpha_adjust = args_default()$.approach_alpha_adjust,
 .nr_comparisons        = args_default()$.nr_comparisons
)
}
\arguments{
\item{.alpha}{An integer or a numeric vector of significance levels.
Defaults to \code{0.05}.}

\item{.approach_alpha_adjust}{Character string. Approach used to adjust the significance level.
One of: "\emph{none}", "\emph{bonferroni}",}

\item{.nr_comparisons}{Numeric. Indicating the number of comparisons. Defaults to \code{NULL}.}
}
\value{
A vector of (adjusted) significance levels.
}
\description{
Adjust a given significance level .alpha to accomodate multiple testing.
}
\keyword{internal}
