% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_arguments.R
\name{args_default}
\alias{args_default}
\title{List all arguments and their defaults}
\usage{
args_default(.only_dots = FALSE)
}
\arguments{
\item{csem_arguments}{}
}
\value{
A named list of argument names and their defaults.
}
\description{
The function \code{args_default()} lists all possible arguments and their defaults.
For a description of the arguments see \link{csem_arguments}.
}
\details{
This is the complete list of arguments, their current defaults, and their
possible alternatives. Each default may be changed by passing a \code{arg_name = value}
pair to the function that uses the argument. Changes to the default arguments
are checked and processed by the \link{handleArgs} function.

If \code{.only_dots = TRUE} all possible arguments to be passed to lower level functions via the
\code{...} argument of the \link{csem} or \link{cca} function are returned.
}
\examples{

args_default() # List all possible arguments and their defaults.
args_default(.only_dots = TRUE) # list only those accepted by `...`.

}
\seealso{
\link{handleArgs}, \link{csem_arguments}, \link{cca}, \link{csem}, \link{foreman}
}
