% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_arguments.R
\name{args_default}
\alias{args_default}
\title{Internal: Show argument defaults or candidates}
\usage{
args_default(.choices = FALSE)
}
\arguments{
\item{.choices}{Logical. Should candidate values for the arguments be returned?
Defaults to \code{FALSE}.}
}
\value{
A named list of argument names and defaults or accepted candidates.
}
\description{
Show all arguments used by package functions including default or candidate
values. For argument descriptions, see: \link{csem_arguments}.
}
\details{
By default \code{args_default()}returns a list of default values by argument name.
If the list of accepted candidate values is required instead, use \code{.choices = TRUE}.
}
\examples{

# List all possible arguments and their defaults
args_default()

## Show accepted candidates:
args_default(.choices = TRUE)

}
\seealso{
\code{\link[=handleArgs]{handleArgs()}}, \link{csem_arguments}, \code{\link[=csem]{csem()}}, \code{\link[=foreman]{foreman()}}
}
\keyword{internal}
