% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_assess.R
\name{assess}
\alias{assess}
\title{Assess model}
\usage{
assess(.object)
}
\arguments{
\item{.object}{An R object of class \code{cSEM<class>} with corresponding method.}
}
\value{
(TODO)
}
\description{
Assess model using common evaluation criteria and fit measures.
}
\details{
The following evaulation criteria and fit measures are used (TODO).
}
\seealso{
\code{\link[=csem]{csem()}}, \code{\link[=cca]{cca()}}, \code{\link[=foreman]{foreman()}}, \link{cSEMResults}
}
