% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_assess.R
\name{assess}
\alias{assess}
\alias{assess.cSEMResults_default}
\alias{assess.cSEMResults_multi}
\alias{assess.cSEMResults_2ndorder}
\title{Assess model}
\usage{
assess(.object, .only_common_factors = TRUE, ...)

\method{assess}{cSEMResults_default}(.object,
  .only_common_factors = TRUE, ...)

\method{assess}{cSEMResults_multi}(.object, ...)

\method{assess}{cSEMResults_2ndorder}(.object, ...)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}

\item{.only_common_factors}{Logical. Should only constructs modeled as common
factors be included? Defaults to \code{TRUE}.}

\item{...}{Further arguments passed to functions called by \code{assess()}.}
}
\value{
(TODO)
}
\description{
Assess model using common evaluation criteria and fit measures.
}
\details{
For details see the vignette on assess.
}
\section{Methods (by class)}{
\itemize{
\item \code{cSEMResults_default}: (TODO)

\item \code{cSEMResults_multi}: (TODO)

\item \code{cSEMResults_2ndorder}: (TODO)
}}

\seealso{
\code{\link[=csem]{csem()}}, \code{\link[=foreman]{foreman()}}, \link{cSEMResults}
}
