% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_new.R
\name{AVE}
\alias{AVE}
\title{AVE}
\usage{
AVE(
 .object              = args_default()$.object,
 .only_common_factors = args_default()$.only_common_factors
)
}
\arguments{
\item{.object}{An R object of class \code{cSEM<class>} with corresponding method.}

\item{.only_common_factors}{Logical. Should only common factors be included?
Defaults to \code{FALSE}.}
}
\description{
Computes the average variance extracted \insertCite{Raykov1997} based on the standardized loadings.
}
\examples{
\dontrun{
# still to implement
}

}
\references{
\insertAllCited{}
Fornell, C. & Larcker, D. F. (1981). Evaluating structural equation models with unobservable variables and measurement error. \emph{Journal of Marketing Research}, 18, 39–50.
}
\seealso{
\link{csem}, \link{cSEMResults}
}
