% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_doTrees.R
\name{bdiagFit}
\alias{bdiagFit}
\title{Block-diagonalized model-implied correlation matrices}
\usage{
bdiagFit(
  .object    = NULL,
  .n_blocks  = args_default()$.n_blocks,
  .type_vcv  = args_default()$.type_vcv,
  .saturated = args_default()$.saturated
  )
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}

\item{.n_blocks}{Integer. The number of times the model-implied correlation
matrix of a single-group model is repeated along the block diagonal by
\code{\link[=bdiagFit]{bdiagFit()}}. Ignored for multigroup objects, where the number of blocks
equals the number of groups. Defaults to \code{1}.}

\item{.type_vcv}{Character string. Which model-implied correlation
matrix should be calculated?
One of "\emph{indicator}" or "\emph{construct}". Defaults to "\emph{indicator}".}

\item{.saturated}{Logical. Should a saturated structural model be used?
Defaults to \code{FALSE}.}
}
\value{
A block-diagonal \code{matrix}.
}
\description{
Make it easy to compare the matrix distance between the model-implied
correlation matrix of a single-group model and the model-implied matrices
of a multigroup model: the single-group model's implied matrix is
duplicated block-diagonally a user-chosen number of times, while for a
multigroup model each group's implied matrix is placed along the block
diagonal. The block-diagonal matrices of a single-group model (repeated
as many times as the multigroup model has groups) and of the multigroup
model then have identical dimensions and can be compared with any matrix
distance.
}
\seealso{
\code{\link[=fit]{fit()}}, \code{\link[=csem]{csem()}}
}
