% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_plotModel.R
\name{buildDotCode}
\alias{buildDotCode}
\title{Internal: buildDotCode}
\usage{
buildDotCode(
title, 
graph_attrs,
constructs, 
r2_values,
measurement_edge_fun,
path_coefficients, 
path_p_values,
correlations,
plot_significances,
plot_indicator_correlations,
plot_structural_model_only,
is_second_order = FALSE
)
}
\arguments{
\item{title}{Character string. Title of an object. Defaults to \emph{NULL}.}

\item{graph_attrs}{Character String. Additional attributes that should be passed
to the DiagrammeR syntax, e.g., c("rankdir=LR", "ranksep=1.0"). Defaults to \emph{NULL}.}

\item{path_coefficients}{List. A list that contains the resampled and the original
path coefficient estimates. Typically a part of a \code{cSEMResults_resampled} object.
Defaults to \code{NULL}.}

\item{plot_significances}{Logical. Should p-values in the form of stars be plotted? Defaults to \code{TRUE}.}

\item{plot_indicator_correlations}{Logical. Should the correlations between the exogenous constructs
or indicators be plotted? Defaults to \code{exo}.}

\item{plot_structural_model_only}{Logical. Should only the structural model,
i.e., the constructs and their relationships be plotted? Defaults to \code{FALSE}.}
}
\value{
Character string. .
}
\description{
Build DOT code from common components.
}
\keyword{internal}
