% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_csem.R
\name{calculate2ndStage}
\alias{calculate2ndStage}
\title{Second order constructs}
\usage{
calculate2ndStage(
 .csem_model          = args_default()$.csem_model,
 .first_stage_results = args_default()$.first_stage_results,
 .original_arguments  = args_default()$.original_arguments,
 .approach_2ndorder   = args_default()$.approach_2ndorder
  )
}
\arguments{
\item{.csem_model}{A (possibly incomplete) \link{cSEMModel}-list.}

\item{.original_arguments}{The list of arguments used within \code{\link[=csem]{csem()}}.}

\item{.approach_2ndorder}{Character string. Approach used for models containing
second order constructs. One of: "\emph{2stage}", or "\emph{mixed}". Defaults to "\emph{2stage}".}
}
\value{
A cSEMResults object.
}
\description{
Performs the second and third stage for a model containing second order
constructs.
}
\keyword{internal}
