% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_assess.R
\name{calculateAVE}
\alias{calculateAVE}
\title{Internal: AVE}
\usage{
calculateAVE(
 .object              = NULL,
 .only_common_factors = TRUE,
)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}

\item{.only_common_factors}{Logical. Should only constructs modeled as common
factors be included? Defaults to \code{TRUE}.}
}
\description{
Calculate the average variance extracted (AVE). See (TODO) for details.
}
\details{
The function is only applicable to objects of class \code{cSEMResults_default}.
For other object classes use \code{\link[=assess]{assess()}}.
}
\references{
\insertAllCited{}
}
\seealso{
\link{assess}, \link{csem}, \link{cSEMResults}
}
\keyword{internal}
