% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_auxillary.R
\name{calculateConstructVCV}
\alias{calculateConstructVCV}
\title{Internal: Calculate construct VCV matrix}
\usage{
calculateConstructVCV(.C, .Q, .csem_model)
}
\value{
The (J x J) construct VCV matrix. Disattenuated if requested.
}
\description{
Calculates the covariance matrix of the constructs.
}
\details{
More description here
}
