% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_correction_factor.R
\name{calculateCorrectionFactors}
\alias{calculateCorrectionFactors}
\title{Calculate the correction factor}
\usage{
calculateCorrectionFactors(.S = NULL, .W = NULL, .modes = NULL,
  .csem_model = NULL, .PLS_approach_cf = NULL)
}
\arguments{
\item{.S}{The (K x K) empirical indicator correlation matrix.}

\item{.W}{A list with elements \code{$W}, \code{$E}, \code{Modes}, \code{Conv_status}, and
\code{Iterations}.}

\item{.modes}{A named vector giving the mode used to obtain new outer weights.}

\item{.csem_model}{A (possibly incomplete) \link{cSEMModel}-list.}

\item{.PLS_approach_cf}{Character string. Approach used to obtain the correction
factors for PLSc. One of: "\emph{dist_euclid}", "\emph{dist_euclid_weighted}",
"\emph{fisher_transformed}", "\emph{mean_arithmetic}", "\emph{mean_geometric}", "\emph{mean_harmonic}",
"\emph{geo_of_harmonic}". Defaults to "\emph{dist_euclid}".
Ignored if \code{.disattenuate = FALSE} or if \code{.approach_weights} is not PLS.}
}
\value{
A numeric vector of correction factors with element names equal
to the names of the J constructs used in the measurement model
}
\description{
Calculates the correction factor of PLSc.
}
\details{
Currently seven methods are available:

\itemize{
\item "dist_euclid" (default)
\item "dist_euclid_weighted"
\item "fisher_transformed"
\item "mean_geometric"
\item "mean_harmonic"
\item "mean_arithmetic"
\item "geo_of_harmonic" (not yet implemented)
}
}
