% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_assess.R
\name{calculateDf}
\alias{calculateDf}
\title{Internal: Degrees of freedom}
\usage{
calculateDf(
  .object     = NULL,
  .null_model = FALSE
  )
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}
}
\value{
A single numeric value.
}
\description{
Calculates the degrees of freedom for a given model from a \link{cSEMResults} object.
}
\seealso{
\link{assess}, \link{cSEMResults}
}
\keyword{internal}
