% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_test_MGD.R
\name{calculateDistance}
\alias{calculateDistance}
\title{Internal: Calculate matrix difference using distance measure}
\usage{
calculateDistance(
  .matrices = args_default()$.matrices, 
  .distance = args_default()$.distance
  )
}
\arguments{
\item{.matrices}{A list of at least two matrices.}

\item{.distance}{Character string. A distance measure. One of: "\emph{geodesic}"
or "\emph{squared_euclidian}". Defaults to "\emph{geodesic}".}
}
\value{
A numeric vector of length one containing the (arithmetic mean) of
the differences between all possible combinations of matrices.
}
\description{
Calculates the differences between symmetric matrices using a given
distance measure. This is typically used to calculate the difference between
the model-implied and the empirical indicator covariance matrix.
C
\code{.matrices} must be a list of at least two matrices. If more than two matrices
are supplied the arithmetic mean over all possible matrix
distances is computed. Hence, supplying a large number of matrices will
quickly become computationally challenging. Use with care.
}
\details{
Currently two distance measures are supported:
\describe{
\item{\code{geodesic}}{The geodesic distance}
\item{\code{squared_euclidian}}{The squared Euclidian distance}
}
}
\keyword{internal}
