% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_new.R
\name{calculateEffectSize}
\alias{calculateEffectSize}
\title{Calculate effect size}
\usage{
calculateEffectSize(
 .object              = args_default()$.object,
)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}
}
\description{
(TODO)
}
\examples{
\dontrun{
# still to implement
}

}
\seealso{
\link{csem}, \link{cSEMResults}
}
