% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_test_MGD.R
\name{calculateFR}
\alias{calculateFR}
\title{Internal: ANOVA F-test statistic}
\usage{
calculateFR(.resample_sarstedt)
}
\arguments{
\item{.resample_sarstedt}{A matrix containing the parameter estimates that
could potentially be compared and an id column indicating the group adherance
of each row.}
}
\value{
A named scaler, the test statistic of the ANOVA F-test
}
\description{
Calculate the ANOVA F-test statistic suggested by
\insertCite{Sarstedt2011;textual}{cSEM}.
}
\references{
\insertAllCited{}
}
\keyword{internal}
