% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_assess.R
\name{calculateHTMT}
\alias{calculateHTMT}
\title{Internal: HTMT}
\usage{
calculateHTMT(
 .object              = NULL,
 .only_common_factors = args_default()$.only_common_factors
)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}

\item{.only_common_factors}{Logical. Should only constructs modeled as common
factors be included? Defaults to \code{TRUE}.}
}
\description{
Compute the heterotrait-monotrait ratio of correlations
\insertCite{Henseler2015}{cSEM}.
}
\details{
The function is only applicable to objects of class \code{cSEMResults_default}.
For other object classes use \code{\link[=assess]{assess()}}.
}
\seealso{
\link{assess}, \link{csem}, \link{cSEMResults}
}
\keyword{internal}
