% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_assess.R
\name{calculateHTMT}
\alias{calculateHTMT}
\title{HTMT}
\usage{
calculateHTMT(
 .object              = NULL,
 .absolute            = TRUE,
 .alpha               = 0.5,
 .inference           = TRUE,
 .only_common_factors = TRUE,
 .R                   = 499
)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}

\item{.absolute}{Logical. Should the absolute HTMT values be returned?
Defaults to \code{TRUE} .}

\item{.alpha}{An integer or a numeric vector of significance levels.
Defaults to \code{0.05}.}

\item{.only_common_factors}{Logical. Should only concepts modeled as common
factors be included when calculating one of the following quality critera:
AVE, the Fornell-Larcker criterion, HTMT, and all reliability estimates.
Defaults to \code{TRUE}.}

\item{.R}{Integer. The number of bootstrap replications. Defaults to \code{499}.}
}
\value{
A lower tringular matrix of HTMT values.
}
\description{
Compute the heterotrait-monotrait ratio of correlations (HTMT) based on
\insertCite{Henseler2015;textual}{cSEM}.
}
\details{
The HTMT is a consistent estimator for the
construct correlations in case of tau equivalent measurement models.

The HTMT is used to assess discriminant validity.
}
\seealso{
\code{\link[=assess]{assess()}}, \link{csem}, \link{cSEMResults}
}
\keyword{internal}
