% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_foreman.R
\name{calculateIndicatorCor}
\alias{calculateIndicatorCor}
\title{Internal: Calculate indicator correlation matrix}
\usage{
calculateIndicatorCor(
  .X_cleaned    = args_default()$.X_cleaned, 
  .approach_cor = args_default()$.approach_cor
 )
}
\arguments{
\item{.X_cleaned}{A matrix of processed data (cleaned and ordered). Note: \code{X_cleaned}
may not be scaled!}

\item{.approach_cor}{Character string. Approach used to obtain the indicator
correlation matrix. One of: "\emph{bravais-pearson}" or "\emph{theil-sen}".
Defaults to "\emph{bravais-pearson}".}
}
\value{
The (K x K) indicator correlation matrix S.
}
\description{
Calculate the indicator correlation matrix using Bravais-Pearson
or Theil-Sen correlation.
}
\keyword{internal}
