% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_foreman.R
\name{calculateIndicatorCor}
\alias{calculateIndicatorCor}
\title{Internal: Calculate indicator correlation matrix}
\usage{
calculateIndicatorCor(
  .X_cleaned           = args_default()$.X_cleaned, 
  .approach_cor_robust = args_default()$.approach_cor
 )
}
\arguments{
\item{.X_cleaned}{A data.frame of processed data (cleaned and ordered). Note: \code{X_cleaned}
may not be scaled!}

\item{.approach_cor_robust}{Character string. Approach used to obtain a robust
indicator correlation matrix. One of: "\emph{none}" in which case the standard
bravais-person correlation is used,
"\emph{spearman}" for the spearman correlation, or
"\emph{mcd}" via \code{\link[MASS:cov.rob]{MASS::cov.rob()}} for a robust correlation matrix.
Defaults to "\emph{none}".}
}
\value{
A list containing the (K x K) indicator correlation matrix S.
}
\description{
Calculate the indicator correlation matrix using conventional or robust methods.
}
\details{
If \code{"none"} the method depends on the type of column of \code{.X_cleaned}:
\describe{
\item{\code{Numeric-numeric}}{Bravais-Pearson product-moment correlation}
\item{\code{Numeric-factor}}{Polyserial correlation}
\item{\code{Factor-factor}}{Polychoric correlation}
}
Note: logical input is treated as a 0-1 factor variable.

If  \code{"mcd"} (= minimum covariance determinant) a robust covariance is estimated
via \code{MASS::cov.rob()}. See \code{?MASS::cov.rob()} for details.
}
\keyword{internal}
