% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators_weights.R
\name{calculateInnerWeightsPLS}
\alias{calculateInnerWeightsPLS}
\title{Calculate the inner weights for PLS}
\usage{
calculateInnerWeightsPLS(
  .S                        = NULL,
  .W                        = NULL,
  .csem_model               = NULL,
  .PLS_weight_scheme_inner  = NULL,
  .ignore_structural_model  = NULL
)
}
\arguments{
\item{.S}{The (scaled) empirical (K x K) indicator covariance/correlation matrix,
where K is the number of indicators.}

\item{.W}{The (J x K) weight matrix, where J is the number of constructs and
K the number of indicators.}

\item{.csem_model}{A (possibly incomplete) \link{cSEMModel} list.}

\item{.PLS_weight_scheme_inner}{Character string. The inner weighting scheme
used in PLS. Possible choices are: "\emph{centroid}", "\emph{factorial}", or "\emph{path}".
Defaults to "\emph{centroid}". The argument is ignored if \code{.approach_weight} is not PLS.}

\item{.ignore_structural_model}{Logical. Should the structural (path) model be ignored
when calculating the inner weights of the PLS algorithm? Defaults to \code{FALSE}}
}
\value{
The (J x J) matrix of inner weights.
}
\description{
Calculates inner weights
}
\details{
More description here.
}
