% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_foreman.R
\name{calculateLoadings}
\alias{calculateLoadings}
\title{Internal: Calculate factor, composite, and cross-loadings}
\usage{
calculateLoadings(
 .S              = args_default()$.S,
 .W              = args_default()$.W,
 .Q              = args_default()$.Q,
 .csem_model     = args_default()$.csem_model,
 .disattenuate   = args_default()$.disattenuate,
 .modes          = args_default()$.modes
 )
}
\arguments{
\item{.S}{The (K x K) empirical indicator correlation matrix.}

\item{.W}{A (J x K) matrix of weights.}

\item{.Q}{A vector of composite-construct correlations with element names equal to
the names of the J construct names used in the measurement model. Note
Q^2 is also called the reliability coefficient.}

\item{.csem_model}{A (possibly incomplete) \link{cSEMModel}-list.}

\item{.disattenuate}{Logical. If possible, should composite correlations be disattenuated
if the construct is modeled as a common factor? Defaults to \code{TRUE}.}

\item{.modes}{A vector giving the mode for each construct in the form \code{"name" = "mode"}.
Only used internally.}
}
\value{
The (J x K) matrix of loadings and cross-loadings (attenuated if required).
}
\description{
Calculates factor loadings (for constructs modeled as common factors),
composite loadings (for constructs modeled as composites), and cross-loadings.
}
\keyword{internal}
