% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_loadings.R
\name{calculateLoadings}
\alias{calculateLoadings}
\title{Calculate factor and composite loadings}
\usage{
calculateLoadings(.S = NULL, .W = NULL, .csem_model = NULL,
  .disattenuate = NULL, .modes = NULL, .correction_factors = NULL)
}
\arguments{
\item{.S}{The (scaled) empirical (K x K) indicator covariance/correlation matrix,
where K is the number of indicators.}

\item{.W}{The (J x K) weight matrix, where J is the number of constructs and
K the number of indicators.}

\item{.csem_model}{A (possibly incomplete) \link{cSEMModel} list.}

\item{.disattenuate}{Logical. Should proxy correlations be disattenuated
if the construct is modeled as a common factor? Defaults to \code{TRUE}.}
}
\value{
The (J x K) matrix of loadings (disattenuated if requested).
}
\description{
Calculates factor loadings (for constructs modeled as common factors) and
composite loadings (for constructs modeled as composites).
}
\details{
Here more details
}
