% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_estimators_weights.R
\name{calculateOuterWeightsPLS}
\alias{calculateOuterWeightsPLS}
\title{Internal: Calculate the outer weights for PLS}
\usage{
calculateOuterWeightsPLS(
   .S      = args_default()$.S,
   .W      = args_default()$.W,
   .E      = args_default()$.E,
   .modes  = args_default()$.modes
   )
}
\arguments{
\item{.S}{The (K x K) empirical indicator correlation matrix.}

\item{.W}{A (J x K) matrix of weights.}

\item{.E}{A (J x J) matrix of inner weights.}

\item{.modes}{A vector giving the mode for each construct in the form \code{"name" = "mode"}.
Only used internally.}
}
\value{
A (J x K) matrix of outer weights.
}
\description{
Calculates outer weights using Mode A or Mode B.
}
\keyword{internal}
