% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_estimators_weights.R
\name{calculateOuterWeightsPLS}
\alias{calculateOuterWeightsPLS}
\title{Internal: Calculate the outer weights for PLS-PM}
\usage{
calculateOuterWeightsPLS(
   .data   = args_default()$.data  
   .S      = args_default()$.S,
   .W      = args_default()$.W,
   .E      = args_default()$.E,
   .modes  = args_default()$.modes
   )
}
\arguments{
\item{.data}{A \code{data.frame} or a \code{matrix} of standardized or unstandarized data.
Possible column types or classes of the data provided are: logical,
numeric (double or integer), factor (ordered and unordered)
or a mix of several types. The data may also include
\emph{one} character column whose column name must be given to \code{.id}.
This column is assumed to contain group identifiers used to split
the data into groups.}

\item{.S}{The (K x K) empirical indicator correlation matrix.}

\item{.W}{A (J x K) matrix of weights.}

\item{.E}{A (J x J) matrix of inner weights.}

\item{.modes}{A vector giving the mode for each construct in the form \code{"name" = "mode"}.
Only used internally.}
}
\value{
A (J x K) matrix of outer weights.
}
\description{
Calculates outer weights using default or user supplied modes.
}
\keyword{internal}
