% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_auxillary.R
\name{calculateProxies}
\alias{calculateProxies}
\title{Internal: Calculate proxies}
\usage{
calculateProxies(.X = NULL, .W = NULL)
}
\arguments{
\item{.W}{The (J x K) weight matrix, where J is the number of constructs and
K the number of indicators.}
}
\value{
The (N x J) matrix of proxy values with column names equal to
the names of the J constructs.
}
\description{
Calculates proxy values for the J constructs of the structural model.
}
\details{
More on that
}
