% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_auxillary.R
\name{calculateProxyConstructCV}
\alias{calculateProxyConstructCV}
\title{Internal: Calculate proxy-construct covariance}
\usage{
calculateProxyConstructCV(.W = NULL, .csem_model = NULL,
  .disattenuate = NULL, .modes = NULL, .correction_factors = NULL,
  .reliabilities = NULL)
}
\arguments{
\item{.W}{A list with elements \code{$W}, \code{$E}, \code{Modes}, \code{Conv_status}, and
\code{Iterations}.}

\item{.csem_model}{A (possibly incomplete) \link{cSEMModel}-list.}

\item{.disattenuate}{Logical. If possible, should composite correlations be disattenuated
if the construct is modelled as a common factor? Defaults to \code{TRUE}.}

\item{.modes}{A named vector giving the mode used to obtain new outer weights.}

\item{.reliabilities}{A character vector of \code{"name" = value} pairs,
where \code{value} is a number between 0 and 1 and \code{"name"} a character string
of the corresponding construct name, or \code{NULL}. Reliabilities
may be given for a subset of the constructs. Defaults to \code{NULL} in which case
reliabilites are estimated by \code{csem()}.}
}
\value{
A vector of proxy-construct covariances.
}
\description{
Calculates the proxy-covariance covariance (Q)
(also called reliability coefficient).
}
\details{
More on that
}
