% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_auxillary.R
\name{calculateProxyConstructCV}
\alias{calculateProxyConstructCV}
\title{Internal: Calculate proxy-construct covariance}
\usage{
calculateProxyConstructCV(.W = NULL, .csem_model = NULL,
  .disattenuate = NULL, .modes = NULL, .correction_factors = NULL,
  .reliabilities = NULL)
}
\arguments{
\item{.W}{A list with elements \code{$W}, \code{$E}, \code{PLS_mode} and xxx.}

\item{.csem_model}{A (possibly incomplete) \link{cSEMModel} list.}

\item{.disattenuate}{Logical. Should proxy correlations be disattenuated
if the construct is modeled as a common factor? Defaults to \code{TRUE}.}

\item{.reliabilities}{A vector of \code{name = value} pairs of reliability coefficients.
Element names are equal to the names of the J construct names. Reliabilities
may be given for a subset of the constructs. Defaults to \code{NULL} in which case
reliabilites are estimated by \code{cSEM}.}
}
\value{
A vector of proxy-construct covariances.
}
\description{
Calculates the proxy-covariance covariance (Q)
(also called reliability coefficient).
}
\details{
More on that
}
