% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_auxillary.R
\name{calculateProxyConstructCV}
\alias{calculateProxyConstructCV}
\title{Internal: Calculate proxy-construct covariance}
\usage{
calculateProxyConstructCV(.S = NULL, .W = NULL, .csem_model = NULL,
  .disattenuate = NULL, .modes = NULL, .correction_factors = NULL)
}
\arguments{
\item{.S}{The (scaled) empirical (K x K) indicator covariance/correlation matrix,
where K is the number of indicators.}

\item{.W}{The (J x K) weight matrix, where J is the number of constructs and
K the number of indicators.}

\item{.csem_model}{A (possibly incomplete) \link{cSEMModel} list.}

\item{.disattenuate}{Logical. Should proxy correlations be disattenuated
if the construct is modeled as a common factor? Defaults to \code{TRUE}.}
}
\value{
A vector of proxy-construct covariances.
}
\description{
Calculates the proxy-covariance covariance (Q)
(also called reliability coefficient).
}
\details{
More on that
}
