% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_auxillary.R
\name{calculateProxyVCV}
\alias{calculateProxyVCV}
\title{Internal: Calculate proxy variance-covariance matrix}
\usage{
calculateProxyVCV(.S = NULL, .W = NULL)
}
\arguments{
\item{.S}{The (scaled) empirical (K x K) indicator covariance/correlation matrix,
where K is the number of indicators.}

\item{.W}{A list with elements \code{$W}, \code{$E}, \code{PLS_mode} and xxx.}
}
\value{
A (J x J) proxy VCV matrix.
}
\description{
Calculates the variance-covariance (VCV) matrix of the proxies.
}
\details{
More on that
}
