% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_assess.R
\name{calculateVIFModeB}
\alias{calculateVIFModeB}
\title{Internal: Calculate VIF values for PLS Mode B}
\usage{
calculateVIFModeB(.object = NULL)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}
}
\value{
A named list of vectors containing the VIF values. Each list name
is the name of a construct whose weights were obtained using Mode B.
The vectors contain the VIF values obtained from a regression of each
explanatory variable of a given construct on the remaining explanatory
variables of that construct.

If the weightning approach was not \code{"PLS-PM"} or non of the constructs had Mode B,
the function silently returns \code{NA}.
}
\description{
Calculate the variance inflation factor (VIF) for indicators of constructs whose
outer PLS estimation scheme is Mode B.
}
\details{
Weights obtained using Mode B often suffer from multicollinearity. VIF values
are commonly used to assess the severity of multicollinearity.

The function is only applicable to objects of class \code{cSEMResults_default}.
For other object classes use \code{\link[=assess]{assess()}}.
}
\references{
\insertAllCited{}
}
\seealso{
\link{assess}, \link{cSEMResults}
}
\keyword{internal}
