% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators_weights.R
\name{calculateWeightsFixed}
\alias{calculateWeightsFixed}
\title{Calculate composite weights using fixed weights}
\usage{
calculateWeightsFixed(.data, .model)
}
\arguments{
\item{.data}{A \code{data.frame}, a \code{matrix} or a list containing data of either type.
Possible column types or classes of the data provided are:
logical, numeric (double or integer), factor (ordered and unordered)
or a mix of several types. The data may also include
\emph{one} character column whose column name must be given to \code{.id}.
This column is assumed to contain group identifiers used to split the data into groups.}

\item{.model}{A model in \code{\link[lavaan:model.syntax]{lavaan model syntax}}
or a \link{cSEMModel}-list.}
}
\value{
A list with the elements
\describe{
\item{\code{$W}}{A (J x K) matrix of estimated weights.}
\item{\code{$E}}{A (J x J) matrix of inner weights.}
\item{\code{$Modes}}{A named vector of Modes used for the outer estimation.}
\item{\code{$Conv_status}}{The convergence status. \code{TRUE} if the algorithm has converged
and \code{FASLE} otherwise. If one-step weights are used via \code{.iter_max = 1}
or a non-iterative procedure was used, the convergence status is set to \code{NULL}.}
\item{\code{$Iterations}}{The number of iterations used.}
}
}
\description{
Calculates weights...
}
\details{
Some more description...
}
