% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators_weights.R
\name{calculateWeightsFixed}
\alias{calculateWeightsFixed}
\title{Calculate weights using fixed weights}
\usage{
calculateWeightsFixed(.data, .model)
}
\arguments{
\item{.data}{A \code{data.frame} or a \code{matrix} containing the raw data.}

\item{.model}{A model in \href{http://lavaan.ugent.be/tutorial/syntax1.html}{lavaan model syntax}
or a \link{cSEMModel} object.}
}
\value{
A list with the following elements:
\describe{
\item{\code{$W}}{A (J x K) matrix of estimated weights.}
\item{\code{$E}}{A (J x J) matrix of inner weights.}
\item{\code{$Modes}}{A named vector of Modes used for the outer estimation.}
\item{\code{$Conv_status}}{The convergence status. One of \code{TRUE}or \code{FASLE}. If
one-step weights are used via \code{.iter_max = 1} the convergence status
is set to \code{NULL}.}
\item{\code{$Iterations}}{The number of iterations used.}
}
}
\description{
Calculates weights...
}
\details{
Some more description...
}
