% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators_weights.R, R/gsca.R
\name{calculateWeightsGSCA}
\alias{calculateWeightsGSCA}
\title{Calculate weights using GSCA}
\usage{
calculateWeightsGSCA(.data, .model)

calculateWeightsGSCA(.data, .model)
}
\arguments{
\item{.data}{A \code{data.frame} or a \code{matrix} containing the raw data. Possible
data column types or classes are: logical, numeric (double or integer), factor
(ordered and unordered) or a mix of several types.}

\item{.S}{The (K x K) empirical indicator correlation matrix.}

\item{.model}{A model in \code{\link[lavaan:model.syntax]{lavaan model syntax}}
or a \link{cSEMModel}-list.}

\item{.iter_max}{Integer. The maximum number of iterations allowed.
If \code{iter_max = 1} and \code{.approach_weights = "PLS"} one-step weights are returned.
If the algorithm exceeds the specified number, weights of iteration step
\code{.iter_max - 1}  will be returned with a warning. Defaults to \code{100}.}

\item{.tolerance}{Double. The tolerance criterion for convergence.
Defaults to \code{1e-05}.}
}
\value{
A list with the elements
\describe{
\item{\code{$W}}{A (J x K) matrix of estimated weights.}
\item{\code{$E}}{A (J x J) matrix of inner weights.}
\item{\code{$Modes}}{A named vector of Modes used for the outer estimation.}
\item{\code{$Conv_status}}{The convergence status. \code{TRUE} if the algorithm has converged
and \code{FASLE} otherwise. If one-step weights are used via \code{.iter_max = 1}
or a non-iterative procedure was used, the convergence status is set to \code{NULL}.}
\item{\code{$Iterations}}{The number of iterations used.}
}
}
\description{
Calculates weights...

Calculates weights...
}
\details{
Some more description...

Some more description...
}
