% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators_weights.R
\name{calculateWeightsGSCA}
\alias{calculateWeightsGSCA}
\title{Calculate weights using GSCA}
\usage{
calculateWeightsGSCA(
  .X                           = args_default()$.X,
  .S                           = args_default()$.S,
  .csem_model                  = args_default()$.csem_model,
  .conv_criterion              = args_default()$.conv_criterion,
  .iter_max                    = args_default()$.iter_max,
  .tolerance                   = args_default()$.tolerance
   )
}
\arguments{
\item{.X}{A matrix of processed data (scaled, cleaned and ordered).}

\item{.S}{The (K x K) empirical indicator correlation matrix.}

\item{.csem_model}{A (possibly incomplete) \link{cSEMModel}-list.}

\item{.conv_criterion}{Character string. The criterion to use for the convergence check.
One of: "\emph{diff_absolute}", "\emph{diff_squared}", or "\emph{diff_relative}". Defaults
to "\emph{diff_absolute}".}

\item{.iter_max}{Integer. The maximum number of iterations allowed.
If \code{iter_max = 1} and \code{.approach_weights = "PLS-PM"} one-step weights are returned.
If the algorithm exceeds the specified number, weights of iteration step
\code{.iter_max - 1}  will be returned with a warning. Defaults to \code{100}.}

\item{.tolerance}{Double. The tolerance criterion for convergence.
Defaults to \code{1e-05}.}
}
\value{
A list with the elements
\describe{
\item{\code{$W}}{A (J x K) matrix of estimated weights.}
\item{\code{$E}}{A (J x J) matrix of inner weights.}
\item{\code{$Modes}}{A named vector of Modes used for the outer estimation, for GSCA
the mode is automatically set to 'gsca'.}
\item{\code{$Conv_status}}{The convergence status. \code{TRUE} if the algorithm has converged
and \code{FALSE} otherwise. If one-step weights are used via \code{.iter_max = 1}
or a non-iterative procedure was used, the convergence status is set to \code{NULL}.}
\item{\code{$Iterations}}{The number of iterations used.}
}
}
\description{
Calculates weights of a structural model using the GSCA-procedure. This approach
is an alternative to PLS in structural equation modeling with composites.
The first version of this approach was presented in \insertCite{Hwang2004;textual}{cSEM}.
Since then, there had been several advancements, such that the latest version
of GSCA can been found in \insertCite{Hwang2014;textual}{cSEM}. This is also
the GSCA version the implementation in this package relies on.
}
\details{
GSCA models latent variables (constructs) as composites, i.e. as an exact linear
combination of its corresponding indicators. The main idea of GSCA is to
divide the overall structural model into three submodels: the structural model,
the measurement model and the weighted relation model. These sub-models and their
corresponding matrices are provided by the \code{.csem_model} argument.

Weights as well as structural coefficients and loadings are estimated with the
Alternating Least Squares Algorithm.
This algorithm consists of two main steps. In the first step, structural coefficients
and loadings are updated for fixed weights. Then, in the second step, weights
are updated keeping the other parameters constant. Doing this, the value of a
global optimization decreases.

These two steps are iterated until convergence is reached or the prescribed
maximum number of iteration steps, '.iter_max', was carried out. In their example
in \insertCite{Hwang2014;textual}{cSEM}, p. 75, the authors set '.iter_max'
equal to 100, but this is an arbitrary choice. The user might try several values
for '.iter_max' and eventually increase its value if convergence happened previously
due to reaching the maximum number of steps (indicated by a \code{Conv_status} equal
to \code{FALSE}).

Several convergence criterion are possible, all consisting in calculating some
difference of the weight matrices W of two subsequent steps. Checking for convergence
this way results in the same interpretation as calculating the differences
the global optimization criterion values as proposed by Hwang and Takane.
The \code{.tolerance} argument provides the value below which the difference has to
fall such that the algorithm converges. The default value of \code{1e-05} is the same
value which Hwang and Takane use in their example in \insertCite{Hwang2014;textual}{cSEM}
, p. 75. Nevertheless, the argument \code{.tolerance} can take any other value as well.

The ALS algorithm as outlined in \insertCite{Hwang2014;textual}{cSEM} essentially
consists of regressions of endogenous constructs (resp. their proxies) on
exogenous constructs for the structural model coefficients and in regressions
of all common factor constructs on their related indicators for the loadings
(steps 1a and 1b). If a construct is not of common factor type but a pure composite
the corresponding "construct" loadings do not need to be calculated because the
constructs are exact linear combinations of their indicators. Instead, "composite"
loadings are calculated and provided as the estimated loadings. These "composite"
loadings are the "true" correlations between a construct and indicator. This correlation
always exists, no matter which is the type of the construct under consideration.
Considering a construct of common factor type, estimated ("construct") loadings
are the correlations between indicators and proxies, but not the "true" correlation
between indicator and underlying but unknown construct.

In the case that there is at least one construct which is not modeled as a
common factor but as a pure composite, the user imperatively has to use GSCA
and GSCA_M is no option. The reason is that the implementation of GSCA_M in
the cSEM-package is based on \insertCite{Hwang2017;textual}{cSEM}.
The GSCA_M-approach presented by the authors only works out if all constructs
are of common factor type. Otherwise, i.e., if there is at least one construct
of composite type estimation using GSCA_M fails because calculating weight estimates
with GSCA_M leads to a product involving the measurement matrix.
This matrix does not have full rank in the case of at least one pure composite construct.
It has a zero row for every construct which is a pure composite (i.e. all related loadings are zero)
and, therefore, leads to a non-invertible matrix when multiplying it with its transposed.

Therefore estimation must be carried out with GSCA if there is at least one composite
construct. Otherwise, i.e., if there are only constructs which are of common factor type,
calling \code{\link[=csem]{csem()}} will lead to an estimation via GSCA_M except in the case that
the user explicitly sets the argument \code{.disattenuate} to \code{FALSE}. estimation is always done
by 'standard' GSCA.
}
\references{
\insertAllCited{}
}
