% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators_weights.R
\name{calculateWeightsGSCA}
\alias{calculateWeightsGSCA}
\title{Calculate weights using GSCA}
\usage{
calculateWeightsGSCA(.data, .model)
}
\arguments{
\item{.data}{A \code{data.frame} or a \code{matrix} containing the raw data. Possible
data column types or classes are: logical, numeric (double or integer), factor
(ordered and unordered) or a mix of several types.}

\item{.model}{A model in \code{\link[lavaan:model.syntax]{lavaan model syntax}}
or a \link{cSEMModel}-list.}
}
\value{
A list with the elements
\describe{
\item{\code{$W}}{A (J x K) matrix of estimated weights.}
\item{\code{$E}}{A (J x J) matrix of inner weights.}
\item{\code{$Modes}}{A named vector of Modes used for the outer estimation.}
\item{\code{$Conv_status}}{The convergence status. One of \code{TRUE}or \code{FASLE}. If
one-step weights are used via \code{.iter_max = 1} or no iterative algorithm was used,
the convergence status is set to \code{NULL}.}
\item{\code{$Iterations}}{The number of iterations used.}
}
}
\description{
Calculates weights...
}
\details{
Some more description...
}
