% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators_weights.R, R/gsca.R
\name{calculateWeightsGSCA}
\alias{calculateWeightsGSCA}
\title{Calculate weights using GSCA}
\usage{
calculateWeightsGSCA(.data, .model)

calculateWeightsGSCA(
   .data        = args_default()$.data,
   .S           = args_default()$.S,
   .model       = args_default()$.model,
   .iter_max    = args_default()$.iter_max,
   .tolerance   = args_default()$.tolerance
   )
}
\arguments{
\item{.data}{A \code{data.frame} or a \code{matrix} containing the raw data. Possible
data column types or classes are: logical, numeric (double or integer), factor
(ordered and unordered) or a mix of several types.}

\item{.S}{The (K x K) empirical indicator correlation matrix.}

\item{.model}{A model in \code{\link[lavaan:model.syntax]{lavaan model syntax}}
or a \link{cSEMModel}-list.}

\item{.iter_max}{Integer. The maximum number of iterations allowed.
If \code{iter_max = 1} and \code{.approach_weights = "PLS-PM"} one-step weights are returned.
If the algorithm exceeds the specified number, weights of iteration step
\code{.iter_max - 1}  will be returned with a warning. Defaults to \code{100}.}

\item{.tolerance}{Double. The tolerance criterion for convergence.
Defaults to \code{1e-05}.}
}
\value{
An object of class \code{cSEMResults}. The result is a list with the elements:
\describe{
\item{\code{$Estimates}}{A list containing a list of estimated quantities.}
\item{\code{$Information}}{A list containing a list of additional information.}
}

The estimated quantities in the \code{$Estimates} list are:
\describe{
\item{\code{$Path_estimates}}{A (J x J) matrix of estimated path coefficients.}
\item{\code{$Loading_estimates}}{A (K x J) matrix of loading estimates.}
\item{\code{$Weight_estimates}}{A (J x K) matrix of weight estimates.}
\item{\code{$Construct_scores}}{A (N x 1) matrix of construct scores calculated
using the estimated weights via the weighted relation model.}
\item{\code{$Indicator_VCV}}{The empirical correlation matrix of the indicators.}
\item{\code{$Proxy_VCV}}{The correlation matrix of the proxies.}
\item{\code{$FIT}}{The \code{FIT} measure of global model fit.}
\item{\code{$AFIT}}{The \code{adjusted FIT} measure of global model fit.}
\item{\code{$FIT_M}}{The \code{FIT_M} measure of local model fit in the measurement model.}
\item{\code{$FIT_S}}{The \code{FIT_S} measure of local model fit in the structural model.}
}

The additional information in the list \code{information} are:
\describe{
\item{\code{$Data}}{The (standardized) data used for the estimation.}
\item{\code{$Model}}{The specified model as given by the function \code{parseModel}.}
\item{\code{$Arguments}}{A list containing all arguments used and their respective value.}
\item{\code{$Weight_info}}{A list with the elements \code{Number_iterations} and
\code{Convergence_status}.}
}
}
\description{
Calculates weights...

This function calculates weights, path coefficients and loadings of a specified
strucural equation model with the GSCA procedure (\insertCite{Hwang2004}{cSEM}
and \insertCite{Hwang2014}{cSEM}). For more details concerning
the GSCA approach, the estimation routine and its embedding in the cSEM-package
see vignette.
}
\details{
Some more description...
}
\examples{
model <- "
# Structural model
QUAL ~ EXPE
EXPE ~ IMAG
SAT  ~ IMAG + EXPE + QUAL + VAL
LOY  ~ IMAG + SAT
VAL  ~ EXPE + QUAL 
# Measurement model
EXPE <~ expe1 + expe2 + expe3 + expe4 + expe5
IMAG <~ imag1 + imag2 + imag3 + imag4 + imag5
LOY  =~ loy1  + loy2  + loy3  + loy4
QUAL =~ qual1 + qual2 + qual3 + qual4 + qual5
SAT  <~ sat1  + sat2  + sat3  + sat4
VAL  <~ val1  + val2  + val3  + val4
"
gsca_results <- csem(.data = satisfaction, .model = model, .approach_weights = "GSCA")
 
}
\references{
\insertAllCited{}
}
