% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators_weights.R
\name{calculateWeightsGSCAm}
\alias{calculateWeightsGSCAm}
\title{Calculate weights using GSCAm}
\usage{
calculateWeightsGSCAm(.X = args_default()$.X,
  .csem_model = args_default()$.csem_model,
  .conv_criterion = args_default()$.conv_criterion,
  .iter_max = args_default()$.iter_max,
  .tolerance = args_default()$.tolerance,
  .starting_values = args_default()$.starting_values)
}
\description{
Calculates weights...
}
\details{
(TODO)
}
