% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators_weights.R
\name{calculateWeightsIGSCA}
\alias{calculateWeightsIGSCA}
\title{Calculate weights using Integrated Generalised Structured Component Analysis (IGSCA)}
\usage{
calculateWeightsIGSCA(
  .data,
  .S = args_default()$.S,
  .csem_model = args_default()$.csem_model,
  .conv_criterion = args_default()$.conv_criterion,
  .GSCA_modes = args_default()$.GSCA_modes,
  .iter_max = args_default()$.iter_max,
  .tolerance = args_default()$.tolerance,
  .starting_values = args_default()$.starting_values
)
}
\arguments{
\item{.data}{A \code{data.frame} or a \code{matrix} of standardized or unstandardized
data (indicators/items/manifest variables).
Additionally, a \code{list} of data sets (data frames or matrices) is accepted in which
case estimation is repeated for each data set. Possible column types or classes
of the data provided are: "\code{logical}", "\code{numeric}" ("\code{double}" or "\code{integer}"),
"\code{factor}" ("\code{ordered}" and/or "\code{unordered}"), "\code{character}" (will be converted to factor),
or a mix of several types.}

\item{.S}{The (K x K) empirical indicator correlation matrix.}

\item{.csem_model}{A (possibly incomplete) \link{cSEMModel}-list.}

\item{.conv_criterion}{Character string. The criterion to use for the convergence check.
One of: "\emph{diff_absolute}", "\emph{diff_squared}", "\emph{diff_relative}", or "\emph{sum_diff_absolute}". Defaults
to "\emph{diff_absolute}".}

\item{.GSCA_modes}{Either a named list specifying the mode that should be
used for each composite in the form \code{"composite_name" = mode}, a single
character string giving the mode that should be used for all composites.
Possible single character string choices for \code{mode} are: "\emph{CCMP}" for
canonical composites, or "\emph{NCMP}" for nomological composites, or \code{NULL} for
default behavior. Default behavior is to estimate canonical composites.
Passed to (I-)GSCA estimating functions in \code{\link[=calculateWeightsGSCA]{calculateWeightsGSCA()}}
or \code{\link[=calculateWeightsIGSCA]{calculateWeightsIGSCA()}}. Defaults to \code{NULL}.}

\item{.iter_max}{Integer. The maximum number of iterations allowed.
If \code{iter_max = 1} and \code{.approach_weights = "PLS-PM"} one-step weights are returned.
If the algorithm exceeds the specified number, weights of iteration step
\code{.iter_max - 1}  will be returned with a warning. Defaults to \code{100}.}

\item{.tolerance}{Double. The tolerance criterion for convergence.
Defaults to \code{1e-05}.}

\item{.starting_values}{A named list of vectors where the
list names are the construct names whose indicator weights the user
wishes to set. The vectors must be named vectors of \code{"indicator_name" = value}
pairs, where \code{value} is the (scaled or unscaled) starting weight. Defaults to \code{NULL}.}
}
\value{
List of 4 matrices that make up a fitted I-GSCA Model:
\itemize{
\item (1) Weights
\item (2) Loadings
\item (3) Squared Unique Loadings D^2
\item (4) Path Coefficients.
\item (5) Unique Component of Indicators (for common factors) DU
}
}
\description{
This R implementation of I-GSCA is based on the Matlab implementation in igsca_sim.m by Dr. Heungsun Hwang.
}
\details{
In the example section, the specified model is based on the tutorial I-GSCA model associated with GSCA Pro \insertCite{hwangetal2023StructuralEquationModelingAMultidisciplinaryJournal}{cSEM}.

\strong{Note}: Here, we assume that there is only one unique loading per indicator.
}
\examples{
\dontrun{
# Specify the model according to GSCA Pro's example
tutorial_igsca_model <- "
# Composite Model
NetworkingBehavior <~ Behavior1 + Behavior2 + Behavior3 + Behavior5 +
                      Behavior7 + Behavior8 +  Behavior9
Numberofjobinterviews <~ Interview1 + Interview2
Numberofjoboffers <~ Offer1 + Offer2

# Reflective Measurement Model
HonestyHumility =~ Honesty1 + Honesty2 + Honesty3 + Honesty4 + Honesty5 +
                    Honesty6 + Honesty7 + Honesty8 + Honesty9 + Honesty10
Emotionality =~ Emotion1 + Emotion2 + Emotion3 + Emotion4 +
                Emotion5 + Emotion6 + Emotion8 + Emotion10
Extraversion =~ Extraver2 + Extraver3 + Extraver4 + Extraver5 +
                Extraver6 + Extraver7 + Extraver8 + Extraver9 + Extraver10
Agreeableness =~ Agreeable1 + Agreeable3 + Agreeable4 + Agreeable5 +
                 Agreeable7 + Agreeable8 + Agreeable9 + Agreeable10
Conscientiousness =~ Conscientious1 + Conscientious3 + Conscientious4 +
                     Conscientious6 + Conscientious7 + Conscientious8 +
                     Conscientious9 + Conscientious10
OpennesstoExperience =~ Openness1 + Openness2 + Openness3 + Openness5 +
                        Openness7 + Openness8 + Openness9 + Openness10

# Structural Model
NetworkingBehavior ~ HonestyHumility + Emotionality + Extraversion +
                     Agreeableness + Conscientiousness + OpennesstoExperience
Numberofjobinterviews ~ NetworkingBehavior
Numberofjoboffers ~ NetworkingBehavior
"

data(LeDang2022)

csem(.data = LeDang2022, tutorial_igsca_model, .approach_weights = "GSCA",
.dominant_indicators = NULL, .tolerance = 0.0001, .conv_criterion =
"sum_diff_absolute")
}
}
\references{
\insertAllCited{}
}
\author{
Michael S. Truong
}
