% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators_weights.R
\name{calculateWeightsIGSCA}
\alias{calculateWeightsIGSCA}
\title{Calculate weights using Integrated Generalised Structured Component Analysis (IGSCA)}
\usage{
calculateWeightsIGSCA(
  .data,
  .S = args_default()$.S,
  .csem_model = args_default()$.csem_model,
  .conv_criterion = args_default()$.conv_criterion,
  .GSCA_modes = args_default()$.GSCA_modes,
  .iter_max = args_default()$.iter_max,
  .tolerance = args_default()$.tolerance,
  .starting_values = args_default()$.starting_values
)
}
\arguments{
\item{.data}{A \code{data.frame} or a \code{matrix} of standardized or unstandardized
data (indicators/items/manifest variables). Possible column types or classes
of the data provided are: "\code{logical}", "\code{numeric}" ("\code{double}" or "\code{integer}"),
"\code{factor}" ("\code{ordered}" and/or "\code{unordered}"), "\code{character}" (converted to factor),
or a mix of several types.}

\item{.S}{The (K x K) empirical indicator correlation matrix.}

\item{.csem_model}{A (possibly incomplete) \link{cSEMModel}-list.}

\item{.conv_criterion}{Character string. The criterion to use for the convergence check.
One of: "\emph{diff_absolute}", "\emph{diff_squared}", "\emph{diff_relative}", or "\emph{sum_diff_absolute}". Defaults
to "\emph{diff_absolute}".}

\item{.GSCA_modes}{Either a named list specifying the mode that should be
used for each composite in the form \code{"composite_name" = mode}, a single
character string giving the mode that should be used for all composites.
Possible single character string choices for \code{mode} are: "\emph{CCMP}" for
canonical composites, or "\emph{NCMP}" for nomological composites, or \code{NULL} for
default behavior. Default behavior is to estimate canonical composites.
Passed to (I-)GSCA estimating functions in \code{\link[=calculateWeightsGSCA]{calculateWeightsGSCA()}}
or \code{\link[=calculateWeightsIGSCA]{calculateWeightsIGSCA()}}. Defaults to \code{NULL}.}

\item{.iter_max}{Integer. The maximum number of iterations allowed.
If \code{iter_max = 1} and \code{.approach_weights = "PLS-PM"} one-step weights are returned.
If the algorithm exceeds the specified number, weights of iteration step
\code{.iter_max - 1}  will be returned with a warning. Defaults to \code{100}.}

\item{.tolerance}{Double. The tolerance criterion for convergence.
Defaults to \code{1e-05}.}

\item{.starting_values}{A named list of vectors where the
list names are the construct names whose indicator weights the user
wishes to set. The vectors must be named vectors of \code{"indicator_name" = value}
pairs, where \code{value} is the (scaled or unscaled) starting weight. Defaults to \code{NULL}.}
}
\value{
List of matrices of the fitted IGSCA Model
}
\description{
Calculate weights using Integrated Generalised Structured Component Analysis (IGSCA)
}
