% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators_weights.R
\name{calculateWeightsIGSCA}
\alias{calculateWeightsIGSCA}
\title{Calculate weights using Integrated Generalised Structured Component Analysis (I-GSCA)}
\usage{
calculateWeightsIGSCA()
}
\description{
Calculate weights using Integrated Generalised Structured Component Analysis (I-GSCA)
}
