% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators_weights.R
\name{calculateWeightsIGSCA}
\alias{calculateWeightsIGSCA}
\title{Calculate weights using Integrated Generalised Structured Component Analysis (I-GSCA)}
\usage{
calculateWeightsIGSCA(
  .data,
  .csem_model = args_default()$.csem_model,
  .tolerance = args_default()$.tolerance,
  .iter_max = args_default()$.iter_max,
  .dominant_indicators = args_default()$.dominant_indicators,
  .conv_criterion = args_default()$.conv_criterion
)
}
\arguments{
\item{.data}{A \code{data.frame} or a \code{matrix} of standardized or unstandardized
data (indicators/items/manifest variables).
Additionally, a \code{list} of data sets (data frames or matrices) is accepted in which
case estimation is repeated for each data set. Possible column types or classes
of the data provided are: "\code{logical}", "\code{numeric}" ("\code{double}" or "\code{integer}"),
"\code{factor}" ("\code{ordered}" and/or "\code{unordered}"), "\code{character}" (will be converted to factor),
or a mix of several types.}

\item{.csem_model}{A (possibly incomplete) \link{cSEMModel}-list.}

\item{.tolerance}{Minimum amount of absolute change in the estimates of the
path-coefficients (if B0 is non-zero) or the loadings (if B0 is all zero,
meaning there are no path-coefficients) between ALS iterations before ending
the optimization.}

\item{.iter_max}{Maximum number of iterations of the Alternating Least Squares
(ALS) algorithm.}

\item{.dominant_indicators}{A named vector that indices the dominant
indicator for each construct variable. \emph{It is to be clarified whether this
should only apply to factor latent variables or also composite latent
variables.} This is important for ensuring that the signs of the
path-coefficients and loadings are consistent. It is sometimes the case in
composite-based structural equation modelling methods that
loadings/path-coefficients may have the opposite sign. The length of this
vector should be equal to the number of construct variables and each value
should represent the row number of the dominant indicator for that
construct variable.}

\item{.conv_criterion}{Character string. The criterion to use for the convergence check.
One of: "\emph{diff_absolute}", "\emph{diff_squared}", "\emph{diff_relative}", or "\emph{sum_diff_absolute}". Defaults
to "\emph{diff_absolute}".}
}
\value{
List of matrices of the fitted I-GSCA Model
}
\description{
Calculate weights using Integrated Generalised Structured Component Analysis (I-GSCA)
}
