% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators_weights.R
\name{calculateWeightsKettenring}
\alias{calculateWeightsKettenring}
\title{Calculate composite weights using one of Kettenrings's approaches}
\usage{
calculateWeightsKettenring(.data, .model, .approach)
}
\arguments{
\item{.data}{A \code{data.frame} or a \code{matrix} containing the raw data.}

\item{.model}{A model in \code{\link[lavaan:model.syntax]{lavaan model syntax}}
or a \link{cSEMModel}-list.}

\item{.approach}{Character string. The Kettenring approach to use. One of
"\emph{SUMCORR}", "\emph{MAXVAR}", "\emph{SSQCORR}", "\emph{MINVAR}" or "\emph{GENVAR}". Defaults to
"\emph{SUMCORR}".}
}
\value{
A list with the elements
\describe{
\item{\code{$W}}{A (J x K) matrix of estimated weights.}
\item{\code{$E}}{A (J x J) matrix of inner weights.}
\item{\code{$Modes}}{A named vector of Modes used for the outer estimation.}
\item{\code{$Conv_status}}{The convergence status. One of \code{TRUE}or \code{FASLE}. If
one-step weights are used via \code{.iter_max = 1} or no iterative algorithm was used,
the convergence status is set to \code{NULL}.}
\item{\code{$Iterations}}{The number of iterations used.}
}
}
\description{
Calculates weights according to on of the the five criteria
"\emph{SUMCORR}", "\emph{MAXVAR}", "\emph{SSQCORR}", "\emph{MINVAR}", and "\emph{GENVAR}"
suggested by \insertCite{Kettenring1971;textual}{cSEM}.
}
\details{
Some more description...
}
