% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators_weights.R
\name{calculateWeightsKettenring}
\alias{calculateWeightsKettenring}
\title{Calculate composite weights using one of Kettenrings's approaches}
\usage{
calculateWeightsKettenring(
  .S           = args_default()$.S, 
  .csem_model  = args_default()$.csem_model,   
  .approach    = args_default()$.approach
  )
}
\arguments{
\item{.S}{The (K x K) empirical indicator correlation matrix.}

\item{.csem_model}{A (possibly incomplete) \link{cSEMModel}-list.}

\item{.approach}{Character string. The Kettenring approach to use. One of
"\emph{SUMCORR}", "\emph{MAXVAR}", "\emph{SSQCORR}", "\emph{MINVAR}" or "\emph{GENVAR}". Defaults to
"\emph{SUMCORR}".}
}
\value{
A list with the elements
\describe{
\item{\code{$W}}{A (J x K) matrix of estimated weights.}
\item{\code{$E}}{A (J x J) matrix of inner weights.}
\item{\code{$Modes}}{A named vector of modes used for the outer estimation.}
\item{\code{$Conv_status}}{The convergence status. \code{TRUE} if the algorithm has converged
and \code{FALSE} otherwise. If one-step weights are used via \code{.iter_max = 1}
or a non-iterative procedure was used, the convergence status is set to \code{NULL}.}
\item{\code{$Iterations}}{The number of iterations used.}
}
}
\description{
Calculates weights according to on of the the five criteria
"\emph{SUMCORR}", "\emph{MAXVAR}", "\emph{SSQCORR}", "\emph{MINVAR}", and "\emph{GENVAR}"
suggested by \insertCite{Kettenring1971;textual}{cSEM}.
}
\details{
Some more description...
}
