% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators_weights.R
\name{calculateWeightsKettenring}
\alias{calculateWeightsKettenring}
\title{Calculate weights using one of Kettenrings's approaches}
\usage{
calculateWeightsKettenring(.data, .model, .approach)
}
\arguments{
\item{.data}{A \code{data.frame} or a \code{matrix} containing the raw data.}

\item{.model}{A model in \href{http://lavaan.ugent.be/tutorial/syntax1.html}{lavaan model syntax}
or a \link{cSEMModel} object.}
}
\value{
A list with the following elements:
\describe{
\item{\code{$W}}{A (J x K) matrix of estimated weights.}
\item{\code{$E}}{A (J x J) matrix of inner weights.}
\item{\code{$Modes}}{A named vector of Modes used for the outer estimation.}
}
}
\description{
Calculates weights according to on of the the five criteria
"\emph{SUMCORR}", "\emph{MAXVAR}", "\emph{SSQCORR}", "\emph{MINVAR}", and "\emph{GENVAR}"
suggested by \insertCite{Kettenring1971;textual}{cSEM}.
}
\details{
Some more description...
}
