% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators_weights.R
\name{calculateWeightsPLS}
\alias{calculateWeightsPLS}
\title{Calculate composite weights using PLS}
\usage{
calculateWeightsPLS(
  .S                           = args_default()$.S,
  .csem_model                  = args_default()$.csem_model,
  .conv_criterion              = args_default()$.conv_criterion,
  .iter_max                    = args_default()$.iter_max,
  .PLS_ignore_structural_model = args_default()$.PLS_ignore_structural_model
  .PLS_modes                   = args_default()$.PLS_modes,
  .PLS_weight_scheme_inner     = args_default()$.PLS_weight_scheme_inner,
  .tolerance                   = args_default()$.tolerance
   )
}
\arguments{
\item{.S}{The (K x K) empirical indicator correlation matrix.}

\item{.csem_model}{A (possibly incomplete) \link{cSEMModel}-list.}

\item{.conv_criterion}{Character string. The criterion to use for the convergence check.
One of: "\emph{diff_absolute}", "\emph{diff_squared}", or "\emph{diff_relative}". Defaults
to "\emph{diff_absolute}".}

\item{.iter_max}{Integer. The maximum number of iterations allowed.
If \code{iter_max = 1} and \code{.approach_weights = "PLS"} one-step weights are returned.
If the algorithm exceeds the specified number, weights of iteration step
\code{.iter_max - 1}  will be returned with a warning. Defaults to \code{100}.}

\item{.PLS_ignore_structural_model}{Logical. Should the structural model be ignored
when calculating the inner weights of the PLS algorithm? Defaults to \code{FALSE}.
Ignored if \code{.approach_weights} is not PLS.}

\item{.PLS_modes}{Either a named vector specifying the mode that should be used for
each construct in the form \code{"name" = "mode"}, a single character
string giving the mode that should be used for all constructs, or \code{NULL}.
Possible choices are: "\emph{ModeA}" or "\emph{ModeB}". Defaults to \code{NULL}.
If \code{NULL}, \code{csem()} will choose the appropriate mode according to the type
of construct used. Ignored if \code{.approach_weight} is not PLS.}

\item{.PLS_weight_scheme_inner}{Character string. The inner weighting scheme
used in PLS. One of: "\emph{centroid}", "\emph{factorial}", or "\emph{path}".
Defaults to "\emph{path}". Ignored if \code{.approach_weight} is not PLS.}

\item{.tolerance}{Double. The tolerance criterion for convergence.
Defaults to \code{1e-05}.}
}
\value{
A list with the elements
\describe{
\item{\code{$W}}{A (J x K) matrix of estimated weights.}
\item{\code{$E}}{A (J x J) matrix of inner weights.}
\item{\code{$Modes}}{A named vector of Modes used for the outer estimation.}
\item{\code{$Conv_status}}{The convergence status. \code{TRUE} if the algorithm has converged
and \code{FASLE} otherwise. If one-step weights are used via \code{.iter_max = 1}
or a non-iterative procedure was used, the convergence status is set to \code{NULL}.}
\item{\code{$Iterations}}{The number of iterations used.}
}
}
\description{
Calculates composite weights using the PLS algorithm. (TODO)
}
\details{
More details here. (TODO)
}
