% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators_weights.R
\name{calculateWeightsUnit}
\alias{calculateWeightsUnit}
\title{Calculate composite weights using unit weights}
\usage{
calculateWeightsUnit(
 .S                 = args_default()$.S,
 .csem_model        = args_default()$.csem_model
  )
}
\arguments{

\item{.S}{The (K x K) empirical indicator correlation matrix.}

\item{.csem_model}{A (possibly incomplete) \link{cSEMModel}-list.}
}
\value{
A list with the elements
\describe{
\item{\code{$W}}{A (J x K) matrix of estimated weights.}
\item{\code{$E}}{A (J x J) matrix of inner weights.}
\item{\code{$Modes}}{A named vector of Modes used for the outer estimation.}
\item{\code{$Conv_status}}{The convergence status. One of \code{TRUE}or \code{FASLE}. If
one-step weights are used via \code{.iter_max = 1} or no iterative algorithm was used,
the convergence status is set to \code{NULL}.}
\item{\code{$Iterations}}{The number of iterations used.}
}
}
\description{
Calculate unit weights for all blocks, i.e., each indicator of a block is
equally weighted.
}
