% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators_weights.R
\name{calculateWeightsUnit}
\alias{calculateWeightsUnit}
\title{Calculate unit weights for all blocks, i.e., each indicator of a block is equally weighted
This approach might be useful for Kroon correction.}
\usage{
calculateWeightsUnit(.data, .model)
}
\arguments{
\item{.data}{A \code{data.frame} or a \code{matrix} containing the raw data.}

\item{.model}{A model in \href{http://lavaan.ugent.be/tutorial/syntax1.html}{lavaan model syntax}
or a \link{cSEMModel} object.}
}
\value{
A list with the following elements:
\describe{
\item{\code{$W}}{A (J x K) matrix of estimated weights.}
\item{\code{$E}}{A (J x J) matrix of inner weights.}
\item{\code{$Modes}}{A named vector of Modes used for the outer estimation.}
}
}
\description{
Calculates weights...
}
\details{
Some more description...
}
