% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_doTrees.R
\name{coef.igsca_tree}
\alias{coef.igsca_tree}
\title{Get coefficients of tree object}
\usage{
\method{coef}{igsca_tree}(object, node = NULL, drop = TRUE, ...)
}
\arguments{
\item{...}{}
}
\value{
List of Vectors
}
\description{
Get coefficients of tree object
}
