% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_new.R
\name{CR}
\alias{CR}
\alias{CR.cSEMResults_default}
\alias{CR.cSEMResults_multi}
\alias{CR.cSEMResults_2ndorder}
\title{CR}
\usage{
CR(
 .object              = args_default()$.object,
 .only_common_factors = args_default()$.only_common_factors
)

\method{CR}{cSEMResults_default}(.object = args_default()$.object,
  .only_common_factors = args_default()$.only_common_factors)

\method{CR}{cSEMResults_multi}(.object = args_default()$.object,
  .only_common_factors = args_default()$.only_common_factors)

\method{CR}{cSEMResults_2ndorder}(.object = args_default()$.object,
  .only_common_factors = args_default()$.only_common_factors)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}

\item{.only_common_factors}{Logical. Should only common factors be included?
Defaults to \code{FALSE}.}
}
\description{
Computes the composite reliability \insertCite{Raykov1997}{cSEM} based on standardized loading.
}
\section{Methods (by class)}{
\itemize{
\item \code{cSEMResults_default}: (TODO)

\item \code{cSEMResults_multi}: (TODO)

\item \code{cSEMResults_2ndorder}: (TODO)
}}

\examples{
\dontrun{
# still to implement
}

}
\references{
\insertAllCited{}
}
\seealso{
\link{csem}, \link{cSEMResults}
}
