% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_new.R
\name{Cronbach_alpha}
\alias{Cronbach_alpha}
\alias{Cronbach_alpha.cSEMResults_default}
\alias{Cronbach_alpha.cSEMResults_multi}
\alias{Cronbach_alpha.cSEMResults_2ndorder}
\title{Cronbach_alpha}
\usage{
Cronbach_alpha(
 .object              = args_default()$.object,
 .only_common_factors = args_default()$.only_common_factors
)


  \method{Cronbach_alpha}{cSEMResults_default}(.object = args_default()$.object,
  .only_common_factors = args_default()$.only_common_factors,
  .alpha = args_default()$.alpha)


  \method{Cronbach_alpha}{cSEMResults_multi}(.object = args_default()$.object,
  .only_common_factors = args_default()$.only_common_factors,
  .alpha = args_default()$.alpha)


  \method{Cronbach_alpha}{cSEMResults_2ndorder}(.object = args_default()$.object,
  .only_common_factors = args_default()$.only_common_factors)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}

\item{.only_common_factors}{Logical. Should only constructs modeled as common
factors be included? Defaults to \code{TRUE}.}

\item{.alpha}{An integer or a numeric vector of significance levels.
Defaults to \code{0.05}.}
}
\description{
Computes Cronbach's alpha \insertCite{Cronbach1951}{cSEM} based on the correlation matrix
}
\section{Methods (by class)}{
\itemize{
\item \code{cSEMResults_default}: (TODO)

\item \code{cSEMResults_multi}: (TODO)

\item \code{cSEMResults_2ndorder}: (TODO)
}}

\examples{
\dontrun{
# still to implement
}

}
\references{
\insertAllCited{}
}
\seealso{
\link{csem}, \link{cSEMResults}
}
