% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_package.R
\docType{package}
\name{cSEM-package}
\alias{cSEM}
\alias{cSEM-package}
\title{cSEM: A package for composite based structural equation modeling}
\description{
The package provides functions to analyze structural equation models using composite based approaches.
}
\section{General}{

Some comments on the package structure.
}

\section{The cca function}{

Some text on the cca function
}

\section{The csem function}{

Some text on the csem function
}

\section{The foreman function}{

Some text on the workhorse function
}

\author{
\strong{Maintainer}: Manuel Rademaker \email{manuel.rademaker@uni-wuerzburg.de} (0000-0002-8902-3561)

Authors:
\itemize{
  \item Florian Schuberth \email{f.schuberth@utwente.nl} (0000-0002-2110-9086)
}

Other contributors:
\itemize{
  \item Tamara Schamberger \email{tamara.schamberger@uni-wuerzburg.de} [contributor]
}

}
\keyword{internal}
