% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_package.R
\docType{package}
\name{cSEM-package}
\alias{cSEM}
\alias{cSEM-package}
\title{cSEM: A package for composite-based structural equation modeling}
\description{
Estimate, analyse, test, and study linear, nonlinear, hierachical and
multigroup structural equation models using composite-based approaches and procedures including estimation
techniques such as partial least squares path modelling (PLS) and its derivatives
(PLSc, ordPLSc, robustPLSc), generalized structured component analysis (GSCA),
generalized structured component analyis with uniqueness terms (GSCAm),
generalized canonical correlation analysis (GCCA) unit weights (sum score) and
fixed weights, as well as several tests and typical postestimation
procedures (e.g., assess the model fit, compute direct, indirect and total effects).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/M-E-Rademaker/cSEM}
  \item \url{https://m-e-rademaker.github.io/cSEM/}
  \item Report bugs at \url{https://github.com/M-E-Rademaker/cSEM/issues}
}

}
\author{
\strong{Maintainer}: Manuel Rademaker \email{manuel.rademaker@uni-wuerzburg.de} (0000-0002-8902-3561)

Authors:
\itemize{
  \item Florian Schuberth \email{f.schuberth@utwente.nl} (0000-0002-2110-9086)
}

Other contributors:
\itemize{
  \item Tamara Schamberger \email{tamara.schamberger@uni-wuerzburg.de} [contributor]
  \item Michael Klesel [contributor]
  \item Theo K. Dijkstra [contributor]
}

}
\keyword{internal}
