% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_csem.R
\name{csem}
\alias{csem}
\title{Composite-based SEM}
\usage{
csem(
  .data                        = NULL,
  .model                       = NULL,
  .approach_nl                 = c("sequential", "replace"),
  .approach_paths              = c("OLS", "2SLS"),
  .approach_weights            = c("PLS-PM", "SUMCORR", "MAXVAR", "SSQCORR", "MINVAR", "GENVAR", 
                                   "GSCA", "PCA", "unit", "bartlett", "regression"),
  .disattenuate                = TRUE,
  .id                          = NULL,
  .normality                   = TRUE,
  .reliabilities               = NULL,
  .resample_method             = c("none", "bootstrap", "jackknife"),
  .resample_method2            = c("none", "bootstrap", "jackknife"),
  .R                           = 499,
  .R2                          = 199,
  .handle_inadmissibles        = c("drop", "ignore", "replace"),
  .user_funs                   = NULL,
  .eval_plan                   = c("sequential", "multiprocess"),
  .seed                        = NULL,
  ...
  )
}
\arguments{
\item{.data}{A \code{data.frame} or a \code{matrix} of standardized or unstandarized
data (indicators/items/manifest variables).
Additionally, a \code{list} of \code{data.frame}(s) or \code{matrice}(s) is accepted in which
case estimation is repeated for each data set. Possible column types or classes
of the data provided are: "logical", "numeric" ("double" or "integer"),
"factor" ("ordered" and/or "unordered") or a mix of several types.
The data may also include \emph{one} character column whose column name must be
given to \code{.id}. This column is assumed to contain group identifiers used to split
the data into groups.}

\item{.model}{A model in \link[lavaan:model.syntax]{lavaan model syntax}
or a \link{cSEMModel} list.}

\item{.approach_2ndorder}{Character string. Approach used for models containing
second order constructs. One of: "\emph{3stage}" or "\emph{repeated_indicators}".
Defaults to "\emph{3stage}".}

\item{.approach_nl}{Character string. Approach used to estimate nonlinear
structural relationships. One of: "\emph{sequential}" or "\emph{replace}".
Defaults to "\emph{sequential}".}

\item{.approach_paths}{Character string. Approach used to estimate the
structural coefficients. One of: "\emph{OLS}" or "\emph{2SLS}" (not yet implemented).
Defaults to "\emph{OLS}".}

\item{.approach_weights}{Character string. Approach used to
obtain composite weights. One of: "\emph{PLS-PM}", "\emph{SUMCORR}", "\emph{MAXVAR}",
"\emph{SSQCORR}", "\emph{MINVAR}", "\emph{GENVAR}", "\emph{GSCA}", "\emph{PCA}", "\emph{unit}", "\emph{bartlett}",
or "\emph{regression}".
Defaults to "\emph{PLS-PM}".}

\item{.disattenuate}{Logical. Should composite/proxy correlations
be disattenuated to yield consisten loadings and path estimates if at least
one of the construct is modeled as a common factor? Defaults to \code{TRUE}.}

\item{.id}{Character string or integer. The name or position of the column of
\code{.data} used to split the data into groups.
Defaults to \code{NULL}.}

\item{.normality}{Logical. Should joint normality of
\eqn{[\eta_{1:p}; \zeta; \epsilon]}{[\eta_(1:p); \zeta; \epsilon]}
be assumed in the nonlinear model? See \insertCite{Dijkstra2014}{cSEM} for details.
Defaults to \code{TRUE}. Ignored if the model is linear.}

\item{.reliabilities}{A character vector of \code{"name" = value} pairs,
where \code{value} is a number between 0 and 1 and \code{"name"} a character string
of the corresponding construct name, or \code{NULL}. Reliabilities
may be given for a subset of the constructs. Defaults to \code{NULL} in which case
reliabilities are estimated by \code{csem()}. Currently, only supported for
\code{.approach_weights = "PLS-PM"}.}

\item{.resample_method}{Character string. The resampling method to use. One of:
"\emph{bootstrap}" or "\emph{jackknife}". Defaults to "\emph{bootstrap}".}

\item{.resample_method2}{Character string. The resampling method to use when resampling
from a resample. One of: "\emph{none}", "\emph{bootstrap}" or "\emph{jackknife}". For
"\emph{bootstrap}" the number of draws is provided via \code{.R2}. Currently,
resampling from each resample is only required for the studentized confidence
intervall computed by the \code{\link[=infer]{infer()}} function. Defaults to "\emph{none}".}

\item{.R}{Integer. The number of bootstrap replications. Defaults to \code{499}.}

\item{.R2}{Integer. The number of bootstrap replications to use when
resampling from a resample. Defaults to \code{199}.}

\item{.handle_inadmissibles}{Character string. How should inadmissible results
be treated? One of "\emph{drop}", "\emph{ignore}", or "\emph{replace}". If "\emph{drop}", all
replications/resamples yielding an inadmissible result will be dropped
(i.e. the number of results returned will be less than .R). For "\emph{ignore}"
all results are returned even if they are inadmissible (i.e.
number of results returned = .R). For "\emph{replace}" resampling continues until
there are exactly .R admissible solutions.
Defaults to "\emph{drop}".}

\item{.user_funs}{A function or a (named) list of functions to apply to every
resample. Takes \code{.object} as an input (e.g., \code{myFun <- function(.object) {...}}).
Output should preferably be a (named)
vector but matrices are also accepted. However, the output will be
vectorized (columnwise) in this case. See the examples section for details.}

\item{.eval_plan}{Character string. The evaluation plan to use. One of
"\emph{sequential}" or "\emph{multiprocess}". In the latter case
all available cores will be used. Defaults to "\emph{sequential}".}

\item{.seed}{Integer. The random seed to use. Defaults to \code{NULL} in which
case an arbitrary seed is choosen.}

\item{...}{Further arguments to be passed down to lower level functions of \code{csem()}.
See \link{csem_dotdotdot} for a complete list of available arguments.}
}
\value{
An object of class \code{cSEMResults} with methods for all postestimation generics.
Note that, technically, a call to \code{\link[=csem]{csem()}} results in an object with at least
two class attributes. The first class attribute is always \code{cSEMResults}.
The second is one of \code{cSEMResults_default}, \code{cSEMResults_multi}, or
\code{cSEMResults_2ndorder} and depends on the estimated model and/or the type of
data provided to the \code{.model} and \code{.data} arguments. The third class attribute
\code{cSEMResults_resampled} is only added if resampling was conducted.
Technically, method dispatch for all postestimation
functions is based on the second class attribute. For a details see the
\link[=cSEMResults]{cSEMResults helpfile }.
}
\description{
Estimate linear, nonlinear, hierachical or multigroup structural equation
models using a composite-based approach. In \pkg{cSEM}
any method or approach that involves linear compounts (scores/proxies/composites)
of observables (indicators/items/manifest variables) is defined as composite-based.
In \pkg{cSEM} linear compounts are usually refered to as proxies and observables as
indicators.
}
\details{
\code{csem()} estimates linear, nonlinear, hierachical or multigroup structural
equation models using a composite-based approach.

In order to use the \pkg{cSEM} package effectivly users should internalize the
following fundamental principles which form the basis of the package and
its terminolgy.
\enumerate{
\item{The approach to obtain scores (i.e. how weights that form a score are build/estimated)
and the approach to obtain estimates of the structural/path coefficients based on these
scores (e.g. by ordinary least squares ("OLS") or two-stage least squares ("2SLS"))
are strictly distinct. This conceptual separation is reflected by the arguments
\code{.approach_weights} and \code{.approach_path} which allow users to combine any
given weight approach with any path approach.}
\item{The research objective or question (explanatory vs. confirmatory and/or
prediction vs. causal inference) defines the appropriate model. }
\item{How a certain concept/construct is modeled is strictly distguished
from the way quantities related to this construct (e.g weights, scores, loadings,
path coefficients) are estimated. Technically,  \strong{any} valid
model can be estimated using \strong{any} weightning approach, however,
it is the choice of the model that definies the appropriateness of the
method and potential steps that are potentially necessary to yield statistically
sound results (e.g. consistency of the structural estiamtes
if constructs are modeled as common factors when the objective is confirmatory).
In \pkg{cSEM} constructs may be modeled as common factors or
composites.
}
}
Currently, this includes
partial least squares path modeling (PLS-PM) and
its derivatives (PLSc, ordPLSc, robustPLSc), generalized structured
component analysis (GSCA), generalized structured component analysis with
uniqueness terms (GSCAm), generalized canonical correlation analysis (GCCA),
principal component analysis (PCA), factor score regression (FSR) using sum
scores, regression or bartlett scores. If constructs are modeled as common
factors, proxy correlations, loading and path estimates are disattenuated  using \insertCite{Croon2002;textual}{cSEM}'s
for scores obtained by GCCA, PCA, or FSR. \code{csem()} is designed for quick and flexible
use by providing the user with default options except for
the mandatory \code{.data} and \code{.model} argument.

\subsection{Data and model:}{
The \code{.data} and \code{.model} arguments are required. Data must be
provided as either a \code{matrix} or a \code{data.frame} with column names matching
the indicator names used in the model description of the measurement model.
Alternatively, a list of matrices or \code{data.frame}s may be provided
in which case estimation is repeated for each data set.
The data provided via \code{.data} may contain \strong{one} character column whose column name
must be provided to \code{.id}. Values of this column are interpreted as group
identifiers and \code{csem()} will split the data by levels of that column and run
the estimation for each level separately.

To provide a model use the \link[lavaan:model.syntax]{lavaan model syntax}
with two notable extensions/changes. First: the "\code{<~}" operator in \code{cSEM} is
used to define a composite instead of a causal formative common factor. Second:
the "\code{.}" is used to indicate interactions between constructs as in e.g.,
\code{construct1.construct2}. Alternatively a standardized (possibly incomplete)
\link{cSEMModel}-list may be supplied.
}

\subsection{Weights and path coefficients:}{
By default weights are estimated using the partial least squares (path) algorithm (\code{"PLS-PM"}).
A broad range of alternative weightning algorithms may be supplied to \code{.appraoch_weights}.
Currently the following approaches are implemented
\enumerate{
\item{(Default) Partial least squares path modeling (\code{"PLS"}). The algorithm
can be customized. See \code{\link[=calculateWeightsPLS]{calculateWeightsPLS()}} for details.}
\item{Generalized structured component analysis (\code{"GSCA"})}
\item{Generalized canoncial correlation analysis (\emph{GCCA}), including
\code{"SUMCORR"}, \code{"MAXVAR"}, \code{"SSQCORR"}, \code{"MINVAR"}, \code{"GENVAR"}}
\item{Principal component analysis (\code{"PCA"})}
\item{Factor score regression using sum scores (\code{"unit"}),
regression (\code{"regression"}) or bartlett scores (\code{"bartlett"})}
}

Composite-indicator and composite-composite correlations are properly
disattenuated by default to yield consistent loadings, construct correlations,
and path coefficients if any of the constructs in the model are modeled as a
common factor.

For \emph{PLS-PM} disattenuation is done using \emph{PLSc} \insertCite{Dijkstra2015}{cSEM}.
For \emph{GSCA} disattenuation is done implicitly by using \emph{GSCAm}. Weights obtained
by GCCA, unit weights and fixed weights are disattenuated using a
Disattenuation my be suppressed by setting
\code{.disattenuate = FALSE}. Note, however, that quantities in this case are inconsistent
estimates for their construct level counterparts if any of the constructs in the structural model is
modeled as a common factor!
}

\subsection{Nonlinear models:}{
If the model is nonlinear \code{csem()} estimates a polynomial structural equation model
using a non-iterative method of moments approach described in
\insertCite{Dijkstra2014}{cSEM}. Non linear terms include interactions and
exponential terms. The latter is described in model syntax as an
"interaction with itself", e.g., \code{x_1^3 = x1.x1.x1}. Currently only exponential
terms up to a power of three (i.e. three-way interactions) are allowed.

The current version of the package allows two kinds of estimation:
estimation of the reduced form equation (\code{.approach_nl = "reduced"}) and
sequential estimation (\code{.approach_nl = "sequential"}). The latter does not
not allow for multivariate normality of all exogenous variables, i.e.,
the latent variables and the error terms.

Distributional assumptions are kept to a minimum (an i.i.d. sample from a
population with finite moments for the relevant order); for higher order models,
that go beyond interaction, we work in this version with the assumption that
as far as the relevant moments are concerned certain combinations of
measurement errors behave as if they were Gaussian.
For details see: \insertCite{Dijkstra2014;textual}{cSEM}.
}

\subsection{Inference:}{
Inference is done via resampling. See \link{resamplecSEMResults} for details.
}
}
\examples{

}
\references{
\insertAllCited{}
}
\seealso{
\link{args_default}, \link{cSEMArguments}, \link{cSEMResults}, \link{foreman}
}
