% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_csem.R
\name{csem}
\alias{csem}
\title{Composite based SEM}
\usage{
csem(
  .data                = NULL,
  .model               = NULL,
  .approach_weights    = c("PLS", "SUMCOR", "MAXVAR", "SSQCOR", "MINVAR", "GENVAR", "GSCA", "fixed", "unit"),
  .approach_path       = c("OLS", "2SLS", "3SLS"),
  .approach_nl         = c("none", "replace"),
  .disattenuate        = TRUE,
  .PLS_weight_scheme_inner = c("centroid", "factorial", "path"),
  .PLS_mode            = NULL,
  .estimate_structural = TRUE,
  .reliabilities       = NULL
  ...)
}
\arguments{
\item{.data}{A data frame or a matrix containing the raw data.}

\item{.model}{A model in \href{http://lavaan.ugent.be/tutorial/syntax1.html}{lavaan model syntax}
or a \link{cSEMModel} object.}

\item{.id}{Charachter string. The name of the column of \code{.data} used to split
the date into groups.}

\item{.approach_weights}{Character string. The name of the approach used to
obtain composite weights. Possible choices are: "\emph{PLS}", "\emph{SUMCOR}", "\emph{MAXVAR}",
"\emph{SSQCOR}", "\emph{MINVAR}", "\emph{GENVAR}", "\emph{GSCA}", "\emph{fixed}", or "\emph{unit}".
Defaults to "\emph{PLS}".}

\item{.approach_paths}{Character string. Approach used to estimate the
structural coefficients. Possible choices are: "\emph{OLS}", "\emph{2SLS}", or "\emph{3SLS}".
Defaults to "\emph{OLS}".}

\item{.approach_nl}{Character string. The approach used to handle non-linear
structural relationships. Possible choices are: "\emph{none}", or "\emph{replace}".
Defaults to "\emph{none}".}

\item{.disattenuate}{Logical. Should proxy correlations be disattenuated
if the construct is modeled as a common factor? Defaults to \code{TRUE}.}

\item{.PLS_weight_scheme_inner}{Character string. The inner weighting scheme
used in PLS. Possible choices are: "\emph{centroid}", "\emph{factorial}", or "\emph{path}".
Defaults to "\emph{centroid}". The argument is ignored if \code{.approach_weight} is not PLS.}

\item{.PLS_mode}{Either a named vector specifying the mode that should be used for
each construct in the form \code{name = "mode"}, a single character
string giving the mode that should be used for all constructs or \code{NULL}.
Possible choices are: "\emph{ModeA}" or "\emph{ModeB}". Defaults to \code{NULL}.
If \code{NULL}, \code{cSEM} will choose the appropriate mode according to the type
of construct used. The argument is ignored if \code{.approach_weight} is not PLS.}

\item{.estimate_structural}{Logical. Should the structural (path) coefficients
be estimated? Defaults to \code{TRUE}.}

\item{.reliabilities}{A vector of \code{name = value} pairs of reliability coefficients.
Element names are equal to the names of the J construct names. Reliabilities
may be given for a subset of the constructs. Defaults to \code{NULL} in which case
reliabilites are estimated by \code{cSEM}.}

\item{...}{Further arguments to be passed down to other functions.
See \link{args_default} for a complete list of possible arguments and their defaults.}
}
\value{
An object of class \code{cSEMResults} for which the following methods exist:
\describe{
\item{\code{print.cSEMResults}}{Prints a message to inform the user
whether estimation has been successful and what functions may be used
to examine the object.}
\item{\code{summary.cSEMResults}}{Print and return a comprehensive summary of the results.}
}
Technically \code{cSEMResults} is a named list containing the following list elements:
\describe{
\item{\code{$Estimates}}{A list containing the estimated quantities.}
\item{\code{$Information}}{A list of additional information. (incomplete)}
}
}
\description{
Estimates linear and nonlinear structural equation models using a
composite based approach.
}
\details{
The \code{csem} function is a wrapper around the more general \link{workhorse} function.
It is designed for quick, easy, and flexible use by providing the user with
defaults choices for all relevant arguments except the mandatory \code{.data}
and \code{.model} argument.

To get started the \code{.data} and \code{.model} arguments are required. Data must be
provided as either a matrix or a data frame with column names matching
the variable/indicator names used in the model description of the measurement model.

To provide a model use the \href{http://lavaan.ugent.be/tutorial/syntax1.html}{lavaan model syntax}
with two notable extensions/changes. First: the "\code{<~}" operator in \code{cSEM} is
used to define a composite instead of a formative common factor. Second:
the "\code{.}" is used to indicate interactions between constructs as in e.g.,
\code{construct1.construct2}.

By default weights are estimated using \emph{PLS}. Alternative approaches include
n, "\emph{fixed weights}"
or "\emph{unit weight}". \emph{Generalized Structured Component Analysis} (\emph{GSCA}) may
also be chosen as a weighing approach although technically GSCA obtains weight
and structural coefficient estimates simultaneously. Hence, setting
\code{.approach_weights = "GSCA"} automatically sets \code{.approach_path = "GSCA"} (and
vice-versa).

The weights are properly rescaled to get consistent estimates for the factor
loadings as well as for the correlations between the proxies and their
corresponding factors unless \code{.disattenuate} is set to \code{FALSE}.
Consistent estimates are calculated for the entries of the moment equations
that define the structural parameters.
The solutions to the moment equations are reported as consistent estimates
for the structural parameters.

If the model is nonlinear \link{csem} estimates a polynomial structural equation model
using a non-iterative method of moments approach described in
\insertCite{Dijkstra2014}{cSEM}. Non linear terms include interactions and
exponential terms. The latter is described in model syntax as an
"interaction with itself", e.g., \code{x_1^3 = x1.x1.x1}. Currently only exponential
terms up to a power of three (i.e. three-way interactions) are allowed.

The current version of the package allows two kinds of estimation:
Estimation of the reduced form equation and estimation each equation by not
inserting the other equations. The latter does not not allow for assuming multivariate normality of all
exogenous variables, i.e., the latent variables and the error terms.

Distributional assumptions are kept to a minimum (an i.i.d. sample from a population with finite
moments for the relevant order); for higher order models, that go beyond interaction, we work in
this version with the assumption that as far as the relevant moments are concerned certain
combinations of measurement errors behave as if they were Gaussian.
}
\examples{
\dontrun{
# still to implement
}

}
\references{
\insertAllCited{}
}
\seealso{
\link{cca}, \link{workhorse}
}
