% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_arguments.R
\name{csem_arguments}
\alias{csem_arguments}
\alias{cSEMArguments}
\title{cSEMArguments}
\arguments{
\item{.alpha}{An integer or a numeric vector of significance levels.
Defaults to \code{0.05}.}

\item{.approach}{Character string. The Kettenring approach to use. One of
"\emph{SUMCORR}", "\emph{MAXVAR}", "\emph{SSQCORR}", "\emph{MINVAR}" or "\emph{GENVAR}". Defaults to
"\emph{SUMCORR}".}

\item{.approach_2ndorder}{Character string. Approach used for models containing
second order constructs. One of: "\emph{3stage}" or "\emph{repeated_indicators}".
Defaults to "\emph{3stage}".}

\item{.approach_cor_robust}{Character string. Approach used to obtain a robust
indicator correlation matrix. One of: "\emph{none}" in which case the standard
Bravais-Person correlation is used,
"\emph{spearman}" for the Spearman rank correlation, or
"\emph{mcd}" via \code{\link[MASS:cov.rob]{MASS::cov.rob()}} for a robust correlation matrix.
Defaults to "\emph{none}".}

\item{.approach_nl}{Character string. Approach used to estimate nonlinear
structural relationships. One of: "\emph{sequential}" or "\emph{replace}".
Defaults to "\emph{sequential}".}

\item{.approach_paths}{Character string. Approach used to estimate the
structural coefficients. One of: "\emph{OLS}" or "\emph{2SLS}" (not yet implemented).
Defaults to "\emph{OLS}".}

\item{.approach_weights}{Character string. Approach used to
obtain composite weights. One of: "\emph{PLS-PM}", "\emph{SUMCORR}", "\emph{MAXVAR}",
"\emph{SSQCORR}", "\emph{MINVAR}", "\emph{GENVAR}", "\emph{GSCA}", "\emph{PCA}", "\emph{unit}", "\emph{bartlett}",
or "\emph{regression}".
Defaults to "\emph{PLS-PM}".}

\item{.args_used}{A list of function argument names to \code{fun()} whose value
was modified by the user.}

\item{.bias_corrected}{Logical. Should the standard and the tStat
confidence intervall be bias-corrected using the bootstraped bias estimate?
If \code{TRUE} the confidence intervall for some estimated parameter \code{theta}
is centered at \code{2*theta - theta*_hat},
where \code{theta*_hat} is the average over all .R bootstrap estimates of \code{theta}.
Defaults to \code{TRUE}}

\item{.C}{A (J x J) composite variance-covariance matrix.}

\item{.choices}{Logical. Should candidate values for the arguments be returned?
Defaults to \code{FALSE}.}

\item{.ci}{A vector of character strings naming the confidence interval to compute.
For possible choices see \code{\link[=infer]{infer()}}.}

\item{.closed_form_ci}{Logical. Should a closed-form confidence interval be computed?
Defaults to \code{FALSE}.}

\item{.conv_criterion}{Character string. The criterion to use for the convergence check.
One of: "\emph{diff_absolute}", "\emph{diff_squared}", or "\emph{diff_relative}". Defaults
to "\emph{diff_absolute}".}

\item{.csem_model}{A (possibly incomplete) \link{cSEMModel}-list.}

\item{.csem_resample}{A list resulting from a call to \code{\link[=resamplecSEMResults]{resamplecSEMResults()}}.}

\item{.cv_folds}{Integer. The number of cross-validation folds to use. Setting
\code{.cv_folds} to \code{N} (the number of observations) produces
leave-one-out cross-validation samples. Defaults to \code{10}.}

\item{.data}{A \code{data.frame} or a \code{matrix} of standardized or unstandarized data.
Possible column types or classes of the data provided are: logical,
numeric (double or integer), factor (ordered and unordered)
or a mix of several types. The data may also include
\emph{one} character column whose column name must be given to \code{.id}.
This column is assumed to contain group identifiers used to split
the data into groups.}

\item{.disattenuate}{Logical. Should composite/proxy correlations
be disattenuated to yield consisten loadings and path estimates if at least
one of the construct is modeled as a common factor? Defaults to \code{TRUE}.}

\item{.dist}{Character string. The distribution to use for the critical value.
One of \emph{"t"} for Student's t-distribution or \emph{"z"} for the standard normal distribution.
Defaults to \emph{"z"}.}

\item{.distance}{Character string. A distance measure. One of: "\emph{geodesic}"
or "\emph{squared_euclidian}". Defaults to "\emph{geodesic}".}

\item{.df}{Character string. The method for obtaining the degrees of freedom.
Choices are "\emph{type1}" and "\emph{type2}". Defaults to "\emph{type1}" .}

\item{.dominant_indicators}{A character vector of \code{"name" = "value"} pairs,
where \code{"value"} is a character string giving the name of the dominant indicator
and \code{"name"} a character string of the corresponding construct name.
Dominant indicators may be specified for a subset of the constructs.}

\item{.E}{A (J x J) matrix of inner weights.}

\item{.estimate_structural}{Logical. Should the structural coefficients
be estimated? Defaults to \code{TRUE}.}

\item{.eval_plan}{Character string. The evaluation plan to use. One of
"\emph{sequential}" or "\emph{multiprocess}". In the latter case
all available cores will be used. Defaults to "\emph{sequential}".}

\item{.first_resample}{A list containing the \code{.R} resamples based on the original
data obtained by resamplecSEMResults()}

\item{.second_resample}{A list containing \code{.R2} resamples for each of the \code{.R}
resamples of the first run.}

\item{.H}{The (N x J) matrix of construct scores.}

\item{.handle_inadmissibles}{Character string. How should inadmissible results
be treated? One of "\emph{drop}", "\emph{ignore}", or "\emph{replace}". If "\emph{drop}", all
replications/resamples yielding an inadmissible result will be dropped
(i.e. the number of results returned will be less than .R). For "\emph{ignore}"
all results are returned even if they are inadmissible (i.e.
number of results returned = .R). For "\emph{replace}" resampling continues until
there are exactly .R admissible solutions.
Defaults to "\emph{drop}".}

\item{.id}{Character string or integer. The name or position of the column of
\code{.data} used to split the data into groups.
Defaults to \code{NULL}.}

\item{.iter_max}{Integer. The maximum number of iterations allowed.
If \code{iter_max = 1} and \code{.approach_weights = "PLS-PM"} one-step weights are returned.
If the algorithm exceeds the specified number, weights of iteration step
\code{.iter_max - 1}  will be returned with a warning. Defaults to \code{100}.}

\item{.n}{Integer. The number of observations of the original data.}

\item{.matrix1}{A \code{matrix} to compare.}

\item{.matrix2}{A \code{matrix} to compare.}

\item{.matrices}{A list of at least two matrices.}

\item{.model}{A model in \link[lavaan:model.syntax]{lavaan model syntax}
or a \link{cSEMModel} list.}

\item{.modes}{A vector giving the mode for each construct in the form \code{"name" = "mode"}.
Only used internally.}

\item{.normality}{Logical. Should joint normality of
\eqn{[\eta_{1:p}; \zeta; \epsilon]}{[\eta_(1:p); \zeta; \epsilon]}
be assumed in the nonlinear model? See \insertCite{Dijkstra2014}{cSEM} for details.
Defaults to \code{TRUE}. Ignored if the model is linear.}

\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}

\item{.only_common_factors}{Logical. Should only constructs modeled as common
factors be included? Defaults to \code{TRUE}.}

\item{.P}{A (J x J) construct variance-covariance matrix (possibly disattenuated).}

\item{.PLS_approach_cf}{Character string. Approach used to obtain the correction
factors for PLSc. One of: "\emph{dist_squared_euclid}", "\emph{dist_euclid_weighted}",
"\emph{fisher_transformed}", "\emph{mean_arithmetic}", "\emph{mean_geometric}", "\emph{mean_harmonic}",
"\emph{geo_of_harmonic}". Defaults to "\emph{dist_squared_euclid}".
Ignored if \code{.disattenuate = FALSE} or if \code{.approach_weights} is not PLS-PM.}

\item{.PLS_ignore_structural_model}{Logical. Should the structural model be ignored
when calculating the inner weights of the PLS-PM algorithm? Defaults to \code{FALSE}.
Ignored if \code{.approach_weights} is not PLS-PM.}

\item{.PLS_modes}{Either a named list specifying the mode that should be used for
each construct in the form \code{"name" = "mode"}, a single character
string giving the mode that should be used for all constructs, or \code{NULL}.
Possible choices for \code{"mode"} are: "\emph{modeA}", "\emph{modeB}", "\emph{modeBNNLS}",
"\emph{unit}", "\emph{PCA}", a single integer or
a vector of fixed weights of the same length as there are indicators for the
construct given by \code{"name"}. If only a single number is provided this is identical to
using unit weights, as weights are rescaled such that the related composite
has unit variance.  Defaults to \code{NULL}.
If \code{NULL} the appropriate mode according to the type
of construct used is choosen. Ignored if \code{.approach_weight} is not PLS-PM.}

\item{.PLS_weight_scheme_inner}{Character string. The inner weighting scheme
used in PLS-PM. One of: "\emph{centroid}", "\emph{factorial}", or "\emph{path}".
Defaults to "\emph{path}". Ignored if \code{.approach_weight} is not PLS-PM.}

\item{.probs}{A vector of probabilities.}

\item{.quality_criterion}{Character string. A single character string or a
vector of character strings naming the quality criterion to compute. See
the details section for a list of possible candidates. Defaults to "\emph{all}".}

\item{.quantity}{Character string. Which statistic should be returned?
One of (TODO)
Defaults to (TODO).}

\item{.Q}{A vector of composite-construct correlations with element names equal to
the names of the J construct names used in the measurement model. Note
Q^2 is also called the reliability coefficient.}

\item{.reliabilities}{A character vector of \code{"name" = value} pairs,
where \code{value} is a number between 0 and 1 and \code{"name"} a character string
of the corresponding construct name, or \code{NULL}. Reliabilities
may be given for a subset of the constructs. Defaults to \code{NULL} in which case
reliabilities are estimated by \code{csem()}. Currently, only supported for
\code{.approach_weights = "PLS-PM"}.}

\item{.resample_method}{Character string. The resampling method to use. One of:
"\emph{bootstrap}" or "\emph{jackknife}". Defaults to "\emph{bootstrap}".}

\item{.resample_method2}{Character string. The resampling method to use when resampling
from a resample. One of: "\emph{none}", "\emph{bootstrap}" or "\emph{jackknife}". For
"\emph{bootstrap}" the number of draws is provided via \code{.R2}. Currently,
resampling from each resample is only required for the studentized confidence
intervall computed by the \code{\link[=infer]{infer()}} function. Defaults to "\emph{none}".}

\item{`.resample_object`}{An R object of class \code{cSEMResults_resampled}
obtained from \code{\link[=resamplecSEMResults]{resamplecSEMResults()}} or by setting \code{.resample_method = "bootstrap"}
or \code{"jackknife"} when calling \code{\link[=csem]{csem()}}.}

\item{.R}{Integer. The number of bootstrap replications. Defaults to \code{499}.}

\item{.R2}{Integer. The number of bootstrap replications to use when
resampling from a resample. Defaults to \code{199}.}

\item{.S}{The (K x K) empirical indicator correlation matrix.}

\item{.saturated}{Logical. Should a saturated structural model be used?
Defaults to \code{FALSE}.}

\item{.sign_change_option}{Character string. Which sign change option should
be used to handle flipping signs when resampling? One of "\emph{none}","\emph{individual}",
"\emph{individual_reestimate}", "\emph{construct_reestimate}". Defaults to "\emph{none}".}

\item{.seed}{Integer or FALSE. The random seed to use. Defaults to \code{FALSE} in which
case an arbitrary seed is choosen.}

\item{.stage}{Character string. The stage the model is need for.
One of "\emph{first}" or "\emph{second}". Defaults to "\emph{first}".}

\item{.starting_values}{list of named vectors containig the starting values.
Defaults to \code{NULL}.}

\item{.terms}{A vector of construct names to be classified.}

\item{.tolerance}{Double. The tolerance criterion for convergence.
Defaults to \code{1e-05}.}

\item{.type_vcv}{Character string. Indicates which model-implied correlation matrix is calcuted
One of "\emph{indicator}" or "\emph{construct}". Defaults to "\emph{indicator}".}

\item{.verbose}{Logical. Should information be printed to the console? Defaults
to \code{TRUE}.}

\item{.user_funs}{A function or a (named) list of functions to apply to every
resample. Takes \code{.object} as an input (e.g., \code{myFun <- function(.object) {...}}).
Output should preferably be a (named)
vector but matrices are also accepted. However, the output will be
vectorized (columnwise) in this case. See the examples section for details.}

\item{.vector1}{A vector of numeric values.}

\item{.vector2}{A vector of numeric values.}

\item{.W}{A (J x K) matrix of weights.}

\item{.W_new}{A (J x K) matrix of weights.}

\item{.W_old}{A (J x K) matrix of weights.}

\item{.X}{A matrix of processed data (scaled, cleaned and ordered).}

\item{.X_cleaned}{A data.frame of processed data (cleaned and ordered). Note: \code{X_cleaned}
may not be scaled!}
}
\description{
An alphabetical list of all arguments used by functions of the \code{cSEM} package,
including their description and defaults.
Mainly used for internal purposes (parameter inheritance). To list all arguments
and their defaults, use \code{\link[=args_default]{args_default()}}. To list all arguments and
their possible choices, use \code{args_default(.choices = TRUE)}.
}
\keyword{internal}
