% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_arguments.R
\name{csem_arguments}
\alias{csem_arguments}
\alias{cSEMArguments}
\alias{csem_parameters}
\title{cSEMArguments}
\arguments{
\item{.data}{A data frame or a matrix containing the raw data.}

\item{.model}{A model in \href{http://lavaan.ugent.be/tutorial/syntax1.html}{lavaan model syntax}
or a \link{cSEMModel} object.}

\item{.alpha}{A numeric vector of significance levels. Defaults to \code{c(0.1, 0.05, 0.01)}.}

\item{.approach_cf}{Character string. The approach to obtain the correction
factor for PLSc. Possible choices are: "\emph{dist_euclid}", "\emph{dist_euclid_weighted}",
"\emph{fisher_transformed}", "\emph{mean_arithmetic}", "\emph{mean_geometric}", "\emph{mean_harmonic}",
"\emph{geo_of_harmonic}". Defaults to "\emph{dist_euclid}". Ignored if \code{.disattenuate = FALSE}.}

\item{.approach_nl}{Character string. The approach used to handle non-linear
structural relationships. Possible choices are: "\emph{none}", or "\emph{replace}".
Defaults to "\emph{none}".}

\item{.approach_paths}{Character string. Approach used to estimate the
structural coefficients. Possible choices are: "\emph{OLS}", "\emph{2SLS}", or "\emph{3SLS}".
Defaults to "\emph{OLS}".}

\item{.approach_weights}{Character string. The name of the approach used to
obtain composite weights. Possible choices are: "\emph{PLS}", "\emph{SUMCOR}", "\emph{MAXVAR}",
"\emph{SSQCOR}", "\emph{MINVAR}", "\emph{GENVAR}", "\emph{GSCA}", "\emph{fixed}", or "\emph{unit}".
Defaults to "\emph{PLS}".}

\item{.args_used}{A list of argument names and user picked values. Usually captured by
\code{as.list(match_call())}.}

\item{.C}{A (J x J) proxy variance-covariance matrix.}

\item{.criteria}{Character string. The criteria to use. One of "\emph{SUMCOR}", "\emph{MAXVAR}",
"\emph{SSQCOR}", "\emph{MINVAR}" or "\emph{GENVAR}".}

\item{.csem_model}{A (possibly incomplete) \link{cSEMModel} list.}

\item{.disattenuate}{Logical. Should proxy correlations be disattenuated
if the construct is modeled as a common factor? Defaults to \code{TRUE}.}

\item{.dominant_indicators}{A character vector of \code{name = value} pairs, where \code{value} is
a character string giving the name of the dominant indicator and \code{name}
the corresponding construct name. Dominant indicators may be specified for
a subset of the constructs.}

\item{.E}{A (J x J) matrix of inner weights.}

\item{.estimate_structural}{Logical. Should the structural (path) coefficients
be estimated? Defaults to \code{TRUE}.}

\item{.group_var}{Character string. The name of the column used to split the data into groups.}

\item{.H}{The (N x J) matrix of proxy values.}

\item{.id}{Charachter string. The name of the column of \code{.data} used to split
the date into groups.}

\item{.ignore_structural_model}{Logical. Should the structural (path) model be ignored
when calculating the inner weights of the PLS algorithm? Defaults to \code{FALSE}}

\item{.iter_max}{Integer. The maximum number of iterations of the PLS algorithm.
If \code{iter_max = 1} one-step weights are returned. If the algorithm exceeds
the specified number, weights of iteration step \code{.iter_max - 1}  will be returned
with a warning. Defaults to \code{100}. The argument is ignored if
\code{.approach_weight} is not PLS.}

\item{.mode}{A named vector giving the mode used to obtain new outer weights.}

\item{.normality}{Logical. Should joint normality be assumed if the model
contains non-linear terms. For details see: \insertCite{Dijkstra2014;textual}{cSEM}.
Defaults to \code{TRUE}.}

\item{.object}{An R object of class \link{cSEMResults}.}

\item{.only_dots}{Logical. Should only those arguments to be passed to lower level functions via the
\code{...} argument of the \link{csem} or \link{cca} function be returned. Defaults to \code{FALSE}.}

\item{.permutations}{Integer. The number permutations used for step 2 and 3 of the test.
Defaults to \code{100}. Note however that the number should typically be a lot higher.}

\item{.PLS_weight_scheme_inner}{Character string. The inner weighting scheme
used in PLS. Possible choices are: "\emph{centroid}", "\emph{factorial}", or "\emph{path}".
Defaults to "\emph{centroid}". The argument is ignored if \code{.approach_weight} is not PLS.}

\item{.PLS_mode}{Either a named vector specifying the mode that should be used for
each construct in the form \code{name = "mode"}, a single character
string giving the mode that should be used for all constructs or \code{NULL}.
Possible choices are: "\emph{ModeA}" or "\emph{ModeB}". Defaults to \code{NULL}.
If \code{NULL}, \code{cSEM} will choose the appropriate mode according to the type
of construct used. The argument is ignored if \code{.approach_weight} is not PLS.}

\item{.Q}{A vector of proxy-construct correlations with element names equal to
the names of the J construct names used in the measurement model. Note
Q^2 is also called the reliability coefficient.}

\item{.reliabilities}{A vector of \code{name = value} pairs of reliability coefficients.
Element names are equal to the names of the J construct names. Reliabilities
may be given for a subset of the constructs. Defaults to \code{NULL} in which case
reliabilites are estimated by \code{cSEM}.}

\item{.S}{The (scaled) empirical (K x K) indicator covariance/correlation matrix,
where K is the number of indicators.}

\item{.terms}{A vector of construct names to be classified.}

\item{.tolerance}{Double. The tolerance criterion for convergence of the PLS
algorithm. Defaults to \code{1e-05}.
The argument is ignored if \code{.approach_weight} is not PLS.}

\item{.verbose}{Logical. Should information be printed to the console? Defaults
to \code{TRUE}.}

\item{.W}{A list with elements \code{$W}, \code{$E}, \code{PLS_mode} and xxx.}

\item{...}{Further arguments to be passed down to other functions.
See \link{args_default} for a complete list of possible arguments and their defaults.}
}
\description{
A list of all arguments including a description of their use and current defaults.
Mainly used for interal purposes (parameter inheritance). To list all arguments
and their defaults, use \code{\link[=args_default]{args_default()}}. To list only those arguments to be
passed down to lower level functions via the \code{...} argument of the \code{\link[=csem]{csem()}} function,
set the \code{.only_dots} argument of the \code{\link[=args_default]{args_default()}} function to \code{TRUE}.
}
