% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_utils.R
\name{csem_model}
\alias{csem_model}
\alias{cSEMModel}
\title{cSEMModel}
\description{
A standardized list containing model-related information. To convert a
a model written in \href{http://lavaan.ugent.be/tutorial/syntax1.html}{lavaan model syntax}
to a cSEMModel list use \link{parseModel}.
}
\details{
An object of class cSEMModel is a standardized list containing the following
components (assume in the following that there are J constructs and K indicators)
\describe{
\item{\code{$structural}}{A matrix mimicking the structural relationship between
constructs. If constructs are only linearly related, \code{structural} is
of dimension (J x J) with row- and column names equal to the construct
names. If the structural model contains non-linear relationships
\code{structural} is (J x (J + J*)) where J* is the number of
non-linear terms.}
\item{\code{$structural_ordered}}{A matrix of the same dimension as \code{structural}
with rows rearranged to satisfy the following criteria. 1. the construct
in the first row depends on exogenous constructs only. 2. the constructs
of the following rows depend only on exogenous constructs or those
of previous rows. This is required to estimate non-linear structural equation
model relationships using the replacement approach.
}
\item{\code{$measurement}}{A (J x K) matrix mimicking the measurement relationship
between constructs and their related indicators with row names equal to
the construct names and column names equal to the indicator names.}
\item{\code{$error_cor}}{A (K x K) matrix mimicking the measurement error
correlation relationship.}
\item{\code{$construct_type}}{A data frame with two columns \emph{"Name"} and \emph{"Type"}
containing the name of each construct and their respective type
(\strong{"Common factor"} or \strong{"Composite"}).}
\item{\code{$vars_endo}}{A vector of names of the endogenous constructs.}
\item{\code{$vars_exo}}{A vector of names of the exogenous constructs (incudes
possible interaction and exponential terms).}
\item{\code{$vars_explana}}{ A vector of names of the constructs that appear as
explanatory variables in at least one structural equation (incudes
possible interaction and exponential terms).}
\item{\code{$explained_by_exo}}{A vector of names of the constructs that are
solely explained by exogenous constructs.}
}
Note: it is possible to supply an incomplete cSEMModel list
to all functions that require \code{.csem_model} as a mandatory argument. Currently,
only the structural and the measurement matrix are required.
However, specifying an incomplete cSEMModel list may lead to unexpected behaviour
and errors so do use this technique with caution.
}
\seealso{
\link{parseModel}
}
