% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_utils.R
\name{csem_results}
\alias{csem_results}
\alias{cSEMResults}
\title{cSEMResults}
\value{
An object of class \code{cSEMResults} for which the following methods exist:
\describe{
\item{\code{print.cSEMResults}}{Prints a message to inform the user
whether estimation has been successful and what functions may be used
to examine the object.}
\item{\code{summary.cSEMResults}}{Print and return a comprehensive summary of the results.}
}
Technically \code{cSEMResults} is a named list containing the following list elements:
\describe{
\item{\code{$Estimates}}{A list containing the estimated quantities.}
\item{\code{$Tests}}{A list containing test results. (not yet implemented)}
\item{\code{$Meta_information}}{A list of additional information. (incomplete)}
}
}
\description{
cSEMResults
}
