% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_templates.R
\name{csem_summary}
\alias{csem_summary}
\alias{cSEMSummary}
\title{cSEMSummary}
\value{
An object of class \code{cSEMSummary}.
Technically \code{cSEMSummary} is a named list containing the following list elements:
\describe{
\item{`...}{Not finished yet.}
}
}
\description{
cSEMSummary
}
\keyword{internal}
