% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_assess.R
\name{distance_measures}
\alias{distance_measures}
\alias{calculateSRMR}
\alias{calculateDG}
\alias{calculateDL}
\alias{calculateDML}
\title{Internal: Calculate difference between S and Sigma_hat}
\usage{
calculateSRMR(.object = NULL)

calculateDG(.object = NULL)

calculateDL(.object = NULL)

calculateDML(.object = NULL)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}
}
\description{
Calculates the difference between the empirical
and the model-implied indicator variance-covariance matrix
using different distance measures. See vignette for assess.
}
\details{
The functions are only applicable to objects of class \code{cSEMResults_default}.
For other object classes use \code{\link[=assess]{assess()}}.
}
\section{Functions}{
\itemize{
\item \code{calculateSRMR}: The standardized root means squared residual (SRMR).

\item \code{calculateDG}: The geodesic distance (dG).

\item \code{calculateDL}: The squared euclidian distance

\item \code{calculateDML}: The distance measure used by FIML
}}

\keyword{internal}
