% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_assess.R
\name{distance_measures}
\alias{distance_measures}
\alias{calculateDG}
\alias{calculateDL}
\alias{calculateDML}
\title{Internal: Calculate difference between S and Sigma_hat}
\usage{
calculateDG(.object = NULL, .saturated = args_default()$.saturated,
  .type_vcv = args_default()$.type_vcv, ...)

calculateDL(.object = NULL, .saturated = args_default()$.saturated,
  .type_vcv = args_default()$.type_vcv, ...)

calculateDML(.object = NULL, .saturated = args_default()$.saturated,
  .type_vcv = args_default()$.type_vcv, ...)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}

\item{.saturated}{Logical. Should a saturated structural model be used?
Defaults to \code{FALSE}.}

\item{.type_vcv}{Character string. Indicates which model-implied correlation matrix is calcuted
One of "\emph{indicator}" or "\emph{construct}". Defaults to "\emph{indicator}".}

\item{...}{Ignored.}
}
\value{
A single numeric value.
}
\description{
Calculate the difference between the empirical (S)
and the model-implied indicator variance-covariance matrix (Sigma_hat)
using different distance measures. See vignette assess.
}
\details{
The functions are only applicable to objects inheriting class \code{cSEMResults_default}.
For objects of class \code{cSEMResults_multi} and \code{cSEMResults_2ndorder} use \code{\link[=assess]{assess()}}.
}
\section{Functions}{
\itemize{
\item \code{calculateDG}: The geodesic distance (dG).

\item \code{calculateDL}: The squared Euclidian distance

\item \code{calculateDML}: The distance measure used by FIML
}}

\keyword{internal}
