% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_doModelSearch.R
\name{doModelSearch}
\alias{doModelSearch}
\title{Do a model search}
\usage{
doModelSearch(.object = NULL)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}
}
\value{
A named numeric vector of correlations. If
the weighting approach used to obtain \code{.object} is not \code{"PLS-PM"} or
non of the PLS outer modes was mode B, the function silently returns \code{NA}.
}
\description{
\lifecycle{stable}
}
\details{
Perform a model search \insertCite{Hair2016;textual}{cSEM}
}
\references{
\insertAllCited{}
}
\seealso{
\link{cSEMResults}
}
