% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_doModelSearch.R
\name{doModelSearch}
\alias{doModelSearch}
\title{Do a model specification search}
\usage{
dfunction(.object = NULL, 
.n_exogenous = 3,
.popsize = 20,
.maxiter=20,
.mutation_prob = 0.5,
.seeds = c(2, 4, 5),
.only_structural = FALSE)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}

\item{.only_structural}{Should the the log-likelihood be based on the
structural model? Ignored if \code{.by_equation == TRUE}. Defaults to \code{TRUE}.}
}
\value{
The mean matrix generated by the AGAS-PLS runs
}
\description{
\lifecycle{"experimental"}
}
\details{
Perform a model search \insertCite{Hair2016;textual}{cSEM}
}
\references{
\insertAllCited{}
}
\seealso{
\link{cSEMResults}
}
