% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_doModelSearch.R
\name{doModelSearch}
\alias{doModelSearch}
\title{Do a model specification search}
\usage{
doModelSearch(.object = NULL, 
.popsize = 20,
.iter_max= 20,
.mutation_prob = 0.5,
.cross_prob = 0.8,
.only_structural = FALSE,
.ms_criterion = 'bic')
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}

\item{.iter_max}{Integer. The maximum number of iterations allowed.
If \code{iter_max = 1} and \code{.approach_weights = "PLS-PM"} one-step weights are returned.
If the algorithm exceeds the specified number, weights of iteration step
\code{.iter_max - 1}  will be returned with a warning. Defaults to \code{100}.}

\item{.only_structural}{Should the the log-likelihood be based on the
structural model? Ignored if \code{.by_equation == TRUE}. Defaults to \code{TRUE}.}

\item{.ms_criterion}{Character string. Either a single character string or a vector
of character strings naming the model selection criterion to compute.
Defaults to \code{"all"}.}

\item{.seed}{Integer or \code{NULL}. The random seed to use. Defaults to \code{NULL} in which
case an arbitrary seed is chosen. Note that the scope of the seed is limited
to the body of the function it is used in. Hence, the global seed will
not be altered!}
}
\value{
A list containing a list of model(s) showing the 'best' value of the model
selection criterion, the value of the model selection criteria, and a list of cSEM_models
}
\description{
\lifecycle{experimental}
}
\details{
Perform a model search
}
\references{
\insertAllCited{}
}
\seealso{
\link{cSEMResults}
}
