% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_doSimpleEffectsAnalysis.R
\name{doSimpleEffectsAnalysis}
\alias{doSimpleEffectsAnalysis}
\title{Do a simple effects analysis}
\usage{
doSimpleEffectsAnalysis(
 .object           = NULL,
 .dependent        = NULL, 
 .independent      = NULL,
 .moderator        = NULL,
 .n_steps          = 100,
 .values_moderator = c(-2,-1,0,1,2)
 )
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}

\item{.dependent}{Character string. The name of the dependent variable. Defaults to \code{NULL}.}

\item{.independent}{Character string. The name of the independent variable. Defaults to \code{NULL}.}

\item{.moderator}{Character string. The name of the moderator variable. Defaults to \code{NULL}.}

\item{.n_steps}{Integer. A numeric value giving the number of steps, e.g., in
surface analysis or floodlight analysis the spotlights (= values of .moderator)
between min(.moderator) and max(.moderator) to use. Defaults to \code{100}.}

\item{.values_moderator}{A numeric vector. Provides the values of the moderator, e.g.,
in the simple effects analysis. Defaults to c(-2,-1,0,1,2).}
}
\value{
A list of class \code{cSEMSimpleEffects} with a corresponding method for \code{plot()}.
See: \code{\link[=plot.cSEMFloodlight]{plot.cSEMFloodlight()}}.
}
\description{
Calculate the effect of an independent variable on a dependent variable
conditional on the values of a (continous) moderator variable. All other variables in the model
are assumed to be zero, i.e., at their mean level.
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=csem]{csem()}}, \link{cSEMResults}, \code{\link[=plot.cSEMFloodlight]{plot.cSEMFloodlight()}}
}
