% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_new.R
\name{effect_moderator_two_way}
\alias{effect_moderator_two_way}
\title{Output for floodlight analysis}
\usage{
effect_moderator_two_way=function(.object=args_default()$.object,
                                         .steps = seq(-1.5,1.5,0.01),
                                         .alpha = args_default()$.alpha, 
                                         .independent = NULL,
                                         .moderator = NULL,
                                         .dependent = NULL )
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}

\item{.alpha}{An integer or a numeric vector of significance levels.
Defaults to \code{0.05}.}
}
\description{
Calculate the the effect of an independent variable depending on different values
of its moderator variable to perform a floodlight analysis \insertCite{Spiller2013;textual}{cSEM}.
Moreover, the Johnson-Neyman point(s) is calculated. In doing so, it is considered
whether there is a sign switch in the lower and upper confidence intervals.
If so, the position of the switch is returned.
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=csem]{csem()}}, \code{\link[=foreman]{foreman()}}, \link{cSEMResults}
}
