% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_methods.R
\name{effects.cSEMResults}
\alias{effects.cSEMResults}
\title{Effects}
\usage{
effects(.object)
}
\arguments{
\item{.object}{An R object of class \code{cSEM<class>} with corresponding method.}
}
\description{
Compute direct, indirect and total effects.
}
\examples{
\dontrun{
# still to implement
}

}
\seealso{
\link{csem}, \link{cSEMResults}
}
