% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators_paths.R
\name{estimatePathOLS}
\alias{estimatePathOLS}
\title{Estimate the structural coefficients}
\usage{
estimatePathOLS(
  .H           = args_default()$.H,
  .Q           = args_default()$.Q,
  .P           = args_default()$.P,
  .csem_model  = args_default()$.csem_model,
  .normality   = args_default()$.normality,
  .approach_nl = args_default()$.approach_nl
  )
}
\arguments{
\item{.H}{The (N x J) matrix of construct scores.}

\item{.Q}{A vector of composite-construct correlations with element names equal to
the names of the J construct names used in the measurement model. Note
Q^2 is also called the reliability coefficient.}

\item{.P}{A (J x J) construct variance-covariance matrix (possibly disattenuated).}

\item{.csem_model}{A (possibly incomplete) \link{cSEMModel}-list.}

\item{.normality}{Logical. Should joint normality of
\eqn{[\eta_{1:p}; \zeta; \epsilon]}{[\eta_(1:p); \zeta; \epsilon]}
be assumed in the nonlinear model? See \insertCite{Dijkstra2014}{cSEM} for details.
Defaults to \code{TRUE}. Ignored if the model is linear.}

\item{.approach_nl}{Character string. Approach used to estimate nonlinear
structural relationships. One of: "\emph{sequential}" or "\emph{replace}".
Defaults to "\emph{sequential}".}
}
\value{
A named list containing the estimated structural coefficients and the
r^2 for each regression.
}
\description{
Estimates the coefficients of the structural model (nonlinear and linear) using
OLS. (TODO)
}
\details{
More details here (TODO).
}
