% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators_paths.R
\name{estimatePathOLS}
\alias{estimatePathOLS}
\title{Estimate the structural coefficients}
\usage{
estimatePathOLS(.H = NULL, .W = NULL, .Q = NULL, .P = NULL,
  .csem_model = NULL, .normality = NULL, .approach_nl = NULL)
}
\arguments{
\item{.H}{The (N x J) matrix of proxy values.}

\item{.W}{A list with elements \code{$W}, \code{$E}, \code{PLS_mode} and xxx.}

\item{.Q}{A vector of proxy-construct correlations with element names equal to
the names of the J construct names used in the measurement model. Note
Q^2 is also called the reliability coefficient.}

\item{.csem_model}{A (possibly incomplete) \link{cSEMModel} list.}

\item{.normality}{Logical. Should joint normality be assumed if the model
contains non-linear terms. For details see: \insertCite{Dijkstra2014;textual}{cSEM}.
Defaults to \code{TRUE}.}

\item{.approach_nl}{Character string. The approach used to handle non-linear
structural relationships. Possible choices are: "\emph{none}", or "\emph{replace}".
Defaults to "\emph{none}".}
}
\value{
A named list containing the estimated structural coefficients and the
r^2 for each regression.
}
\description{
Estimates the coefficients of the structural model (non-linear and linear)
}
\details{
More details here
}
