% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportToExcel.R
\name{exportToExcel}
\alias{exportToExcel}
\title{Export to Excel (.xlsx)}
\usage{
exportToExcel(
  .object    = NULL, 
  .file_name = "results.xlsx"
  )
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}

\item{.file_name}{Character string. The file name.}
}
\description{
Export results from a postestimation function such as \code{\link[=assess]{assess()}} or
\code{\link[=summarize]{summarize()}} to an .xlsx (Excel) file. The function uses the openxlsx
package which does not depend on Java!
}
\details{
The function is deliberately kept simple: all it does is to take all the
relevant elements in \code{.object} and write them (worksheet by worksheet) into
an .xlsx file named \code{.file_name} in your current working directory.
The rest is left to be done in Excel.

Note: rerunning \code{\link[=exportToExcel]{exportToExcel()}} without changing \code{.file_name} overwrites
the existing file.
}
\seealso{
\code{\link[=assess]{assess()}}, \code{\link[=summarize]{summarize()}}, \code{\link[=predict]{predict()}}, \code{\link[=verify]{verify()}}
}
