% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igsca_translation.R
\name{extract_parseModel}
\alias{extract_parseModel}
\title{A parseModel extractor function for the purposes of running I-GSCA code example}
\usage{
extract_parseModel(model, data, ind_domi_as_first = TRUE)
}
\arguments{
\item{model}{Specified Model in lavaan style}

\item{data}{Dataframe}

\item{ind_domi_as_first}{Boolean for whether the first indicator for each latent factor should be chosen as the dominant indicator}
}
\value{
z0, W0, B0, C0, lv_type, ov_type, ind_domi
}
\description{
A parseModel extractor function for the purposes of running I-GSCA code example
}
\examples{

require(here)
require(readxl)

tutorial_igsca_model <- "
# Composite Model
NetworkingBehavior <~ Behavior1 + Behavior2 + Behavior3 + Behavior5 + Behavior7 + Behavior8 + Behavior9
Numberofjobinterviews <~ Interview1 + Interview2
Numberofjoboffers <~ Offer1 + Offer2 

# Reflective Measurement Model
HonestyHumility =~ Honesty1 + Honesty2 + Honesty3 + Honesty4 + Honesty5 + Honesty6 + Honesty7 + Honesty8 + Honesty9 + Honesty10
Emotionality =~ Emotion1 + Emotion2 + Emotion3 + Emotion4 + Emotion5 + Emotion6 + Emotion8 + Emotion10
Extraversion =~ Extraver2 + Extraver3 + Extraver4 + Extraver5 + Extraver6 + Extraver7 + Extraver8 + Extraver9 + Extraver10
Agreeableness =~ Agreeable1 + Agreeable3 + Agreeable4 + Agreeable5 + Agreeable7 + Agreeable8 + Agreeable9 + Agreeable10
Conscientiousness =~ Conscientious1 + Conscientious3 + Conscientious4 + Conscientious6 + Conscientious7 + Conscientious8 + Conscientious9 + Conscientious10
OpennesstoExperience =~ Openness1 + Openness2 + Openness3 + Openness5 + Openness7 + Openness8 + Openness9 + Openness10

# Structural Model
NetworkingBehavior ~ HonestyHumility + Emotionality + Extraversion + Agreeableness + Conscientiousness + OpennesstoExperience
Numberofjobinterviews ~ NetworkingBehavior
Numberofjoboffers ~ NetworkingBehavior
"

dat <- readxl::read_excel(here::here("dev", "Notes", "data", "mmc1.xlsx")) 

extract_parseModel(model = tutorial_igsca_model, data = dat, ind_domi_as_first = TRUE)
}
