% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_polycor.R
\name{fastIntTab}
\alias{fastIntTab}
\title{Internal: Fast (cross-)tabulation of integer-coded categorical variables}
\usage{
fastIntTab(x, y = NULL)
}
\description{
A faster, more specialized alternative to \code{\link[base:table]{base::table()}} for
(cross-)tabulating one or two variables that are (or can be coerced to)
small positive integer codes, as used by \code{\link[=polychor]{polychor()}} and
\code{\link[=polyserial]{polyserial()}}. Rows with a missing value in either \code{x} or \code{y} are
dropped before tabulating.Implemented by Kjell S. Slupphaug.
}
\references{
\insertAllCited{}
}
\keyword{internal}
