% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_plot.cSEMResults.R
\name{firstOrderMeasurementEdges}
\alias{firstOrderMeasurementEdges}
\title{Internal: secondOrderMeasurementEdges}
\usage{
firstOrderMeasurementEdges(
    construct, 
    weights,
    loadings, 
    weight_p_values, 
    loading_p_values, 
    plot_signif, 
    constructTypes
    )
}
\value{
Character string. .
}
\description{
Build measurement edges for a first–order model.
}
\keyword{internal}
