% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csem_resample.R, R/postestimate_fit.R
\name{resamplecSEMResults.cSEMResults_default}
\alias{resamplecSEMResults.cSEMResults_default}
\alias{resamplecSEMResults.cSEMResults_multi}
\alias{resamplecSEMResults.cSEMResults_2ndorder}
\alias{fit}
\alias{fit.cSEMResults_default}
\alias{fit.cSEMResults_multi}
\alias{fit.cSEMResults_2ndorder}
\title{Model-implied indicator covariance matrix}
\usage{

  \method{resamplecSEMResults}{cSEMResults_default}(.object = args_default()$.object,
  .resample_method = args_default()$.resample_method,
  .resample_method2 = args_default()$.resample_method2,
  .R = args_default()$.R, .R2 = args_default()$.R2,
  .handle_inadmissibles = args_default()$.handle_inadmissibles,
  .user_funs = args_default()$.user_funs,
  .eval_plan = args_default()$.eval_plan)


  \method{resamplecSEMResults}{cSEMResults_multi}(.object = args_default()$.object,
  .resample_method = args_default()$.resample_method,
  .resample_method2 = args_default()$.resample_method2,
  .R = args_default()$.R, .R2 = args_default()$.R2,
  .handle_inadmissibles = args_default()$.handle_inadmissibles,
  .user_funs = args_default()$.user_funs,
  .eval_plan = args_default()$.eval_plan)


  \method{resamplecSEMResults}{cSEMResults_2ndorder}(.object = args_default()$.object,
  .resample_method = args_default()$.resample_method,
  .resample_method2 = args_default()$.resample_method2,
  .R = args_default()$.R, .R2 = args_default()$.R2,
  .handle_inadmissibles = args_default()$.handle_inadmissibles,
  .user_funs = args_default()$.user_funs,
  .eval_plan = args_default()$.eval_plan)

fit(
  .object    = args_default()$.object, 
  .saturated = args_default()$.saturated,
  .type_vcv  = args_default()$.type_vcv
  )

\method{fit}{cSEMResults_default}(.object = args_default()$.object,
  .saturated = args_default()$.saturated,
  .type_vcv = args_default()$.type_vcv)

\method{fit}{cSEMResults_multi}(.object = args_default()$.object,
  .saturated = args_default()$.saturated,
  .type_vcv = args_default()$.type_vcv)

\method{fit}{cSEMResults_2ndorder}(.object = args_default()$.object,
  .saturated = args_default()$.saturated,
  .type_vcv = args_default()$.type_vcv)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}

\item{.resample_method}{Character string. The resampling method to use. One of:
"\emph{bootstrap}" or "\emph{jackknife}". Defaults to "\emph{bootstrap}".}

\item{.resample_method2}{Character string. The resampling method to use when resampling
from a resample. One of: "\emph{none}", "\emph{bootstrap}" or "\emph{jackknife}". For
"\emph{bootstrap}" the number of draws is provided via \code{.R2}. Currently,
resampling from each resample is only required for the studentized confidence
intervall computed by the \code{\link[=infer]{infer()}} function. Defaults to "\emph{none}".}

\item{.R}{Integer. The number of bootstrap replications. Defaults to \code{499}.}

\item{.R2}{Integer. The number of bootstrap replications to use when
resampling from a resample. Defaults to \code{199}.}

\item{.handle_inadmissibles}{Character string. How should inadmissible results
be treated? One of "\emph{drop}", "\emph{ignore}", or "\emph{replace}". If "\emph{drop}", all
replications/resamples yielding an inadmissible result will be dropped (
the number of results returned will be less than .R). For "\emph{ignore}" all results are returned
even if they are inadmissible (number of results returned = .R). For "\emph{replace}"
resampling continues until there are exactly .R admissible solutions.
Defaults to "\emph{drop}".}

\item{.user_funs}{A function or a (named) list of functions to apply to every
resample. Takes \code{.object} as an input (e.g., \code{myFun <- function(.object) {...}}).
Output should preferably be a (named)
vector but matrices are also accepted. However, the output will be
vectorized (columnwise) in this case. See the examples section for details.}

\item{.eval_plan}{Character string. The evaluation plan to use. One of
"\emph{sequential}" or "\emph{multiprocess}". In the latter case
all available cores will be used. Defaults to "\emph{sequential}".}

\item{.saturated}{Logical. Should a saturated structural model be used? Defaults to \code{FALSE}.}

\item{.type_vcv}{Character string. Indicates which model-implied correlation matrix is calcuted
One of "\emph{indicator}" or "\emph{construct}". Defaults to "\emph{indicator}".}
}
\description{
Calculate the model-implied indicator or construct variance-covariance (VCV)
matrix. Currently only the model-implied VCV for linear model is implemented.
}
\details{
Notation is taken from \insertCite{Bollen1989;textual}{cSEM}.
By default the model-implied VCV matrix is based on the structural model.
If \code{.saturated = TRUE} the structural model is ignored (i.e. a full graph is estimated).
Hence: V(eta) = WSW' (possibly disattenuated).
}
\section{Methods (by class)}{
\itemize{
\item \code{cSEMResults_default}: (TODO)

\item \code{cSEMResults_multi}: (TODO)

\item \code{cSEMResults_2ndorder}: (TODO)

\item \code{cSEMResults_default}: (TODO)

\item \code{cSEMResults_multi}: (TODO)

\item \code{cSEMResults_2ndorder}: (TODO)
}}

\references{
\insertAllCited{}
}
\seealso{
\code{\link[=csem]{csem()}}, \code{\link[=cca]{cca()}}, \code{\link[=foreman]{foreman()}}, \link{cSEMResults}
}
