% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_fit.R
\name{fit}
\alias{fit}
\title{Model-implied indicator covariance matrix}
\usage{
fit(
  .object    = args_default()$.object, 
  .saturated = args_default()$.saturated
  )
}
\arguments{
\item{.object}{An R object of class \code{cSEM<class>} with corresponding method.}

\item{.saturated}{Logical. Should a saturated structural model be used? Defaults to \code{FALSE}.}
}
\description{
Calculate the model-implied indicator variance-covariance (VCV) matrix.
Currently only the model-implied VCV for linear model is implemented.
}
\details{
Notation is taken from \insertCite{Bollen1989;textual}{cSEM}.
By default the model-implied VCV matrix is based on the structural model.
If \code{.saturated = TRUE} the structural model is ignored (i.e. a full graph is estimated).
Hence: V(eta) = WSW' (possibly disattenuated).
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=csem]{csem()}}, \code{\link[=cca]{cca()}}, \code{\link[=foreman]{foreman()}}, \link{cSEMResults}
}
