% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csem_fit.R
\name{fit}
\alias{fit}
\alias{fit.cSEMResults_default}
\alias{fit.cSEMResults_multi}
\alias{fit.cSEMResults_2ndorder}
\title{Model-implied indicator covariance matrix}
\usage{
fit(
  .object    = args_default()$.object, 
  .saturated = args_default()$.saturated,
  .type_vcv  = args_default()$.type_vcv
  )

\method{fit}{cSEMResults_default}(.object = args_default()$.object,
  .saturated = args_default()$.saturated,
  .type_vcv = args_default()$.type_vcv)

\method{fit}{cSEMResults_multi}(.object = args_default()$.object,
  .saturated = args_default()$.saturated,
  .type_vcv = args_default()$.type_vcv)

\method{fit}{cSEMResults_2ndorder}(.object = args_default()$.object,
  .saturated = args_default()$.saturated,
  .type_vcv = args_default()$.type_vcv)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}

\item{.saturated}{Logical. Should a saturated structural model be used? Defaults to \code{FALSE}.}

\item{.type_vcv}{Character string. Indicates which model-implied correlation matrix is calcuted
One of "\emph{indicator}" or "\emph{construct}". Defaults to "\emph{indicator}".}
}
\description{
Calculate the model-implied indicator or construct variance-covariance (VCV)
matrix. Currently only the model-implied VCV for linear model is implemented.
}
\details{
Notation is taken from \insertCite{Bollen1989;textual}{cSEM}.
By default the model-implied VCV matrix is based on the structural model.
If \code{.saturated = TRUE} the structural model is ignored (i.e. a full graph is estimated).
Hence: V(eta) = WSW' (possibly disattenuated).
}
\section{Methods (by class)}{
\itemize{
\item \code{cSEMResults_default}: (TODO)

\item \code{cSEMResults_multi}: (TODO)

\item \code{cSEMResults_2ndorder}: (TODO)
}}

\references{
\insertAllCited{}
}
\seealso{
\code{\link[=csem]{csem()}}, \code{\link[=foreman]{foreman()}}, \link{cSEMResults}
}
