% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csem_fit.R
\name{fit}
\alias{fit}
\alias{fit.cSEMResults_default}
\alias{fit.cSEMResults_multi}
\alias{fit.cSEMResults_2ndorder}
\title{Model-implied indicator and construct variance-covariance matrix}
\usage{
fit(
  .object    = args_default()$.object, 
  .saturated = args_default()$.saturated,
  .type_vcv  = args_default()$.type_vcv
  )

\method{fit}{cSEMResults_default}(.object = args_default()$.object,
  .saturated = args_default()$.saturated,
  .type_vcv = args_default()$.type_vcv)

\method{fit}{cSEMResults_multi}(.object = args_default()$.object,
  .saturated = args_default()$.saturated,
  .type_vcv = args_default()$.type_vcv)

\method{fit}{cSEMResults_2ndorder}(.object = args_default()$.object,
  .saturated = args_default()$.saturated,
  .type_vcv = args_default()$.type_vcv)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}

\item{.saturated}{Logical. Should a saturated structural model be used?
Defaults to \code{FALSE}.}

\item{.type_vcv}{Character string. Indicates which model-implied correlation matrix is calcuted
One of "\emph{indicator}" or "\emph{construct}". Defaults to "\emph{indicator}".}
}
\value{
Either a (K x K) matrix or a (J x J) matrix depending on the \code{*type_vcv*}.
}
\description{
Calculate the model-implied indicator or construct variance-covariance (VCV)
matrix. Currently only the model-implied VCV for linear model is implemented.
}
\details{
Notation is taken from \insertCite{Bollen1989;textual}{cSEM}.
If \code{.saturated = TRUE} the model-implied variance-covariance matrix is calculated
for a saturated structural model (i.e., the VCV of the constructs is replaced
by their correlation matrix). Hence: V(eta) = WSW' (possibly disattenuated).
}
\section{Methods (by class)}{
\itemize{
\item \code{cSEMResults_default}: (TODO)

\item \code{cSEMResults_multi}: (TODO)

\item \code{cSEMResults_2ndorder}: (TODO)
}}

\examples{
\dontrun{
res <- csem(.data = dataset, .model = model)

fit(.object = res, .saturated = FALSE, .type_vcv = "indicator")
}
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=csem]{csem()}}, \code{\link[=foreman]{foreman()}}, \link{cSEMResults}
}
