% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_doTrees.R
\name{fit_csem}
\alias{fit_csem}
\title{\code{.id = NULL} yields the pooled single-group fit; \code{.id = "group"} the MGA fit.}
\usage{
fit_csem(.data, .model, .id = NULL)
}
\description{
\code{.id = NULL} yields the pooled single-group fit; \code{.id = "group"} the MGA fit.
}
