% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_assess.R
\name{fit_measures}
\alias{fit_measures}
\alias{calculateGFI}
\alias{calculateNFI}
\alias{calculateRMSEA}
\alias{calculateSRMR}
\alias{calculateRMSTheta}
\title{Internal: Fit measures}
\usage{
calculateGFI(.object)

calculateNFI(.object)

calculateRMSEA(.object)

calculateSRMR(.object = NULL, .saturated = args_default()$.saturated,
  .type_vcv = args_default()$.type_vcv, ...)

calculateRMSTheta(.object)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}

\item{.saturated}{Logical. Should a saturated structural model be used?
Defaults to \code{FALSE}.}

\item{.type_vcv}{Character string. Indicates which model-implied correlation matrix is calcuted
One of "\emph{indicator}" or "\emph{construct}". Defaults to "\emph{indicator}".}
}
\value{
A single numeric value.
}
\description{
Calculate common fit measures
}
\details{
The functions are only applicable to objects inheriting class \code{cSEMResults_default}.
For objects of class \code{cSEMResults_multi} and \code{cSEMResults_2ndorder} use \code{\link[=assess]{assess()}}.
}
\section{Functions}{
\itemize{
\item \code{calculateGFI}: The goodness of fit index (GFI).

\item \code{calculateNFI}: The normed fit index (NFI).

\item \code{calculateRMSEA}: The root mean square error of approximation (RMSEA).

\item \code{calculateSRMR}: The standardized root mean square residual (SRMR).

\item \code{calculateRMSTheta}: The RMS theta (SRMR).
}}

\keyword{internal}
