% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_methods.R
\name{fitted.cSEMResults}
\alias{fitted.cSEMResults}
\title{Model implied indicator variance-covariance matrix}
\usage{
fitted(object)
}
\arguments{
\item{object}{An R object of class cSEMResults.}
}
\description{
Compute the model implied indicator variance-covariance (VCV) matrix. Usually
this is called \eqn{\hat\Sigma}. Currently only the model implied VCV for
linear model for constructs modeled as common factors or composites are
implemented. An error is given if the model is not linear.
}
\examples{
\dontrun{
# still to implement
}

}
\seealso{
\link{csem}, \link{cSEMResults}
}
