% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igsca.R
\name{flip_signs_ind_domi}
\alias{flip_signs_ind_domi}
\title{Flip signs of Gamma, Loadings and Path-Coefficients Cells Based on Dominant Indicator}
\usage{
flip_signs_ind_domi(n_constructs, Z, .dominant_indicators, Gamma, C, B)
}
\arguments{
\item{n_constructs}{Number of construct variables}

\item{Z}{Standardized data matrix}

\item{.dominant_indicators}{Named vector of the dominant indicator for each construct variable}

\item{Gamma}{Matrix of construct scores}

\item{C}{Loadings matrix}

\item{B}{Path Coefficients matrix}
}
\value{
List of matrices: Gamma, Loadings (C) and Path-Coefficients (B)
}
\description{
Flip signs of Gamma, Loadings and Path-Coefficients Cells Based on Dominant Indicator
}
