% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igsca.R
\name{flip_signs_ind_domi}
\alias{flip_signs_ind_domi}
\title{Flip signs of Gamma, Loadings and Path-Coefficients Cells Based on Dominant Indicator}
\usage{
flip_signs_ind_domi(n_constructs, Z, ind_domi, j, Gamma, C, B)
}
\arguments{
\item{n_constructs}{}

\item{Z}{}

\item{ind_domi}{}

\item{j}{}

\item{Gamma}{}

\item{C}{}

\item{B}{}
}
\value{
List of matrices: Gamma, Loadings (C) and Path-Coefficients (B)
}
\description{
Flip signs of Gamma, Loadings and Path-Coefficients Cells Based on Dominant Indicator
}
