% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_foreman.R
\name{foreman}
\alias{foreman}
\title{Composite based SEM}
\usage{
foreman(
    .data                        = args_default()$.data,
    .model                       = args_default()$.model,
    .approach_cor_robust         = args_default()$.approach_cor_robust,
    .approach_nl                 = args_default()$.approach_nl,
    .approach_paths              = args_default()$.approach_paths,
    .approach_weights            = args_default()$.approach_weights,
    .disattenuate                = args_default()$.disattenuate,
    .dominant_indicators         = args_default()$.dominant_indicators,
    .estimate_structural         = args_default()$.estimate_structural,
    .iter_max                    = args_default()$.iter_max,
    .normality                   = args_default()$.normality,
    .PLS_approach_cf             = args_default()$.PLS_approach_cf,
    .PLS_ignore_structural_model = args_default()$.PLS_ignore_structural_model,
    .PLS_modes                   = args_default()$.PLS_modes,
    .PLS_weight_scheme_inner     = args_default()$.PLS_weight_scheme_inner,
    .reliabilities               = args_default()$.reliabilities,
    .tolerance                   = args_default()$.tolerance
    )
}
\arguments{
\item{.data}{A \code{data.frame}, a \code{matrix} or a list containing data of either type.
Possible column types or classes of the data provided are:
logical, numeric (double or integer), factor (ordered and unordered)
or a mix of several types. The data may also include
\emph{one} character column whose column name must be given to \code{.id}.
This column is assumed to contain group identifiers used to split the data into groups.}

\item{.model}{A model in \code{\link[lavaan:model.syntax]{lavaan model syntax}}
or a \link{cSEMModel}-list.}

\item{.approach_cor_robust}{Character string. Approach used to obtain a robust
indicator correlation matrix. One of: "\emph{none}" in which case nothing is done or
"\emph{mcd}" via \code{MASS::cor.rob()}. Defaults to "\emph{none}".}

\item{.approach_nl}{Character string. Approach used to estimate nonlinear
structural relationships. One of: "\emph{sequential}" or "\emph{replace}".
Defaults to "\emph{sequential}".}

\item{.approach_paths}{Character string. Approach used to estimate the
structural coefficients. One of: "\emph{OLS}" or "\emph{2SLS}".
Defaults to "\emph{OLS}".}

\item{.approach_weights}{Character string. Approach used to
obtain composite weights. One of: "\emph{PLS-PM}", "\emph{SUMCORR}", "\emph{MAXVAR}",
"\emph{SSQCORR}", "\emph{MINVAR}", "\emph{GENVAR}", "\emph{GSCA}", "\emph{fixed}", or "\emph{unit}".
Defaults to "\emph{PLS-PM}".}

\item{.conv_criterion}{Character string. The criterion to use for the convergence check.
One of: "\emph{diff_absolute}", "\emph{diff_squared}", or "\emph{diff_relative}". Defaults
to "\emph{diff_absolute}".}

\item{.disattenuate}{Logical. If possible, should composite correlations be disattenuated
if the construct is modeled as a common factor? Defaults to \code{TRUE}.}

\item{.dominant_indicators}{A character vector of \code{"name" = "value"} pairs,
where \code{"value"} is a character string giving the name of the dominant indicator
and \code{"name"} a character string of the corresponding construct name.
Dominant indicators may be specified for a subset of the constructs.}

\item{.estimate_structural}{Logical. Should the structural coefficients
be estimated? Defaults to \code{TRUE}.}

\item{.id}{Character string or integer. The name or position of the column of
\code{.data} used to split the data into groups. Ignored if \code{.object} is provided.
Defaults to \code{NULL}.}

\item{.iter_max}{Integer. The maximum number of iterations allowed.
If \code{iter_max = 1} and \code{.approach_weights = "PLS-PM"} one-step weights are returned.
If the algorithm exceeds the specified number, weights of iteration step
\code{.iter_max - 1}  will be returned with a warning. Defaults to \code{100}.}

\item{.normality}{Logical. Should joint normality be assumed in the nonlinear model?
For details see: \insertCite{Dijkstra2014;textual}{cSEM}.
Defaults to \code{TRUE}. Ignored if the model is linear.}

\item{.PLS_approach_cf}{Character string. Approach used to obtain the correction
factors for PLSc. One of: "\emph{dist_squared_euclid}", "\emph{dist_euclid_weighted}",
"\emph{fisher_transformed}", "\emph{mean_arithmetic}", "\emph{mean_geometric}", "\emph{mean_harmonic}",
"\emph{geo_of_harmonic}". Defaults to "\emph{dist_euclid}".
Ignored if \code{.disattenuate = FALSE} or if \code{.approach_weights} is not PLS-PM.}

\item{.PLS_ignore_structural_model}{Logical. Should the structural model be ignored
when calculating the inner weights of the PLS-PM algorithm? Defaults to \code{FALSE}.
Ignored if \code{.approach_weights} is not PLS-PM.}

\item{.PLS_modes}{Either a named vector specifying the mode that should be used for
each construct in the form \code{"name" = "mode"}, a single character
string giving the mode that should be used for all constructs, or \code{NULL}.
Possible choices are: "\emph{ModeA}" or "\emph{ModeB}". Defaults to \code{NULL}.
If \code{NULL}, \code{csem()} will choose the appropriate mode according to the type
of construct used. Ignored if \code{.approach_weight} is not PLS-PM.}

\item{.PLS_weight_scheme_inner}{Character string. The inner weighting scheme
used in PLS-PM. One of: "\emph{centroid}", "\emph{factorial}", or "\emph{path}".
Defaults to "\emph{path}". Ignored if \code{.approach_weight} is not PLS-PM.}

\item{.reliabilities}{A character vector of \code{"name" = value} pairs,
where \code{value} is a number between 0 and 1 and \code{"name"} a character string
of the corresponding construct name, or \code{NULL}. Reliabilities
may be given for a subset of the constructs. Defaults to \code{NULL} in which case
reliabilities are estimated by \code{csem()}.}

\item{.tolerance}{Double. The tolerance criterion for convergence.
Defaults to \code{1e-05}.}
}
\value{
An object of class \code{cSEMResults} with methods for all postestimation generics:
\itemize{
\item{\code{assess.cSEMResults}   Asses results using common fit measures and indices.}
\item{\code{summarize.cSEMResults}   Summarize the results.}
\item{\code{test.cSEMResults}   Run tests.}
\item{\code{verify.cSEMResults}   Verify admissibility (e.g., loadings larger than one).}
}

The structure and class of the resulting \code{cSEMResults} object is generally
determined by the type of dataset(s) provided (\code{matrix}, \code{data.frame}, \code{list}
of data)

\describe{
\item{Default}{If data is a single \code{matrix} or \code{data.frame} with no id-column,
the result is a \code{list} with elements:
\describe{
\item{\code{$Estimates}}{A list containing a list of estimated quantities.}
\item{\code{$Information}}{A list containing a list of additional information.}
}
The resulting object has classes \code{cSEMResults} and \code{cSEMResults_default}.
}
\item{Multi}{If the data provided contains an id-column to split the data by \code{G} group levels
or if a list of \code{G} datasets is provided, the resulting object is a list of \code{G}
lists, where \code{G} is equal to the number of groups or the number of datasets
in the list of datasets provided. Each of the \code{G} list elements are itself
a \code{cSEMResults_default} object. Hence its structure is identical to
the structure described in \code{Default}.

The resulting object has classes \code{cSEMResults} and \code{cSEMResults_multi}.
}
\item{2ndorder}{
A special output is generated if the model to estimate contains hierachical constructs
\strong{and} the 2step (3step) approach is used to estimate the model. In this case
the resulting object is a list containing two elements \code{First_stage} and
\code{Second_stage}.

Each list element is itself a \code{cSEMResults_default} object. Hence its structure is identical to
the structure described in \code{Default}.
}
}

If \code{.resample_method = "bootstrap"} or \code{.resample_method = "jackknife"} resamples
are attached to each object. For objects of class \code{cSEMResults_default} the resamples are
attached to \code{.object$Estimates$Estimates_resamples}. For objects of type
\code{cSEMResults_multi} the same is done by group. All objects containing
these elements gain the \code{cSEMResults_resampled} class.

As should be evident from the above, the core structure of the resulting object
is by and large similiar across all cases. Practically, most users wont have to worry
about the structure as postestimation functions have corresponding methods to
accomodate all cases.
}
\description{
The function is the central hub of the \code{cSEM}package. It acts like a
foreman by collecting all (estimation) tasks, distributing them to lower
level package functions, and eventually recollecting all of their results.
It is called by \code{\link[=csem]{csem()}} to manage the actual calculations.
It may be called directly by the user, however, in most cases it will likely
be more convenient to use \code{\link[=csem]{csem()}} or \code{\link[=cca]{cca()}} instead.
}
\details{
More details here (TODO).
}
\examples{
\dontrun{
# TODO
}

}
\seealso{
\link{csem}, \link{cca}, \link{cSEMResults}
}
