% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_foreman.R
\name{foreman}
\alias{foreman}
\title{Composite based SEM}
\usage{
foreman(
    .data                        = args_default()$.data,
    .model                       = args_default()$.model,
    .approach_cor_robust         = args_default()$.approach_cor_robust,
    .approach_nl                 = args_default()$.approach_nl,
    .approach_paths              = args_default()$.approach_paths,
    .approach_weights            = args_default()$.approach_weights,
    .disattenuate                = args_default()$.disattenuate,
    .dominant_indicators         = args_default()$.dominant_indicators,
    .estimate_structural         = args_default()$.estimate_structural,
    .iter_max                    = args_default()$.iter_max,
    .normality                   = args_default()$.normality,
    .PLS_approach_cf             = args_default()$.PLS_approach_cf,
    .PLS_ignore_structural_model = args_default()$.PLS_ignore_structural_model,
    .PLS_modes                   = args_default()$.PLS_modes,
    .PLS_weight_scheme_inner     = args_default()$.PLS_weight_scheme_inner,
    .reliabilities               = args_default()$.reliabilities,
    .tolerance                   = args_default()$.tolerance
    )
}
\arguments{
\item{.data}{A \code{data.frame} or a \code{matrix} containing the raw data. Possible
data column types or classes are: logical, numeric (double or integer), factor
(ordered and unordered) or a mix of several types.}

\item{.model}{A model in \code{\link[lavaan:model.syntax]{lavaan model syntax}}
or a \link{cSEMModel}-list.}

\item{.approach_cor_robust}{Character string. Approach used to obtain a robust
indicator correlation matrix. One of: "\emph{none}" in which case nothing is done,
"\emph{theil-sen}" or (TODO)
Defaults to "\emph{none}".}

\item{.approach_nl}{Character string. Approach used to estimate nonlinear
structural relationships. One of: "\emph{sequential}" or "\emph{replace}".
Defaults to "\emph{sequential}".}

\item{.approach_paths}{Character string. Approach used to estimate the
structural coefficients. One of: "\emph{OLS}" or "\emph{2SLS}".
Defaults to "\emph{OLS}".}

\item{.approach_weights}{Character string. Approach used to
obtain composite weights. One of: "\emph{PLS}", "\emph{SUMCORR}", "\emph{MAXVAR}",
"\emph{SSQCORR}", "\emph{MINVAR}", "\emph{GENVAR}", "\emph{GSCA}", "\emph{fixed}", or "\emph{unit}".
Defaults to "\emph{PLS}".}

\item{.conv_criterion}{Character string. The criterion to use for the convergence check.
One of: "\emph{diff_absolute}", "\emph{diff_squared}", or "\emph{diff_relative}". Defaults
to "\emph{diff_absolute}".}

\item{.disattenuate}{Logical. If possible, should composite correlations be disattenuated
if the construct is modeled as a common factor? Defaults to \code{TRUE}.}

\item{.dominant_indicators}{A character vector of \code{"name" = "value"} pairs,
where \code{"value"} is a character string giving the name of the dominant indicator
and \code{"name"} a character string of the corresponding construct name.
Dominant indicators may be specified for a subset of the constructs.}

\item{.estimate_structural}{Logical. Should the structural coefficients
be estimated? Defaults to \code{TRUE}.}

\item{.id}{Character string. The name of the column of \code{.data} used to split
the data into groups. Defaults to \code{NULL}.}

\item{.iter_max}{Integer. The maximum number of iterations allowed.
If \code{iter_max = 1} and \code{.approach_weights = "PLS"} one-step weights are returned.
If the algorithm exceeds the specified number, weights of iteration step
\code{.iter_max - 1}  will be returned with a warning. Defaults to \code{100}.}

\item{.normality}{Logical. Should joint normality be assumed in the nonlinear model?
For details see: \insertCite{Dijkstra2014;textual}{cSEM}.
Defaults to \code{TRUE}. Ignored if the model is linear.}

\item{.PLS_approach_cf}{Character string. Approach used to obtain the correction
factors for PLSc. One of: "\emph{dist_squared_euclid}", "\emph{dist_euclid_weighted}",
"\emph{fisher_transformed}", "\emph{mean_arithmetic}", "\emph{mean_geometric}", "\emph{mean_harmonic}",
"\emph{geo_of_harmonic}". Defaults to "\emph{dist_euclid}".
Ignored if \code{.disattenuate = FALSE} or if \code{.approach_weights} is not PLS.}

\item{.PLS_ignore_structural_model}{Logical. Should the structural model be ignored
when calculating the inner weights of the PLS algorithm? Defaults to \code{FALSE}.
Ignored if \code{.approach_weights} is not PLS.}

\item{.PLS_modes}{Either a named vector specifying the mode that should be used for
each construct in the form \code{"name" = "mode"}, a single character
string giving the mode that should be used for all constructs, or \code{NULL}.
Possible choices are: "\emph{ModeA}" or "\emph{ModeB}". Defaults to \code{NULL}.
If \code{NULL}, \code{csem()} will choose the appropriate mode according to the type
of construct used. Ignored if \code{.approach_weight} is not PLS.}

\item{.PLS_weight_scheme_inner}{Character string. The inner weighting scheme
used in PLS. One of: "\emph{centroid}", "\emph{factorial}", or "\emph{path}".
Defaults to "\emph{centroid}". Ignored if \code{.approach_weight} is not PLS.}

\item{.reliabilities}{A character vector of \code{"name" = value} pairs,
where \code{value} is a number between 0 and 1 and \code{"name"} a character string
of the corresponding construct name, or \code{NULL}. Reliabilities
may be given for a subset of the constructs. Defaults to \code{NULL} in which case
reliabilities are estimated by \code{csem()}.}

\item{.tolerance}{Double. The tolerance criterion for convergence.
Defaults to \code{1e-05}.}
}
\value{
An object of class \code{cSEMResults} with methods for all postestimation generics:
\describe{
\item{\code{check.cSEMResults}}{Checks results using common indices and fit measures.}
\item{\code{fit.cSEMResults}}{Compute the model-implied covariance matrix.}
\item{\code{summarize.cSEMResults}}{Summarize the results.}
\item{\code{test.cSEMResults}}{Run tests.}
\item{\code{verify.cSEMResults}}{Verify admissibility (e.g., loadings larger than one).}
}
The structure of the \code{cSEMResults} object is determined by the data set provided
or more precisely the number of estimation runs performed.

If data is a single matrix or data frame with no id-column, the result is a list with elements:
\describe{
\item{\code{$Estimates}}{A list containing a list of estimated quantities.}
\item{\code{$Information}}{A list containing a list of additional information.}
}

If \code{.data} contains an id-column to split the data by or if a list of
data sets is provided, the results contains \code{n} list with elements \code{$Estimates}
and \code{$Informations} where \code{n} is equal to the number of groups or the number of
data sets in the list of data sets provided.
}
\description{
The function is the central hub of the \code{cSEM}package. It acts like a
foreman by collecting all (estimation) tasks, distributing them to lower
level package functions, and eventually recollecting all of their results.
It is called by \code{\link[=csem]{csem()}} to manage the actual calculations.
It may be called directly by the user, however, in most cases it will likely
be more convenient to use \code{\link[=csem]{csem()}} or \code{\link[=cca]{cca()}} instead.
}
\details{
More details here (TODO).
}
\examples{
\dontrun{
# TODO
}

}
\seealso{
\link{csem}, \link{cca}, \link{cSEMResults}
}
