% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_workhorse.R
\name{foreman}
\alias{foreman}
\title{Composite based estimation}
\usage{
foreman(
    .data                    = args_default()$.data,
    .model                   = args_default()$.model,
    .approach_cf             = args_default()$.approach_cf,
    .approach_nl             = args_default()$.approach_nl,
    .approach_paths          = args_default()$.approach_paths,
    .approach_weights        = args_default()$.approach_weights,
    .disattenuate            = args_default()$.disattenuate,
    .dominant_indicators     = args_default()$.dominant_indicators,
    .estimate_structural     = args_default()$.estimate_structural,
    .ignore_structural_model = args_default()$.ignore_structural_model,
    .iter_max                = args_default()$.iter_max,
    .normality               = args_default()$.normality,
    .PLS_mode                = args_default()$.PLS_mode,
    .PLS_weight_scheme_inner = args_default()$.PLS_weight_scheme_inner,
    .reliabilities           = args_default()$.reliabilities,
    .tolerance               = args_default()$.tolerance
    )
}
\arguments{
\item{.data}{A \code{data.frame} or a \code{matrix} containing the raw data.}

\item{.model}{A model in \href{http://lavaan.ugent.be/tutorial/syntax1.html}{lavaan model syntax}
or a \link{cSEMModel} object.}

\item{.approach_cf}{Character string. The approach to obtain the correction
factor for PLSc. Possible choices are: "\emph{dist_euclid}", "\emph{dist_euclid_weighted}",
"\emph{fisher_transformed}", "\emph{mean_arithmetic}", "\emph{mean_geometric}", "\emph{mean_harmonic}",
"\emph{geo_of_harmonic}". Defaults to "\emph{dist_euclid}". Ignored if \code{.disattenuate = FALSE}.}

\item{.approach_nl}{Character string. The approach used to handle non-linear
structural relationships. Possible choices are: "\emph{none}", or "\emph{replace}".
Defaults to "\emph{none}".}

\item{.approach_paths}{Character string. Approach used to estimate the
structural coefficients. Possible choices are: "\emph{OLS}", "\emph{2SLS}", or "\emph{3SLS}".
Defaults to "\emph{OLS}".}

\item{.approach_weights}{Character string. The name of the approach used to
obtain composite weights. Possible choices are: "\emph{PLS}", "\emph{SUMCORR}", "\emph{MAXVAR}",
"\emph{SSQCORR}", "\emph{MINVAR}", "\emph{GENVAR}", "\emph{GSCA}", "\emph{fixed}", or "\emph{unit}".
Defaults to "\emph{PLS}".}

\item{.disattenuate}{Logical. Should proxy correlations be disattenuated
if the construct is modeled as a common factor? Defaults to \code{TRUE}.}

\item{.dominant_indicators}{A character vector of \code{name = value} pairs, where \code{value} is
a character string giving the name of the dominant indicator and \code{name}
the corresponding construct name. Dominant indicators may be specified for
a subset of the constructs.}

\item{.estimate_structural}{Logical. Should the structural (path) coefficients
be estimated? Defaults to \code{TRUE}.}

\item{.id}{Charachter string. The name of the column of \code{.data} used to split
the date into groups.}

\item{.ignore_structural_model}{Logical. Should the structural (path) model be ignored
when calculating the inner weights of the PLS algorithm? Defaults to \code{FALSE}.}

\item{.iter_max}{Integer. The maximum number of iterations of the PLS algorithm.
If \code{iter_max = 1} one-step weights are returned. If the algorithm exceeds
the specified number, weights of iteration step \code{.iter_max - 1}  will be returned
with a warning. Defaults to \code{100}. The argument is ignored if
\code{.approach_weight} is not PLS.}

\item{.normality}{Logical. Should joint normality be assumed if the model
contains non-linear terms. For details see: \insertCite{Dijkstra2014;textual}{cSEM}.
Defaults to \code{TRUE}.}

\item{.PLS_mode}{Either a named vector specifying the mode that should be used for
each construct in the form \code{name = "mode"}, a single character
string giving the mode that should be used for all constructs or \code{NULL}.
Possible choices are: "\emph{ModeA}" or "\emph{ModeB}". Defaults to \code{NULL}.
If \code{NULL}, \code{cSEM} will choose the appropriate mode according to the type
of construct used. The argument is ignored if \code{.approach_weight} is not PLS.}

\item{.PLS_weight_scheme_inner}{Character string. The inner weighting scheme
used in PLS. Possible choices are: "\emph{centroid}", "\emph{factorial}", or "\emph{path}".
Defaults to "\emph{centroid}". The argument is ignored if \code{.approach_weight} is not PLS.}

\item{.reliabilities}{A vector of \code{name = value} pairs of reliability coefficients.
Element names are equal to the names of the J construct names. Reliabilities
may be given for a subset of the constructs. Defaults to \code{NULL} in which case
reliabilites are estimated by \code{cSEM}.}

\item{.tolerance}{Double. The tolerance criterion for convergence of the PLS
algorithm. Defaults to \code{1e-05}.
The argument is ignored if \code{.approach_weight} is not PLS.}
}
\value{
An object of class \code{cSEMResults} for which the following methods exist:
\describe{
\item{\code{print.cSEMResults}}{Prints a message to inform the user
whether estimation has been successful and what functions may be used
to examine the object.}
\item{\code{summary.cSEMResults}}{Print and return a comprehensive summary of the results.}
}
Technically \code{cSEMResults} is a named list containing the following list elements:
\describe{
\item{\code{$Estimates}}{A list containing the estimated quantities.}
\item{\code{$Information}}{A list of additional information. (incomplete)}
}
}
\description{
This function is the backbone of the \code{cSEM} package. It is called by \link{csem},
\link{cca}, and test functions such as \link{testMICOM} to do the actual calculations.
It may be called directly by the user, however, in most cases it will likely
be more convenient to use \link{csem} or \link{cca} instead.
}
\examples{
\dontrun{
# still to implement
}

}
\seealso{
\link{csem}, \link{cca}, \link{testMICOM}
}
