% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igsca.R, R/igsca_translation.R
\name{get_lavaan_table_igsca_gscapro}
\alias{get_lavaan_table_igsca_gscapro}
\title{Takes GSCAPro input and Creates a Lavaan-style Table}
\usage{
get_lavaan_table_igsca_gscapro(gscapro_in, model)

get_lavaan_table_igsca_gscapro(gscapro_in, model)
}
\arguments{
\item{gscapro_in}{}

\item{model}{}
}
\value{
Table of GSCA Pro results of Weights, Loadings, Path Coefficients and Uniqueness Terms in lavaan style.

Table of GSCA Pro results of Weights, Loadings, Path Coefficients and Uniqueness Terms in lavaan style.
}
\description{
TODO: Rename the functions here

Assumes that every indicator loads onto only one latent variable (composite/factor)
}
\details{
Assumes that every indicator loads onto only one latent variable (composite/factor)

Expects output from parse_GSCAPro_FullResults.

Expects output from parse_GSCAPro_FullResults.
}
