% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igsca_translation.R
\name{get_lavaan_table_igsca_matrix}
\alias{get_lavaan_table_igsca_matrix}
\title{Converts Output of igsca functions into a table to facilitate comparisons}
\usage{
get_lavaan_table_igsca_matrix(model, weights, loadings, uniqueD, paths)
}
\arguments{
\item{model}{}

\item{weights}{}

\item{loadings}{}

\item{uniqueD}{}

\item{paths}{}
}
\value{
Table of Weights, Loadings, Path-Coefficients and Uniqueness terms from i-gsca algorithms in Matlab or R.
}
\description{
Assumes that indicators only load onto one factor and that there are no cross-factor loadings
}
