% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_csem-tidiers.R
\name{glance.cSEMResults}
\alias{glance.cSEMResults}
\title{glance Method for cSEMResults Objects}
\usage{
\method{glance}{cSEMResults}(x, ...)
}
\arguments{
\item{x}{Fitted \code{csem} model with class \code{cSEMResults} as returned from \code{\link[=csem]{csem()}}}
}
\value{
A one-row \code{data.frame} with columns:
}
\description{
Placeholder function for future development.
}
\details{
TODO: This method should be used to obtain a one row \code{data.frame} with a different column for every fit-statistic/estimation information.

This method and documentation is based on the \code{glance} method for \code{lavaan}, as found in \code{broom:::glance.lavaan()}.
}
