% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igsca.R
\name{gsca_inione}
\alias{gsca_inione}
\title{One-to-One R Translation of gsca_inione.m from Heungsun Hwang}
\usage{
gsca_inione(z0, W0, B0)
}
\arguments{
\item{z0}{Data Matrix}

\item{W0}{Indicator matrix of weights}

\item{B0}{Indicator matrix of Path Coefficients.}
}
\value{
When used in the context of igsca_sim(), it returns a list of the starting values for Weights, Loadings and Path Coefficients. In principle, otherwise, it is a slightly modified implementation of ordinary Generalised Structured Component Analysis (GSCA).
}
\description{
Internal I-GSCA Function
}
\details{
Initializes the values for I-GSCA
}
\examples{


tutorial_gsca_model <- "
# Composite Model
NetworkingBehavior <~ Behavior1 + Behavior2 + Behavior3 + Behavior5 + Behavior7 + Behavior8 + Behavior9
Numberofjobinterviews <~ Interview1 + Interview2
Numberofjoboffers <~ Offer1 + Offer2 

# Reflective Measurement Model Forced into Composite
HonestyHumility <~ Honesty1 + Honesty2 + Honesty3 + Honesty4 + Honesty5 + Honesty6 + Honesty7 + Honesty8 + Honesty9 + Honesty10
Emotionality <~ Emotion1 + Emotion2 + Emotion3 + Emotion4 + Emotion5 + Emotion6 + Emotion8 + Emotion10
Extraversion <~ Extraver2 + Extraver3 + Extraver4 + Extraver5 + Extraver6 + Extraver7 + Extraver8 + Extraver9 + Extraver10
Agreeableness <~ Agreeable1 + Agreeable3 + Agreeable4 + Agreeable5 + Agreeable7 + Agreeable8 + Agreeable9 + Agreeable10
Conscientiousness <~ Conscientious1 + Conscientious3 + Conscientious4 + Conscientious6 + Conscientious7 + Conscientious8 + Conscientious9 + Conscientious10
OpennesstoExperience <~ Openness1 + Openness2 + Openness3 + Openness5 + Openness7 + Openness8 + Openness9 + Openness10

# Structural Model
NetworkingBehavior ~ HonestyHumility + Emotionality + Extraversion + Agreeableness + Conscientiousness + OpennesstoExperience
Numberofjobinterviews ~ NetworkingBehavior
Numberofjoboffers ~ NetworkingBehavior
"

data("LeDang2022")


gsca_in <- extract_parseModel(model = tutorial_gsca_model,
                                   data = LeDang2022,
                                   ind_domi_as_first = TRUE)
                         
(gsca_inione_test <- with(gsca_in,
                          gsca_inione(z0 = z0,
                                      W0 = W0,
                                      B0 = B0)
                                      )
                                      )
}
\author{
Michael S. Truong
}
