% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igsca.R
\name{gsca_inione}
\alias{gsca_inione}
\title{R Implementation of gsca_inione.m from Heungsun Hwang}
\usage{
gsca_inione(Z0, W0, B0)
}
\arguments{
\item{Z0}{Data matrix of N cases (measurements) x J indicators with named columns, unstandardized.}

\item{W0}{Indicator matrix of weights: J indicators (rows) and their corresponding Gamma construct variables (columns).}

\item{B0}{Square indicator matrix of path coefficients: from-construct-variable (rows) and to-construct-variable (columns). The order of Gamma construct variables should match the order in C0 and W0.}
}
\value{
Returns a list of starting values for:
\itemize{
\item Weights (W)
\item Loadings (C)
\item Path Coefficients (B)
}
}
\description{
Internal I-GSCA function that is a slightly modified implementation of ordinary Generalised Structured Component Analysis (GSCA).
}
\details{
Initializes the values for I-GSCA
}
\author{
Michael S. Truong
}
