% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_arguments.R
\name{handleArgs}
\alias{handleArgs}
\title{Internal: Handle arguments}
\usage{
handleArgs(.args_used)
}
\arguments{
\item{.args_used}{A list of argument names and user picked values.}
}
\value{
The \link{args_default} list, with default values
changed to the values given by the user.
}
\description{
Internal helper function to handle arguments passed to any function within \code{cSEM}.
}
\keyword{internal}
