% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_assess.R
\name{HTMT}
\alias{HTMT}
\title{HTMT}
\usage{
HTMT(.object)
}
\arguments{
\item{.object}{An R object of class \code{cSEM<class>} with corresponding method.}
}
\description{
Compute the heterotrait-monotrait ratio of correlations (TODO)
}
\examples{
\dontrun{
# still to implement
}

}
\seealso{
\link{csem}, \link{cSEMResults}
}
