% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_igsca.R
\name{igsca}
\alias{igsca}
\title{R Implementation of igsca_sim.m}
\usage{
igsca(
  Z0,
  W0,
  C0,
  B0,
  con_type,
  indicator_type,
  .dominant_indicators,
  .iter_max = 100,
  .tolerance = 1e-04,
  .conv_criterion,
  .S = args_default()$.S
)
}
\arguments{
\item{Z0}{Data matrix of N cases (measurements) x J indicators with named
columns, unstandardized.}

\item{W0}{Indicator matrix of weights: J indicators (rows) and their
corresponding Gamma construct variables (columns).}

\item{C0}{Indicator matrix of loadings: J indicators (rows) and their
corresponding Gamma construct variable (columns).}

\item{B0}{Square indicator matrix of path coefficients:
from-construct-variable (rows) and to-construct-variable (columns). The
order of Gamma construct variables should match the order in C0 and W0.}

\item{con_type}{A vector that denotes whether each construct variable
(columns in W0 and C0) is a common factor or composite. Its length should
be equal to the number of columns of W0 and C0.}

\item{indicator_type}{An indicator vector that indices whether a j indicator
(rows of W0 and C0) corresponds to a common factor variable (1) or a
composite variable (0). This vector is important for computing the
uniqueness terms (D) because it zeros the entries for composite indicators.}

\item{.dominant_indicators}{A named vector that indices the dominant
indicator for each construct variable. \emph{It is to be clarified whether this
should only apply to factor latent variables or also composite latent
variables.} This is important for ensuring that the signs of the
path-coefficients and loadings are consistent. It is sometimes the case in
composite-based structural equation modelling methods that
loadings/path-coefficients may have the opposite sign. The length of this
vector should be equal to the number of construct variables and each value
should represent the row number of the dominant indicator for that
construct variable.}

\item{.iter_max}{Maximum number of iterations of the Alternating Least Squares
(ALS) algorithm.}

\item{.tolerance}{Minimum amount of absolute change in the estimates of the
path-coefficients (if B0 is non-zero) or the loadings (if B0 is all zero,
meaning there are no path-coefficients) between ALS iterations before ending
the optimization.}

\item{.conv_criterion}{Character string. The criterion to use for the convergence check.
One of: "\emph{diff_absolute}", "\emph{diff_squared}", "\emph{diff_relative}", or "\emph{sum_diff_absolute}". Defaults
to "\emph{diff_absolute}".}

\item{.S}{The (K x K) empirical indicator correlation matrix.}
}
\value{
List of 4 matrices that make up a fitted I-GSCA Model:
\itemize{
\item (1) Weights
\item (2) Loadings
\item (3) Squared Unique Loadings D^2
\item (4) Path Coefficients.
\item (5) Unique Component of Indicators (for common factors) DU
}
}
\description{
This R implementation of I-GSCA is based on the Matlab implementation in igsca_sim.m by Dr. Heungsun Hwang.
}
\details{
In the example section, the specified model is based on the tutorial I-GSCA model associated with GSCA Pro \insertCite{hwangetal2023StructuralEquationModelingAMultidisciplinaryJournal}{cSEM}.
}
\examples{

# Specify the model according to GSCA Pro's example
tutorial_igsca_model <- "
# Composite Model
NetworkingBehavior <~ Behavior1 + Behavior2 + Behavior3 + Behavior5 + 
                      Behavior7 + Behavior8 +  Behavior9
Numberofjobinterviews <~ Interview1 + Interview2
Numberofjoboffers <~ Offer1 + Offer2 

# Reflective Measurement Model
HonestyHumility =~ Honesty1 + Honesty2 + Honesty3 + Honesty4 + Honesty5 + 
                    Honesty6 + Honesty7 + Honesty8 + Honesty9 + Honesty10
Emotionality =~ Emotion1 + Emotion2 + Emotion3 + Emotion4 + 
                Emotion5 + Emotion6 + Emotion8 + Emotion10
Extraversion =~ Extraver2 + Extraver3 + Extraver4 + Extraver5 + 
                Extraver6 + Extraver7 + Extraver8 + Extraver9 + Extraver10
Agreeableness =~ Agreeable1 + Agreeable3 + Agreeable4 + Agreeable5 +
                 Agreeable7 + Agreeable8 + Agreeable9 + Agreeable10
Conscientiousness =~ Conscientious1 + Conscientious3 + Conscientious4 + 
                     Conscientious6 + Conscientious7 + Conscientious8 + 
                     Conscientious9 + Conscientious10
OpennesstoExperience =~ Openness1 + Openness2 + Openness3 + Openness5 + 
                        Openness7 + Openness8 + Openness9 + Openness10

# Structural Model
NetworkingBehavior ~ HonestyHumility + Emotionality + Extraversion + 
                     Agreeableness + Conscientiousness + OpennesstoExperience
Numberofjobinterviews ~ NetworkingBehavior
Numberofjoboffers ~ NetworkingBehavior
"

data(LeDang2022)

csem(.data = LeDang2022, tutorial_igsca_model, .approach_weights = "GSCA",
.dominant_indicators = NULL, .tolerance = 0.0001, .conv_criterion =
"sum_diff_absolute")
}
\references{
\insertAllCited{}
}
\author{
Michael S. Truong
}
