% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_doTrees.R
\name{igsca_tree_control}
\alias{igsca_tree_control}
\title{Title}
\usage{
igsca_tree_control(
  alpha = 0.05,
  bonferroni = FALSE,
  minbucket = 30L,
  minsplit = 2L * minbucket,
  maxdepth = 3L,
  max_cuts = 20L,
  R_test = 500L,
  coin_distribution = c("approximate", "asymptotic")
)
}
\arguments{
\item{coin_distribution}{}
}
\description{
Title
}
