% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csem_resample.R
\name{infer}
\alias{infer}
\title{Internal: Inference}
\usage{
infer(
 .object        = args_default()$.object
)
}
\arguments{
\item{.object}{An R object of class \code{cSEM<class>} with corresponding method.}

\item{.runs}{Integer. How many runs should be performed? Defaults to \code{499}.}

\item{.handle_inadmissibles}{Character string. How should inadmissible results
be treated? One of "\emph{drop}", "\emph{ignore}", or "\emph{replace}". If "\emph{drop}", all
replications yielding an inadmissible result will be dropped (=
number of results shown <= .runs). For "\emph{ignore}" all results are returned
even if they are inadmissible (= number of results = .runs). For "\emph{replace}"
bootstrapping continues until there are .runs admissible solutions.
Defaults to "\emph{drop}".}

\item{.verbose}{Logical. Should information be printed to the console? Defaults
to \code{TRUE}.}
}
\description{
Compute quantities for inference (TODO)
}
\examples{
\dontrun{
# still to implement
}

}
\seealso{
\link{csem}, \link{cSEMResults}
}
