% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csem_resample.R
\name{infer}
\alias{infer}
\alias{MeanResample}
\alias{VarResample}
\alias{SdResample}
\alias{BiasResample}
\alias{StandardCIResample}
\alias{PercentilCIResample}
\alias{BasicCIResample}
\alias{TStatCIResample}
\alias{BcCI}
\alias{BcaCI}
\title{Internal: Inference}
\usage{
infer(
 .object        = args_default()$.object
)

MeanResample(.x)

VarResample(.x, .method, .n)

SdResample(.x, .method, .n)

BiasResample(.x, .method, .n)

StandardCIResample(.x, .bias_corrected, .df = c("type1", "type2"),
  .dist = c("z", "t"), .method, .n, .probs)

PercentilCIResample(.x, .probs)

BasicCIResample(.x, .bias_corrected, .probs)

TStatCIResample(.x1, .x2, .bias_corrected, .method, .method2, .n, .probs)

BcCI(.x, .probs)

BcaCI(.object, .x, .probs)
}
\arguments{
\item{.csem_resample}{A list resulting from a call to \code{\link[=resamplecSEMResults]{resamplecSEMResults()}}.}

\item{.alpha}{An integer or a numeric vector of significance levels.
Defaults to \code{0.05}.}

\item{.bias_corrected}{Logical. Should the standard and the tStat
confidence intervall be bias-corrected using the bootstraped bias estimate?
If \code{TRUE} the confidence intervall for some estimated parameter \code{theta}
is centered at \code{2*theta - theta*_hat},
where \code{theta*_hat} is the average over all .R bootstrap estimates of \code{theta}.
Defaults to \code{TRUE}}

\item{.statistic}{Character string. Which statistic should be returned?
One of (TODO)
Defaults to (TODO).}

\item{.object}{An R object of class \code{cSEM<class>} with corresponding method.}

\item{.method}{Character string. The resampling method to use. One of:
"\emph{bootstrap}" or "\emph{jackknife}". Defaults to "\emph{bootstrap}".}

\item{.method2}{Character string. The resampling method to use when resampling
from a resample. One of: "\emph{none}", "\emph{bootstrap}" or "\emph{jackknife}". For
\emph{bootstrap} the number of draws may be provided via \code{.R2}.
Defaults to "\emph{none}".}
}
\description{
Compute quantities for inference (TODO)
}
\section{Functions}{
\itemize{
\item \code{MeanResample}: The bootstraped mean

\item \code{VarResample}: (TODO)

\item \code{SdResample}: (TODO)

\item \code{BiasResample}: (TODO)

\item \code{StandardCIResample}: (TODO)

\item \code{PercentilCIResample}: (TODO)

\item \code{BasicCIResample}: (TODO)

\item \code{TStatCIResample}: (TODO)

\item \code{BcCI}: (TODO)

\item \code{BcaCI}: (TODO)
}}

\examples{
\dontrun{
# still to implement
}

}
\seealso{
\link{csem}, \link{cSEMResults}
}
