% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_datasets.R
\docType{data}
\name{LeDang2022}
\alias{LeDang2022}
\title{Data: LeDang2022}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 773 rows and 107 columns.
}
\source{
The data were made available in \insertCite{ledang2022DataBrief;textual}{cSEM}
}
\usage{
LeDang2022
}
\description{
Data: LeDang2022
}
\details{
The data comes from a European manufacturer of durable consumer goods and was
studied by \insertCite{ledang2022DataBrief;textual}{cSEM}
}
\references{
\insertAllCited{}
}
\keyword{datasets}
