% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_datasets.R
\docType{data}
\name{LeDang2022}
\alias{LeDang2022}
\title{Data: LeDang2022}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 773 rows and 107 columns.
}
\source{
The data were made available in \insertCite{ledang2022DataBrief;textual}{cSEM}
}
\usage{
LeDang2022
}
\description{
A data frame of 107 variables and 773 participants.
}
\details{
This data is based on self-report surveys collected from graduates of universities in Northern Vietnam \insertCite{ledang2022DataBrief;textual}{cSEM}. The data has been used to analyze the relationship between measures of HEXACO personality traits and two job-search outcomes: (1) number of job interviews; and (2) job offers through job search networking behavior \insertCite{ledang2022DataBrief;textual}{cSEM}.

The data were also analyzed using an Integrated-Generalised Structued Component Analysis (I-GSCA) approach in GSCAPro \insertCite{hwangetal2023StructuralEquationModelingAMultidisciplinaryJournal}{cSEM}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
