% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igsca.R
\name{model_multigroup_igsca}
\alias{model_multigroup_igsca}
\title{Title}
\usage{
model_multigroup_igsca(model, group)
}
\arguments{
\item{model}{}

\item{group}{}
}
\value{
Multi-group igsca model
}
\description{
Title
}
