% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igsca.R
\name{model_multigroup_igsca}
\alias{model_multigroup_igsca}
\title{Compute multi-group model for I-GSCA}
\usage{
model_multigroup_igsca(model, group, data)
}
\arguments{
\item{model}{Specified Model in lavaan style}

\item{group}{Column name of group of input data}

\item{data}{Input dataframe}
}
\value{
Multi-group igsca model
}
\description{
TBD
}
