% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_model.R
\name{parseModel}
\alias{parseModel}
\title{Parse lavaan model}
\usage{
parseModel(.model)
}
\arguments{
\item{.model}{A model in \href{http://lavaan.ugent.be/tutorial/syntax1.html}{lavaan model syntax}
or a \link{cSEMModel} object.}
}
\value{
A \link{cSEMModel} list that may be passed to any function requiring
\code{.csem_model} as a mandatory argument.
}
\description{
Turns a model written in \link[lavaan:model.syntax]{lavaan model syntax} into a
\link{cSEMModel} list.
}
\examples{
model <- '
# Structural model
y1 ~ y2 + y3

# Measurement model
y1 =~ x1 + x2 + x3
y2 =~ x4 + x5
y3 =~ x6 + x7

# Error correlation
x1 ~~ x2
'

(m <- parseModel(model))

# If the model is already a cSEMModel object, the model is returned as is.

identical(m, parseModel(m)) # TRUE
}
