% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_assess.R
\name{plot.cSEMFloodlight}
\alias{plot.cSEMFloodlight}
\title{\code{cSEMFloodlight} method for \code{plot()}}
\usage{
\method{plot}{cSEMFloodlight}(x, ...)
}
\arguments{
\item{x}{An R object of class \code{cSEMFloodlight} resulting from a call to \link{doFloodlightAnalysis}.}

\item{...}{ignored.}
}
\description{
Plot the direct effect of an independent variable (z) on a dependent variable
(y) conditional on the values of a moderator variable (x), including
the confidence interval and the Johnson-Neyman points.
}
