% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cSEMPredict.R
\name{plot.cSEMPredict}
\alias{plot.cSEMPredict}
\title{\code{cSEMPredict} method for \code{plot()}}
\usage{
\method{plot}{cSEMPredict}(x, ...)
}
\arguments{
\item{x}{An R object of class \code{cSEMPredict}.}

\item{...}{Currently ignored.}
}
\description{
The \code{cSEMPredict} method for the generic function \code{\link[=plot]{plot()}}.
}
\seealso{
\code{\link[=predict]{predict()}}
}
\keyword{internal}
