% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cSEMSimpleEffects.R
\name{plot.cSEMSimpleEffects}
\alias{plot.cSEMSimpleEffects}
\title{\code{cSEMSimpleEffects} method for \code{plot()}}
\usage{
\method{plot}{cSEMSimpleEffects}(x, ...)
}
\arguments{
\item{x}{An R object of class \code{cSEMSimpleEffects}.}

\item{...}{Currently ignored.}
}
\description{
Creates a simple effects plot. The expected value of the dependent variable is plotted
for different values of the independent variable and the moderator. As levels for the moderator
-2, -1, 0, 1, and 2 are used. Since the constructs are standardized the values of the moderator
equals the deviation from its mean measured in standard deviations.
}
