% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cSEMSurface.R
\name{plot.cSEMSurface}
\alias{plot.cSEMSurface}
\title{\code{cSEMSurface} method for \code{plot()}}
\usage{
\method{plot}{cSEMSurface}(x, .plot_type = c("plotly"), ...)
}
\arguments{
\item{x}{An R object of class \code{cSEMSurface}.}

\item{.plot_type}{A character vector indicating the plot package used. Options are
plotly, persp, and rsm. Defaults to '\emph{plotly}'.}

\item{...}{Ignored.}
}
\description{
Plot the predicted values of an independent variable (z)
The values are predicted based on a certain moderator and a certain
independent variable including all their higher-order terms.
}
