% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_doTrees.R
\name{plot.igsca_tree}
\alias{plot.igsca_tree}
\title{Title}
\usage{
\method{plot}{igsca_tree}(x, terminal_panel = NULL, FUN = NULL, tp_args = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
