% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_plotModel.R
\name{plotModel}
\alias{plotModel}
\title{Plot model}
\usage{
plotModel(
.object, 
.show_stars = TRUE, 
.title = "",
.run_correlation_lines = FALSE,
.show_structural_model_only = FALSE,
.node_ranks = NULL, 
.graph_attrs = NULL,
.remove_colors = FALSE, 
.edge_lengths = NULL,
.stage = c("second", "first")
)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}

\item{.show_stars}{Logical, whether to show significance stars based on p-values.}

\item{.title}{Character string. Plot title.}

\item{.run_correlation_lines}{Logical. Should the correlations between the indicators displayed? Defaults to \code{FALSE}.}

\item{.show_structural_model_only}{Logical. Should only the structural model,
i.e., the constructs and their relationships be displayed? Default to \code{FALSE}.}

\item{.node_ranks}{List specifying the ranks for nodes (e.g., list(rank1 = c("Node1", "Node2"))), DiagrammeR syntax.}

\item{.graph_attrs}{Character vector of custom graph attributes for the DOT graph (e.g., c("rankdir=LR", "ranksep=1.0")). DiagrammeR syntax.}

\item{.remove_colors}{Logical. Should the plot be colored? Defaults to \code{FALSE}.}

\item{.edge_lengths}{Integer. Length of edges in the DOT graph.}

\item{.stage}{Character string. Which stage of the two-stage approach for models containing second-order constructs should be displayed?
Defaults to \code{second}.}
}
\value{
A DiagrammeR graph object or a list of DiagrammeR graph objects in case of a multi analysis.
}
\description{
\lifecycle{experimental}
}
\details{
Create a plot of a cSEM model
}
\examples{
\dontrun{
model_Bergami_int="
  # Common factor and composite models
  OrgPres <~ cei1 + cei2 + cei3 + cei4 + cei5 + cei6 + cei7 + cei8 
  OrgIden =~ ma1 + ma2 + ma3 + ma4 + ma5 + ma6
  AffJoy =~ orgcmt1 + orgcmt2 + orgcmt3 + orgcmt7
  AffLove  =~ orgcmt5 + orgcmt6 + orgcmt8

  # Structural model 
  OrgIden ~ OrgPres 
  AffLove ~ OrgPres+OrgIden+OrgPres.OrgIden
  AffJoy  ~ OrgPres+OrgIden
  "
  
  outBergamiInt <- csem(.data = BergamiBagozzi2000,.model = model_Bergami_int,
                        .disattenuate = T,
                        .PLS_weight_scheme_inner = 'factorial',
                        .tolerance = 1e-6,
                        .resample_method = 'none')
  
  plotModel(outBergamiInt)
}
}
\seealso{
\code{\link[=csem]{csem()}}, \link{cSEMResults}
}
