% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_polycor.R
\name{polychor}
\alias{polychor}
\title{Internal: Fast polychoric correlation}
\usage{
polychor(
  x,
  y,
  control = list(),
  maxrho = 0.999,
  start = rawcor(x, y),
  thresholds = FALSE
)
}
\description{
Estimate the polychoric correlation between two ordinal (categorical)
variables \code{x} and \code{y}, i.e., the correlation between the two continuous,
bivariate normal latent variables assumed to underlie \code{x} and \code{y}
\insertCite{Drasgow1988}{cSEM}. Implemented by Kjell S. Slupphaug.
}
\references{
\insertAllCited{}
}
\keyword{internal}
