% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_polycor.R
\name{polyserial}
\alias{polyserial}
\title{Internal: Fast polyserial correlation}
\usage{
polyserial(
  x,
  y,
  control = list(),
  maxrho = 0.999,
  start = rawcor(x, y),
  thresholds = FALSE
)
}
\description{
Estimate the polyserial correlation between a continuous variable \code{x}
and an ordinal (categorical) variable \code{y}, i.e., the correlation between
\code{x} and the continuous latent variable assumed to underlie \code{y}
\insertCite{Drasgow1988}{cSEM}. Implemented by Kjell S. Slupphaug.
}
\references{
\insertAllCited{}
}
\keyword{internal}
