% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_new.R
\name{predict}
\alias{predict}
\title{Implementation of PLS predict adopted from Shmueli et al. (2016) Table 1}
\usage{
predict(.object, testDataset)
}
\description{
Implementation of PLS predict adopted from Shmueli et al. (2016) Table 1
}
