% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igsca.R
\name{prepare_for_ALS}
\alias{prepare_for_ALS}
\title{Prepare for ALS Algorithm}
\usage{
prepare_for_ALS(Z0, W0, B0, n_indicators, n_case, n_constructs, indicator_type)
}
\arguments{
\item{Z0}{}

\item{W0}{}

\item{B0}{}

\item{n_indicators}{}

\item{n_case}{}

\item{n_constructs}{}

\item{indicator_type}{}
}
\value{
List of matrices to put through the Alternating Least Squared (ALS) algorithm:
}
\description{
Internal I-GSCA function
}
