% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igsca.R
\name{prepare_for_ALS}
\alias{prepare_for_ALS}
\title{Prepare for ALS Algorithm}
\usage{
prepare_for_ALS(Z0, W0, B0, n_indicators, n_case, n_constructs, indicator_type)
}
\arguments{
\item{Z0}{Raw data matrix}

\item{W0}{Indicator matrix of weights}

\item{B0}{Indicator matrix of path coefficients}

\item{n_indicators}{Number of indicators}

\item{n_case}{Number of measurements}

\item{n_constructs}{Number of constructs}

\item{indicator_type}{Whether the indicator corresponds to a common factor or composite}
}
\value{
List of matrices to put through the Alternating Least Squared (ALS) algorithm:
}
\description{
Internal I-GSCA function
}
