% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cSEMResults.R
\name{print.cSEMResults}
\alias{print.cSEMResults}
\title{\code{cSEMResults} method for \code{print()}}
\usage{
\method{print}{cSEMResults}(x, ...)
}
\arguments{
\item{x}{Character string. The name of the moderator variable. Defaults to \code{NULL}.}
}
\description{
The \link{cSEMResults} method for the generic function \code{\link[=print]{print()}}.
}
\seealso{
\code{\link[=csem]{csem()}}, \link{cSEMResults}
}
\keyword{internal}
