% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_methods.R
\name{print.cSEMResults}
\alias{print.cSEMResults}
\title{\code{cSEMResults} method for \code{print()}}
\usage{
print(object)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults}.}
}
\description{
The \link{cSEMResults} method for the generic function \code{\link[=print]{print()}}.
}
\details{
Prints basic information on the \link{cSEMResults} object and suggested next
steps to the console.
}
\seealso{
\link{csem}, \link{cSEMResults}
}
