% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_methods.R
\name{print.cSEMResultssummary}
\alias{print.cSEMResultssummary}
\title{\code{cSEMResultssummary} method for \code{print()}}
\usage{
print(.object)
}
\arguments{
\item{...}{Further arguments to be passed down to other functions.
See \code{\link[=args_default]{args_default()}} with \code{.only_dots = TRUE} for a complete list
of possible arguments and their defaults.}
}
\description{
The \link{cSEMResultssummary} method for the generic function \code{\link[=print]{print()}} prints a
a formated summary of the output obtained from running \code{\link[=summary]{summary()}} on
a \link{cSEMResults} object.
}
\seealso{
\link{csem}, \link{cSEMResults}
}
