% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_methods.R
\name{print.cSEMResultssummary}
\alias{print.cSEMResultssummary}
\title{\code{cSEMResultssummary} method for \code{print()}}
\usage{
print(.object)
}
\arguments{
\item{...}{Further arguments to be passed down to other functions.
See \link{args_default} for a complete list of possible arguments and their defaults.}
}
\description{
The \link{cSEMResultssummary} method for the generic function \code{\link[=print]{print()}}.
}
\details{
Prints a summary of the results obtained from runinng \link{csem}, \link{cca}, or
\link{workhorse}.
}
\seealso{
\link{csem}, \link{cSEMResults}
}
