% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_print.R
\name{print.cSEMSummarize}
\alias{print.cSEMSummarize}
\title{\code{cSEMSummarize} method for \code{print()}}
\usage{
\method{print}{cSEMSummarize}(x, .full_output = TRUE, ...)
}
\arguments{
\item{.full_output}{Logical. Should the full ouput of summarize be printed.
Defaults to \code{TRUE}.}
}
\description{
The \link{cSEMSummary} method for the generic function \code{\link[=print]{print()}}.
}
\seealso{
\code{\link[=csem]{csem()}}, \code{\link[=foreman]{foreman()}}, \link{cSEMResults}, \code{\link[=summarize]{summarize()}}
}
\keyword{internal}
