% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_print.R
\name{print.cSEMSummarize_default}
\alias{print.cSEMSummarize_default}
\title{\code{cSEMSummarize_default} method for \code{print()}}
\usage{
print(.object)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}
}
\description{
The \link{cSEMSummary} method for the generic function \code{\link[=print]{print()}}.
}
\seealso{
\code{\link[=csem]{csem()}}, \code{\link[=foreman]{foreman()}}, \link{cSEMResults}, \code{\link[=summarize]{summarize()}}
}
\keyword{internal}
