% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cSEMTestMGD.R
\name{print.cSEMTestMGD}
\alias{print.cSEMTestMGD}
\title{\code{cSEMTestMGD} method for \code{print()}}
\usage{
\method{print}{cSEMTestMGD}(x, ...)
}
\arguments{
\item{x}{Character string. The name of the moderator variable. Defaults to \code{NULL}.}
}
\description{
The \code{cSEMTestMGD} method for the generic function \code{\link[=print]{print()}}.
}
\seealso{
\code{\link[=csem]{csem()}}, \link{cSEMResults}, \code{\link[=testMGD]{testMGD()}}
}
\keyword{internal}
