% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_print.R
\name{print.cSEMTestMICOM}
\alias{print.cSEMTestMICOM}
\title{\code{cSEMTestMICOM} method for \code{print()}}
\usage{
\method{print}{cSEMTestMICOM}(x, ...)
}
\description{
The \code{cSEMTestMICOM} method for the generic function \code{\link[=print]{print()}}.
}
\seealso{
\code{\link[=csem]{csem()}}, \code{\link[=foreman]{foreman()}}, \link{cSEMResults}
}
\keyword{internal}
