% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cSEMVerify.R
\name{print.cSEMVerify}
\alias{print.cSEMVerify}
\title{\code{cSEMVerify} method for \code{print()}}
\usage{
\method{print}{cSEMVerify}(x, ...)
}
\arguments{
\item{x}{Character string. The name of the moderator variable. Defaults to \code{NULL}.}
}
\description{
The \code{cSEMVerify} method for the generic function \code{\link[=print]{print()}}.
}
\seealso{
\code{\link[=csem]{csem()}}, \link{cSEMResults}, \code{\link[=verify]{verify()}}
}
\keyword{internal}
