% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_print.R
\name{print.cSEMVerify_2ndorder}
\alias{print.cSEMVerify_2ndorder}
\title{\code{cSEMVerify_2ndorder} method for \code{print()}}
\usage{
print(.object)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}
}
\description{
The \code{cSEMVerify_default} method for the generic function \code{\link[=print]{print()}}.
}
\seealso{
\code{\link[=csem]{csem()}}, \code{\link[=cca]{cca()}}, \code{\link[=foreman]{foreman()}}, \link{cSEMResults}
}
