% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csem_data.R
\name{processData}
\alias{processData}
\title{Process data}
\usage{
processData(.data, .model)
}
\arguments{
\item{.data}{A \code{data.frame}, a \code{matrix} or a list containing data of either type.
Possible column types or classes of the data provided are:
logical, numeric (double or integer), factor (ordered and unordered)
or a mix of several types. The data may also include
\emph{one} character column whose column name must be given to \code{.id}.
This column is assumed to contain group identifiers used to split the data into groups.}

\item{.model}{A model in \code{\link[lavaan:model.syntax]{lavaan model syntax}}
or a \link{cSEMModel}-list.}
}
\value{
A (N x K) data.frame containing the standardized data with columns ordered
according to the order they appear in the measurement model equations provided
via the \code{.model} argument.
}
\description{
Prepare, standardize, check, and clean data provided via the \code{.data} argument.
}
\examples{
dat <- data.frame(x2 = rnorm(100),
                  x1 = rnorm(100),
                  b2 = rnorm(100),
                  b1 = rnorm(100),
                  a1 = rnorm(100),
                  a2 = rnorm(100)
                  )

model <- '
# Structural model
y1 ~ y2 + y3

# Measurement model
y1 =~ x1 + x2
y2 =~ a1 + a2
y3 =~ b1 + b2
'

dat <- processData(dat, model)

is.matrix(dat) # TRUE
colnames(dat)  # "x1", "x2", "a1", "a2", "b1", "b2"

}
