% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csem_resample.R
\name{resamplecSEMResults}
\alias{resamplecSEMResults}
\title{Resample cSEMResults}
\usage{
resamplecSEMResults(
 .object                = NULL,
 .resample_method       = c("bootstrap", "jackknife), 
 .resample_method2      = c("none", "bootstrap", "jackknife), 
 .R                     = 499,
 .R2                    = 199,
 .handle_inadmissibles  = c("drop", "ignore", "replace"),
 .user_funs             = NULL,
 .eval_plan             = c("sequential", "multiprocess")
)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}

\item{.resample_method}{Character string. The resampling method to use. One of:
"\emph{bootstrap}" or "\emph{jackknife}". Defaults to "\emph{bootstrap}".}

\item{.resample_method2}{Character string. The resampling method to use when resampling
from a resample. One of: "\emph{none}", "\emph{bootstrap}" or "\emph{jackknife}". For
\emph{bootstrap} the number of draws may be provided via \code{.R2}.
Defaults to "\emph{none}".}

\item{.R}{Integer. The number of bootstrap replications. Defaults to \code{499}.}

\item{.R2}{Integer. The number of bootstrap replications to use when
resampling from a resample. Defaults to \code{199}.}

\item{.handle_inadmissibles}{Character string. How should inadmissible results
be treated? One of "\emph{drop}", "\emph{ignore}", or "\emph{replace}". If "\emph{drop}", all
replications/resamples yielding an inadmissible result will be dropped (
number of results shown <= .R). For "\emph{ignore}" all results are returned
even if they are inadmissible (number of results = .R). For "\emph{replace}"
resampling continues until there are exactly .R admissible solutions.
Defaults to "\emph{drop}".}

\item{.user_funs}{A function or a (named) list of functions to apply to every
resample. Takes \code{.object} as an input:
\code{myFun <- function(.object) {...}}.  Output should preferably be a (named)
vector but matrices are also accepted. Note however that output will be
vectorized (columnwise) in this case.}

\item{.eval_plan}{Character string. The evaluation plan to use. One of
\emph{sequential} or \emph{multiprocess}. In the latter case
all available cores will be used. Defaults to \emph{sequential}.}
}
\description{
The function resamples a \link{cSEMResults} object using bootstrap or jackknife resampling.
}
\details{
The function essentially calls \code{\link[=csem]{csem()}} on each of the \emph{M} resamples (created via
\code{\link[=resampleData]{resampleData()}}) and returns M estimates for each of a subset of practically useful
resampled parameters/statistics computed by \code{\link[=csem]{csem()}}. Currently, the following
quantities are computed and returned based on each resample:
\describe{
\item{Parameters}{Path estimates, Weight estimates, Loading estimates}
\item{Statistics}{The heterotrait-monotrait ratio (HTMT)}
}

If the user needs to resample a statistic that is not returned by default,
this statistic can be provided by a function \code{f(.object)} via the \code{.user_fun} argument.
The only accepted argument of this function is \code{.object} which must be an
object of class \link{cSEMResults}. Internally, the function will be applied on each
cSEMResults resample to produce the desired statistic. As long as this is
the only argument of the function provided, arbitrary complicated statistics
may be resampled.

Both resampling the origianl \link{cSEMResults} object (first resample) and resampling based on
a resampled \link{cSEMResults} object (second resample) are supported. Choices for the former
are "bootstrap" and "jackknife". Resampling based on a resample is turned of
by default (\code{.resample_method2 = "none"}) as this significantly
increases computation time (there are now \code{.R * .R2} resamples to compute)
However, in order to compute e.g., the studentized confidence interval via \code{\link[=infer]{infer()}},
resampling from a resample is required.

The number of bootstrap runs for the first and second run are given by \code{.R} and \code{.R}.
The default is \code{499} for the first and \code{199} for the second run
but should be increased in real applications. See e.g.,
\insertCite{Hesterberg2015;textual}{cSEM}, p.380 for recommendations.
For jackknife \code{.R} are \code{.R2} are ignored.

Resampling may produce inadmissble results (as checked by \code{\link[=verify]{verify()}}).
By default these results are dropped however users may choose to \code{"ignore"}
or \code{"replace"} inadmissble results in which case resampling continous until
the necessary number of admissble results is reached.

The cSEM package supports (multi)processing via the \href{https://github.com/HenrikBengtsson/future}{future}
\insertCite{Bengtsson2018}{cSEM} framework. User may simply choose an evaluation plan
via \code{.eval_plan} and the package takes care of all the complicated backend
issues. Currently, users may chose between standard single-core evaluation
(\code{"sequential"}) and multiprocessing (\code{"multiprocess"}). The future package
provides other options (e.g. \code{"cluster"} or \code{"remote"}), however, they probably
wont be needed in the context of the cSEM package. Depeding on the platform, the future
package will manage to distribute tasks to multiple R sessions (Windows)
or multiple cores. Note that multiprocessing is not necessary always faster
when only a small naumber of replications is required.
}
\examples{
\dontrun{
# still to implement
}

}
\references{
\insertAllCited{}
}
\seealso{
\link{csem}, \link{cSEMResults}
}
