% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csem_resample.R
\name{resamplecSEMResults}
\alias{resamplecSEMResults}
\title{Resample cSEMResults}
\usage{
resampleResults(
 .object        = args_default()$.object,
 .data          = NULL,
 .method        = c("bootstrap", "jackknife", "permutation", "cross-validation"),
 .draws         = 499,
 .id            = NULL,
 .cv_folds      = 10      
)
}
\arguments{
\item{.object}{An R object of class \code{cSEM<class>} with corresponding method.}

\item{.draws}{Integer. The number of resamples to take. Defaults to \code{499}.}

\item{.handle_inadmissibles}{Character string. How should inadmissible results
be treated? One of "\emph{drop}", "\emph{ignore}", or "\emph{replace}". If "\emph{drop}", all
replications yielding an inadmissible result will be dropped (=
number of results shown <= .runs). For "\emph{ignore}" all results are returned
even if they are inadmissible (= number of results = .runs). For "\emph{replace}"
bootstrapping continues until there are .runs admissible solutions.
Defaults to "\emph{drop}".}

\item{.method}{Character string. The resampling method to use. One of:
"\emph{bootstrap}" or "\emph{jackknife}". Defaults to "\emph{bootstrap}".}

\item{.verbose}{Logical. Should information be printed to the console? Defaults
to \code{TRUE}.}

\item{.user_funs}{A function or a (named) list of functions to apply to every
resample. Takes \code{.object} as an input:
\code{myFun <- function(.object) {...}}.  Output should preferably be a (named)
vector but matrices are also accepted. Note however that output will be
vectorized (columnwise) in this case.}

\item{.draws2}{Integer. The number of resamples to take when resampling from
a resample. Defaults to \code{NULL}.}

\item{.handle_inadmissibles2}{Character string. How should inadmissible results
be treated when resampling from a resample? For details see
the description for \code{.handle_inadmissibles}.}

\item{.method2}{Character string. The resampling method to use when resampling
from a resample. One of: "\emph{bootstrap}" or "\emph{jackknife}".
Defaults to "\emph{bootstrap}".}
}
\description{
Resample (TODO)
}
\examples{
\dontrun{
# still to implement
}

}
\seealso{
\link{csem}, \link{cSEMResults}
}
