% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_savePlot.R
\name{savePlot}
\alias{savePlot}
\title{savePlot}
\usage{
savePlot(
 .plot_object,
 .filename,
 .path = NULL)
}
\arguments{
\item{.plot_object}{Object returned by one of the following functions \code{\link[=plot.cSEMResults_default]{plot.cSEMResults_default()}}, \code{\link[=plot.cSEMResults_multi]{plot.cSEMResults_multi()}}, or
\code{\link[=plot.cSEMResults_2ndorder]{plot.cSEMResults_2ndorder()}}.}

\item{.filename}{Character string. The name of the file to save the plot to (supports 'pdf', 'png', 'svg', and 'dot' formats).}

\item{.path}{Character string. Path of the directory to save the file to. Defaults
to the current working directory.}
}
\description{
This function saves a given plot of a cSEMResults object to a specified file format.
}
\seealso{
\code{\link[=plot.cSEMResults_default]{plot.cSEMResults_default()}} \code{\link[=plot.cSEMResults_multi]{plot.cSEMResults_multi()}} \code{\link[=plot.cSEMResults_2ndorder]{plot.cSEMResults_2ndorder()}}
}
