% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_auxillary.R
\name{scaleWeights}
\alias{scaleWeights}
\title{Internal: Scale weights}
\usage{
scaleWeights(.S = NULL, .W = NULL)
}
\arguments{
\item{.S}{The (scaled) empirical (K x K) indicator covariance/correlation matrix,
where K is the number of indicators.}

\item{.W}{The (J x K) weight matrix, where J is the number of constructs and
K the number of indicators.}
}
\value{
The (J x K) matrix of scaled weights.
}
\description{
Scales weights to ensure that the proxies have variance of one.
}
\details{
More description here
}
