% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_estimators_weights.R
\name{scaleWeights}
\alias{scaleWeights}
\title{Internal: Scale weights}
\usage{
scaleWeights(
  .S = args_default()$.S, 
  .W = args_default()$.W
  )
}
\arguments{
\item{.S}{The (K x K) empirical indicator correlation matrix.}

\item{.W}{A (J x K) matrix of weights.}
}
\value{
The (J x K) matrix of scaled weights.
}
\description{
Scale weights such that the composite variance is one.
}
\keyword{internal}
