% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_plotModel.R
\name{secondOrderMeasurementEdges}
\alias{secondOrderMeasurementEdges}
\title{Internal: secondOrderMeasurementEdges}
\usage{
secondOrderMeasurementEdges(
 construct,
 weights_first,
 loadings_first,
 weight_p_first,
 loading_p_first,
 weights_second,
 weight_p_second,
 plot_signif, 
 constructTypes,
 only_second_stage = FALSE
 )
}
\value{
Character string. .
}
\description{
Build measurement edges for a second–order model.
}
\keyword{internal}
