% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_foreman.R
\name{setDominantIndicator}
\alias{setDominantIndicator}
\title{Internal: Set the dominant indicator}
\usage{
setDominantIndicator(
 .W                   = W,
 .dominant_indicators = .dominant_indicators
 )
}
\arguments{
\item{.W}{A (J x K) matrix of weights.}

\item{.dominant_indicators}{A character vector of \code{"name" = "value"} pairs,
where \code{"value"} is a character string giving the name of the dominant indicator
and \code{"name"} a character string of the corresponding construct name.
Dominant indicators may be specified for a subset of the constructs.}
}
\value{
The (J x K) matrix of weights with the dominant indicator set.
}
\description{
Set the dominant indicator for each construct.
}
\keyword{internal}
