% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_estimators_weights.R
\name{setStartingValues}
\alias{setStartingValues}
\title{Internal: Set Starting values}
\usage{
setStartingValues(
  .W = args_default()$.W,
  .starting_values = args_default()$.starting_values
  )
}
\arguments{
\item{.W}{A (J x K) matrix of weights.}

\item{.starting_values}{list of named vectors containig the starting values.
Defaults to \code{NULL}.}
}
\value{
The (J x K) matrix of scaled weights.
}
\description{
Returns a matrix with that contains the starting values
}
\keyword{internal}
