% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_estimators_weights.R
\name{setStartingValues}
\alias{setStartingValues}
\title{Internal: Set starting values}
\usage{
setStartingValues(
  .W               = args_default()$.W,
  .starting_values = args_default()$.starting_values
  )
}
\arguments{
\item{.W}{A (J x K) matrix of weights.}

\item{.starting_values}{A named list of vectors where the
list names are the construct names whose indicator weights the user
wishes to set. The vectors must be named vectors of \code{"indicator_name" = value}
pairs, where \code{value} is the (scaled or unscaled) starting weight. Defaults to \code{NULL}.}
}
\value{
The (J x K) matrix of starting values.
}
\description{
Set the starting values.
}
\keyword{internal}
