% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_assess.R
\name{SRMR}
\alias{SRMR}
\title{SRMR}
\usage{
SRMR(.object)
}
\arguments{
\item{.matrix1}{A \code{matrix} to compare.}

\item{.matrix2}{A \code{matrix} to compare.}
}
\description{
Compute the SRMR (TODO)
}
\examples{
\dontrun{
# still to implement
}

}
\seealso{
\link{csem}, \link{cSEMResults}
}
