% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_methods.R
\name{status}
\alias{status}
\alias{status.cSEMResults}
\title{\code{cSEMResults} method to check the validity of the results}
\usage{
status(object)

\method{status}{cSEMResults}(.object)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults}.}
}
\value{
A NULL vector or a vector of status codes pointing to specific problems.
Problems are:
\itemize{
\item 1 Algorithm has not not converged
\item 2: at least one absolute standardized loading estimate is larger than 1,
which implies either a negative veriance of the measurement error or
a correlation larger than 1
\item 3: construct VCV is not positive semi-definit
\item 4 model-implied indicators VCV is not positive semi-definit
}
}
\description{
Check the validity of the estimated quantities.
}
\section{Methods (by class)}{
\itemize{
\item \code{cSEMResults}: Some text
}}

\seealso{
\link{csem}, \link{cSEMResults}
}
