% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Status.R
\name{status}
\alias{status}
\title{Function to check the status of a cSEMResults object}
\usage{
status(object)
}
\arguments{
\item{object}{An R object of class \link{cSEMResults}.}

\item{...}{Further arguments to be passed down to other functions.
See \link{args_default} for a complete list of possible arguments and their defaults.}
}
\description{
It returns a vector where the number indicates the following:
\itemize{
\item 0: Everything is fine
\item 1 Algorithm has not not converged
\item 2: at least one absolute standardized loading estimate is larger than 1,
which implies either a negative veriance of the measurement error or
a correlation larger than 1
\item 3: construct VCV is not positive semi-definit
\item 4 model-implied indicators VCV is not positive semi-definit
}
}
\seealso{
\link{csem}, \link{cSEMResults}
}
