% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_summarize.R
\name{summarize}
\alias{summarize}
\alias{summarize.cSEMResults_default}
\alias{summarize.cSEMResults_multi}
\alias{summarize.cSEMResults_2ndorder}
\title{Summarize model}
\usage{
summarize(.object)

\method{summarize}{cSEMResults_default}(.object = NULL,
  .csem_resample = args_default()$.csem_resample,
  .alpha = args_default()$.alpha,
  .bias_corrected = args_default()$.bias_corrected,
  .statistic = args_default()$.statistic,
  .verbose = args_default()$.verbose)

\method{summarize}{cSEMResults_multi}(.object)

\method{summarize}{cSEMResults_2ndorder}(.object)
}
\arguments{
\item{.object}{An R object of class \code{cSEM<class>} with corresponding method.}

\item{.csem_resample}{A list resulting from a call to \code{\link[=resamplecSEMResults]{resamplecSEMResults()}}.}

\item{.alpha}{An integer or a numeric vector of significance levels.
Defaults to \code{0.05}.}

\item{.bias_corrected}{Logical. Should the standard and the tStat
confidence intervall be bias-corrected using the bootstraped bias estimate?
If \code{TRUE} the confidence intervall for some estimated parameter \code{theta}
is centered at \code{2*theta - theta*_hat},
where \code{theta*_hat} is the average over all .R bootstrap estimates of \code{theta}.
Defaults to \code{TRUE}}

\item{.statistic}{Character string. Which statistic should be returned?
One of (TODO)
Defaults to (TODO).}

\item{.verbose}{Logical. Should information be printed to the console? Defaults
to \code{TRUE}.}
}
\description{
Summarize the model (TODO).
}
\details{
Summary (TODO)
}
\section{Methods (by class)}{
\itemize{
\item \code{cSEMResults_default}: (TODO)

\item \code{cSEMResults_multi}: (TODO)

\item \code{cSEMResults_2ndorder}: (TODO)
}}

\seealso{
\link{csem}, \link{cSEMResults}
}
