% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_summarize.R
\name{summarize}
\alias{summarize}
\title{Summarize model}
\usage{
summarize(.object)
}
\arguments{
\item{.object}{An R object of class \code{cSEM<class>} with corresponding method.}
}
\description{
Summarize the model (TODO).
}
\details{
Summary (TODO)
}
\seealso{
\link{csem}, \link{cSEMResults}
}
