% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_summarize.R
\name{summarize}
\alias{summarize}
\alias{summarize.cSEMResults_default}
\alias{summarize.cSEMResults_multi}
\alias{summarize.cSEMResults_2ndorder}
\title{Summarize model}
\usage{
summarize(
 .object = NULL, 
 .alpha  = args_default()$.alpha,
 .ci     = NULL,
 ...
 )

\method{summarize}{cSEMResults_default}(.object = NULL,
  .alpha = args_default()$.alpha, .ci = NULL, ...)

\method{summarize}{cSEMResults_multi}(.object = NULL,
  .alpha = args_default()$.alpha, .ci = NULL, ...)

\method{summarize}{cSEMResults_2ndorder}(.object = NULL,
  .alpha = args_default()$.alpha, .ci = NULL, ...)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}

\item{.alpha}{An integer or a numeric vector of significance levels.
Defaults to \code{0.05}.}

\item{.ci}{A vector of character strings naming the confidence interval to compute.
For possible choices see \code{\link[=infer]{infer()}}.}
}
\description{
Summarize the model (TODO).
}
\details{
Summary (TODO)
}
\section{Methods (by class)}{
\itemize{
\item \code{cSEMResults_default}: (TODO)

\item \code{cSEMResults_multi}: (TODO)

\item \code{cSEMResults_2ndorder}: (TODO)
}}

\seealso{
\link{csem}, \link{cSEMResults}
}
