% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_methods.R
\name{summary.cSEMResults}
\alias{summary.cSEMResults}
\title{\code{cSEMResults} method for \code{summary()}}
\usage{
summary(.object, .what = NULL)
}
\arguments{
\item{.object}{An R object of class \code{cSEM<class>} with corresponding method.}
}
\value{
An object of class \code{cSEMResultssummary} for which the following methods exist:
\describe{
\item{\code{print.cSEMResultssummary}}{Prints a summary.}
}
Technically \code{cSEMResultssummary} is a named list containing the following list elements:
\describe{
\item{`...}{Not finished yet.}
}

An object of class \code{cSEMResultssummary} for which the following methods exist:
\describe{
\item{\code{print.cSEMResultssummary}}{Prints a summary.}
}
Technically \code{cSEMResultssummary} is a named list containing the following list elements:
\describe{
\item{`...}{Not finished yet.}
}
}
\description{
The \link{cSEMResults} method for the generic function \code{\link[=summary]{summary()}}.
}
\details{
Computes a summary of the results obtained from runinng \link{csem}, \link{cca}, or
\link{foreman}.
}
\seealso{
\link{csem}, \link{cSEMResults}
}
