% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_datasets.R
\docType{data}
\name{Switching}
\alias{Switching}
\title{Data: Switching}
\format{
An object of class \code{data.frame} with 767 rows and 26 columns.
}
\source{
The dataset is provided by Jörg Henseler.
}
\usage{
Switching
}
\description{
A data frame containing 26 variables with 767 observations.
}
\details{
The data contains variables about the consumers’ intention to switch a
service provider. It is also used in \insertCite{Henseler2021;textual}{cSEM}
for demonstration purposes, see the corresponding tutorial.
}
\examples{
#============================================================================
# Example is taken from Henseler (2021)
#============================================================================
model_Int <-"
# Measurement models
INV =~ INV1 + INV2 + INV3 +INV4
SAT =~ SAT1 + SAT2 + SAT3
INT =~ INT1 + INT2

# Structural model containing an interaction term.
INT ~ INV + SAT + INV.SAT
"

out <- csem(.data = Switching, .model = model_Int,
            .PLS_weight_scheme_inner = 'factorial',
            .tolerance = 1e-06)

}
\references{
\insertAllCited{}
}
\keyword{datasets}
