% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_test_hausman.R
\name{testHausman}
\alias{testHausman}
\alias{testHausman.cSEMResults_default}
\alias{testHausman.cSEMResults_multi}
\alias{testHausman.cSEMResults_2ndorder}
\title{Hausman test}
\usage{
testHausman(
 .object         = NULL,
 .seed           = args_default()$.seed,
 .alpha          = args_default()$.alpha,
 .R              = args_default()$.R,
 .R2             = args_default()$.R2,
 .vcv_asymptotic = args_default()$.vcv_asymptotic
 )

\method{testHausman}{cSEMResults_default}(.object = NULL,
  .seed = args_default()$.seed, .alpha = args_default()$.alpha,
  .R = args_default()$.R, .R2 = args_default()$.R2,
  .vcv_asymptotic = args_default()$.vcv_asymptotic)

\method{testHausman}{cSEMResults_multi}(.object = NULL,
  .seed = args_default()$.seed, .alpha = args_default()$.alpha,
  .R = args_default()$.R, .R2 = args_default()$.R2,
  .vcv_asymptotic = args_default()$.vcv_asymptotic)

\method{testHausman}{cSEMResults_2ndorder}(.object = NULL,
  .seed = args_default()$.seed, .alpha = args_default()$.alpha,
  .R = args_default()$.R, .R2 = args_default()$.R2,
  .vcv_asymptotic = args_default()$.vcv_asymptotic)
}
\arguments{
\item{.object}{A 2SLS or 3SLS object}

\item{.seed}{Integer or NULL. The random seed to use. Defaults to \code{NULL} in which
case an arbitrary seed is choosen.}

\item{.alpha}{An integer or a numeric vector of significance levels.
Defaults to \code{0.05}.}

\item{.R}{Integer. The number of bootstrap replications. Defaults to \code{499}.}

\item{.R2}{Integer. The number of bootstrap replications to use when
resampling from a resample. Defaults to \code{199}.}

\item{.vcv_asymptotic}{Logical. Should the asymptotic variance-covariance matrix be used, i.e.,
VCV(b0) - VCV(b1)= VCV(b1-b0), or should VCV(b1-b0) directly computed.
Defaults to \code{FALSE}.}
}
\value{
A standardized list of class \code{cSEMTest}. Technically \code{cSEMTest} is a named
list containing the following list elements:
\describe{
\item{\code{$Test_statistic}}{The value of test statistic(s).}
\item{\code{$Critical_value}}{The critical value(s).}
\item{\code{$Decision}}{The test decision. One of: \strong{Reject} or \strong{Do not reject}}
\item{\code{$Number_admissibles}}{The number of admissible runs. See \code{\link[=verify]{verify()}} for
what constitues and inadmissible run.}
}
}
\description{
Calculates a bootstrap-based Hausman test that can be used to compare
OLS to 2SLS estimates
or 2SLS to 3SLS estimates \insertCite{Hausman1978,Wong1996}{cSEM}
(TODO: Needs to be checked whether the implemention is correct, I doubt).
}
\section{Methods (by class)}{
\itemize{
\item \code{cSEMResults_default}: (TODO)

\item \code{cSEMResults_multi}: (TODO)

\item \code{cSEMResults_2ndorder}: (TODO)
}}

\examples{
\dontrun{
# TODO
}

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=csem]{csem()}}, \code{\link[=foreman]{foreman()}}, \link{cSEMResults}
}
