% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_test_hausman.R
\name{testHausman}
\alias{testHausman}
\title{Regression-based Hausman test}
\usage{
testHausman(
 .object               = NULL,
 .alpha                = args_default()$.alpha,
 .eval_plan            = args_default()$.eval_plan,
 .handle_inadmissibles = args_default()$.handle_inadmissibles,
 .R                    = args_default()$.R,
 .resample_method      = args_default()$.resample_method,
 .seed                 = args_default()$.seed
 )
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}

\item{.alpha}{An integer or a numeric vector of significance levels.
Defaults to \code{0.05}.}

\item{.eval_plan}{Character string. The evaluation plan to use. One of
"\emph{sequential}" or "\emph{multiprocess}". In the latter case
all available cores will be used. Defaults to "\emph{sequential}".}

\item{.handle_inadmissibles}{Character string. How should inadmissible results
be treated? One of "\emph{drop}", "\emph{ignore}", or "\emph{replace}". If "\emph{drop}", all
replications/resamples yielding an inadmissible result will be dropped
(i.e. the number of results returned will be less than .R). For "\emph{ignore}"
all results are returned even if they are inadmissible (i.e.
number of results returned = .R). For "\emph{replace}" resampling continues until
there are exactly .R admissible solutions.
Defaults to "\emph{drop}".}

\item{.R}{Integer. The number of bootstrap replications. Defaults to \code{499}.}

\item{.resample_method}{Character string. The resampling method to use. One of:
"\emph{none}", "\emph{bootstrap}" or "\emph{jackknife}". Defaults to "\emph{none}".}

\item{.seed}{Integer or NULL. The random seed to use. Defaults to \code{NULL} in which
case an arbitrary seed is choosen.}
}
\description{
Calculates the regression-based Hausman test to be used to compare
OLS to 2SLS estimates or 2SLS to 3SLS estimates.
}
\examples{
\dontrun{
# TODO
}

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=csem]{csem()}}, \code{\link[=foreman]{foreman()}}, \link{cSEMResults}
}
