% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_test_MGD.R
\name{testMGD}
\alias{testMGD}
\title{Tests for multi-group comparisons}
\usage{
testMGD(
 .object                = NULL,
 .alpha                 = 0.05,
 .approach_p_adjust     = "none",
 .approach_mgd          = c("all", "Klesel", "Chin", "Sarstedt", "Keil", "Nitzl"),
 .parameters_to_compare = NULL,
 .handle_inadmissibles  = c("replace", "drop", "none"),
 .R_permutation         = 499,
 .R_bootstrap           = 499,
 .saturated             = FALSE,
 .seed                  = NULL,
 .type_vcv              = c("indicator", "construct"),
 .verbose               = TRUE
 )
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}

\item{.alpha}{An integer or a numeric vector of significance levels.
Defaults to \code{0.05}.}

\item{.approach_p_adjust}{Character string or a vector of character strings.
Approach used to adjust the p value in multiple testing.
See the \code{methods} argument of \code{\link[stats:p.adjust]{stats::p.adjust()}} for a list of choices and
their description. Defaults to "\emph{none}".}

\item{.approach_mgd}{Character string or a vector of character strings.
Approach used for the multi-group comparison. One: "\emph{all}", "\emph{Klesel}", "\emph{Chin}",
"\emph{Sarstedt}", "\emph{Keil}, or "\emph{Nitzl}". Default to "\emph{all}" in which case all approaches are
computed (if possible).}

\item{.parameters_to_compare}{A model in \link[lavaan:model.syntax]{lavaan model syntax} indicating which
parameters (i.e, path (\code{~}), loadings (\code{=~}), or weights (\code{<~})) should be
compared across groups. Defaults to \code{NULL} in which case all parameters of the
originally specified model are compared.}

\item{.handle_inadmissibles}{Character string. How should inadmissible results
be treated? One of "\emph{drop}", "\emph{ignore}", or "\emph{replace}". If "\emph{drop}", all
replications/resamples yielding an inadmissible result will be dropped
(i.e. the number of results returned will potentially be less than \code{.R}).
For "\emph{ignore}" all results are returned even if all or some of the replications
yieled inadmissible results (i.e. number of results returned is equal to \code{.R}).
For "\emph{replace}" resampling continues until there are exactly \code{.R} admissible solutions.
Defaults to "\emph{replace}" to accomodate all approaches.}

\item{.R_permutation}{Integer. The number of permutations. Defaults to \code{499}}

\item{.R_bootstrap}{Integer. The number of bootstrap runs. Ignored if \code{.object}
contains resamples. Defaults to \code{499}}

\item{.saturated}{Logical. Should a saturated structural model be used?
Defaults to \code{FALSE}.}

\item{.seed}{Integer or \code{NULL}. The random seed to use. Defaults to \code{NULL} in which
case an arbitrary seed is choosen. Note that the scope of the seed is limited
to the body of the function it is used in. Hence, the global seed will
not be altered!}

\item{.type_vcv}{Character string. Indicates which model-implied correlation matrix is calcuted
One of "\emph{indicator}" or "\emph{construct}". Defaults to "\emph{indicator}".}

\item{.verbose}{Logical. Should information (e.g., progress bar) be printed to the console? Defaults
to \code{TRUE}.}
}
\value{
A list of class \code{cSEMTestMGD}. Technically, \code{cSEMTestMGD} is a
named list containing the following list elements:

\describe{
\item{\code{$Information}}{Additional information.}
\item{\code{$Klesel}}{A list with elements, \code{Test_statistic}, \code{P_value}, and \code{Decision}}
\item{\code{$Chin}}{A list with elements, \code{Test_statistic}, \code{P_value}, \code{Decision}, and \code{Decision_overall}}
\item{\code{$Sarstedt}}{A list with elements, \code{Test_statistic}, \code{P_value}, \code{Decision}, and \code{Decision_overall}}
\item{\code{$Keil}}{A list with elements, \code{Test_statistic}, \code{P_value}, \code{Decision}, and \code{Decision_overall}}
\item{\code{$Nitzl}}{A list with elements, \code{Test_statistic}, \code{P_value}, \code{Decision}, and \code{Decision_overall}}
}
}
\description{
This function performs several permutation tests, i.e., the reference distribution
of the test statistic is obtained by permutation.
}
\details{
The following tests are implemented:
\describe{
\item{Approach suggested by \insertCite{Klesel2019;textual}{cSEM}}{
The model-implied variance-covariance matrix (either indicator
\code{.type_vcv = "indicator"} or construct \code{.type_vcv = "construct"})
is compared across groups.

To measure the distance between the model-implied variance-covariance matrices,
the geodesic distance (dG) and the squared Euclidean distance (dL) are used.
If more than two groups are compared, the average distance over all groups
is used.}
\item{Approach suggested by \insertCite{Sarstedt2011;textual}{cSEM}}{
Groups are compared in terms of parameter differences across groups.
\insertCite{Sarstedt2011;textual}{cSEM} tests if parameter k is equal
across all groups. If several parameters are tested simulaneously
it is recommended to adjust the signficance level or the p-values (in \pkg{cSEM} correction is
done by p-value). By default
no multiple testing correction is done, however, several common
adjustments are available via \code{.approach_p_adjust}. See
\code{\link[stats:p.adjust]{stats::p.adjust()}} for details.
}
\item{Approach suggested by \insertCite{Chin2010;textual}{cSEM}}{
Groups are compared in terms of parameter differences across groups.
\insertCite{Chin2010;textual}{cSEM} tests if parameter k is equal
between two groups. If more than two groups are tested the parameter is compared
between all pairs of groups. In this case, it is recommended
to adjust the signficance level or the p-values (in \pkg{cSEM} correction is
done by p-value). If several parameters are tested simultaneously, correction
is by group and number of parameters. By default
no multiple testing correction is done, however, several common
adjustments are available via \code{.approach_p_adjust}. See
\code{\link[stats:p.adjust]{stats::p.adjust()}} for details.
}
\item{Approach suggested by \insertCite{Keil2000;textual}{cSEM}}{
Groups are compared in terms of parameter differences across groups.
\insertCite{Keil2000;textual}{cSEM} tests if parameter k is equal
between two groups. It is assumed, that the standard errors of the coefficients are
equal across groups. The calculation of the standard error of the parameter
difference is adjusted as proposed by \insertCite{Henseler2009;textual}{cSEM}.
If more than two groups are tested the parameter is compared
between all pairs of groups. In this case, it is recommended
to adjust the signficance level or the p-values (in \pkg{cSEM} correction is
done by p-value). If several parameters are tested simultaneously, correction
is by group and number of parameters. By default
no multiple testing correction is done, however, several common
adjustments are available via \code{.approach_p_adjust}. See
\code{\link[stats:p.adjust]{stats::p.adjust()}} for details.
}
\item{Approach suggested by \insertCite{Nitzl2010;textual}{cSEM}}{
Groups are compared in terms of parameter differences across groups.
Similarly to \insertCite{Keil2000;textual}{cSEM}, a single parameter k is tested
whether it is equal between two groups. In contrast to \insertCite{Keil2000;textual}{cSEM},
it is assumed, that the standard errors of the coefficients are
equal across groups \insertCite{Sarstedt2011}{cSEM}.
If more than two groups are tested the parameter is compared
between all pairs of groups. In this case, it is recommended
to adjust the signficance level or the p-values (in \pkg{cSEM} correction is
done by p-value). If several parameters are tested simultaneously, correction
is by group and number of parameters. By default
no multiple testing correction is done, however, several common
adjustments are available via \code{.approach_p_adjust}. See
\code{\link[stats:p.adjust]{stats::p.adjust()}} for details.
}
}

Use \code{.approach_mgd} to choose the approach. By default all approaches are computed
(\code{.approach_mgd = "all"}).

By default, approaches based on parameter differences across groups compare
all parameters (\code{.parameters_to_compare = NULL}). To compare only
a subset of parameters provide the parameters in lavaan model syntax just like
providing the model. Note that the "model" provided to \code{.parameters_to_compare}
does not have to be an estimatable model! See the example below.

Note that compared to all other functions in \pkg{cSEM}, \code{.handle_inadmissibles}
defaults to \code{"replace"} to accomdate the Sarstedt et al. (2011) approach.
}
\examples{
# ===========================================================================
# Basic usage
# ===========================================================================
model <- "
# Structural model
QUAL ~ EXPE
EXPE ~ IMAG
SAT  ~ IMAG + EXPE + QUAL + VAL
LOY  ~ IMAG + SAT
VAL  ~ EXPE + QUAL

# Measurement model

EXPE <~ expe1 + expe2 + expe3 + expe4 + expe5
IMAG <~ imag1 + imag2 + imag3 + imag4 + imag5
LOY  =~ loy1  + loy2  + loy3  + loy4
QUAL =~ qual1 + qual2 + qual3 + qual4 + qual5
SAT  <~ sat1  + sat2  + sat3  + sat4
VAL  <~ val1  + val2  + val3  + val4
"

## Create list of virtually identical data sets
dat <- list(satisfaction[-3,], satisfaction[-5, ], satisfaction[-10, ])
out <- csem(dat, model, .resample_method = "bootstrap", .R = 20) 

## Test 
testMGD(out, .R_permutation = 20)

# Notes: 
#  1. .R_permutation (and .R in the call to csem) is small to save time; should be
#     higher in real applications.
#  2. Test will not reject their respective H0s since the groups are virtually
#     identical.
#  3. Only exception is the approach suggested by Sarstedt et al. (2011), a
#     sign that the test is unreliable.
#  4. As opposed to other functions involing the argument, 
#     '.handle_inadmissibles' the default is "replace" as this is
#     required by Sarstedt et al. (2011)'s approach.

# ===========================================================================
# Extended usage
# ===========================================================================
### Test only a subset ------------------------------------------------------
# By default all parameters are compared. Select a subset by providing a 
# model in lavaan model syntax:

to_compare <- "
# Path coefficients
QUAL ~ EXPE

# Loadings
EXPE <~ expe1 + expe2 + expe3 + expe4 + expe5
"

## Test 
testMGD(out, .parameters_to_compare = to_compare, .R_permutation = 20, .R_bootstrap = 20)

### Different p_adjustments --------------------------------------------------
# To adjust p-values to accomodate multiple testing use .approach_p_adjust. 
# The number of tests to use for adjusting depends on the approach choosen. For
# the Chin approach for example it is the number of parameters to test times the
# number of possible group comparisons. To compare the results for different
# adjustments, a vector of p-adjustments may be choosen.

## Test 
testMGD(out, .parameters_to_compare = to_compare, 
        .approach_p_adjust = c("none", "bonferroni"),
        .R_permutation = 20, .R_bootstrap = 20)

}
\references{
\insertAllCited{}
}
\seealso{
\link{cSEMResults}
}
