% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_test_MGD.R
\name{testMGD}
\alias{testMGD}
\title{Estimates the distance between multiple groups.}
\usage{
testOH(
 .object             = args_default()$.model,
 .alpha              = args_default()$.alpha,
 .drop_inadmissibles = args_default()$.drop_inadmissibles,
 .parallel           = args_default()$.parallel,
 .runs               = args_default()$.runs,
 .show_progress      = args_default()$.show_progress
 )
}
\arguments{
\item{.object}{An R object of class \code{cSEM<class>} with corresponding method.}

\item{.alpha}{An integer or a numeric vector of significance levels.
Defaults to \code{0.05}.}

\item{.drop_inadmissibles}{Logical. Should inadmissible solutions be dropped?
Defaults to \code{TRUE}.}

\item{.parallel}{Logical. Use parallel computing. Defaults to \code{FALSE}. Note:
requires the \code{doSNOW} and the \code{parallel} package to be installed.}

\item{.runs}{Integer. How many runs should be performed? Defaults to \code{499}.}

\item{.show_progress}{Logical. Show progress bar. Defaults to \code{TRUE}.}
}
\value{
A standardized list of class \code{cSEMTest}. Technically \code{cSEMTest} is a named
list containing the following list elements:
\describe{
\item{\code{Hypothesis}}{The null hypothesis.}
\item{\code{$Test_statistic}}{The value of test statistic(s).}
\item{\code{$Critical_value}}{The critical value(s).}
\item{\code{$Decision}}{The test decision. One of: \strong{Reject} or \strong{Do not reject}}
\item{\code{$Number_admissibles}}{The number of admissible runs. See \code{\link[=verify]{verify()}} for
what constitues and inadmissble run.}
}
}
\description{
The geodesic distance (dG) and the Euclidean distance (dE) is used
to assess group differences. Permutation is used to generate a
reference distribtuion.
}
\examples{
\dontrun{
require(cSEM)
data(satisfaction)

model <- "
# Structural model
QUAL ~ EXPE
EXPE ~ IMAG
SAT  ~ IMAG + EXPE + QUAL + VAL
LOY  ~ IMAG + SAT
VAL  ~ EXPE + QUAL

# Measurement model

EXPE <~ expe1 + expe2 + expe3 + expe4 + expe5
IMAG <~ imag1 + imag2 + imag3 + imag4 + imag5
LOY  =~ loy1  + loy2  + loy3  + loy4
QUAL =~ qual1 + qual2 + qual3 + qual4 + qual5
SAT  <~ sat1  + sat2  + sat3  + sat4
VAL  <~ val1  + val2  + val3  + val4
"

listData <- list(satisfaction[-3,], satisfaction[-5, ], satisfaction[-10, ])
out.cSEM <- csem(listData, model) 

testMGD(.object = out.cSEM, .runs = 20, .parallel = TRUE)
}

}
\seealso{
\link{cSEMResults}
}
