% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_test_MGD.R
\name{testMGD}
\alias{testMGD}
\alias{testMGD.cSEMResults_default}
\alias{testMGD.cSEMResults_multi}
\alias{testMGD.cSEMResults_2ndorder}
\title{Estimates the distance between multiple groups.}
\usage{
testMGD(
 .object             = args_default()$.object,
 .alpha              = args_default()$.alpha,
 .drop_inadmissibles = args_default()$.drop_inadmissibles,
 .runs               = args_default()$.runs,
 .saturated          = args_default()$.saturated,
 .type_vcv           = args_default()$.type_vcv
 .verbose            = args_default()$.verbose
 )

\method{testMGD}{cSEMResults_default}(.object = args_default()$.object)

\method{testMGD}{cSEMResults_multi}(.object = args_default()$.object,
  .alpha = args_default()$.alpha,
  .handle_inadmissibles = args_default()$.handle_inadmissibles,
  .runs = args_default()$.runs, .saturated = args_default()$.saturated,
  .type_vcv = args_default()$.type_vcv,
  .verbose = args_default()$.verbose)

\method{testMGD}{cSEMResults_2ndorder}(.object = args_default()$.object,
  .alpha = args_default()$.alpha,
  .handle_inadmissibles = args_default()$.handle_inadmissibles,
  .runs = args_default()$.runs, .saturated = args_default()$.saturated,
  .type_vcv = args_default()$.type_vcv,
  .verbose = args_default()$.verbose)
}
\arguments{
\item{.object}{An R object of class \code{cSEM<class>} with corresponding method.}

\item{.alpha}{An integer or a numeric vector of significance levels.
Defaults to \code{0.05}.}

\item{.handle_inadmissibles}{Character string. How should inadmissible results
be treated? One of "\emph{drop}", "\emph{ignore}", or "\emph{replace}". If "\emph{drop}", all
replications yielding an inadmissible result will be dropped (=
number of results shown <= .runs). For "\emph{ignore}" all results are returned
even if they are inadmissible (= number of results = .runs). For "\emph{replace}"
bootstrapping continues until there are .runs admissible solutions.
Defaults to "\emph{drop}".}

\item{.runs}{Integer. How many runs should be performed? Defaults to \code{499}.}

\item{.saturated}{Logical. Should a saturated structural model be used? Defaults to \code{FALSE}.}

\item{.type_vcv}{Character string. Indicates which model-implied correlation matrix is calcuted
One of "\emph{indicator}" or "\emph{construct}". Defaults to "\emph{indicator}".}

\item{.verbose}{Logical. Should information be printed to the console? Defaults
to \code{TRUE}.}
}
\value{
A standardized list of class \code{cSEMTest}. Technically \code{cSEMTest} is a named
list containing the following list elements:
\describe{
\item{\code{$Test_statistic}}{The value of test statistic(s).}
\item{\code{$Critical_value}}{The critical value(s).}
\item{\code{$Decision}}{The test decision. One of: \strong{Reject} or \strong{Do not reject}}
\item{\code{$Number_admissibles}}{The number of admissible runs. See \code{\link[=verify]{verify()}} for
what constitues and inadmissible run.}
}
}
\description{
The geodesic distance (dG) and the Euclidean distance (dE) is used
to assess group differences. Permutation is used to generate a
reference distribtuion.
}
\section{Methods (by class)}{
\itemize{
\item \code{cSEMResults_default}: (TODO)

\item \code{cSEMResults_multi}: (TODO)

\item \code{cSEMResults_2ndorder}: (TODO)
}}

\examples{
\dontrun{
require(cSEM)
data(satisfaction)

model <- "
# Structural model
QUAL ~ EXPE
EXPE ~ IMAG
SAT  ~ IMAG + EXPE + QUAL + VAL
LOY  ~ IMAG + SAT
VAL  ~ EXPE + QUAL

# Measurement model

EXPE <~ expe1 + expe2 + expe3 + expe4 + expe5
IMAG <~ imag1 + imag2 + imag3 + imag4 + imag5
LOY  =~ loy1  + loy2  + loy3  + loy4
QUAL =~ qual1 + qual2 + qual3 + qual4 + qual5
SAT  <~ sat1  + sat2  + sat3  + sat4
VAL  <~ val1  + val2  + val3  + val4
"

listData <- list(satisfaction[-3,], satisfaction[-5, ], satisfaction[-10, ])
out.cSEM <- csem(listData, model) 

testMGD(.object = out.cSEM, .runs = 20, .type_vcv= 'construct')
}

}
\seealso{
\link{cSEMResults}
}
