% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_test_MICOM.R
\name{testMICOM}
\alias{testMICOM}
\title{Test measurement invariance of composites}
\usage{
testMICOM(
  .data             = NULL,
  .model            = NULL,
  .group_var        = NULL,
  .approach_weights = c("PLS", "SUMCORR", "MAXVAR", "SSQCORR", "MINVAR", "GENVAR",
                        "GSCA", "fixed", "unit")
  .permutations     = 100,
  .alpha            = c(0.1, 0.05, 0.01),
  .verbose          = TRUE,
  ...)
}
\arguments{
\item{.data}{A \code{data.frame} or a \code{matrix} containing the raw data.}

\item{.model}{A model in \code{\link[lavaan:model.syntax]{lavaan model syntax}}
or a \link{cSEMModel}-list.}

\item{.approach_weights}{Character string. Approach used to
obtain composite weights. One of: "\emph{PLS}", "\emph{SUMCORR}", "\emph{MAXVAR}",
"\emph{SSQCORR}", "\emph{MINVAR}", "\emph{GENVAR}", "\emph{GSCA}", "\emph{fixed}", or "\emph{unit}".
Defaults to "\emph{PLS}".}

\item{.PLS_weight_scheme_inner}{Character string. The inner weighting scheme
used in PLS. One of: "\emph{centroid}", "\emph{factorial}", or "\emph{path}".
Defaults to "\emph{centroid}". Ignored if \code{.approach_weight} is not PLS.}

\item{.runs}{Integer. How many runs should be performed? Defaults to \code{499}.}

\item{.alpha}{An integer or a numeric vector of significance levels.
Defaults to \code{0.05}.}

\item{.verbose}{Logical. Should information be printed to the console? Defaults
to \code{TRUE}.}
}
\value{
An object of class "testMICOM"

An object of class "testMICOM" is a named list containing the following list element:
\describe{
\item{\strong{Step2}}{A list containing the results of the test for compositional invariance (Step 2).}
\item{\strong{Step3}}{A list containing the results of the test for mean and variance equality (Step 3).}
\item{\strong{Meta_information}}{A list of additional information on the test.}
}
}
\description{
This functions performs the test for measurement invariance of composites
proposed by Henseler et al. (2016).
}
\details{
The test is only meaningful for composite models.

If more than two groups are to be compared issues related to multiple testing
should be taken into account.

The number of permutation runs is defaults to 100 for performance reasons.
According to Henseler et al. (2016) the number of permutations should be at least 5000 for
assessment to be reliable.
}
\examples{

still to implement

}
