% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_test_MICOM.R
\name{testMICOM}
\alias{testMICOM}
\alias{testMICOM.cSEMResults_default}
\alias{testMICOM.cSEMResults_multi}
\alias{testMICOM.cSEMResults_2ndorder}
\title{Test measurement invariance of composites}
\usage{
testMICOM(
 .object               = args_default()$.object,
 .alpha                = args_default()$.alpha,
 .handle_inadmissibles = args_default()$.handle_inadmissibles,
 .R                    = args_default()$.R,
 .seed                 = args_default()$.seed,
 .verbose              = args_default()$.verbose
 )


  \method{testMICOM}{cSEMResults_default}(.object = args_default()$.object,
  .alpha = args_default()$.alpha,
  .handle_inadmissibles = args_default()$.handle_inadmissibles,
  .R = args_default()$.R, .seed = args_default()$.seed,
  .verbose = args_default()$.verbose)

\method{testMICOM}{cSEMResults_multi}(.object = args_default()$.object,
  .alpha = args_default()$.alpha,
  .handle_inadmissibles = args_default()$.handle_inadmissibles,
  .R = args_default()$.R, .seed = args_default()$.seed,
  .verbose = args_default()$.verbose)


  \method{testMICOM}{cSEMResults_2ndorder}(.object = args_default()$.object,
  .alpha = args_default()$.alpha,
  .handle_inadmissibles = args_default()$.handle_inadmissibles,
  .R = args_default()$.R, .seed = args_default()$.seed,
  .verbose = args_default()$.verbose)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}

\item{.alpha}{An integer or a numeric vector of significance levels.
Defaults to \code{0.05}.}

\item{.handle_inadmissibles}{Character string. How should inadmissible results
be treated? One of "\emph{drop}", "\emph{ignore}", or "\emph{replace}". If "\emph{drop}", all
replications/resamples yielding an inadmissible result will be dropped
(i.e. the number of results returned will be less than .R). For "\emph{ignore}"
all results are returned even if they are inadmissible (i.e.
number of results returned = .R). For "\emph{replace}" resampling continues until
there are exactly .R admissible solutions.
Defaults to "\emph{drop}".}

\item{.R}{Integer. The number of bootstrap replications. Defaults to \code{499}.}

\item{.seed}{Integer or NULL. The random seed to use. Defaults to \code{NULL} in which
case an arbitrary seed is choosen.}

\item{.verbose}{Logical. Should information be printed to the console? Defaults
to \code{TRUE}.}
}
\value{
An object of class "testMICOM"

An object of class "testMICOM" is a named list containing the following list element:
\describe{
\item{\strong{Step2}}{A list containing the results of the test for compositional invariance (Step 2).}
\item{\strong{Step3}}{A list containing the results of the test for mean and variance equality (Step 3).}
\item{\strong{Meta_information}}{A list of additional information on the test.}
}
}
\description{
This functions performs the test for measurement invariance of composites
proposed by \insertCite{Henseler2016}{cSEM}.
}
\details{
The test is only meaningful for concepts modeled as composites.

If more than two groups are to be compared issues related to multiple testing
should be taken into account.

The number of permutation runs defaults to \code{args_default()$.R} for performance reasons.
According to \insertCite{Henseler2016}{cSEM} the number of permutations should
be at least 5000 for assessment to be sufficiently reliable.
}
\section{Methods (by class)}{
\itemize{
\item \code{cSEMResults_default}: (TODO)

\item \code{cSEMResults_multi}: (TODO)

\item \code{cSEMResults_2ndorder}: (TODO)
}}

\references{
\insertAllCited{}
}
