% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_MICOM.R
\name{testMICOM}
\alias{testMICOM}
\title{Test measurement invariance of composites}
\usage{
testMICOM(
  .data             = NULL,
  .model            = NULL,
  .group_var        = NULL,
  .approach_weights = c("PLS", "SUMCOR", "MAXVAR", "SSQCOR", "MINVAR", "GENVAR",
                        "GSCA", "fixed", "unit")
  .permutations     = 100,
  .alpha            = c(0.1, 0.05, 0.01),
  .verbose          = TRUE,
  ...)
}
\arguments{
\item{.data}{A data frame or a matrix containing the raw data.}

\item{.model}{A model in \href{http://lavaan.ugent.be/tutorial/syntax1.html}{lavaan model syntax}
or a \link{cSEMModel} object.}

\item{.group_var}{Character string. The name of the column used to split the data into groups.}

\item{.approach_weights}{Character string. The name of the approach used to
obtain composite weights. Possible choices are: "\emph{PLS}", "\emph{SUMCOR}", "\emph{MAXVAR}",
"\emph{SSQCOR}", "\emph{MINVAR}", "\emph{GENVAR}", "\emph{GSCA}", "\emph{fixed}", or "\emph{unit}".
Defaults to "\emph{PLS}".}

\item{.PLS_weight_scheme_inner}{Character string. The inner weighting scheme
used in PLS. Possible choices are: "\emph{centroid}", "\emph{factorial}", or "\emph{path}".
Defaults to "\emph{centroid}". The argument is ignored if \code{.approach_weight} is not PLS.}

\item{.permutations}{Integer. The number permutations used for step 2 and 3 of the test.
Defaults to \code{100}. Note however that the number should typically be a lot higher.}

\item{.alpha}{A numeric vector of significance levels. Defaults to \code{c(0.1, 0.05, 0.01)}.}

\item{.verbose}{Logical. Should information be printed to the console ?}

\item{...}{Further arguments to be passed down to other functions.
See \link{args_default} for a complete list of possible arguments and their defaults.}
}
\value{
An object of class "testMICOM"

An object of class "testMICOM" is a named list containing the following list elemement:
\describe{
\item{\strong{Step2}}{A list containing the results of the test for composistional invariance (Step 2).}
\item{\strong{Step3}}{A list containing the results of the test for mean and variance equality (Step 3).}
\item{\strong{Meta_information}}{A list of additional information on the test.}
}
}
\description{
This functions performs the test for measurement invariance of composites
proposed by Henseler et al. (2016).
}
\details{
The test is only meaningful for composite models.

If more than two groups are to be compared issus related to multiple testing
should be taken into account.

The number of permutations is defaults to 100 for performance reasons.
According to Henseler et al. (2016) the number of permutations should be at least 5000 for
assessement to be reliable.
}
\examples{

still to implement

}
