% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_test_MICOM.R
\name{testMICOM}
\alias{testMICOM}
\alias{testMICOM.cSEMResults_default}
\alias{testMICOM.cSEMResults_multi}
\alias{testMICOM.cSEMResults_2ndorder}
\title{Test measurement invariance of composites}
\usage{
testMICOM(
 .object               = args_default()$.object,
 .alpha                = args_default()$.alpha,
 .handle_inadmissibles = args_default()$.handle_inadmissibles,
 .R                    = args_default()$.R,
 .verbose              = args_default()$.verbose
 )


  \method{testMICOM}{cSEMResults_default}(.object = args_default()$.object,
  .alpha = args_default()$.alpha,
  .handle_inadmissibles = args_default()$.handle_inadmissibles,
  .R = args_default()$.R, .verbose = args_default()$.verbose)

\method{testMICOM}{cSEMResults_multi}(.object = args_default()$.object,
  .alpha = args_default()$.alpha,
  .handle_inadmissibles = args_default()$.handle_inadmissibles,
  .R = args_default()$.R, .verbose = args_default()$.verbose)


  \method{testMICOM}{cSEMResults_2ndorder}(.object = args_default()$.object,
  .alpha = args_default()$.alpha,
  .handle_inadmissibles = args_default()$.handle_inadmissibles,
  .R = args_default()$.R, .verbose = args_default()$.verbose)
}
\arguments{
\item{.object}{An R object of class \code{cSEM<class>} with corresponding method.}

\item{.alpha}{An integer or a numeric vector of significance levels.
Defaults to \code{0.05}.}

\item{.handle_inadmissibles}{Character string. How should inadmissible results
be treated? One of "\emph{drop}", "\emph{ignore}", or "\emph{replace}". If "\emph{drop}", all
replications/resamples yielding an inadmissible result will be dropped (
number of results shown <= .R). For "\emph{ignore}" all results are returned
even if they are inadmissible (number of results = .R). For "\emph{replace}"
resampling continues until there are exactly .R admissible solutions.
Defaults to "\emph{drop}".}

\item{.R}{Integer. The number of bootstrap or permuation replications
to use. Defaults to \code{499}.}

\item{.verbose}{Logical. Should information be printed to the console? Defaults
to \code{TRUE}.}
}
\value{
An object of class "testMICOM"

An object of class "testMICOM" is a named list containing the following list element:
\describe{
\item{\strong{Step2}}{A list containing the results of the test for compositional invariance (Step 2).}
\item{\strong{Step3}}{A list containing the results of the test for mean and variance equality (Step 3).}
\item{\strong{Meta_information}}{A list of additional information on the test.}
}
}
\description{
This functions performs the test for measurement invariance of composites
proposed by Henseler et al. (2016).
}
\details{
The test is only meaningful for composite models.

If more than two groups are to be compared issues related to multiple testing
should be taken into account.

The number of permutation runs defaults to \code{args_default()$.R} for performance reasons.
According to Henseler et al. (2016) the number of permutations should be at least 5000 for
assessment to be reliable.
}
\section{Methods (by class)}{
\itemize{
\item \code{cSEMResults_default}: (TODO)

\item \code{cSEMResults_multi}: (TODO)

\item \code{cSEMResults_2ndorder}: (TODO)
}}

\examples{
\dontrun{
# TODO
}

}
