% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_test_MICOM.R
\name{testMICOM}
\alias{testMICOM}
\title{Test measurement invariance of composites}
\usage{
testMICOM(
  .object=args_default()$.object,
  .runs      = args_default()$.runs,
  .alpha        = args_default()$.alpha,
  .show_progress = args_default()$.show_progress,
  .handle_inadmissibles = args_default()$.handle_inadmissibles
}
\arguments{
\item{.object}{An R object of class \code{cSEM<class>} with corresponding method.}

\item{.runs}{Integer. How many runs should be performed? Defaults to \code{499}.}

\item{.alpha}{An integer or a numeric vector of significance levels.
Defaults to \code{0.05}.}

\item{.show_progress}{Logical. Show progress bar. Defaults to \code{TRUE}.}

\item{.handle_inadmissibles}{Character string. How should inadmissible results
be treated? One of "\emph{drop}", "\emph{ignore}", or "\emph{replace}". If "\emph{drop}", all
replications yielding an inadmissible result will be dropped (=
number of results shown <= .runs). For "\emph{ignore}" all results are returned
even if they are inadmissible (= number of results = .runs). For "\emph{replace}"
bootstrapping continues until there are .runs admissible solutions.
Defaults to "\emph{drop}".}
}
\value{
An object of class "testMICOM"
}
\description{
This functions performs the test for measurement invariance of composites
proposed by Henseler et al. (2016).
}
\details{
The test is only meaningful for composite models.

If more than two groups are to be compared issues related to multiple testing
should be taken into account.

The number of permutation runs is defaults to 100 for performance reasons.
According to Henseler et al. (2016) the number of permutations should be at least 5000 for
assessment to be reliable.
}
\examples{

still to implement

}
