% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_test_MICOM.R
\name{testMICOM}
\alias{testMICOM}
\title{Test measurement invariance of composites}
\usage{
testMICOM(
 .object               = args_default()$.object,
 .alpha                = args_default()$.alpha,
 .approach_p_adjust    = args_default()$.approach_p_adjust,
 .handle_inadmissibles = args_default()$.handle_inadmissibles,
 .R                    = args_default()$.R,
 .seed                 = args_default()$.seed,
 .verbose              = args_default()$.verbose
 )
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}

\item{.alpha}{An integer or a numeric vector of significance levels.
Defaults to \code{0.05}.}

\item{.approach_p_adjust}{Character string or a vector of character strings.
Approach used to adjust the p value in multiple testing.
See the \code{methods} argument of \code{\link[stats:p.adjust]{stats::p.adjust()}} for a list of choices and
their description. Defaults to "\emph{none}".}

\item{.handle_inadmissibles}{Character string. How should inadmissible results
be treated? One of "\emph{drop}", "\emph{ignore}", or "\emph{replace}". If "\emph{drop}", all
replications/resamples yielding an inadmissible result will be dropped
(i.e. the number of results returned will potentially be less than \code{.R}).
For "\emph{ignore}" all results are returned even if all or some of the replications
yieled inadmissible results (i.e. number of results returned is equal to \code{.R}).
For "\emph{replace}" resampling continues until there are exactly \code{.R} admissible solutions.
Defaults to "\emph{drop}".}

\item{.R}{Integer. The number of bootstrap replications. Defaults to \code{499}.}

\item{.seed}{Integer or \code{NULL}. The random seed to use. Defaults to \code{NULL} in which
case an arbitrary seed is choosen. Note that the scope of the seed is limited
to the body of the function it is used in. Hence, the global seed will
not be altered!}

\item{.verbose}{Logical. Should information (e.g., progress bar) be printed to the console? Defaults
to \code{TRUE}.}
}
\value{
A named list of class "testMICOM" containing the following list element:
\describe{
\item{\strong{Step2}}{A list containing the results of the test for compositional invariance (Step 2).}
\item{\strong{Step3}}{A list containing the results of the test for mean and variance equality (Step 3).}
\item{\strong{Meta_information}}{A list of additional information on the test.}
}
}
\description{
This functions performs the test for measurement invariance of composites
proposed by \insertCite{Henseler2016}{cSEM}.
}
\details{
The test is only meaningful for concepts modeled as composites.

If more than two groups are to be compared issues related to multiple testing
should be taken into account.

The number of permutation runs defaults to \code{args_default()$.R} for performance reasons.
According to \insertCite{Henseler2016}{cSEM} the number of permutations should
be at least 5000 for assessment to be sufficiently reliable.
}
\references{
\insertAllCited{}
}
