% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_test_OMF.R
\name{testOMF}
\alias{testOMF}
\title{Test for overall model fit}
\usage{
testOMF(
 .object              = args_default()$.model, 
 .alpha               = args_default()$.alpha, 
 .drop_inadmissibles  = args_default()$.drop_inadmissibles, 
 .runs                = args_default()$.runs
)
}
\arguments{
\item{.object}{An R object of class \code{cSEM<class>} with corresponding method.}

\item{.alpha}{An integer or a numeric vector of significance levels.
Defaults to \code{0.05}.}

\item{.drop_inadmissibles}{Logical. Should inadmissible solutions be dropped?
Defaults to \code{TRUE}.}

\item{.runs}{Integer. How many runs should be performed? Defaults to \code{499}.}
}
\value{
A standardized list of class \code{cSEMTest}. Technically \code{cSEMTest} is a named
list containing the following list elements:
\describe{
\item{\code{Hypothesis}}{The null hypothesis.}
\item{\code{$Test_statistic}}{The value of test statistic(s).}
\item{\code{$Critical_value}}{The critical value(s).}
\item{\code{$Decision}}{The test decision. One of: \strong{Reject} or \strong{Do not reject}}
\item{\code{$Number_admissibles}}{The number of admissible runs. See \code{\link[=verify]{verify()}} for
what constitues and inadmissble run.}
}
}
\description{
Test the overall model fit.
}
\details{
Description: TODO.
Test is based on \insertCite{Dijkstra2015;textual}{cSEM}.
Data is transformed according to \insertCite{Beran1985;textual}{cSEM}.
}
\examples{
\dontrun{
# TODO
}

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=csem]{csem()}}, \code{\link[=cca]{cca()}}, \code{\link[=foreman]{foreman()}}, \link{cSEMResults}
}
