% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_test_OMF.R
\name{testOMF}
\alias{testOMF}
\alias{testOMF.cSEMResults_default}
\alias{testOMF.cSEMResults_multi}
\alias{testOMF.cSEMResults_2ndorder}
\title{Test for overall model fit}
\usage{
testOMF(
 .object                = args_default()$.object, 
 .alpha                 = args_default()$.alpha, 
 .handle_inadmissibles  = args_default()$.handle_inadmissibles, 
 .R                     = args_default()$.R, 
 .saturated             = args_default()$.saturated,
 .verbose               = args_default()$.verbose
)

\method{testOMF}{cSEMResults_default}(.object = args_default()$.object,
  .alpha = args_default()$.alpha,
  .handle_inadmissibles = args_default()$.handle_inadmissibles,
  .R = args_default()$.R, .saturated = args_default()$.saturated,
  .verbose = args_default()$.verbose)

\method{testOMF}{cSEMResults_multi}(.object = args_default()$.object,
  .alpha = args_default()$.alpha,
  .handle_inadmissibles = args_default()$.handle_inadmissibles,
  .R = args_default()$.R, .saturated = args_default()$.saturated,
  .verbose = args_default()$.verbose)

\method{testOMF}{cSEMResults_2ndorder}(.object = args_default()$.object,
  .alpha = args_default()$.alpha,
  .handle_inadmissibles = args_default()$.handle_inadmissibles,
  .R = args_default()$.R, .saturated = args_default()$.saturated,
  .verbose = args_default()$.verbose)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}

\item{.alpha}{An integer or a numeric vector of significance levels.
Defaults to \code{0.05}.}

\item{.handle_inadmissibles}{Character string. How should inadmissible results
be treated? One of "\emph{drop}", "\emph{ignore}", or "\emph{replace}". If "\emph{drop}", all
replications/resamples yielding an inadmissible result will be dropped (
the number of results returned will be less than .R). For "\emph{ignore}" all results are returned
even if they are inadmissible (number of results returned = .R). For "\emph{replace}"
resampling continues until there are exactly .R admissible solutions.
Defaults to "\emph{drop}".}

\item{.R}{Integer. The number of bootstrap replications. Defaults to \code{499}.}

\item{.saturated}{Logical. Should a saturated structural model be used? Defaults to \code{FALSE}.}

\item{.verbose}{Logical. Should information be printed to the console? Defaults
to \code{TRUE}.}
}
\value{
A standardized list of class \code{cSEMTest}. Technically \code{cSEMTest} is a named
list containing the following list elements:
\describe{
\item{\code{$Test_statistic}}{The value of test statistic(s).}
\item{\code{$Critical_value}}{The critical value(s).}
\item{\code{$Decision}}{The test decision. One of: \strong{Reject} or \strong{Do not reject}}
\item{\code{$Number_admissibles}}{The number of admissible runs. See \code{\link[=verify]{verify()}} for
what constitues and inadmissible run.}
}
}
\description{
Test the overall model fit.
}
\details{
Description: TODO.
Test is based on \insertCite{Dijkstra2015;textual}{cSEM}.
Data is transformed according to \insertCite{Beran1985;textual}{cSEM}.
After 10000 iterations it stops automatically
}
\section{Methods (by class)}{
\itemize{
\item \code{cSEMResults_default}: (TODO)

\item \code{cSEMResults_multi}: (TODO)

\item \code{cSEMResults_2ndorder}: (TODO)
}}

\examples{
\dontrun{
# TODO
}

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=csem]{csem()}}, \code{\link[=cca]{cca()}}, \code{\link[=foreman]{foreman()}}, \link{cSEMResults}
}
