% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_test_OMF.R
\name{testOMF}
\alias{testOMF}
\title{Test for overall model fit}
\usage{
testOMF(
 .object                = args_default()$.object, 
 .alpha                 = args_default()$.alpha, 
 .handle_inadmissibles  = args_default()$.handle_inadmissibles, 
 .runs                  = args_default()$.runs, 
 .saturated             = args_default()$.saturated,
)
}
\arguments{
\item{.object}{An R object of class \code{cSEM<class>} with corresponding method.}

\item{.alpha}{An integer or a numeric vector of significance levels.
Defaults to \code{0.05}.}

\item{.handle_inadmissibles}{Character string. How should inadmissible results
be treated? One of "\emph{drop}", "\emph{ignore}", or "\emph{replace}". If "\emph{drop}", all
replications yielding an inadmissible result will be dropped (=
number of results shown <= .runs). For "\emph{ignore}" all results are returned
even if they are inadmissible (= number of results = .runs). For "\emph{replace}"
bootstrapping continues until there are .runs admissible solutions.
Defaults to "\emph{drop}".}

\item{.runs}{Integer. How many runs should be performed? Defaults to \code{499}.}

\item{.saturated}{Logical. Should a saturated structural model be used? Defaults to \code{FALSE}.}
}
\value{
A standardized list of class \code{cSEMTest}. Technically \code{cSEMTest} is a named
list containing the following list elements:
\describe{
\item{\code{$Test_statistic}}{The value of test statistic(s).}
\item{\code{$Critical_value}}{The critical value(s).}
\item{\code{$Decision}}{The test decision. One of: \strong{Reject} or \strong{Do not reject}}
\item{\code{$Number_admissibles}}{The number of admissible runs. See \code{\link[=verify]{verify()}} for
what constitues and inadmissible run.}
}
}
\description{
Test the overall model fit.
}
\details{
Description: TODO.
Test is based on \insertCite{Dijkstra2015;textual}{cSEM}.
Data is transformed according to \insertCite{Beran1985;textual}{cSEM}.
After 10000 iterations it stops automatically
}
\examples{
\dontrun{
# TODO
}

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=csem]{csem()}}, \code{\link[=cca]{cca()}}, \code{\link[=foreman]{foreman()}}, \link{cSEMResults}
}
