% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_test_OMF.R
\name{testOMF}
\alias{testOMF}
\title{Test for overall model fit}
\usage{
testOMF(
 .object                = NULL, 
 .alpha                 = 0.05, 
 .handle_inadmissibles  = c("drop", "ignore", "replace"), 
 .R                     = 499, 
 .saturated             = FALSE,
 .seed                  = NULL,
 .verbose               = TRUE
)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}

\item{.alpha}{An integer or a numeric vector of significance levels.
Defaults to \code{0.05}.}

\item{.handle_inadmissibles}{Character string. How should inadmissible results
be treated? One of "\emph{drop}", "\emph{ignore}", or "\emph{replace}". If "\emph{drop}", all
replications/resamples yielding an inadmissible result will be dropped
(i.e. the number of results returned will potentially be less than \code{.R}).
For "\emph{ignore}" all results are returned even if all or some of the replications
yieled inadmissible results (i.e. number of results returned is equal to \code{.R}).
For "\emph{replace}" resampling continues until there are exactly \code{.R} admissible solutions.
Depending on the frequency of inadmissible solutions this may significantly increase
computing time. Defaults to "\emph{drop}".}

\item{.R}{Integer. The number of bootstrap replications. Defaults to \code{499}.}

\item{.saturated}{Logical. Should a saturated structural model be used?
Defaults to \code{FALSE}.}

\item{.seed}{Integer or \code{NULL}. The random seed to use. Defaults to \code{NULL} in which
case an arbitrary seed is choosen. Note that the scope of the seed is limited
to the body of the function it is used in. Hence, the global seed will
not be altered!}

\item{.verbose}{Logical. Should information (e.g., progress bar) be printed to the console? Defaults
to \code{TRUE}.}
}
\value{
A list of class \code{cSEMTestOMF} containing the following list elements:
\describe{
\item{\code{$Test_statistic}}{The value of the test statistics.}
\item{\code{$Critical_value}}{The correponding critical values obtained by the bootstrap.}
\item{\code{$Decision}}{The test decision. One of: \code{FALSE} (\strong{Reject}) or \code{TRUE} (\strong{Do not reject}).}
\item{\code{$Information}}{The \code{.R} bootstrap values; The number of admissible results;
The seed used and the number of total runs.}
}
}
\description{
Bootstrap-based test for overall model fit originally proposed on \insertCite{Beran1985;textual}{cSEM}.
See also \insertCite{Dijkstra2015;textual}{cSEM} who first suggested the test in
the context of PLS-PM.
}
\details{
\code{testOMF()} tests the null hypothesis that the population indicator
correlation matrix equals the population model-implied indicator correlation matrix.
Several potential test statistics may be used. \code{testOMF()} uses the
geodesic distance and the squared Euclidean distance between the
sample indicator correlation matrix and the estimated model-implied indicator correlation matrix
as well as the standardized root mean square residual (SRMR).
The reference distribution for each test statistic is obtained by
the bootstrap as proposed by \insertCite{Beran1985;textual}{cSEM}.
}
\examples{
\donttest{# ===========================================================================
# Basic usage
# ===========================================================================
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
out <- csem(threecommonfactors, model, .approach_weights = "PLS-PM")

## Test (.R is small to save time; should be higher in real applications)
testOMF(out, .R = 50)}
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=csem]{csem()}}, \link{cSEMResults}
}
