% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_MGA.R
\name{testOverallMGA}
\alias{testOverallMGA}
\title{Test for group differences}
\usage{
testOverallMGA(.object = args_default()$.model,
  .drop_inadmissibles = args_default()$.drop_inadmissibles,
  .alpha = args_default()$.alpha, .runs = args_default()$.runs, ...)
}
\arguments{
\item{.object}{An R object of class \code{cSEM<class>} with corresponding method.}

\item{.drop_inadmissibles}{Logical. Should inadmissible solutions be dropped?
Defaults to \code{TRUE}.}

\item{.alpha}{An integer or a numeric vector of significance levels.
Defaults to \code{0.05}.}

\item{.runs}{Integer. How many runs should be performed? Defaults to \code{499}.}

\item{...}{Further arguments to be passed down to other functions.
See \code{\link[=args_default]{args_default()}} with \code{.only_dots = TRUE} for a complete list
of possible arguments and their defaults.}
}
\value{
A standarized list of class \code{cSEMTestResults} for which the following methods exists:
\describe{
\item{\code{print.cSEMTestResults}}{A formated summary of the test.}
\item{\code{summary.cSEMTestResults}}{A formated summary of the test (identical to \code{print.cSEMTestResults}).}
}
Technically \code{cSEMTestResults} is a named list containing the following list elements:
\describe{
\item{\code{$Test_statistic}}{The value of test statistic.}
\item{\code{$Critical_value}}{The critical value.}
\item{\code{$Decision}}{The test decision. One of: \strong{Reject} or \emph{Do not reject}}
\item{\code{$Number_addmissibles}}{The number of admissible runs. See xxx for what
constitutes an \strong{inadmissible run}.}
}
}
\description{
What it does.
}
\details{
More details
}
\examples{
\dontrun{
# still to implement
}

}
\seealso{
\link{cSEMResults}
}
