% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_MGA.R
\name{testOverallMGA}
\alias{testOverallMGA}
\title{Test for group differences}
\usage{
testOverallMGA(.object = args_default()$.model,
  .dropInadmissibles = args_default()$.dropInadmissibles,
  .alpha = args_default()$.alpha, .runs = args_default()$.runs, ...)
}
\arguments{
\item{csem_arguments}{}
}
\value{
csem_testresults
}
\description{
Test for group differences.
}
\details{
Some more details here including references.
}
\examples{
\dontrun{
# still to implement
}

}
\references{
\insertAllCited{}
}
\seealso{
\link{csem}, \link{foreman}
}
