% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_csem-tidiers.R
\name{tidy.cSEMResults}
\alias{tidy.cSEMResults}
\title{tidy Method for cSEMResults Objects}
\usage{
\method{tidy}{cSEMResults}(
  x,
  parameters = c("all", "Path_estimates", "Loadings_estimates", "Weight_estimates", "D2",
    "Effect_estimates", "Indicator_correlation", "Exo_construct_correlation"),
  ...
)
}
\arguments{
\item{x}{Fitted \code{csem} model with class \code{cSEMResults} as returned from \code{\link[=csem]{csem()}}}

\item{parameters}{Character string. Selects the desired parameter estimates as outputted from \code{\link[=summarize]{summarize()}}. Defaults to \code{all}.}

\item{...}{
  Arguments passed on to \code{\link[=summarize]{summarize}}
  \describe{
    \item{\code{.alpha}}{An integer or a numeric vector of significance levels.
Defaults to \code{0.05}.}
    \item{\code{.ci}}{A vector of character strings naming the confidence interval to compute.
For possible choices see \code{\link[=infer]{infer()}}.}
    \item{\code{.object}}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}
  }}
}
\value{
A \code{data.frame} with one row for each estimated parameter. The
description of the column names are as follows:

\item{term}{The result of \code{paste(lhs, op, rhs)}}
\item{op}{The operator in the model syntax (e.g., \verb{~~} for correlation, \verb{<~} for weights, \verb{=~} for loadings, and \code{~} for path estimates)}
\item{estimate}{The parameter estimate}
\item{construct.type}{Whether the modelled construct is a common factor or composite variable}
\item{std.error}{}
\item{statistic}{The t-statistic as returned from \code{\link[=summarize]{summarize()}} and \code{cSEM:::addInfer()}}
\item{p_value}{}
}
\description{
This method and documentation is based on the \code{tidy} method for \code{lavaan}, as found in \code{broom:::tidy.lavaan()}.
}
\details{
\strong{Cautionary Note}: As with all \code{tidy()} methods, mis-specified arguments may be silently ignored.
}
