% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_matrix.R
\name{trace}
\alias{trace}
\title{Internal: Trace of a matrix}
\usage{
trace(.matrix)
}
\arguments{
\item{.matrix}{Square numeric matrix.}
}
\value{
A numeric scalar: value with the trace of \code{.matrix}.
}
\description{
Compute the trace of \code{.matrix}, i.e., the sum of the diagonal elements of \code{.matrix}.
}
\keyword{internal}
