% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igsca.R
\name{update_C_B_D}
\alias{update_C_B_D}
\title{Update Loadings, Path-Coefficients and Uniqueness Terms After Updating Latent Variables}
\usage{
update_C_B_D(
  X,
  n_indicators,
  Gamma,
  cindex,
  C,
  n_constructs,
  bindex,
  B,
  n_case,
  D,
  Z,
  ov_type
)
}
\arguments{
\item{X}{}

\item{n_indicators}{}

\item{Gamma}{}

\item{cindex}{}

\item{C}{}

\item{n_constructs}{}

\item{bindex}{}

\item{B}{}

\item{n_case}{}

\item{D}{}

\item{Z}{}

\item{ov_type}{}
}
\value{
List of matrices:
\enumerate{
\item Uniqueness terms (D) and (\eqn{D^2})
\item Non-zero Path-Coefficients (if Paths are specified in B0) or non-zero Loadings
\item Path-Coefficients (B) and Loadings (C)
\item U, Utilde, v, F_o, Q, U \emph{something}
TODO: Name these accessory matrices
}
}
\description{
Update Loadings, Path-Coefficients and Uniqueness Terms After Updating Latent Variables
}
