% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igsca.R
\name{update_C_B_D_U}
\alias{update_C_B_D_U}
\title{Update Loadings, Path-Coefficients and Uniqueness Terms After Updating Latent Variables}
\usage{
update_C_B_D_U(
  X,
  n_indicators,
  Gamma,
  c_index,
  C,
  n_constructs,
  b_index,
  B,
  n_case,
  D,
  Z,
  indicator_type
)
}
\arguments{
\item{X}{Pseudo-weights for composites}

\item{n_indicators}{Number of indicators}

\item{Gamma}{Construct Scores}

\item{c_index}{Index of loadings}

\item{C}{Loadings matrix}

\item{n_constructs}{Number of constructs}

\item{b_index}{Index of Path Coefficients}

\item{B}{Path Coefficients Matrix}

\item{n_case}{Number of Cases}

\item{D}{Related to Unique Error}

\item{Z}{Standardized Data}

\item{indicator_type}{Vector of whether each indicator corresponds to a common factor or composite}
}
\value{
List of matrices:
\enumerate{
\item Uniqueness terms (D) and (U)
\item Path-Coefficients (B) and Loadings (C)
}
}
\description{
Update Loadings, Path-Coefficients and Uniqueness Terms After Updating Latent Variables
}
