% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igsca.R
\name{update_C_B_D_U}
\alias{update_C_B_D_U}
\title{Update Loadings, Path-Coefficients and Uniqueness Terms After Updating Latent Variables}
\usage{
update_C_B_D_U(
  X,
  n_indicators,
  Gamma,
  c_index,
  C,
  n_constructs,
  b_index,
  B,
  n_case,
  D,
  Z,
  ov_type
)
}
\arguments{
\item{X}{}

\item{n_indicators}{}

\item{Gamma}{}

\item{c_index}{}

\item{C}{}

\item{n_constructs}{}

\item{b_index}{}

\item{B}{}

\item{n_case}{}

\item{D}{}

\item{Z}{}

\item{ov_type}{}
}
\value{
List of matrices:
\enumerate{
\item Uniqueness terms (D) and (U)
\item Path-Coefficients (B) and Loadings (C)
}
}
\description{
Update Loadings, Path-Coefficients and Uniqueness Terms After Updating Latent Variables
}
