% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igsca.R
\name{update_common_factor}
\alias{update_common_factor}
\title{Update Common Factor Variable}
\usage{
update_common_factor(WW, windex_gamma_idx, gamma_idx)
}
\arguments{
\item{WW}{Pseudo-weights for Common Factors}

\item{windex_gamma_idx}{Index of weights related to the indicators for the construct of interest}

\item{gamma_idx}{Index of which construct we are examining}
}
\value{
theta: Used to update factor latent variables -- after accounting for loadings and path-coefficients.
}
\description{
Update Common Factor Variable
}
