% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_igsca.R
\name{update_composite}
\alias{update_composite}
\title{Update Composite Variables}
\usage{
update_composite(
  W,
  A,
  V,
  X,
  windex_gamma_idx,
  n_total_var,
  tot,
  n_constructs,
  gamma_idx
)
}
\arguments{
\item{W}{Weights matrix}

\item{A}{Stacked matrix of loadings and path coefficients}

\item{V}{Unclear meaning}

\item{X}{Pseudo-weights for composite variables}

\item{windex_gamma_idx}{Index of weights related to the indicators for the construct of interest}

\item{n_total_var}{Number of indicators and constructs}

\item{tot}{Index dependent on which construct variable we are examining}

\item{n_constructs}{Number of constructs}

\item{gamma_idx}{Index of which construct we are examining}
}
\value{
theta: A matrix that will later be used to update the weights for the composite variable.
}
\description{
Update Composite Variables
}
