% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igsca.R, R/igsca_translation.R
\name{update_factor_LV}
\alias{update_factor_LV}
\title{Update Factor Latent Variable}
\usage{
update_factor_LV(WW, windex_j, j)

update_factor_LV(WW, windex_j, j)
}
\arguments{
\item{WW}{}

\item{windex_j}{}

\item{j}{}
}
\value{
theta: Used to update factor latent variables -- after accounting for loadings and path-coefficients.

TODO: Double check that this is correct.

theta: Used to update factor latent variables -- after accounting for loadings and path-coefficients.

TODO: Double check that this is correct.
}
\description{
Update Factor Latent Variable

Update Factor Latent Variable
}
