% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igsca_translation.R
\name{update_X_weights}
\alias{update_X_weights}
\title{Update Weights and X (?)}
\usage{
update_X_weights(Z, U, D, C, nlv, B)
}
\arguments{
\item{Z}{}

\item{U}{}

\item{D}{}

\item{C}{}

\item{nlv}{}

\item{B}{}
}
\description{
Update Weights and X (?)
}
