% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_estimators_weights.R
\name{updateUD}
\alias{updateUD}
\title{Update unique scores and unique loadings}
\usage{
updateUD(
  D,
  Eta_normed,
  .indicator_type,
  n_constructs,
  n_case,
  n_indicators,
  Z_normed
)
}
\arguments{
\item{D}{Unique loadings}

\item{Eta_normed}{Normalized data}

\item{.indicator_type}{Vector of whether each indicator corresponds to a common factor or composite}

\item{n_constructs}{Number of constructs}

\item{n_case}{Number of cases}

\item{n_indicators}{Number of indicators}

\item{Z_normed}{Normalized data}
}
\value{
List of 2 elements, normalized unique scores (\code{U}) and normalized unique loadings (\code{D})
}
\description{
Intended to be used within the alternating least squares algorithm for either GSCA_M or IGSCA. Assumes that the construct scores and data are normalized.
}
\details{
Ensures that the unique scores have a mean of 0.

Following page 879 of \insertCite{trendafilovExploratoryFactorAnalysis2011;textual}{cSEM}, as the ratio of the sample size to the number of indicators decreases, the unique scores and unique loadings approach 0. This means that both GSCAm and IGSCA will reduce to GSCA at small enough sample sizes.
}
\references{
\insertAllCited{}
}
