% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_estimators_weights.R
\name{updateUD}
\alias{updateUD}
\title{Update unique scores and unique loadings}
\usage{
updateUD(
  D,
  Eta_normed,
  .indicator_type,
  n_constructs,
  n_case,
  n_indicators,
  Z_normed
)
}
\arguments{
\item{D}{Unique loadings}

\item{Eta_normed}{Normalized data}

\item{.indicator_type}{An indicator vector that indices whether a j indicator
(rows of W0 and C0) corresponds to a common factor variable (1) or a
composite variable (0). This vector is important for computing the
uniqueness terms (D) because it zeros the entries for composite indicators.}

\item{n_constructs}{Number of constructs}

\item{n_case}{Number of cases}

\item{n_indicators}{Number of indicators}

\item{Z_normed}{Normalized data}
}
\value{
List of 2 elements, normalized unique scores (\code{U}) and normalized unique loadings (\code{D})
}
\description{
Intended to be used within the alternating least squares algorithm for either GSCA_M or IGSCA. Assumes that the construct scores and data are normalized.
}
