% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_verify.R
\name{verify}
\alias{verify}
\title{Verify admissibility}
\usage{
verify(.object)
}
\arguments{
\item{.object}{An R object of class \code{cSEM<class>} with corresponding method.}
}
\value{
A logical vector indicating which (if any) problem occurred. Status codes are:
\itemize{
\item 1: The algorithm has not converged.
\item 2: At least one absolute standardized loading estimate is larger than 1,
\item 3: The construct VCV is not positive semi-definite.
\item 4: At least one construct reliability is larger than 1.
\item 5: The model-implied indicator VCV is not positive semi-definite.
This is only checked for linear models.
}
}
\description{
Verify admissibility of the estimated quantities for a given model.
}
\details{
Results based on an estimated model exhibiting one of the
following defects are deemed inadmissible: non-convergence, loadings and/or
reliabilities larger than 1, a construct VCV and/or a
model-implied VCV matrix that is not positive (semi-)definite.
}
\seealso{
\code{\link[=csem]{csem()}}, \code{\link[=cca]{cca()}}, \code{\link[=foreman]{foreman()}}, \link{cSEMResults}
}
