% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_verify.R
\name{verify}
\alias{verify}
\alias{verify.cSEMResults_default}
\alias{verify.cSEMResults_multi}
\alias{verify.cSEMResults_2ndorder}
\title{Verify admissibility}
\usage{
verify(.object)

\method{verify}{cSEMResults_default}(.object)

\method{verify}{cSEMResults_multi}(.object)

\method{verify}{cSEMResults_2ndorder}(.object)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}
}
\value{
A logical vector indicating which (if any) problem occurred.
A \code{TRUE} indicates that the problem has occured. For models containg second order
constructs a list of two such vectors (one for the first and one for the second stage)
is returned. Status codes are:
\itemize{
\item 1: The algorithm has not converged.
\item 2: At least one absolute standardized loading estimate is larger than 1,
\item 3: The construct VCV is not positive semi-definite.
\item 4: At least one reliability is larger than 1.
\item 5: The model-implied indicator VCV is not positive semi-definite.
This is only checked for linear models.
}
}
\description{
Verify admissibility of the estimated quantities for a given model.
}
\details{
The following is declared as inadmissible : non-convergence, loadings and/or
(congeneric) reliabilities larger than 1, a construct VCV and/or a
model-implied VCV matrix that is not positive semi-definite.

For models containing second order constructs estimation is by default done
in a two/three stage procedure. In this case both the first and the second/third
stage model quantities are checked.

Currently, a model-implied indicator VCV matrix for nonlinear model is not
available. \code{verify()} therefore skips the check for positive definiteness of the
model-implied indicator VCV matrix and return an ok-status by default.
}
\section{Methods (by class)}{
\itemize{
\item \code{cSEMResults_default}: (TODO)

\item \code{cSEMResults_multi}: (TODO)

\item \code{cSEMResults_2ndorder}: (TODO)
}}

\seealso{
\code{\link[=csem]{csem()}}, \code{\link[=foreman]{foreman()}}, \link{cSEMResults}
}
