% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimate_verify.R
\name{verify}
\alias{verify}
\title{Verify admissibility}
\usage{
verify(.object)
}
\arguments{
\item{.object}{An R object of class \link{cSEMResults} resulting from a call to \code{\link[=csem]{csem()}}.}
}
\value{
A logical vector indicating which (if any) problem occurred.
A \code{TRUE} indicates that the problem has occured. For models containg second order
constructs estimated by a two stage approach a list of two such vectors
(one for the first and one for the second stage) is returned. Status codes are:
\itemize{
\item 1: The algorithm has not converged.
\item 2: At least one absolute standardized loading estimate is larger than 1,
\item 3: The construct VCV is not positive semi-definite.
\item 4: At least one reliability is larger than 1.
\item 5: The model-implied indicator VCV is not positive semi-definite.
This is only checked for linear models.
}
}
\description{
Verify admissibility of the estimated quantities in \code{.object}.
}
\details{
Result exhibiting one of the following defects are deemed inadmissible:
non-convergence of the algorithm used to obtain weights, loadings and/or
(congeneric) reliabilities larger than 1, a construct variance-covariance (VCV) and/or
model-implied VCV matrix that is not positive semi-definite.

If the two stage approach was used for models containing second order
constructs, both the first and the second stage is checked separately.

Currently, a model-implied indicator VCV matrix for nonlinear model is not
available. \code{verify()} therefore skips the check for positive definitness of the
model-implied indicator VCV matrix for nonlinear models and returns "ok".
}
\seealso{
\code{\link[=csem]{csem()}}, \link{cSEMResults}
}
